addpoly {metafor}	R Documentation

Add Polygons to Forest Plots

Description

The function addpoly is generic. It can be used to add polygons to a forest plot, for example, to indicate summary estimates for all or subgroups of studies and to indicate fitted/predicted values based on models involving moderators.

Usage

addpoly(x, ...)

Arguments

x
either an object of class "rma" or the values at which polygons should be drawn. See ‘Details’.

...
other arguments.

Details

Currently, methods exist for two types of situations.

In the first case, object x is a fitted model coming from the rma.uni, rma.mh, rma.peto, or rma.glmm functions. The model must either be a fixed- or random-effects model, that is, the model should not contain any moderators. The corresponding method is called addpoly.rma. It can be used to add a polygon to an existing forest plot (often at the bottom), showing the summary estimate with corresponding confidence interval based on the fitted model.

Alternatively, object x can be a vector with values at which one or more polygons should be drawn. The corresponding method is then addpoly.default.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

addpoly.rma, addpoly.default, forest.rma, forest.default

[Package metafor version 2.4-0 Index]