addpoly.default {metafor}	R Documentation

Add Polygons to Forest Plots (Default Method)

Description

Function to add one or more polygons to a forest plot.

Usage

## Default S3 method:
addpoly(x, vi, sei, ci.lb, ci.ub, cr.lb, cr.ub,
        rows=-1, level=95, annotate=TRUE, digits=2, width, mlab,
        transf, atransf, targs, efac=1, col, border, fonts, cex, ...)

Arguments

x
vector with the values at which the polygons should be drawn.

vi
vector with the corresponding variances.

sei
vector with the corresponding standard errors note: only one of the two, vi or sei, needs to be specified).

ci.lb
vector with the corresponding lower confidence interval bounds. Not needed if vi or sei is specified. See ‘Details’.

ci.ub
vector with the corresponding upper confidence interval bounds. Not needed if vi or sei is specified. See ‘Details’.

cr.lb
optional vector with the corresponding lower credibility/prediction interval bounds.

cr.ub
optional vector with the corresponding upper credibility/prediction interval bounds.

rows
vector specifying the rows (or more generally, the horizontal positions) for plotting the polygons (defaults is -1). Can also be a single value specifying the row (horizontal position) of the first polygon (the remaining polygons are then plotted below this starting row).

level
numerical value between 0 and 100 specifying the confidence interval level (the default is 95).

annotate
logical specifying whether annotations should be added to the plot for the polygons that are drawn (the default is TRUE).

digits
integer specifying the number of decimal places to which the annotations should be rounded (the default is 2).

width
optional integer to manually adjust the width of the columns for the annotations.

mlab
optional character vector with the same length as x giving labels for the polygons that are drawn.

transf
optional argument specifying the name of a function that should be used to transform the x values and confidence interval bounds (e.g., transf=exp; see also transf). If unspecified, no transformation is used.

atransf
optional argument specifying the name of a function that should be used to transform the annotations (e.g., atransf=exp; see also transf). If unspecified, no transformation is used.

targs
optional arguments needed by the function specified via transf or atransf.

efac
vertical expansion factor for the polygons. The default value of 1 should usually work okay.

col
optional character string specifying the name of a color to use for the polygons. If unspecified, the function sets a default color.

border
optional character string specifying the name of a color to use for the border of the polygons. If unspecified, the function sets a default color.

fonts
optional character string specifying the font to use for the labels and annotations. If unspecified, the default font is used.

cex
optional symbol expansion factor. If unspecified, the function tries to set this to a sensible value.

...
other arguments.

Details

The function can be used to add one or more polygons to an existing forest plot created with the forest function. For example, summary estimates based on a model involving moderators can be added to the plot this way. See example below.

To use the function, one should specify the values at which the polygons should be drawn (via the x argument) together with the corresponding variances (via the vi argument) or with the corresponding standard errors (via the sei argument). Alternatively, one can specify the values at which the polygons should be drawn together with the corresponding confidence interval bounds (via the ci.lb and ci.ub arguments). Optionally, one can also specify the bounds of the corresponding credibility/prediction interval bounds via the cr.lb and cr.ub arguments.

The arguments transf, atransf, efac, and cex should always be set equal to the same values used to create the forest plot.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

forest.rma, forest.default

Examples

### meta-analysis of the log risk ratios using a mixed-
### effects model with absolute latitude as a moderator
res <- rma(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, mods = ~ ablat,
           slab=paste(author, year, sep=", "), data=dat.bcg)

### forest plot of the observed risk ratios
forest(res, addfit=FALSE, atransf=exp, xlim=c(-8,5), ylim=c(-4.5,16), cex=.8,
       order=order(dat.bcg$ablat), ilab=dat.bcg$ablat, ilab.xpos=-4,
       header="Author(s) and Year")

### predicted average log risk ratios for 10, 30, and 50 degrees absolute latitude
x <- predict(res, newmods=c(10, 30, 50))

### add predicted average risk ratios to forest plot
addpoly(x$pred, sei=x$se, atransf=exp, col="white", rows=-2,
        mlab=c("- at 10 Degrees", "- at 30 Degrees", "- at 50 Degrees"), cex=.8)
abline(h=0)
text(-8, -1, "Model-Based Estimates of RR:", pos=4, cex=.8)
text(-4, 15, "Latitude", cex=.8, font=2)

[Package metafor version 2.4-0 Index]