cint {metafor}	R Documentation

Confidence Intervals for Model Parameters

Description

The function cint is generic. It provides confidence intervals for specific classes of objects.

Usage

cint(object, ...)

Arguments

object 	 an object for which confidence intervals can be calculated. See ‘Details’.

... 	 other arguments.

Details

Currently, there is only a method for objects of class "rma.uni" created by the rma.uni function. Accordingly, the corresponding method is called cint.rma.uni. See the documentation for that function for more details.

Value

Confidence interval bounds of the model parameters.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

cint.rma.uni

[Package metafor version 0.5-7 Index]