cint.rma.uni {metafor}	R Documentation

Confidence Intervals for rma.uni Objects

Description

The function calculates a confidence interval for the amount of (residual) heterogeneity in random- and mixed-effects models that were fitted with the rma.uni function. Confidence intervals for the model coefficients can also be obtained.

Usage

## S3 method for class 'rma.uni':
cint(object, fixed=FALSE, random=TRUE, level=object$level, 
     digits=object$digits, control=list(), ...)

Arguments

object 	 an object of class "rma.uni".

fixed 	 logical indicating whether confidence intervals for the model coefficients should be returned (default is FALSE).

random 	 logical indicating whether a confidence interval for the amount of (residual) heterogeneity should be returned (default is TRUE).

level 	 a numerical value between 0 and 100 specifying the confidence interval level (the default is to take the value from the object).

digits 	 an integer specifying the number of decimal places to which the results should be rounded (the default is to take the value from the object).

control 	 a list of control values for the iterative algorithm. Defaults to an empty list, which means that default values are defined inside the function. See ‘Note’.

... 	 other arguments.

Details

The confidence intervals for the model coefficients are simply the usual Wald-type intervals which are also shown when printing the fitted object or when calling coef.rma.

The confidence interval for the amount of (residual) heterogeneity is obtained iteratively via the Q-profile method as described by Hartung and Knapp (2005) and Viechtbauer (2007). The method provides an exact confidence interval for tau^2 in random- and mixed-effects models. The square root of the interval bounds is also returned for easier interpretation. For random-effects models, confidence intervals for I^2 and H^2 are also provided (Higgins & Thompson, 2002). Since I^2 and H^2 are just monotonic transformation of tau^2, the confidence intervals for I^2 and H^2 are also exact.

Value

Either a single data frame or a list with two data frames (named fixed and random) with the following elements:

estimate 	 estimate of the model coefficient or variance component

ci.lb 	 lower bound of the confidence interval.

ci.ub 	 upper bound of the confidence interval.

For fixed-effects models, the variance component estimates are NA, but the confidence interval bounds are still provided.

Note

The iterative algorithm used as part of the Q-profile method makes use of the uniroot function. By default, the desired accuracy is set equal to .Machine$double.eps^0.25 and the maximum number of iterations to 1000. The upper bound of the interval searched is set to 50 (which should be large enough for most cases). The desired accuracy (tol), maximum number of iterations (maxiter), and upper bound (tau2.max) can be adjusted with control=list(tol=value, maxiter=value, tau2.max=value). One can also adjust the lower bound of the interval searched with control=list(tau2.min=value) (the default is to take that value from the object, which is 0 by default). You should only play around with this latter value if you know what you are doing.

It is possible that the lower and upper confidence interval bounds both fall below zero (or whatever value was chosen for tau2.min). Since both values then fall outside of the parameter space, the confidence interval then just consists of the null set.

Usually, the estimate of tau^2 from the random/mixed-effects model will fall within the confidence interval provided by the Q-profile method. However, this is not guaranteed. Depending on the method used to estimate tau^2 and the width of the confidence interval, it can happen in rare cases that the confidence interval does not actually contain the estimate (trying to explain this to reviewers can be tricky). However, uing the empirical Bayes estimator of tau^2 when fitting the model guarantees that the estimate of tau^2 falls within the confidence interval (as long as level is larger than 26).

The Q-profile method is exact under the assumptions of the random- and mixed-effects models (i.e., normally distributed observed and true outcomes and known sampling variances). These assumptions are usually only approximately true, turning the confidence interval for tau^2 also into an approximation.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

References

Hartung, J. & Knapp, G. (2005) On confidence intervals for the among-group variance in the one-way random effects model with unequal error variances. Journal of Statistical Planning and Inference, 127, 157–177.

Higgins, J. P. T. & Thompson, S. G. (2002) Quantifying heterogeneity in a meta-analysis. Statistics in Medicine, 21, 1539–1558.

Viechtbauer, W. (2007) Confidence intervals for the amount of heterogeneity in meta-analysis. Statistics in Medicine, 26, 37–52.

See Also

rma.uni

Examples

### load BCG vaccine data
data(dat.bcg)

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, 
           data=dat.bcg, measure="RR", method="REML")

### confidence interval for the total amount of heterogeneity
cint(res)

### mixed-effects model with absolute latitude in the model
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, mods=ablat,
           data=dat.bcg, measure="RR", method="REML")

### confidence interval for the residual amount of heterogeneity
cint(res)

[Package metafor version 0.5-7 Index]