Extract the Model Coefficients and Coefficient Table from 'rma' and 'summary.rma' Objects

The coef function extracts the estimated model coefficients from objects of class "rma". For objects of class "summary.rma", the model coefficients, corresponding standard errors, test statistics, p-values, and confidence interval bounds are extracted.

# S3 method for rma
coef(object, ...)

# S3 method for summary.rma
coef(object, ...)

Arguments

object
an object of class "rma" or "summary.rma".

...
other arguments.

Value

Either a vector with the estimated model coefficient(s) or a data frame with the following elements:

estimate

estimated model coefficient(s).

se

corresponding standard error(s).

zval

corresponding test statistic(s).

pval

corresponding p-value(s).

ci.lb

corresponding lower bound of the confidence interval(s).

ci.ub

corresponding upper bound of the confidence interval(s).

When the model was fitted with the Knapp and Hartung (2003) method (i.e., test="knha" in the rma.uni function) or with test="t" in the rma.glmm and rma.mv functions, then zval is called tval in the data frame that is returned by the function.
Author

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://doi.org/10.18637/jss.v036.i03

See also

Examples

### meta-analysis of the log risk ratios using a mixed-effects model
### with two moderators (absolute latitude and publication year)
res <- rma(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg,
           mods = ~ ablat + year, data=dat.bcg)
coef(res)

#>      intrcpt        ablat         year 
#> -3.545505079 -0.028011275  0.001907557 

coef(summary(res))

#>             estimate          se       zval        pval        ci.lb
#> intrcpt -3.545505079 29.09587983 -0.1218559 0.903013130 -60.57238164
#> ablat   -0.028011275  0.01023404 -2.7370689 0.006198931  -0.04806963
#> year     0.001907557  0.01468382  0.1299088 0.896638598  -0.02687219
#>                ci.ub
#> intrcpt 53.481371479
#> ablat   -0.007952924
#> year     0.030687307