cumul.rma.mh {metafor}	R Documentation

Cumulative Meta-Analysis for rma.mh and rma.peto Objects

Description

The functions leave1out.rma.mh and leave1out.rma.peto repeatedly fit the specified model, adding one observation (i.e., 2x2 table) at a time to the model.

Usage

## S3 method for class 'rma.mh':
cumul(x, order=NULL, digits=x$digits, transf=FALSE, ...)
## S3 method for class 'rma.peto':
cumul(x, order=NULL, digits=x$digits, transf=FALSE, ...)

Arguments

x 	 an object of class "rma.mh" or "rma.peto".

order 	 an optional vector with indices giving the desired order for the cumulative meta-analysis.

digits 	 an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

transf 	 logical indicating whether odds ratios or risk ratios (and the corresponding confidence interval bounds) should be given in the transformed (meaning: raw) units or in terms of log units (the default).

... 	 other arguments.

Value

An object of class c("list.rma", "cumul.rma"). The object is a list containing the following components:

estimate 	 estimated coefficients of the model.

se 	 standard errors of the coefficients. NA if transf=TRUE.

zval 	 test statistics of the coefficients.

pval 	 p-values for the test statistics.

ci.lb 	 lower bounds of the confidence intervals for the coefficients.

ci.ub 	 upper bounds of the confidence intervals for the coefficients.

Q 	 test statistics for the tests of heterogeneity.

Qp 	 p-values for the tests of heterogeneity.

The object is formated and printed with print.list.rma. A forest plot showing the results from the cumulative meta-analysis can be obtained with forest.cumul.rma.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

cumul, forest.cumul.rma

Examples

### load BCG vaccine data
data(dat.bcg)

### meta-analysis of the (log) risk ratios using the Mantel-Haenszel method
res <- rma.mh(ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg, measure="RR")

cumul(res, order=order(dat.bcg$year))
cumul(res, order=order(dat.bcg$year), transf=TRUE)

### meta-analysis of the (log) odds ratios using Peto's method
res <- rma.mh(ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

cumul(res, order=order(dat.bcg$year))
cumul(res, order=order(dat.bcg$year), transf=TRUE)

[Package metafor version 0.5-7 Index]