Studies on the Association between Recidivism and Mental Health

Results from 17 studies on the association between recidivism and mental health in delinquent juveniles.

dat.assink2016

Format

The data frame contains the following columns:

study	numeric	study id number

esid	numeric	effect size within study id number

id	numeric	row id number

yi	numeric	standardized mean difference

vi	numeric	corresponding sampling variance

pubstatus	numeric	published study (0 = no; 1 = yes)

year	numeric	publication year of the study (centered)

deltype	character	type of delinquent behavior in which juveniles have recidivated (either general, overt, or covert)

Details

The studies included in this dataset compared the difference in recidivism between delinquent juveniles with a mental health disorder and a comparison group of juveniles without a mental health disorder. Since studies differed in the way recidivism was defined and assessed, results are given in terms of standardized mean differences, with positive values indicating a higher prevalence of recidivism in the group of juveniles with a mental health disorder.

Multiple effect size estimates could be extracted from most studies, necessitating the use of appropriate models/methods for the analysis. Assink and Wibbelink (2016) illustrated the use of multilevel meta-analytic models for these purposes.

Source

Assink, M., & Wibbelink, C. J. M. (2016). Fitting three-level meta-analytic models in R: A step-by-step tutorial. The Quantitative Methods for Psychology, 12(3), 154--174. https://doi.org/10.20982/tqmp.12.3.p154

References

Assink, M., van der Put, C. E., Hoeve, M., de Vries, S. L. A., Stams, G. J. J. M., & Oort, F. J. (2015). Risk factors for persistent delinquent behavior among juveniles: A meta-analytic review. Clinical Psychology Review, 42, 47--61. https://doi.org/10.1016/j.cpr.2015.08.002

Examples

### copy data into 'dat' and examine data
dat <- dat.assink2016
dat

#>     study esid  id      yi     vi pubstatus  year deltype
#> 1       1    1   1  0.9066 0.0740         1   4.5 general
#> 2       1    2   2  0.4295 0.0398         1   4.5 general
#> 3       1    3   3  0.2679 0.0481         1   4.5 general
#> 4       1    4   4  0.2078 0.0239         1   4.5 general
#> 5       1    5   5  0.0526 0.0331         1   4.5 general
#> 6       1    6   6 -0.0507 0.0886         1   4.5 general
#> 7       2    1   7  0.5117 0.0115         1   1.5 general
#> 8       2    2   8  0.4738 0.0076         1   1.5 general
#> 9       2    3   9  0.3544 0.0065         1   1.5 general
#> 10      3    1  10  2.2844 0.3325         1  -8.5 general
#> 11      3    2  11  2.1771 0.3073         1  -8.5 general
#> 12      3    3  12  1.7777 0.2697         1  -8.5 general
#> 13      3    4  13  1.5480 0.4533         1  -8.5 general
#> 14      3    5  14  1.4855 0.1167         1  -6.5 general
#> 15      3    6  15  1.4836 0.1706         1  -6.5 general
#> 16      3    7  16  1.2777 0.1538         1  -8.5 general
#> 17      3    8  17  1.0311 0.3132         1  -8.5 general
#> 18      3    9  18  0.9409 0.1487         1  -6.5 general
#> 19      3   10  19  0.6263 0.2139         1  -6.5 general
#> 20      4    1  20 -0.0447 0.0331         1  -7.5 general
#> 21      5    1  21  1.5490 0.1384         1 -11.5 general
#> 22      6    1  22  0.7516 0.0477         1   5.5 general
#> 23      6    2  23  0.3979 0.0449         1   5.5 general
#> 24      6    3  24  0.2062 0.3533         1   5.5 general
#> 25      6    4  25 -0.2927 0.0758         1   5.5 general
#> 26      6    5  26 -0.5067 0.0314         1   5.5 general
#> 27      7    1  27  1.7083 0.0729         1   3.5 general
#> 28      7    2  28  0.8003 0.0815         1   3.5 general
#> 29      7    3  29  0.7268 0.1394         1   3.5 general
#> 30      7    4  30  0.6569 0.0943         1   3.5 general
#> 31      7    5  31  0.5569 0.1355         1   3.5 general
#> 32      7    6  32  0.3504 0.0763         1   3.5 general
#> 33      8    1  33  0.3695 0.0199         1   1.5 general
#> 34      9    1  34  0.1048 0.0331         1   5.5 general
#> 35      9    2  35  0.1748 0.0346         1   5.5 general
#> 36     10    1  36  0.4549 0.0278         1   4.5 general
#> 37     10    2  37  0.3161 0.0185         1   4.5 general
#> 38     10    3  38  0.0826 0.0328         1   4.5 general
#> 39     10    4  39  0.0221 0.0612         1   4.5 general
#> 40     10    5  40 -0.0199 0.0127         1   4.5 general
#> 41     11    1  41  2.3580 0.1226         0  -6.5 general
#> 42     11    2  42  2.3580 0.1226         0  -6.5 general
#> 43     11    3  43  2.3157 0.2416         0  -6.5 general
#> 44     11    4  44  2.0765 0.0844         0  -6.5 general
#> 45     11    5  45  2.0034 0.0780         0  -6.5 general
#> 46     11    6  46  1.7716 0.0723         0  -6.5 general
#> 47     11    7  47  1.4918 0.0701         0  -6.5 general
#> 48     11    8  48  1.4476 0.0422         0  -6.5 general
#> 49     11    9  49  1.2897 0.0663         0  -6.5   overt
#> 50     11   10  50  1.2637 0.0935         0  -6.5   overt
#> 51     11   11  51  1.2423 0.0392         0  -6.5 general
#> 52     11   12  52  1.1769 0.0387         0  -6.5 general
#> 53     11   13  53  1.0760 0.0391         0  -6.5 general
#> 54     11   14  54  1.0653 0.0380         0  -6.5   overt
#> 55     11   15  55  1.0379 0.1651         0  -6.5 general
#> 56     11   16  56  1.0101 0.0586         0  -6.5   overt
#> 57     11   17  57  1.0087 0.0386         0  -6.5   overt
#> 58     11   18  58  0.8765 0.0893         0  -6.5 general
#> 59     11   19  59  0.8615 0.1591         0  -6.5   overt
#> 60     11   20  60  0.8531 0.0350         0  -6.5 general
#> 61     11   21  61  0.6999 0.0353         0  -6.5 general
#> 62     11   22  62  0.6947 0.0353         0  -6.5 general
#> 63     12    1  63  0.2994 0.0041         0   4.5   overt
#> 64     12    2  64  0.2992 0.0042         0   4.5 general
#> 65     12    3  65  0.2989 0.0041         0   4.5   overt
#> 66     12    4  66  0.2910 0.0042         0   4.5   overt
#> 67     12    5  67  0.2170 0.0041         0   4.5   overt
#> 68     13    1  68  0.3874 0.0138         1  -3.5 general
#> 69     13    2  69  0.2178 0.0137         1  -3.5 general
#> 70     14    1  70  0.4571 0.0100         1  -1.5 general
#> 71     14    2  71  0.1947 0.0059         1  -1.5 general
#> 72     14    3  72  0.0239 0.0025         1  -1.5 general
#> 73     14    4  73 -0.0575 0.0032         1  -1.5 general
#> 74     14    5  74 -0.1190 0.0789         1  -1.5 general
#> 75     14    6  75 -0.2945 0.0179         1  -1.5 general
#> 76     15    1  76  0.7130 0.1537         1 -13.5 general
#> 77     15    2  77  0.3990 0.2334         1 -13.5 general
#> 78     15    3  78 -0.7645 0.1674         1 -13.5 general
#> 79     16    1  79  0.7156 0.0914         1   2.5   overt
#> 80     16    2  80  0.7067 0.0875         1   2.5  covert
#> 81     16    3  81  0.6475 0.0330         1   2.5 general
#> 82     16    4  82  0.6428 0.0861         1   2.5  covert
#> 83     16    5  83  0.6271 0.0400         1   2.5 general
#> 84     16    6  84  0.6238 0.0680         1   2.5 general
#> 85     16    7  85  0.6025 0.1287         1   2.5   overt
#> 86     16    8  86  0.5763 0.0332         1   2.5 general
#> 87     16    9  87  0.5171 0.0517         1   2.5  covert
#> 88     16   10  88 -0.3797 0.0390         1   2.5  covert
#> 89     16   11  89 -0.4228 0.0664         1   2.5  covert
#> 90     16   12  90 -0.4245 0.0809         1   2.5  covert
#> 91     16   13  91 -0.4671 0.0667         1   2.5  covert
#> 92     16   14  92 -0.5230 0.0988         1   2.5   overt
#> 93     16   15  93 -0.5675 0.0340         1   2.5  covert
#> 94     16   16  94 -0.7586 0.0437         1   2.5  covert
#> 95     17    1  95  0.3453 0.0340         1   5.5 general
#> 96     17    2  96  0.1221 0.0158         1   5.5 general
#> 97     17    3  97  0.0906 0.0107         1   5.5 general
#> 98     17    4  98  0.0040 0.0208         1   5.5 general
#> 99     17    5  99 -0.0207 0.0123         1   5.5 general
#> 100    17    6 100 -0.0660 0.0100         1   5.5 general

### fit multilevel model
res <- rma.mv(yi, vi, random = ~ 1 | study/esid, data=dat)
res

#> 
#> Multivariate Meta-Analysis Model (k = 100; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed      factor 
#> sigma^2.1  0.1879  0.4334     17     no       study 
#> sigma^2.2  0.1120  0.3347    100     no  study/esid 
#> 
#> Test for Heterogeneity:
#> Q(df = 99) = 809.4611, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.4268  0.1184  3.6038  0.0003  0.1947  0.6589  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### use cluster-robust inference methods
robust(res, cluster=dat$study)

#> 
#> Number of outcomes:   100
#> Number of clusters:   17
#> Outcomes per cluster: 1-22 (mean: 5.88, median: 5)
#> 
#> Model Results:
#> 
#> estimate      se    tval    pval   ci.lb   ci.ub 
#>   0.4268  0.1183  3.6076  0.0024  0.1760  0.6776  ** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### LRTs for the variance components
res0 <- rma.mv(yi, vi, random = ~ 1 | study/esid, data=dat, sigma2=c(0,NA))
anova(res0, res)

#> 
#>         df      AIC      BIC     AICc    logLik     LRT   pval       QE 
#> Full     3 153.2654 161.0508 153.5180  -73.6327                809.4611 
#> Reduced  2 214.0678 219.2581 214.1928 -105.0339 62.8024 <.0001 809.4611 
#> 

res0 <- rma.mv(yi, vi, random = ~ 1 | study/esid, data=dat, sigma2=c(NA,0))
anova(res0, res)

#> 
#>         df      AIC      BIC     AICc    logLik     LRT   pval       QE 
#> Full     3 153.2654 161.0508 153.5180  -73.6327                809.4611 
#> Reduced  2 233.1313 238.3215 233.2563 -114.5656 81.8658 <.0001 809.4611 
#> 

### examine some potential moderators via meta-regression
rma.mv(yi, vi, mods = ~ pubstatus, random = ~ 1 | study/esid, data=dat)

#> 
#> Multivariate Meta-Analysis Model (k = 100; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed      factor 
#> sigma^2.1  0.1712  0.4138     17     no       study 
#> sigma^2.2  0.1128  0.3359    100     no  study/esid 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 98) = 702.5110, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 1.8444, p-val = 0.1744
#> 
#> Model Results:
#> 
#>            estimate      se     zval    pval    ci.lb   ci.ub 
#> intrcpt      0.8117  0.3056   2.6565  0.0079   0.2128  1.4106  ** 
#> pubstatus   -0.4474  0.3294  -1.3581  0.1744  -1.0930  0.1983     
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

rma.mv(yi, vi, mods = ~ year, random = ~ 1 | study/esid, data=dat)

#> 
#> Multivariate Meta-Analysis Model (k = 100; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed      factor 
#> sigma^2.1  0.1351  0.3675     17     no       study 
#> sigma^2.2  0.1128  0.3359    100     no  study/esid 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 98) = 673.5617, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 5.4633, p-val = 0.0194
#> 
#> Model Results:
#> 
#>          estimate      se     zval    pval    ci.lb    ci.ub 
#> intrcpt    0.4257  0.1040   4.0947  <.0001   0.2219   0.6294  *** 
#> year      -0.0421  0.0180  -2.3374  0.0194  -0.0773  -0.0068    * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

dat$deltype <- factor(dat$deltype)
dat$deltype <- relevel(dat$deltype, ref="general")
rma.mv(yi, vi, mods = ~ deltype, random = ~ 1 | study/esid, data=dat)

#> 
#> Multivariate Meta-Analysis Model (k = 100; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed      factor 
#> sigma^2.1  0.1899  0.4358     17     no       study 
#> sigma^2.2  0.0847  0.2910    100     no  study/esid 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 97) = 761.8270, p-val < .0001
#> 
#> Test of Moderators (coefficients 2:3):
#> QM(df = 2) = 14.9760, p-val = 0.0006
#> 
#> Model Results:
#> 
#>                estimate      se     zval    pval    ci.lb    ci.ub 
#> intrcpt          0.4702  0.1180   3.9858  <.0001   0.2390   0.7015  *** 
#> deltypecovert   -0.7297  0.1923  -3.7941  0.0001  -1.1066  -0.3527  *** 
#> deltypeovert    -0.2219  0.1392  -1.5939  0.1110  -0.4948   0.0510      
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

rma.mv(yi, vi, mods = ~ year + deltype, random = ~ 1 | study/esid, data=dat)

#> 
#> Multivariate Meta-Analysis Model (k = 100; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed      factor 
#> sigma^2.1  0.1493  0.3863     17     no       study 
#> sigma^2.2  0.0853  0.2920    100     no  study/esid 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 96) = 610.2644, p-val < .0001
#> 
#> Test of Moderators (coefficients 2:4):
#> QM(df = 3) = 19.2399, p-val = 0.0002
#> 
#> Model Results:
#> 
#>                estimate      se     zval    pval    ci.lb    ci.ub 
#> intrcpt          0.4656  0.1071   4.3461  <.0001   0.2556   0.6756  *** 
#> year            -0.0380  0.0183  -2.0773  0.0378  -0.0738  -0.0021    * 
#> deltypecovert   -0.7094  0.1914  -3.7069  0.0002  -1.0845  -0.3343  *** 
#> deltypeovert    -0.2040  0.1385  -1.4725  0.1409  -0.4755   0.0675      
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>