dat.bcg {metafor}	R Documentation

Data for BCG Vaccine Studies

Description

Results from 13 clinical trials examining the effectiveness of the bacillus Calmette-Guerin (BCG) vaccine for preventing tuberculosis.

Usage

dat.bcg

Format

The data frame contains the following columns:

trial 	 numeric 	 trial number

author 	 character 	 author(s)

year 	 numeric 	 publication year

tpos 	 numeric 	 number of TB positive cases in the treated (vaccinated) group

tneg 	 numeric 	 number of TB negative cases in the treated (vaccinated) group

cpos 	 numeric 	 number of TB positive cases in the control group

cneg 	 numeric 	 number of TB negative cases in the control group

ablat 	 numeric 	 absolute latitude where the study was conducted

alloc 	 character 	 method of treatment allocation (random, alternate, or systematic)

Details

The 13 studies provide data in terms of 2x2 tables in the form:

TB positive 	 TB negative

vaccinated group 	 tpos 	 tneg

control group 	 cpos 	 cneg

The goal of the meta-analysis was to examine the overall effectiveness of the BCG vaccine for preventing tuberculosis and to examine moderators that may potentially influence the size of the effect.

The data set has been used in several publications to illustrate meta-analytic methods (see ‘References’).

Source

Colditz, G. A., Brewer, T. F., Berkey, C. S., Wilson, M. E., Burdick, E., Fineberg, H. V. & Mosteller, F. (1994) Efficacy of BCG vaccine in the prevention of tuberculosis: Meta-analysis of the published literature. Journal of the American Medical Association, 271, 698–702.

References

Berkey, C. S., Hoaglin, D. C., Mosteller, F. & Colditz, G. A. (1995). A random-effects regression model for meta-analysis. Statistics in Medicine, 14, 395–411.

van Houwelingen, H. C., Arends, L. R. & Stijnen, T. (2002). Advanced methods in meta-analysis: Multivariate approach and meta-regression. Statistics in Medicine, 21, 589–624.

Examples

### load BCG vaccine data
data(dat.bcg)

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, 
           measure="RR", data=dat.bcg, method="DL")
res

### average risk ratio with 95% CI
predict(res, transf=exp)

### mixed-effects model with absolute latitude as a moderator
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, 
           mods=ablat, measure="RR", data=dat.bcg, method="DL")
res

### predicted average risk ratios for 10-60 degrees absolute latitude
predict(res, newmods=c(10, 20, 30, 40, 50, 60), transf=exp)

[Package metafor version 0.5-7 Index]