Studies on Gender Differences in Grant and Fellowship Awards

Results from 21 studies on gender differences in grant and fellowship awards.

dat.bornmann2007

Format

The data frame contains the following columns:

study	character	study reference

obs	numeric	observation within study

doctype	character	document type

gender	character	gender of the study authors

year	numeric	(average) cohort year

org	character	funding organization / program

country	character	country of the funding organization / program

type	character	fellowship or grant application

discipline	character	discipline / field

waward	numeric	number of women who received a grant/fellowship award

wtotal	numeric	number of women who applied for an award

maward	numeric	number of men who received a grant/fellowship award

mtotal	numeric	number of men who applied for an award

Details

The studies in this dataset examine whether the chances of receiving a grant or fellowship award differs for men and women. Note that many studies provide multiple comparisons (e.g., for different years / cohorts / disciplines). A multilevel meta-analytic model can be used to account for the multilevel structure in these data.

Source

Bornmann, L., Mutz, R., & Daniel, H. (2007). Gender differences in grant peer review: A meta-analysis. Journal of Informetrics, 1(3), 226--238. https://doi.org/10.1016/j.joi.2007.03.001

References

Marsh, H. W., Bornmann, L., Mutz, R., Daniel, H.-D., & O'Mara, A. (2009). Gender effects in the peer reviews of grant proposals: A comprehensive meta-analysis comparing traditional and multilevel approaches. Review of Educational Research, 79(3), 1290--1326. https://doi.org/10.3102/0034654309334143

Examples

### copy data into 'dat' and examine data
dat <- dat.bornmann2007
dat

#>                    study obs    doctype gender   year      org        country
#> 1          Ackers (2000)   1       Grey    M&F 1996.0     MSCA         Europe
#> 2          Ackers (2000)   2       Grey    M&F 1996.0     MSCA         Europe
#> 3          Ackers (2000)   3       Grey    M&F 1996.0     MSCA         Europe
#> 4          Ackers (2000)   4       Grey    M&F 1996.0     MSCA         Europe
#> 5          Ackers (2000)   5       Grey    M&F 1996.0     MSCA         Europe
#> 6          Ackers (2000)   6       Grey    M&F 1996.0     MSCA         Europe
#> 7          Ackers (2000)   7       Grey    M&F 1996.0     MSCA         Europe
#> 8    Allmendinger (2002)   1    Article    M&F 1993.0      DFG         Europe
#> 9    Allmendinger (2002)   2    Article    M&F 1994.0      DFG         Europe
#> 10   Allmendinger (2002)   3    Article    M&F 1995.0      DFG         Europe
#> 11   Allmendinger (2002)   4    Article    M&F 1996.0      DFG         Europe
#> 12   Allmendinger (2002)   5    Article    M&F 1997.0      DFG         Europe
#> 13   Allmendinger (2002)   6    Article    M&F 1998.0      DFG         Europe
#> 14   Allmendinger (2002)   7    Article    M&F 1999.0      DFG         Europe
#> 15        Bazeley (1998)   1    Article      F 1995.0      ARC      Australia
#> 16       Bornmann (2005)   1    Article      M 1992.5      BIF         Europe
#> 17         Brouns (2000)   1    Article      F 1993.5 NWO/KNAW         Europe
#> 18         Brouns (2000)   2    Article      F 1993.5 NWO/KNAW         Europe
#> 19         Brouns (2000)   3    Article      F 1993.5 NWO/KNAW         Europe
#> 20         Brouns (2000)   4    Article      F 1993.5 NWO/KNAW         Europe
#> 21         Brouns (2000)   5    Article      F 1993.5 NWO/KNAW         Europe
#> 22         Dexter (2002)   1       Grey      M 2000.5       WT United Kingdom
#> 23         Dexter (2002)   2       Grey      M 2000.5       WT United Kingdom
#> 24         Dexter (2002)   3       Grey      M 2000.5       WT United Kingdom
#> 25          Emery (1992)   1    Article    M&F 1990.0      NIH  United States
#> 26        Friesen (1998)   1     Letter      M 1998.0      MRC         Canada
#> 27        Friesen (1998)   2     Letter      M 1998.0      MRC         Canada
#> 28        Friesen (1998)   3     Letter      M 1998.0      MRC         Canada
#> 29      Goldsmith (2002)   1       Grey      F 1979.0      NSF  United States
#> 30      Goldsmith (2002)   2       Grey      F 1993.0      NSF  United States
#> 31          Grant (1997)   1     Letter      M 1996.0       WT United Kingdom
#> 32          Grant (1997)   2     Letter      M 1995.5       WT United Kingdom
#> 33          Grant (1997)   3     Letter      M 1996.0      MRC United Kingdom
#> 34          Grant (1997)   4     Letter      M 1995.0      MRC United Kingdom
#> 35     Jayasinghe (2001)   1    Article      M 1996.0      ARC      Australia
#> 36            NSF (2005)   1       Grey   <NA> 1997.0      NSF  United States
#> 37            NSF (2005)   2       Grey   <NA> 1998.0      NSF  United States
#> 38            NSF (2005)   3       Grey   <NA> 1999.0      NSF  United States
#> 39            NSF (2005)   4       Grey   <NA> 2000.0      NSF  United States
#> 40            NSF (2005)   5       Grey   <NA> 2001.0      NSF  United States
#> 41            NSF (2005)   6       Grey   <NA> 2002.0      NSF  United States
#> 42            NSF (2005)   7       Grey   <NA> 2003.0      NSF  United States
#> 43            NSF (2005)   8       Grey   <NA> 2004.0      NSF  United States
#> 44           Over (1996)   1    Article      M 1993.0      ARC      Australia
#> 45       Sigelman (1987)   1    Article      M 1985.5      NSF  United States
#> 46        Taplick (2005)   1       Grey      M 2002.5     EMBO         Europe
#> 47        Taplick (2005)   2       Grey      M 2000.0     EMBO         Europe
#> 48        Taplick (2005)   3       Grey      M 2002.0     EMBO         Europe
#> 49          Viner (2004)   1    Article      M 1994.0    EPSRC United Kingdom
#> 50          Viner (2004)   2    Article      M 1998.0    EPSRC United Kingdom
#> 51           Ward (1998)   1    Article    M&F 1995.5    NHMRC      Australia
#> 52 Wellcome Trust (1997)   1       Grey    M&F 1995.0       WT United Kingdom
#> 53       Wenneras (1997)   1 Commentary      F 1995.0      MRC         Europe
#> 54        Willems (2001)   1       Grey      F 1997.0      DFG         Europe
#> 55        Willems (2001)   2       Grey      F 1998.0      DFG         Europe
#> 56        Willems (2001)   3       Grey      F 1999.0      DFG         Europe
#> 57        Willems (2001)   4       Grey      F 2000.0      DFG         Europe
#> 58           Wood (1997)   1       Book      F 1995.0      ARC      Australia
#> 59           Wood (1997)   2       Book      F 1995.0      ARC      Australia
#> 60           Wood (1997)   3       Book      F 1995.0      ARC      Australia
#> 61           Wood (1997)   4       Book      F 1995.0      ARC      Australia
#> 62           Wood (1997)   5       Book      F 1995.0      ARC      Australia
#> 63           Wood (1997)   6       Book      F 1995.0      ARC      Australia
#> 64           Wood (1997)   7       Book      F 1995.0      ARC      Australia
#> 65           Wood (1997)   8       Book      F 1995.0      ARC      Australia
#> 66           Wood (1997)   9       Book      F 1995.0      ARC      Australia
#>          type                discipline waward wtotal maward mtotal
#> 1  Fellowship         Physical Sciences    139    711    274   1029
#> 2  Fellowship         Physical Sciences     45    258    166    908
#> 3  Fellowship         Physical Sciences     44    236    219    928
#> 4  Fellowship         Physical Sciences     63    251     96    507
#> 5  Fellowship Social Sciences / Biology    157    910    252   1118
#> 6  Fellowship         Physical Sciences    114    589    460   2244
#> 7  Fellowship   Life Sciences / Biology    381   2027    489   2275
#> 8       Grant Social Sciences / Biology      8     13     53     72
#> 9       Grant Social Sciences / Biology      5      8     53     82
#> 10      Grant Social Sciences / Biology      6      8     63     97
#> 11      Grant Social Sciences / Biology      8     16     53     94
#> 12      Grant Social Sciences / Biology      4     11     43     92
#> 13      Grant Social Sciences / Biology     20     44     55     93
#> 14      Grant Social Sciences / Biology      5     15     70    116
#> 15      Grant         Multidisciplinary     11     56     82    344
#> 16 Fellowship   Life Sciences / Biology    204   1085    430   1612
#> 17 Fellowship Social Sciences / Biology      8     12     25     33
#> 18 Fellowship Social Sciences / Biology      6     16     26     44
#> 19 Fellowship         Physical Sciences     21     25    127    236
#> 20 Fellowship   Life Sciences / Biology      9     34     77    168
#> 21 Fellowship   Life Sciences / Biology     30     67     82    174
#> 22 Fellowship   Life Sciences / Biology     54    174     84    280
#> 23      Grant   Life Sciences / Biology     84    290    253    791
#> 24      Grant   Life Sciences / Biology      4     17     30     70
#> 25 Fellowship   Life Sciences / Biology     23     91     39    203
#> 26      Grant   Life Sciences / Biology    297   1167   1160   4368
#> 27 Fellowship   Life Sciences / Biology     20    143     64    386
#> 28 Fellowship   Life Sciences / Biology    111    858    222   1361
#> 29 Fellowship         Multidisciplinary    146   1447    378   2841
#> 30 Fellowship         Multidisciplinary    379   3858    491   4776
#> 31      Grant   Life Sciences / Biology     72    268    302   1097
#> 32 Fellowship   Life Sciences / Biology      9    105     12    220
#> 33      Grant   Life Sciences / Biology     49    167    163    617
#> 34 Fellowship   Life Sciences / Biology     11    107     28    169
#> 35      Grant         Multidisciplinary     98    461    537   2520
#> 36      Grant         Multidisciplinary   1950   5396   7859  24532
#> 37      Grant         Multidisciplinary   1938   5627   7323  22513
#> 38      Grant         Multidisciplinary   1682   5315   7428  23022
#> 39      Grant         Multidisciplinary   1949   5509   7778  23671
#> 40      Grant         Multidisciplinary   1894   5839   7867  25510
#> 41      Grant         Multidisciplinary   2012   6704   8203  27500
#> 42      Grant         Multidisciplinary   2090   7335   8495  31238
#> 43      Grant         Multidisciplinary   2118   8427   7923  33300
#> 44      Grant         Multidisciplinary     12     39     62    217
#> 45      Grant Social Sciences / Biology      2     17     31    129
#> 46 Fellowship   Life Sciences / Biology    307    578    274    492
#> 47 Fellowship   Life Sciences / Biology    544   3263    954   4440
#> 48 Fellowship   Life Sciences / Biology     32    268    110    800
#> 49      Grant         Physical Sciences    122    181   3256   5500
#> 50      Grant         Physical Sciences    129    199   3327   5232
#> 51 Fellowship   Life Sciences / Biology     43    120    102    301
#> 52      Grant   Life Sciences / Biology     15     23     55    115
#> 53 Fellowship         Physical Sciences      4     52     16     62
#> 54      Grant   Life Sciences / Biology    392    725   2219   3962
#> 55      Grant   Life Sciences / Biology    443    837   2370   4269
#> 56      Grant   Life Sciences / Biology    424    761   2467   4287
#> 57      Grant   Life Sciences / Biology    483    885   2388   4217
#> 58      Grant   Life Sciences / Biology     10     51     58    241
#> 59      Grant   Life Sciences / Biology      9     56     64    281
#> 60      Grant         Physical Sciences      2     13     48    198
#> 61      Grant         Physical Sciences      1     14     45    177
#> 62      Grant         Physical Sciences      5     25     66    291
#> 63      Grant Social Sciences / Biology     25    104     57    259
#> 64      Grant         Physical Sciences      3     13     63    240
#> 65      Grant Social Sciences / Biology     32    122     92    358
#> 66      Grant         Physical Sciences      6     17     76    346

### calculate log odds ratios and corresponding sampling variances
dat <- escalc(measure="OR", ai=waward, n1i=wtotal, ci=maward, n2i=mtotal, data=dat)

### fit multilevel meta-analytic model
res <- rma.mv(yi, vi, random = ~ 1 | study/obs, data=dat)
res

#> 
#> Multivariate Meta-Analysis Model (k = 66; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed     factor 
#> sigma^2.1  0.0161  0.1268     21     no      study 
#> sigma^2.2  0.0038  0.0613     66     no  study/obs 
#> 
#> Test for Heterogeneity:
#> Q(df = 65) = 221.2850, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval    ci.lb    ci.ub 
#>  -0.1010  0.0417  -2.4196  0.0155  -0.1828  -0.0192  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### estimated average odds ratio (with 95% CI/PI)
predict(res, transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  0.90  0.83  0.98  0.68  1.21 
#> 

### test for a difference between fellowship and grant applications
res <- rma.mv(yi, vi, mods = ~ type, random = ~ 1 | study/obs, data=dat)
res

#> 
#> Multivariate Meta-Analysis Model (k = 66; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed     factor 
#> sigma^2.1  0.0045  0.0670     21     no      study 
#> sigma^2.2  0.0035  0.0596     66     no  study/obs 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 64) = 133.4811, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 11.2312, p-val = 0.0008
#> 
#> Model Results:
#> 
#>            estimate      se     zval    pval    ci.lb    ci.ub 
#> intrcpt     -0.2010  0.0429  -4.6816  <.0001  -0.2852  -0.1169  *** 
#> typeGrant    0.1890  0.0564   3.3513  0.0008   0.0785   0.2995  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, newmods=0:1, transf=exp, digits=2)

#> 
#>   pred ci.lb ci.ub pi.lb pi.ub 
#> 1 0.82  0.75  0.89  0.67  0.99 
#> 2 0.99  0.92  1.06  0.82  1.20 
#>