Studies on the Treatment of Upper Gastrointestinal Bleeding by a Histamine H2 Antagonist

Results from studies examining the effectiveness of histamine H2 antagonists (cimetidine or ranitidine) in treating patients with acute upper gastrointestinal hemorrhage.

dat.collins1985a

Format

The data frame contains the following columns:

id	numeric	study number

trial	character	first author of trial

year	numeric	year of publication

ref	numeric	reference number

trt	character	C = cimetidine, R = ranitidine

ctrl	character	P = placebo, AA = antacids, UT = usual treatment

nti	numeric	number of patients in treatment group

b.xti	numeric	number of patients in treatment group with persistent or recurrent bleedings

o.xti	numeric	number of patients in treatment group in need of operation

d.xti	numeric	number of patients in treatment group that died

nci	numeric	number of patients in control group

b.xci	numeric	number of patients in control group with persistent or recurrent bleedings

o.xci	numeric	number of patients in control group in need of operation

d.xci	numeric	number of patients in control group that died

id	numeric

Details

The data were obtained from Tables 1 and 2 in Collins and Langman (1985). The authors used Peto's (one-step) method for meta-analyzing the 27 trials. This approach is implemented in the rma.peto function. Using the same dataset, van Houwelingen, Zwinderman, and Stijnen (1993) describe some alternative approaches for analyzing these data, including fixed and random-effects conditional logistic models. Those are implemented in the rma.glmm function.

Source

Collins, R., & Langman, M. (1985). Treatment with histamine H2 antagonists in acute upper gastrointestinal hemorrhage. New England Journal of Medicine, 313, 660--666.

References

van Houwelingen, H. C., Zwinderman, K. H., & Stijnen, T. (1993). A bivariate approach to meta-analysis. Statistics in Medicine, 12, 2273--2284.

Examples

### load data
dat <- get(data(dat.collins1985a))

### meta-analysis of log ORs using Peto's method (outcome: persistent or recurrent bleedings)
res <- rma.peto(ai=b.xti, n1i=nti, ci=b.xci, n2i=nci, data=dat)

#> Warning: Tables with NAs omitted from model fitting.

#> Warning: Some yi/vi values are NA.

print(res, digits=2)

#> 
#> Fixed-Effects Model (k = 25)
#> 
#> Test for Heterogeneity: 
#> Q(df = 23) = 37.91, p-val = 0.03
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.12  0.10  -1.22  0.22  -0.32   0.07 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.89   0.73   1.08 
#> 

# NOT RUN {
### meta-analysis of log ORs using a conditional logistic regression model (FE model)
res <- rma.glmm(measure="OR", ai=b.xti, n1i=nti, ci=b.xci, n2i=nci, data=dat,
                model="CM.EL", method="FE")
summary(res)
predict(res, transf=exp, digits=2)

### plot the log-likelihoods of the odds ratios
llplot(measure="OR", ai=b.xti, n1i=nti, ci=b.xci, n2i=nci, data=dat,
       lwd=1, refline=NA, xlim=c(-4,4), drop00=FALSE)

### meta-analysis of log odds ratios using a conditional logistic regression model (RE model)
res <- rma.glmm(measure="OR", ai=b.xti, n1i=nti, ci=b.xci, n2i=nci, data=dat,
                model="CM.EL", method="ML")
summary(res)
predict(res, transf=exp, digits=2)
# }

### meta-analysis of log ORs using Peto's method (outcome: need for surgery)
res <- rma.peto(ai=o.xti, n1i=nti, ci=o.xci, n2i=nci, data=dat)

#> Warning: Tables with NAs omitted from model fitting.

#> Warning: Some yi/vi values are NA.

print(res, digits=2)

#> 
#> Fixed-Effects Model (k = 22)
#> 
#> Test for Heterogeneity: 
#> Q(df = 19) = 25.46, p-val = 0.15
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.25  0.12  -1.97  0.05  -0.49  -0.00 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.78   0.61   1.00 
#> 

### meta-analysis of log ORs using Peto's method (outcome: death)
res <- rma.peto(ai=d.xti, n1i=nti, ci=d.xci, n2i=nci, data=dat)

#> Warning: Tables with NAs omitted from model fitting.

#> Warning: Some yi/vi values are NA.

print(res, digits=2)

#> 
#> Fixed-Effects Model (k = 24)
#> 
#> Test for Heterogeneity: 
#> Q(df = 21) = 27.33, p-val = 0.16
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.36  0.16  -2.21  0.03  -0.68  -0.04 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.70   0.51   0.96 
#>