Studies on the Relationship between Class Attendance and Grades in College Students

Results from 68 studies on the relationship between class attendence and class performance and/or grade point average in college students.

dat.crede2010

Format

The data frame contains the following columns:

studyid	numeric	study number

year	numeric	publication year

source	character	study source (journal, dissertation, other)

sampleid	numeric	sample within study number

criterion	character	criterion variable (grade, gpa)

class	character	class type (science, nonscience)

ni	numeric	sample size

ri	numeric	observed correlation

Details

The 68 studies included in this dataset provide information about the relationship between class attendance of college students and their performance (i.e., grade) in the class and/or their overall grade point average. Some studies included multiple samples and hence the dataset actually contains 97 correlation coefficients.

The dataset was obtained via personal communication. Note that this dataset differs just slightly from the one used by Credé et al. (2010).

Source

Personal communication.

References

Credé, M., Roch, S. G., & Kieszczynka, U. M. (2010). Class attendance in college: A meta-analytic review of the relationship of class attendance with grades and student characteristics. Review of Educational Research, 80(2), 272--295. https://doi.org/10.3102/0034654310362998

Examples

### copy data into 'dat'
dat <- dat.crede2010

### calculate r-to-z transformed correlations and corresponding sampling variances
dat <- escalc(measure="ZCOR", ri=ri, ni=ni, data=dat)

############################################################################

### meta-analysis for the relationship between attendance and grades
res <- rma(yi, vi, data=dat, subset=criterion=="grade")
res

#> 
#> Random-Effects Model (k = 67; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0511 (SE = 0.0104)
#> tau (square root of estimated tau^2 value):      0.2261
#> I^2 (total heterogeneity / total variability):   93.83%
#> H^2 (total variability / sampling variability):  16.21
#> 
#> Test for Heterogeneity:
#> Q(df = 66) = 1068.7213, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval   ci.lb   ci.ub 
#>   0.4547  0.0300  15.1343  <.0001  0.3958  0.5136  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### estimated average correlation with 95% CI/PI
predict(res, transf=transf.ztor, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  0.43  0.38  0.47  0.01  0.72 
#> 

### examine if relationship between attendance and grades differs for nonscience/science classes
res <- rma(yi, vi, mods = ~ class, data=dat, subset=criterion=="grade")
res

#> 
#> Mixed-Effects Model (k = 67; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0515 (SE = 0.0106)
#> tau (square root of estimated tau^2 value):             0.2269
#> I^2 (residual heterogeneity / unaccounted variability): 93.45%
#> H^2 (unaccounted variability / sampling variability):   15.26
#> R^2 (amount of heterogeneity accounted for):            0.00%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 65) = 1009.6949, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 0.6234, p-val = 0.4298
#> 
#> Model Results:
#> 
#>               estimate      se     zval    pval    ci.lb   ci.ub 
#> intrcpt         0.4441  0.0330  13.4603  <.0001   0.3794  0.5087  *** 
#> classscience    0.0640  0.0810   0.7895  0.4298  -0.0948  0.2228      
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### estimated average correlations for nonscience and science classes
predict(res, newmods=c(0,1), transf=transf.ztor, digits=2)

#> 
#>   pred ci.lb ci.ub pi.lb pi.ub 
#> 1 0.42  0.36  0.47 -0.01  0.71 
#> 2 0.47  0.35  0.57  0.04  0.75 
#> 

### examine if relationship between attendance and grades has changed over time
res <- rma(yi, vi, mods = ~ year, data=dat, subset=criterion=="grade")
res

#> 
#> Mixed-Effects Model (k = 67; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0504 (SE = 0.0104)
#> tau (square root of estimated tau^2 value):             0.2246
#> I^2 (residual heterogeneity / unaccounted variability): 93.70%
#> H^2 (unaccounted variability / sampling variability):   15.87
#> R^2 (amount of heterogeneity accounted for):            1.31%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 65) = 1050.5282, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 1.3863, p-val = 0.2390
#> 
#> Model Results:
#> 
#>          estimate      se     zval    pval     ci.lb   ci.ub 
#> intrcpt   -7.0137  6.3432  -1.1057  0.2689  -19.4462  5.4188    
#> year       0.0037  0.0032   1.1774  0.2390   -0.0025  0.0100    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

############################################################################

### meta-analysis for the relationship between attendance and GPA
res <- rma(yi, vi, data=dat, subset=criterion=="gpa")
res

#> 
#> Random-Effects Model (k = 30; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0555 (SE = 0.0165)
#> tau (square root of estimated tau^2 value):      0.2356
#> I^2 (total heterogeneity / total variability):   93.76%
#> H^2 (total variability / sampling variability):  16.02
#> 
#> Test for Heterogeneity:
#> Q(df = 29) = 430.9743, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.3724  0.0459  8.1071  <.0001  0.2824  0.4624  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### estimated average correlation with 95% CI/PI
predict(res, transf=transf.ztor, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  0.36  0.28  0.43 -0.10  0.69 
#> 

### examine if relationship between attendance and GPA has changed over time
res <- rma(yi, vi, mods = ~ year, data=dat, subset=criterion=="gpa")
res

#> 
#> Mixed-Effects Model (k = 30; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0569 (SE = 0.0173)
#> tau (square root of estimated tau^2 value):             0.2384
#> I^2 (residual heterogeneity / unaccounted variability): 93.61%
#> H^2 (unaccounted variability / sampling variability):   15.64
#> R^2 (amount of heterogeneity accounted for):            0.00%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 28) = 392.0912, p-val < .0001
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 0.3796, p-val = 0.5378
#> 
#> Model Results:
#> 
#>          estimate      se     zval    pval    ci.lb   ci.ub 
#> intrcpt    2.5890  3.5979   0.7196  0.4718  -4.4627  9.6407    
#> year      -0.0011  0.0018  -0.6161  0.5378  -0.0047  0.0024    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

############################################################################

### use a multilevel model to examine the relationship between attendance and grades
res <- rma.mv(yi, vi, random = ~ 1 | studyid/sampleid, data=dat, subset=criterion=="grade")
res

#> 
#> Multivariate Meta-Analysis Model (k = 67; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed            factor 
#> sigma^2.1  0.0376  0.1939     54     no           studyid 
#> sigma^2.2  0.0159  0.1259     67     no  studyid/sampleid 
#> 
#> Test for Heterogeneity:
#> Q(df = 66) = 1068.7213, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval   ci.lb   ci.ub 
#>   0.4798  0.0331  14.5167  <.0001  0.4151  0.5446  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=transf.ztor, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  0.45  0.39  0.50  0.02  0.73 
#> 

### use a multilevel model to examine the relationship between attendance and gpa
res <- rma.mv(yi, vi, random = ~ 1 | studyid/sampleid, data=dat, subset=criterion=="gpa")
res

#> 
#> Multivariate Meta-Analysis Model (k = 30; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed            factor 
#> sigma^2.1  0.0572  0.2392     24     no           studyid 
#> sigma^2.2  0.0063  0.0796     30     no  studyid/sampleid 
#> 
#> Test for Heterogeneity:
#> Q(df = 29) = 430.9743, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.3611  0.0538  6.7103  <.0001  0.2556  0.4665  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=transf.ztor, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  0.35  0.25  0.44 -0.14  0.70 
#>