Studies on the Effects of Elevated CO2 Levels on Woody Plant Mass

Results from studies examining the effects of elevated CO2 levels on woody plant mass.

dat.curtis1998

Format

The data frame contains the following columns:

id	numeric	observation number

paper	numeric	paper number

genus	character	genus name

species	character	species name

fungrp	character	plant functional group

co2.ambi	numeric	ambient CO2 level (control group)

co2.elev	numeric	elevated CO2 level (treatment group)

units	character	units for CO2 exposure levels

time	numeric	maximum length of time (days) of CO2 exposure

pot	character	growing method (see ‘Details’)

method	character	CO2 exposure facility (see ‘Details’)

stock	character	planting stock code

xtrt	character	interacting treatment code (see ‘Details’)

level	character	interacting treatment level codes (see ‘Details’)

m1i	numeric	mean plant mass under elevated CO2 level (treatment group)

sd1i	numeric	standard deviation of plant mass underelevated CO2 level (treatment group)

n1i	numeric	number of observations under elevated CO2 level (treatment group)

m2i	numeric	mean plant mass under ambient CO2 level (control group)

sd2i	numeric	standard deviation of plant mass under ambient CO2 level (control group)

n2i	numeric	number of observations under ambient CO2 level (control group)

Details

The studies included in this dataset compared the total above- plus below-ground biomass (in grams) for plants that were either exposed to ambient (around 35 Pa) and elevated CO2 levels (around twice the ambient level). The co2.ambi and co2.elev variables indicate the CO2 levels in the control and treatment groups, respectively (with the units variable specifying the units for the CO2 exposure levels). Many of the studies also varied one or more additional environmental variables (defined by the xtrt and level variables):

NONE = no additional treatment factor

FERT = soil fertility (either a CONTROL, HIGH, or LOW level)

LIGHT = light treatment (always a LOW light level)

FERT+L = soil fertility and light (a LOW light and soil fertility level)

H2O = well watered vs drought (either a WW or DRT level)

TEMP = temperature treatment (either a HIGH or LOW level)

OZONE = ozone exposure (either a HIGH or LOW level)

UVB = ultraviolet-B radiation exposure (either a HIGH or LOW level)

In addition, the studies differed with respect to various design variables, including CO2 exposure duration (time), growing method (pot: number = pot size in liters; GRND = plants rooted in ground; HYDRO = solution or aeroponic culture), CO2 exposure facility (method: GC = growth chamber; GH = greenhouse; OTC = field-based open-top chamber), and planting stock (stock: SEED = plants started from seeds; SAP = plants started from cuttings). The goal of the meta-analysis was to examine the effects of elevated CO2 levels on plant physiology and growth and the interacting effects of the environmental (and design) variables.

Source

Hedges, L. V., Gurevitch, J., & Curtis, P. S. (1999). The meta-analysis of response ratios in experimental ecology. Ecology, 80(4), 1150--1156. https://doi.org/10.1890/0012-9658(1999)080[1150:TMAORR]2.0.CO;2 (data obtained from Ecological Archives, E080-008-S1, at: http://www.esapubs.org/archive/ecol/E080/008/)

References

Curtis, P. S., & Wang, X. (1998). A meta-analysis of elevated CO2 effects on woody plant mass, form, and physiology. Oecologia, 113(3), 299--313. https://doi.org/10.1007/s004420050381

Examples

### copy data into 'dat'
dat <- dat.curtis1998

### calculate (log transformed) ratios of means and corresponding sampling variances
dat <- escalc(measure="ROM", m1i=m1i, sd1i=sd1i, n1i=n1i,
                             m2i=m2i, sd2i=sd2i, n2i=n2i, data=dat)
dat

#>      id paper         genus        species fungrp co2.ambi co2.elev    units 
#> 1    21    44         ALNUS          RUBRA  N2FIX    350.0    650.0     ul/l 
#> 2    22    44         ALNUS          RUBRA  N2FIX    350.0    650.0     ul/l 
#> 3    27   121          ACER         RUBRUM  ANGIO    350.0    700.0      ppm 
#> 4    32   121       QUERCUS         PRINUS  ANGIO    350.0    700.0      ppm 
#> 5    35   121         MALUS      DOMESTICA  ANGIO    350.0    700.0      ppm 
#> 6    38   121          ACER    SACCHARINUM  ANGIO    350.0    700.0      ppm 
#> 7    44   159      CASTANEA         SATIVA  ANGIO    350.0    700.0      ppm 
#> 8    63   183        CITRUS       SINENSIS  ANGIO    395.0    795.0      ppm 
#> 9    86   209      CASTANEA         SATIVA  ANGIO    350.0    700.0      ppm 
#> 10   87   209      CASTANEA         SATIVA  ANGIO    350.0    700.0      ppm 
#> 11   95   210      CASTANEA         SATIVA  ANGIO    350.0    700.0      ppm 
#> 12   96   210      CASTANEA         SATIVA  ANGIO    350.0    700.0      ppm 
#> 13  120   290         PINUS        RADIATA  GYMNO    340.0    640.0     ul/l 
#> 14  125   290    NOTHOFAGUS          FUSCA  ANGIO    340.0    640.0     ul/l 
#> 15  130   290   PSEUDOTSUGA      MENZIESII  GYMNO    340.0    640.0     ul/l 
#> 16  163   468      CASTANEA         SATIVA  ANGIO    350.0    700.0 umol/mol 
#> 17  170   470      CASTANEA         SATIVA  ANGIO    350.0    700.0      ppm 
#> 18  182   503         PINUS       ECHINATA  GYMNO    368.0    695.0     ul/l 
#> 19  208   505       QUERCUS           ALBA  ANGIO    389.0    793.0   cm3/m3 
#> 20  209   505       QUERCUS           ALBA  ANGIO    389.0    793.0   cm3/m3 
#> 21  230   506  LIRIODENDRON     TULIPIFERA  ANGIO    371.0    787.0      ppm 
#> 22  231   506  LIRIODENDRON     TULIPIFERA  ANGIO    371.0    787.0      ppm 
#> 23  236   510       QUERCUS           ALBA  ANGIO    360.0    700.0     ul/l 
#> 24  242   510         PINUS       ECHINATA  GYMNO    360.0    700.0     ul/l 
#> 25  254   550        BETULA        PENDULA  ANGIO    350.0    700.0 umol/mol 
#> 26  256   553         PICEA          ABIES  GYMNO    350.0    750.0   cm3/m3 
#> 27  257   553         PICEA          ABIES  GYMNO    350.0    750.0   cm3/m3 
#> 28  267   582         PIPER        AURITUM  ANGIO    350.0    700.0      ppm 
#> 29  277   582         SENNA      MULTIJUGA  ANGIO    350.0    700.0      ppm 
#> 30  287   582    MYRIOCARPA       LONGIPES  ANGIO    350.0    700.0      ppm 
#> 31  296   582      CECROPIA    OBTUSIFOLIA  ANGIO    350.0    700.0      ppm 
#> 32  306   582 TRICHOSPERMUM      MEXICANUM  ANGIO    350.0    700.0      ppm 
#> 33  313   596        BETULA          LENTA  ANGIO    350.0    700.0     ul/l 
#> 34  320   596        BETULA     PAPYRIFERA  ANGIO    350.0    700.0     ul/l 
#> 35  327   596        BETULA    POPULIFOLIA  ANGIO    350.0    700.0     ul/l 
#> 36  334   596        BETULA ALLEGHANIENSIS  ANGIO    350.0    700.0     ul/l 
#> 37  343   666         PINUS      BANKSIANA  GYMNO    350.0    750.0 umol/mol 
#> 38  344   666         PINUS      BANKSIANA  GYMNO    350.0    750.0 umol/mol 
#> 39  392  2003        BETULA      PUBESCENS  ANGIO    350.0    700.0 umol/mol 
#> 40  393  2003        BETULA      PUBESCENS  ANGIO    350.0    700.0 umol/mol 
#> 41  394  2003        BETULA      PUBESCENS  ANGIO    350.0    560.0 umol/mol 
#> 42  395  2003        BETULA      PUBESCENS  ANGIO    350.0    560.0 umol/mol 
#> 43  410  2026         PINUS      PONDEROSA  GYMNO    350.0    650.0     ubar 
#> 44  411  2026         PINUS      PONDEROSA  GYMNO    350.0    650.0     ubar 
#> 45  441  2037          ACER         RUBRUM  ANGIO    350.0    700.0     ul/l 
#> 46  442  2037          ACER         RUBRUM  ANGIO    350.0    700.0     ul/l 
#> 47  443  2037          ACER         RUBRUM  ANGIO    350.0    700.0     ul/l 
#> 48  444  2037       QUERCUS          RUBRA  ANGIO    350.0    700.0     ul/l 
#> 49  445  2037       QUERCUS          RUBRA  ANGIO    350.0    700.0     ul/l 
#> 50  446  2037       QUERCUS          RUBRA  ANGIO    350.0    700.0     ul/l 
#> 51  447  2037          ACER  PENSYLVANICUM  ANGIO    350.0    700.0     ul/l 
#> 52  448  2037          ACER  PENSYLVANICUM  ANGIO    350.0    700.0     ul/l 
#> 53  449  2037          ACER  PENSYLVANICUM  ANGIO    350.0    700.0     ul/l 
#> 54  450  2037        BETULA    POPULIFOLIA  ANGIO    350.0    700.0     ul/l 
#> 55  451  2037        BETULA    POPULIFOLIA  ANGIO    350.0    700.0     ul/l 
#> 56  452  2037        BETULA    POPULIFOLIA  ANGIO    350.0    700.0     ul/l 
#> 57  453  2037      FRAXINUS      AMERICANA  ANGIO    350.0    700.0     ul/l 
#> 58  454  2037      FRAXINUS      AMERICANA  ANGIO    350.0    700.0     ul/l 
#> 59  455  2037      FRAXINUS      AMERICANA  ANGIO    350.0    700.0     ul/l 
#> 60  456  2037        BETULA ALLEGHANIENSIS  ANGIO    350.0    700.0     ul/l 
#> 61  457  2037        BETULA ALLEGHANIENSIS  ANGIO    350.0    700.0     ul/l 
#> 62  458  2037        BETULA ALLEGHANIENSIS  ANGIO    350.0    700.0     ul/l 
#> 63  479  2039      POPULUSX   EURAMERICANA  ANGIO     34.5     69.3       Pa 
#> 64  480  2039      POPULUSX   EURAMERICANA  ANGIO     34.5     69.3       Pa 
#> 65  523  2045         PICEA        MARIANA  GYMNO    340.0    700.0      ppm 
#> 66  524  2045         PICEA        MARIANA  GYMNO    340.0    700.0      ppm 
#> 67  525  2045         PICEA        MARIANA  GYMNO    340.0    700.0      ppm 
#> 68  526  2045         PICEA        MARIANA  GYMNO    340.0    700.0      ppm 
#> 69  547  2048         PICEA         GLAUCA  GYMNO    350.0    700.0 umol/mol 
#> 70  548  2048         PICEA         GLAUCA  GYMNO    350.0    700.0 umol/mol 
#> 71  561  2048         PICEA        MARIANA  GYMNO    350.0    700.0 umol/mol 
#> 72  562  2048         PICEA        MARIANA  GYMNO    350.0    700.0 umol/mol 
#> 73  575  2048         PINUS      BANKSIANA  GYMNO    350.0    700.0 umol/mol 
#> 74  576  2048         PINUS      BANKSIANA  GYMNO    350.0    700.0 umol/mol 
#> 75  614  2110       POPULUS   EURAMERICANA  ANGIO     34.5     69.3       Pa 
#> 76  615  2110       POPULUS   EURAMERICANA  ANGIO     34.5     69.3       Pa 
#> 77  623  2117         PICEA          ABIES  GYMNO    350.0    700.0 umol/mol 
#> 78  624  2117         PICEA          ABIES  GYMNO    350.0    700.0 umol/mol 
#> 79  636  2117        BETULA        PENDULA  ANGIO    350.0    700.0 umol/mol 
#> 80  637  2117        BETULA        PENDULA  ANGIO    350.0    700.0 umol/mol 
#> 81  721  2217     MARANTHES      CORYMBOSA  ANGIO    350.0    700.0 umol/mol 
#> 82  725  2223          ACER         RUBRUM  ANGIO    350.0    700.0     ul/l 
#> 83  726  2223          ACER         RUBRUM  ANGIO    350.0    700.0     ul/l 
#> 84  727  2223          ACER         RUBRUM  ANGIO    350.0    700.0     ul/l 
#> 85  731  2223       QUERCUS          RUBRA  ANGIO    350.0    700.0     ul/l 
#> 86  732  2223       QUERCUS          RUBRA  ANGIO    350.0    700.0     ul/l 
#> 87  733  2223       QUERCUS          RUBRA  ANGIO    350.0    700.0     ul/l 
#> 88  737  2223          ACER  PENSYLVANICUM  ANGIO    350.0    700.0     ul/l 
#> 89  738  2223          ACER  PENSYLVANICUM  ANGIO    350.0    700.0     ul/l 
#> 90  739  2223          ACER  PENSYLVANICUM  ANGIO    350.0    700.0     ul/l 
#> 91  743  2223        BETULA    POPULIFOLIA  ANGIO    350.0    700.0     ul/l 
#> 92  744  2223        BETULA    POPULIFOLIA  ANGIO    350.0    700.0     ul/l 
#> 93  745  2223        BETULA    POPULIFOLIA  ANGIO    350.0    700.0     ul/l 
#> 94  749  2223      FRAXINUS      AMERICANA  ANGIO    350.0    700.0     ul/l 
#> 95  750  2223      FRAXINUS      AMERICANA  ANGIO    350.0    700.0     ul/l 
#> 96  751  2223      FRAXINUS      AMERICANA  ANGIO    350.0    700.0     ul/l 
#> 97  755  2223        BETULA ALLEGHANIENSIS  ANGIO    350.0    700.0     ul/l 
#> 98  756  2223        BETULA ALLEGHANIENSIS  ANGIO    350.0    700.0     ul/l 
#> 99  757  2223        BETULA ALLEGHANIENSIS  ANGIO    350.0    700.0     ul/l 
#> 100 765  2224       QUERCUS          RUBRA  ANGIO    385.0    642.0     ul/l 
#> 101 774  2224          ACER      SACCHARUM  ANGIO    385.0    642.0     ul/l 
#> 102 783  2224       POPULUS    TREMULOIDES  ANGIO    385.0    642.0     ul/l 
#>     time   pot method stock   xtrt   level        m1i        sd1i n1i 
#> 1     47   0.5     GC  SEED   FERT    HIGH    6.81690   1.7699820   3 
#> 2     47   0.5     GC  SEED   FERT CONTROL    2.59610   0.6674662   5 
#> 3     59   2.6     GH  SEED   NONE       .    2.99000   0.8560000   5 
#> 4     70   2.6     GH  SEED   NONE       .    5.91000   1.7420000   5 
#> 5     64   2.6     GH  SEED   NONE       .    4.61000   1.4070000   4 
#> 6     50   2.6     GH  SEED   NONE       .   10.78000   1.1630000   5 
#> 7    730  GRND     GC   SAP   NONE       .  153.50000  27.1932000   3 
#> 8    365     9     GH   SAP   NONE       . 1439.00000 142.0280000   3 
#> 9    365    24     GH  SEED   FERT    HIGH  183.60000  39.0600000  20 
#> 10   365    24     GH  SEED   FERT CONTROL   71.72000  14.3000000  16 
#> 11   120    12     GH   SAP   NONE       .   45.80000  15.8000000  20 
#> 12   120    12     GH   SAP   NONE       .   72.50000  15.0000000  24 
#> 13   120     4     GC   SAP   NONE       .   59.10000  12.9000000   8 
#> 14   120     4     GC   SAP   NONE       .    4.90000   1.9000000   8 
#> 15   120     4     GC   SAP   NONE       .   19.00000   5.4000000   8 
#> 16  1095    24    OTC   SAP   NONE       .  146.00000  10.0000000   5 
#> 17   180    24    OTC   SAP   NONE       .  129.00000  59.0000000  12 
#> 18   287  0.95     GC  SEED   NONE       .    6.91000   3.2200000   5 
#> 19   168   2.6     GH  SEED   FERT    HIGH   17.89000  10.1850000   5 
#> 20   168   2.6     GH  SEED   FERT CONTROL   14.68000   5.0130000   5 
#> 21   168   3.5     GC  SEED   FERT    HIGH   63.97000   7.6600000   5 
#> 22   168   3.5     GC  SEED   FERT CONTROL    6.01000   1.0200000   5 
#> 23   210  0.95     GC  SEED   NONE       .   14.15000   8.6467000   6 
#> 24   168  0.95     GC  SEED   NONE       .    2.20000   0.4111000  10 
#> 25    40 HYDRO     GC  SEED   NONE       .   23.11000   0.8443000  22 
#> 26   180    50     GC   SAP  OZONE    HIGH  262.10000  16.4000000   6 
#> 27   180    50     GC   SAP  OZONE     LOW  273.30000  14.0000000   6 
#> 28   111  0.67     GH  SEED   NONE       .   10.14000   2.5200000   4 
#> 29   111  0.67     GH  SEED   NONE       .   15.72000   2.7600000   4 
#> 30   111  0.67     GH  SEED   NONE       .    9.01000   2.5800000   4 
#> 31   111  0.67     GH  SEED   NONE       .   21.33000   3.0600000   4 
#> 32   111  0.67     GH  SEED   NONE       .   14.68000   2.8200000   4 
#> 33    90  0.66     GH  SEED   NONE       .    8.36000   3.9150000  10 
#> 34    90  0.66     GH  SEED   NONE       .    7.91000   1.8090000  10 
#> 35    90  0.66     GH  SEED   NONE       .   12.45000   6.9440000  10 
#> 36    90  0.66     GH  SEED   NONE       .    7.84000   3.3240000  10 
#> 37   270  0.04     GH  SEED    UVB    HIGH    0.09316   0.0714040  10 
#> 38   270  0.04     GH  SEED    UVB     LOW    0.59046   0.4103690  10 
#> 39    34   0.2     GC  SEED   TEMP    HIGH    4.77000   0.1980000   2 
#> 40    34   0.2     GC  SEED   TEMP     LOW    3.74000   0.0424000   2 
#> 41    35   0.3     GC  SEED  OZONE    HIGH    3.15000   0.2970000   2 
#> 42    35   0.3     GC  SEED  OZONE     LOW    3.78000   0.4240000   2 
#> 43    60   1.8     GC  SEED   TEMP    HIGH    2.15000   0.3600000  16 
#> 44    60   1.8     GC  SEED   TEMP     LOW    1.76000   0.2800000  16 
#> 45  1095   2.8     GH  SEED   FERT     LOW   47.38000   5.0150000   2 
#> 46  1095   2.8     GH  SEED  LIGHT     LOW  258.53000  50.3150000   2 
#> 47  1095   2.8     GH  SEED FERT+L       .  482.04000  14.8010000   2 
#> 48  1095   2.8     GH  SEED   FERT     LOW   30.59000   2.1520000   2 
#> 49  1095   2.8     GH  SEED  LIGHT     LOW  172.40000  23.6780000   2 
#> 50  1095   2.8     GH  SEED FERT+L       .  277.46000  14.8340000   2 
#> 51  1095   2.8     GH  SEED   FERT     LOW   23.08000   4.2940000   2 
#> 52  1095   2.8     GH  SEED  LIGHT     LOW  196.80000  26.6140000   2 
#> 53  1095   2.8     GH  SEED FERT+L       .  254.39000  32.6240000   2 
#> 54  1095   2.8     GH  SEED   FERT     LOW   31.10000   1.0730000   2 
#> 55  1095   2.8     GH  SEED  LIGHT     LOW  186.00000  17.7290000   2 
#> 56  1095   2.8     GH  SEED FERT+L       .  232.37000   8.8670000   2 
#> 57  1095   2.8     GH  SEED   FERT     LOW   25.19000   2.1510000   2 
#> 58  1095   2.8     GH  SEED  LIGHT     LOW  120.80000  20.7730000   2 
#> 59  1095   2.8     GH  SEED FERT+L       .  262.77000  17.8320000   2 
#> 60  1095   2.8     GH  SEED   FERT     LOW   38.36000   3.2150000   2 
#> 61  1095   2.8     GH  SEED  LIGHT     LOW  178.33000  11.8330000   2 
#> 62  1095   2.8     GH  SEED FERT+L       .  186.26000  11.8340000   2 
#> 63   148  GRND    OTC   SAP   FERT    HIGH  113.00000   4.4720000   5 
#> 64   148  GRND    OTC   SAP   FERT     LOW   75.00000   8.9440000   5 
#> 65   155 0.164     GC  SEED    H2O      WW    4.77300   2.1477000  48 
#> 66   155 0.164     GC  SEED    H2O     DRT    3.27200   0.9215000  48 
#> 67   155 0.164     GC  SEED   FERT    HIGH    5.24900   1.6628000  48 
#> 68   155 0.164     GC  SEED   FERT     LOW    2.79800   0.5958000  48 
#> 69   112   0.2     GC  SEED    UVB    HIGH    0.33300   0.1470000  10 
#> 70   112   0.2     GC  SEED    UVB     LOW    0.64000   0.1540000  10 
#> 71   112   0.2     GC  SEED    UVB    HIGH    0.49800   0.1630000  10 
#> 72   112   0.2     GC  SEED    UVB     LOW    0.82400   0.2340000  10 
#> 73   112   0.2     GC  SEED    UVB    HIGH    0.84300   0.2340000  10 
#> 74   112   0.2     GC  SEED    UVB     LOW    1.53300   0.3950000  10 
#> 75   158  GRND    OTC   SAP   FERT    HIGH  562.80000  62.3863000   5 
#> 76   158  GRND    OTC   SAP   FERT     LOW  374.40000 109.5673000   5 
#> 77    47   0.2     GH  SEED   TEMP    HIGH    1.09200   0.0297000   2 
#> 78    47   0.2     GH  SEED   TEMP     LOW    1.12000   0.0141000   2 
#> 79    35   0.2     GH   SAP   TEMP    HIGH    6.91000   0.3111000   2 
#> 80    35   0.2     GH   SAP   TEMP     LOW    6.94000   0.4243000   2 
#> 81   210    10    OTC  SEED   NONE       .   16.95900   2.1860000   5 
#> 82   165  1.25     GH  SEED   FERT     LOW    2.40000   0.5634000   6 
#> 83   165  1.25     GH  SEED  LIGHT     LOW   16.98000   1.9596000   6 
#> 84   165  1.25     GH  SEED FERT+L       .   28.18000   3.2333000   6 
#> 85   165  1.25     GH  SEED   FERT     LOW    3.09000   0.7348000   6 
#> 86   165  1.25     GH  SEED  LIGHT     LOW   13.80000   2.4250000   6 
#> 87   165  1.25     GH  SEED FERT+L       .   18.20000   3.1843000   6 
#> 88   165  1.25     GH  SEED   FERT     LOW    1.90000   0.1225000   6 
#> 89   165  1.25     GH  SEED  LIGHT     LOW   10.23000   0.6124000   6 
#> 90   165  1.25     GH  SEED FERT+L       .   16.59000   2.9149000   6 
#> 91   165  1.25     GH  SEED   FERT     LOW    2.34000   0.2694000   6 
#> 92   165  1.25     GH  SEED  LIGHT     LOW   16.22000   2.8414000   6 
#> 93   165  1.25     GH  SEED FERT+L       .   24.55000   1.3962000   6 
#> 94   165  1.25     GH  SEED   FERT     LOW    1.51000   0.1715000   6 
#> 95   165  1.25     GH  SEED  LIGHT     LOW    9.12000   1.5922000   6 
#> 96   165  1.25     GH  SEED FERT+L       .   18.62000   2.1556000   6 
#> 97   165  1.25     GH  SEED   FERT     LOW    1.90000   0.2205000   6 
#> 98   165  1.25     GH  SEED  LIGHT     LOW   14.45000   1.6657000   6 
#> 99   165  1.25     GH  SEED FERT+L       .   16.98000   0.5389000   6 
#> 100   60     6     GC  SEED   NONE       .   16.90000   1.7321000   3 
#> 101   60     6     GC  SEED   NONE       .    7.20000   1.7321000   3 
#> 102   60     6     GC  SEED   NONE       .  102.60000   6.2354000   3 
#>            m2i       sd2i n2i      yi     vi 
#> 1      3.94500  1.1157970   5  0.5470 0.0385 
#> 2      2.25120  0.3275839   5  0.1425 0.0175 
#> 3      1.93000  0.5520000   5  0.4378 0.0328 
#> 4      6.62000  1.6310000   5 -0.1134 0.0295 
#> 5      4.10000  1.2570000   4  0.1172 0.0468 
#> 6      6.42000  2.0260000   3  0.5183 0.0355 
#> 7    127.30000 47.4582000   3  0.1872 0.0568 
#> 8   1140.60000 82.9650000   3  0.2324 0.0050 
#> 9    144.10000 25.7000000  20  0.2423 0.0039 
#> 10    59.94000 14.2800000  16  0.1794 0.0060 
#> 11    28.20000 11.9000000  23  0.4850 0.0137 
#> 12    60.50000 14.0000000  24  0.1809 0.0040 
#> 13    46.40000 13.1000000   8  0.2419 0.0159 
#> 14     4.20000  1.0000000   8  0.1542 0.0259 
#> 15    18.50000  8.1000000   8  0.0267 0.0341 
#> 16   136.00000 35.0000000   5  0.0710 0.0142 
#> 17    90.40000 36.0000000  12  0.3556 0.0306 
#> 18     6.81000  1.5700000   5  0.0146 0.0541 
#> 19    14.62000  2.2940000   5  0.2019 0.0697 
#> 20    11.18000  5.7490000   5  0.2724 0.0762 
#> 21    54.03000  1.6200000   5  0.1689 0.0030 
#> 22     4.93000  1.3500000   5  0.1981 0.0208 
#> 23     8.20000  3.3558000   6  0.5456 0.0901 
#> 24     1.33000  0.1897000  10  0.5033 0.0055 
#> 25    14.94000  0.8443000  22  0.4362 0.0002 
#> 26   235.50000 15.0000000   6  0.1070 0.0013 
#> 27   236.40000  7.5000000   6  0.1450 0.0006 
#> 28     8.81000  2.0600000   4  0.1406 0.0291 
#> 29    13.22000  1.9400000   4  0.1732 0.0131 
#> 30     9.80000  2.0600000   4 -0.0840 0.0315 
#> 31    18.60000  2.3400000   4  0.1370 0.0091 
#> 32    12.18000  2.4600000   4  0.1867 0.0194 
#> 33     5.30000  6.6250000  10  0.4558 0.1782 
#> 34     3.60000  3.0260000  10  0.7872 0.0759 
#> 35     4.98000  3.9370000  10  0.9163 0.0936 
#> 36     4.01000  2.1060000  10  0.6704 0.0456 
#> 37     0.07382  0.0832940  10  0.2327 0.1861 
#> 38     0.41117  0.2557020  10  0.3619 0.0870 
#> 39     3.83000  0.1414000   2  0.2195 0.0015 
#> 40     3.40000  0.0566000   2  0.0953 0.0002 
#> 41     2.40000  0.0140000   2  0.2719 0.0045 
#> 42     3.17000  0.4240000   2  0.1760 0.0152 
#> 43     1.70000  0.3200000  16  0.2348 0.0040 
#> 44     1.66000  0.2800000  16  0.0585 0.0034 
#> 45    32.99000  5.3750000   2  0.3620 0.0189 
#> 46   221.00000 32.5420000   2  0.1568 0.0298 
#> 47   396.31000 41.6970000   2  0.1958 0.0060 
#> 48    44.80000  7.8790000   2 -0.3815 0.0179 
#> 49    44.93000 11.8450000   2  1.3447 0.0442 
#> 50   187.67000 26.6350000   2  0.3910 0.0115 
#> 51    18.80000  4.6680000   2  0.2051 0.0481 
#> 52   109.07000 29.5740000   2  0.5902 0.0459 
#> 53   168.88000 20.7340000   2  0.4097 0.0158 
#> 54    29.36000  3.9430000   2  0.0576 0.0096 
#> 55   161.05000  8.8760000   2  0.1440 0.0061 
#> 56   232.69000 17.8970000   2 -0.0014 0.0037 
#> 57    23.20000  1.0780000   2  0.0823 0.0047 
#> 58   114.68000 26.6250000   2  0.0520 0.0417 
#> 59   168.74000 26.8300000   2  0.4429 0.0149 
#> 60    26.49000  1.7900000   2  0.3702 0.0058 
#> 61   161.74000 14.8560000   2  0.0976 0.0064 
#> 62   180.48000 11.8670000   2  0.0315 0.0042 
#> 63    76.00000  6.7080000   5  0.3967 0.0019 
#> 64    60.00000  8.9440000   5  0.2231 0.0073 
#> 65     3.86200  1.4272000  48  0.2118 0.0071 
#> 66     2.49000  0.6582000  48  0.2731 0.0031 
#> 67     4.03100  1.2956000  48  0.2640 0.0042 
#> 68     2.35000  0.5820000  48  0.1745 0.0022 
#> 69     0.24300  0.1180000  10  0.3151 0.0431 
#> 70     0.44300  0.1300000  10  0.3679 0.0144 
#> 71     0.32800  0.1270000  10  0.4176 0.0257 
#> 72     0.55000  0.1120000  10  0.4043 0.0122 
#> 73     0.59900  0.1320000  10  0.3417 0.0126 
#> 74     0.84200  0.1510000  10  0.5992 0.0099 
#> 75   381.60000 64.3988000   5  0.3886 0.0082 
#> 76   298.20000 96.3745000   5  0.2276 0.0380 
#> 77     0.97100  0.0255000   2  0.1174 0.0007 
#> 78     0.97500  0.0127000   2  0.1386 0.0002 
#> 79     5.95000  0.1980000   2  0.1496 0.0016 
#> 80     5.78000  0.0990000   2  0.1829 0.0020 
#> 81    10.57800  1.2100000   5  0.4720 0.0059 
#> 82     2.24000  0.3919000   6  0.0690 0.0143 
#> 83    10.96000  1.9351000   6  0.4378 0.0074 
#> 84    19.50000  2.2535000   6  0.3682 0.0044 
#> 85     2.04000  0.3674000   6  0.4152 0.0148 
#> 86     4.36000  2.4740000   6  1.1522 0.0588 
#> 87    14.79000  3.5028000   6  0.2075 0.0145 
#> 88     1.51000  0.1715000   6  0.2297 0.0028 
#> 89     5.89000  1.3962000   6  0.5521 0.0100 
#> 90    13.80000  1.5922000   6  0.1841 0.0074 
#> 91     2.24000  0.3919000   6  0.0437 0.0073 
#> 92    11.48000  2.0086000   6  0.3456 0.0102 
#> 93    19.95000  2.3025000   6  0.2075 0.0028 
#> 94     2.14000  0.4899000   6 -0.3487 0.0109 
#> 95     5.49000  1.3227000   6  0.5075 0.0148 
#> 96    13.49000  0.7593000   6  0.3223 0.0028 
#> 97     1.48000  0.1715000   6  0.2498 0.0045 
#> 98     8.71000  1.5187000   6  0.5062 0.0073 
#> 99    16.59000  0.9553000   6  0.0232 0.0007 
#> 100    7.20000  2.5981000   3  0.8532 0.0469 
#> 101    4.60000  1.7321000   3  0.4480 0.0666 
#> 102   69.70000  3.6373000   3  0.3866 0.0021 

### meta-analysis using a random-effects model
res <- rma(yi, vi, method="DL", data=dat)
res

#> 
#> Random-Effects Model (k = 102; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0216 (SE = 0.0088)
#> tau (square root of estimated tau^2 value):      0.1471
#> I^2 (total heterogeneity / total variability):   86.87%
#> H^2 (total variability / sampling variability):  7.61
#> 
#> Test for Heterogeneity:
#> Q(df = 101) = 769.0185, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval   ci.lb   ci.ub 
#>   0.2531  0.0185  13.6965  <.0001  0.2168  0.2893  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### average ratio of means with 95% CI
predict(res, transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  1.29  1.24  1.34  0.96  1.72 
#> 

### meta-analysis for plants grown under nutrient stress
res <- rma(yi, vi, method="DL", data=dat, subset=(xtrt=="FERT" & level=="LOW"))
predict(res, transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  1.14  1.04  1.26  0.82  1.59 
#> 

### meta-analysis for plants grown under low light conditions
res <- rma(yi, vi, method="DL", data=dat, subset=(xtrt=="LIGHT" & level=="LOW"))
predict(res, transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  1.57  1.34  1.85  0.94  2.63 
#>