Studies on the Effectiveness of Antihistamines in Reducing Symptoms of the Common Cold

Results from 9 studies on the effectiveness of antihistamines in reducing the severity of runny nose and sneezing in the common cold.

dat.dagostino1998

Format

The data frame contains the following columns:

study	numeric	study id

cold	character	natural or induced cold study

scale.rn	character	scale for measuring runny nose severity

scale.sn	character	scale for measuring sneezing severity

drug	character	type of antihistamine studied

tnt	numeric	total sample size of the treatment group

tnc	numeric	total sample size of the control (placebo) group

outcome	character	outcome variable (see ‘Details’)

mt	numeric	mean in the treatment group

sdt	numeric	SD in the treatment group

mc	numeric	mean in the control group

sdc	numeric	SD in the control group

xt	numeric	number of patients reaching the therapy goal in the treatment group

xc	numeric	number of patients reaching the therapy goal in the control (placebo) group

nt	numeric	sample size of the treatment group for measuring the outcome

nc	numeric	sample size of the control group for measuring the outcome

Details

The studies for this meta-analysis were assembled to examine the effectiveness of antihistamines in reducing the severity of runny nose and sneezing in the common cold. Effectiveness was measured after one and two days of treatment in terms of 4 different outcome variables:

rnic1 and rnic2 (continuous): incremental change (improvement) in runny nose severity at day 1 and day 2,

rngoal1 and rngoal2 (dichotomous): reaching the goal of therapy (of at least a 50% reduction in runny nose severity) at day 1 and day 2,

snic1 and snic2 (continuous): incremental change (improvement) in sneezing severity at day 1 and day 2, and

rngoal1 and rngoal2 (dichotomous): reaching the goal of therapy (of at least a 50% reduction in sneezing severity) at day 1 and day 2.

For the continuous outcomes, standardized mean differences can be computed to quantify the difference between the treatment and control groups. For the dichotomous outcomes, one can compute (log) odds ratios to quantify the difference between the treatment and control groups.

Source

D'Agostino, R. B., Sr., Weintraub, M., Russell, H. K., Stepanians, M., D'Agostino, R. B., Jr., Cantilena, L. R., Jr., Graumlich, J. F., Maldonado, S., Honig, P., & Anello, C. (1998). The effectiveness of antihistamines in reducing the severity of runny nose and sneezing: A meta-analysis. Clinical Pharmacology & Therapeutics, 64(6), 579--596. https://doi.org/10.1016/S0009-9236(98)90049-2

Examples

### copy data into 'dat' and examine data
dat <- dat.dagostino1998
dat

#>    study    cold scale.rn scale.sn             drug tnt tnc outcome    mt   sdt
#> 1      1 natural      0-3      0-3 chlorpheniramine 128 136   rnic1 0.932 0.593
#> 2      1 natural      0-3      0-3 chlorpheniramine 128 136   rnic2 1.433 0.648
#> 3      1 natural      0-3      0-3 chlorpheniramine 128 136 rngoal1    NA    NA
#> 4      1 natural      0-3      0-3 chlorpheniramine 128 136 rngoal2    NA    NA
#> 5      1 natural      0-3      0-3 chlorpheniramine 128 136   snic1 0.927 0.488
#> 6      1 natural      0-3      0-3 chlorpheniramine 128 136   snic2 1.298 0.538
#> 7      1 natural      0-3      0-3 chlorpheniramine 128 136 sngoal1    NA    NA
#> 8      1 natural      0-3      0-3 chlorpheniramine 128 136 sngoal2    NA    NA
#> 9      2 natural      0-3      0-3 chlorpheniramine  63  64   rnic1 0.730 0.745
#> 10     2 natural      0-3      0-3 chlorpheniramine  63  64   rnic2 1.145 0.903
#> 11     2 natural      0-3      0-3 chlorpheniramine  63  64 rngoal1    NA    NA
#> 12     2 natural      0-3      0-3 chlorpheniramine  63  64 rngoal2    NA    NA
#> 13     2 natural      0-3      0-3 chlorpheniramine  63  64   snic1 0.540 0.800
#> 14     2 natural      0-3      0-3 chlorpheniramine  63  64   snic2 0.710 0.876
#> 15     2 natural      0-3      0-3 chlorpheniramine  63  64 sngoal1    NA    NA
#> 16     2 natural      0-3      0-3 chlorpheniramine  63  64 sngoal2    NA    NA
#> 17     3 induced      0-3      0-3 chlorpheniramine  11  16   rnic1 0.273 0.786
#> 18     3 induced      0-3      0-3 chlorpheniramine  11  16   rnic2 0.182 1.328
#> 19     3 induced      0-3      0-3 chlorpheniramine  11  16 rngoal1    NA    NA
#> 20     3 induced      0-3      0-3 chlorpheniramine  11  16 rngoal2    NA    NA
#> 21     3 induced      0-3      0-3 chlorpheniramine  11  16   snic1 0.455 0.522
#> 22     3 induced      0-3      0-3 chlorpheniramine  11  16   snic2 0.545 0.688
#> 23     3 induced      0-3      0-3 chlorpheniramine  11  16 sngoal1    NA    NA
#> 24     3 induced      0-3      0-3 chlorpheniramine  11  16 sngoal2    NA    NA
#> 25     4 induced      0-8    count       doxylamine  16  15   rnic1 0.422 2.209
#> 26     4 induced      0-8    count       doxylamine  16  15   rnic2 0.953 2.961
#> 27     4 induced      0-8    count       doxylamine  16  15 rngoal1    NA    NA
#> 28     4 induced      0-8    count       doxylamine  16  15 rngoal2    NA    NA
#> 29     4 induced      0-8    count       doxylamine  16  15   snic1    NA    NA
#> 30     4 induced      0-8    count       doxylamine  16  15   snic2    NA    NA
#> 31     4 induced      0-8    count       doxylamine  16  15 sngoal1    NA    NA
#> 32     4 induced      0-8    count       doxylamine  16  15 sngoal2    NA    NA
#> 33     5 natural      0-8      0-8       doxylamine  23  23   rnic1 0.350 1.139
#> 34     5 natural      0-8      0-8       doxylamine  23  23   rnic2 0.481 1.299
#> 35     5 natural      0-8      0-8       doxylamine  23  23 rngoal1    NA    NA
#> 36     5 natural      0-8      0-8       doxylamine  23  23 rngoal2    NA    NA
#> 37     5 natural      0-8      0-8       doxylamine  23  23   snic1 0.196 0.521
#> 38     5 natural      0-8      0-8       doxylamine  23  23   snic2 0.587 0.838
#> 39     5 natural      0-8      0-8       doxylamine  23  23 sngoal1    NA    NA
#> 40     5 natural      0-8      0-8       doxylamine  23  23 sngoal2    NA    NA
#> 41     6 natural      0-8    count       doxylamine  39  42   rnic1 0.256 1.666
#> 42     6 natural      0-8    count       doxylamine  39  42   rnic2 1.083 1.663
#> 43     6 natural      0-8    count       doxylamine  39  42 rngoal1    NA    NA
#> 44     6 natural      0-8    count       doxylamine  39  42 rngoal2    NA    NA
#> 45     6 natural      0-8    count       doxylamine  39  42   snic1 3.306 7.690
#> 46     6 natural      0-8    count       doxylamine  39  42   snic2 3.727 9.200
#> 47     6 natural      0-8    count       doxylamine  39  42 sngoal1    NA    NA
#> 48     6 natural      0-8    count       doxylamine  39  42 sngoal2    NA    NA
#> 49     7 natural     0-10      0-8       doxylamine  21  21   rnic1 2.831 1.753
#> 50     7 natural     0-10      0-8       doxylamine  21  21   rnic2 3.431 2.436
#> 51     7 natural     0-10      0-8       doxylamine  21  21 rngoal1    NA    NA
#> 52     7 natural     0-10      0-8       doxylamine  21  21 rngoal2    NA    NA
#> 53     7 natural     0-10      0-8       doxylamine  21  21   snic1 1.439 1.815
#> 54     7 natural     0-10      0-8       doxylamine  21  21   snic2 1.894 2.383
#> 55     7 natural     0-10      0-8       doxylamine  21  21 sngoal1    NA    NA
#> 56     7 natural     0-10      0-8       doxylamine  21  21 sngoal2    NA    NA
#> 57     8 natural     0-10      0-8       doxylamine  13   8   rnic1 2.687 1.607
#> 58     8 natural     0-10      0-8       doxylamine  13   8   rnic2 3.115 1.504
#> 59     8 natural     0-10      0-8       doxylamine  13   8 rngoal1    NA    NA
#> 60     8 natural     0-10      0-8       doxylamine  13   8 rngoal2    NA    NA
#> 61     8 natural     0-10      0-8       doxylamine  13   8   snic1 0.115 0.965
#> 62     8 natural     0-10      0-8       doxylamine  13   8   snic2 0.353 0.977
#> 63     8 natural     0-10      0-8       doxylamine  13   8 sngoal1    NA    NA
#> 64     8 natural     0-10      0-8       doxylamine  13   8 sngoal2    NA    NA
#> 65     9 natural      0-4      0-4       doxylamine 195 193   rnic1 0.490 0.895
#> 66     9 natural      0-4      0-4       doxylamine 195 193   rnic2 0.869 0.879
#> 67     9 natural      0-4      0-4       doxylamine 195 193 rngoal1    NA    NA
#> 68     9 natural      0-4      0-4       doxylamine 195 193 rngoal2    NA    NA
#> 69     9 natural      0-4      0-4       doxylamine 195 193   snic1 0.717 0.927
#> 70     9 natural      0-4      0-4       doxylamine 195 193   snic2 1.147 0.882
#> 71     9 natural      0-4      0-4       doxylamine 195 193 sngoal1    NA    NA
#> 72     9 natural      0-4      0-4       doxylamine 195 193 sngoal2    NA    NA
#>        mc   sdc  xt  xc  nt  nc
#> 1   0.810 0.556  NA  NA 128 136
#> 2   1.182 0.629  NA  NA 128 133
#> 3      NA    NA  54  45 128 136
#> 4      NA    NA  99  94 128 136
#> 5   0.937 0.467  NA  NA 118 116
#> 6   1.104 0.561  NA  NA 118 118
#> 7      NA    NA  87  91 118 118
#> 8      NA    NA 112 104 118 118
#> 9   0.578 0.773  NA  NA  63  64
#> 10  0.891 1.041  NA  NA  62  64
#> 11     NA    NA  10   9  63  64
#> 12     NA    NA  19  19  62  64
#> 13  0.297 0.634  NA  NA  63  64
#> 14  0.469 0.712  NA  NA  62  64
#> 15     NA    NA  16  12  50  49
#> 16     NA    NA  25  20  49  49
#> 17 -0.188 0.834  NA  NA  11  16
#> 18 -0.250 1.125  NA  NA  11  16
#> 19     NA    NA   3   1  11  12
#> 20     NA    NA   4   1  11  12
#> 21  0.000 0.516  NA  NA  11  16
#> 22 -0.125 0.619  NA  NA  11  16
#> 23     NA    NA   4   1   6  10
#> 24     NA    NA   5   1   7   9
#> 25 -0.017 1.374  NA  NA  16  15
#> 26 -0.033 1.436  NA  NA  16  15
#> 27     NA    NA   1   1  16  15
#> 28     NA    NA   3   2  16  15
#> 29     NA    NA  NA  NA  NA  NA
#> 30     NA    NA  NA  NA  NA  NA
#> 31     NA    NA  NA  NA  NA  NA
#> 32     NA    NA  NA  NA  NA  NA
#> 33  0.339 0.744  NA  NA  22  22
#> 34  0.606 0.970  NA  NA  22  22
#> 35     NA    NA  13   7  22  22
#> 36     NA    NA  14  11  22  22
#> 37 -0.170 0.572  NA  NA  23  23
#> 38  0.123 0.750  NA  NA  23  23
#> 39     NA    NA  20  19  23  23
#> 40     NA    NA  21  22  23  23
#> 41  0.537 1.614  NA  NA  39  41
#> 42  1.079 2.223  NA  NA  36  38
#> 43     NA    NA   1   5  39  41
#> 44     NA    NA   9  10  36  38
#> 45  0.976 8.446  NA  NA  36  42
#> 46  2.615 8.213  NA  NA  33  39
#> 47     NA    NA  15  11  36  42
#> 48     NA    NA  17  14  33  39
#> 49  1.396 1.285  NA  NA  21  21
#> 50  1.965 1.813  NA  NA  20  18
#> 51     NA    NA  10   3  21  21
#> 52     NA    NA  14   7  20  18
#> 53  0.998 1.137  NA  NA  21  21
#> 54  1.299 1.385  NA  NA  18  13
#> 55     NA    NA  19  16  21  21
#> 56     NA    NA  20  15  20  18
#> 57  1.625 2.089  NA  NA  13   8
#> 58  2.344 1.897  NA  NA  13   8
#> 59     NA    NA  10   2  13   8
#> 60     NA    NA  11   3  13   8
#> 61  0.286 1.252  NA  NA  13   8
#> 62  0.531 1.175  NA  NA  13   8
#> 63     NA    NA  12   6  13   8
#> 64     NA    NA  13   7  13   8
#> 65  0.264 0.828  NA  NA 194 193
#> 66  0.702 0.930  NA  NA 195 193
#> 67     NA    NA  47  29 194 193
#> 68     NA    NA  62  44 195 193
#> 69  0.663 0.894  NA  NA 166 175
#> 70  0.949 0.939  NA  NA 167 175
#> 71     NA    NA  92  83 166 175
#> 72     NA    NA 126 110 167 175

### compute standardized mean differences and corresponding sampling variances
dat <- escalc(measure="SMD", m1i=mt, m2i=mc, sd1i=sdt, sd2i=sdc, n1i=nt, n2i=nc, data=dat,
              add.measure=TRUE)

### compute log odds ratios and corresponding sampling variances
dat <- escalc(measure="OR",  ai=xt, ci=xc, n1i=nt, n2i=nc, data=dat,
              replace=FALSE, add.measure=TRUE, add=1/2, to="all")

### inspect data for the first study
head(dat, 8)

#>   study    cold scale.rn scale.sn             drug tnt tnc outcome    mt   sdt 
#> 1     1 natural      0-3      0-3 chlorpheniramine 128 136   rnic1 0.932 0.593 
#> 2     1 natural      0-3      0-3 chlorpheniramine 128 136   rnic2 1.433 0.648 
#> 3     1 natural      0-3      0-3 chlorpheniramine 128 136 rngoal1    NA    NA 
#> 4     1 natural      0-3      0-3 chlorpheniramine 128 136 rngoal2    NA    NA 
#> 5     1 natural      0-3      0-3 chlorpheniramine 128 136   snic1 0.927 0.488 
#> 6     1 natural      0-3      0-3 chlorpheniramine 128 136   snic2 1.298 0.538 
#> 7     1 natural      0-3      0-3 chlorpheniramine 128 136 sngoal1    NA    NA 
#> 8     1 natural      0-3      0-3 chlorpheniramine 128 136 sngoal2    NA    NA 
#>      mc   sdc  xt  xc  nt  nc      yi     vi measure 
#> 1 0.810 0.556  NA  NA 128 136  0.2118 0.0153     SMD 
#> 2 1.182 0.629  NA  NA 128 133  0.3920 0.0156     SMD 
#> 3    NA    NA  54  45 128 136  0.3860 0.0647      OR 
#> 4    NA    NA  99  94 128 136  0.4167 0.0781      OR 
#> 5 0.937 0.467  NA  NA 118 116 -0.0209 0.0171     SMD 
#> 6 1.104 0.561  NA  NA 118 118  0.3518 0.0172     SMD 
#> 7    NA    NA  87  91 118 118 -0.1805 0.0905      OR 
#> 8    NA    NA 112 104 118 118  0.8761 0.2413      OR 

### fit a random-effects model for incremental change in runny nose severity at day 1
res <- rma(yi, vi, data=dat, subset=outcome=="rnic1")
res

#> 
#> Random-Effects Model (k = 9; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0000 (SE = 0.0147)
#> tau (square root of estimated tau^2 value):      0.0023
#> I^2 (total heterogeneity / total variability):   0.01%
#> H^2 (total variability / sampling variability):  1.00
#> 
#> Test for Heterogeneity:
#> Q(df = 8) = 9.5056, p-val = 0.3015
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.2325  0.0630  3.6919  0.0002  0.1091  0.3559  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### fit a random-effects model for reaching the goal of therapy for runny nose severity at day 1
res <- rma(yi, vi, data=dat, subset=outcome=="rngoal1")
res

#> 
#> Random-Effects Model (k = 9; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0000 (SE = 0.0829)
#> tau (square root of estimated tau^2 value):      0.0008
#> I^2 (total heterogeneity / total variability):   0.00%
#> H^2 (total variability / sampling variability):  1.00
#> 
#> Test for Heterogeneity:
#> Q(df = 8) = 10.8895, p-val = 0.2080
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.5313  0.1530  3.4726  0.0005  0.2314  0.8312  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp)

#> 
#>    pred  ci.lb  ci.ub  pi.lb  pi.ub 
#>  1.7012 1.2604 2.2961 1.2604 2.2961 
#> 

### construct an approximate V matrix assuming a correlation of 0.7 for sampling errors within studies
V <- lapply(split(dat$vi, dat$study), function(v) {
   S <- diag(sqrt(v), nrow=length(v), ncol=length(v))
   R <- matrix(0.7, nrow=length(v), ncol=length(v))
   diag(R) <- 1
   S %*% R %*% S
})
V <- bldiag(V, order=dat$study)

### fit a model for incremental change in runny nose severity at day 1 and at day 2, allowing for
### correlated sampling errors (no random effects added, since there does not appear to be any
### noteworthy heterogeneity in these data)
res <- rma.mv(yi, V, mods = ~ outcome - 1, data=dat, subset=outcome %in% c("rnic1","rnic2"))

#> Warning: Redundant predictors dropped from the model.

res

#> 
#> Multivariate Meta-Analysis Model (k = 18; method: REML)
#> 
#> Variance Components: none
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 16) = 15.6883, p-val = 0.4749
#> 
#> Test of Moderators (coefficients 1:2):
#> QM(df = 2) = 17.4520, p-val = 0.0002
#> 
#> Model Results:
#> 
#>               estimate      se    zval    pval   ci.lb   ci.ub 
#> outcomernic1    0.2315  0.0630  3.6773  0.0002  0.1081  0.3549  *** 
#> outcomernic2    0.2528  0.0634  3.9896  <.0001  0.1286  0.3769  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### test if there is a difference in effects at day 1 and day 2
anova(res, X=c(1,-1))

#> 
#> Hypothesis:                                   
#> 1: outcomernic1 - outcomernic2 = 0 
#> 
#> Results:
#>    estimate     se    zval   pval 
#> 1:  -0.0212 0.0489 -0.4342 0.6642 
#> 
#> Test of Hypothesis:
#> QM(df = 1) = 0.1885, p-val = 0.6642
#>