Studies on Topical plus Systemic Antibiotics to Prevent Respiratory Tract Infections

Results from 16 studies examining the effectiveness of topical plus systemic antibiotics to prevent respiratory tract infections (RTIs).

dat.damico2009

Format

The data frame contains the following columns:

study	character	first author

year	numeric	publication year

xt	numeric	number of RTIs in the treatment group

nt	numeric	number of patients in the treatment group

xc	numeric	number of RTIs in the control group

nc	numeric	number of patients in the control group

conceal	numeric	allocation concealment (0 = not adequate, 1 = adequate)

blind	numeric	blinding (0 = open, 1 = double-blind)

Details

The dataset includes the results from 16 studies that examined the effectiveness of topical plus systemic antibiotics versus no prophylaxis to prevent respiratory tract infections (RTIs).

Source

D'Amico, R., Pifferi, S., Torri, V., Brazzi, L., Parmelli, E., & Liberati, A. (2009). Antibiotic prophylaxis to reduce respiratory tract infections and mortality in adults receiving intensive care. Cochrane Database of Systematic Reviews, 4, CD000022. https://doi.org/10.1002/14651858.CD000022.pub3

Examples

### copy data into 'dat' and examine data
dat <- dat.damico2009
dat

#>             study year xt  nt  xc  nc conceal blind
#> 1      Abele-Horn 1997 13  58  23  30       0     0
#> 2          Aerdts 1991  1  28  29  60       1     0
#> 3           Blair 1991 12 161  38 170       1     0
#> 4          Boland 1991 14  32  17  32       0     1
#> 5       Cockerill 1992  4  75  12  75       0     0
#> 6           Finch 1991  4  20   7  24       1     0
#> 7          Jacobs 1992  0  45   4  46       0     0
#> 8          Kerver 1988  5  49  31  47       0     0
#> 9         Palomar 1997 10  50  25  49       0     0
#> 10          Rocha 1992  7  47  25  54       0     1
#> 11 Sanchez-Garcia 1992 32 131  60 140       1     1
#> 12    Stoutenbeek 2007 62 201 100 200       1     0
#> 13         Ulrich 1989  7  55  26  57       1     0
#> 14       Verwaest 1997 22 193  40 185       1     0
#> 15         Winter 1992  3  91  17  92       1     0
#> 16        Krueger 2002 91 265 149 262       1     1

### meta-analysis of the (log) odds ratios using the Mantel-Haenszel method
rma.mh(measure="OR", ai=xt, n1i=nt, ci=xc, n2i=nc, data=dat, digits=2)

#> 
#> Fixed-Effects Model (k = 16)
#> 
#> I^2 (total heterogeneity / total variability):  55.61%
#> H^2 (total variability / sampling variability): 2.25
#> 
#> Test for Heterogeneity: 
#> Q(df = 15) = 33.79, p-val < .01
#> 
#> Model Results (log scale):
#> 
#> estimate    se    zval  pval  ci.lb  ci.ub 
#>    -1.14  0.09  -12.74  <.01  -1.31  -0.96 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.32   0.27   0.38 
#> 
#> Cochran-Mantel-Haenszel Test:    CMH = 169.58, df = 1,  p-val < 0.01
#> Tarone's Test for Heterogeneity: X^2 =  37.24, df = 15, p-val < 0.01
#> 

### calculate log odds ratios and corresponding sampling variances
dat <- escalc(measure="OR", ai=xt, n1i=nt, ci=xc, n2i=nc, data=dat)

### meta-analysis using a random-effects model
res <- rma(yi, vi, data=dat, method="DL")
res

#> 
#> Random-Effects Model (k = 16; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.1784 (SE = 0.1362)
#> tau (square root of estimated tau^2 value):      0.4224
#> I^2 (total heterogeneity / total variability):   55.27%
#> H^2 (total variability / sampling variability):  2.24
#> 
#> Test for Heterogeneity:
#> Q(df = 15) = 33.5329, p-val = 0.0040
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval    ci.lb    ci.ub 
#>  -1.2782  0.1572  -8.1295  <.0001  -1.5864  -0.9700  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub pi.lb pi.ub 
#>  0.28  0.20  0.38  0.12  0.67 
#>