Studies on the Effectiveness of Self-Management Education and Regular Medical Review for Adults with Asthma

Results from 15 trials examining the effectiveness of self-management education and regular medical review for adults with asthma.

dat.gibson2002

Format

The data frame contains the following columns:

author	character	first author of study

year	numeric	publication year

n1i	numeric	number of participants in the intervention group

m1i	numeric	mean number of days off work/school in the intervention group

sd1i	numeric	standard deviation of the number of days off work/school in the intervention group

n2i	numeric	number of participants in the control/comparison group

m2i	numeric	mean number of days off work/school in the control/comparison group

sd2i	numeric	standard deviation of the number of days off work/school in the control/comparison group

ai	numeric	number of participants who had one or more days off work/school in the intervention group

bi	numeric	number of participants who no days off work/school in the intervention group

ci	numeric	number of participants who had one or more days off work/school in the control/comparison group

di	numeric	number of participants who no days off work/school in the control/comparison group

type	numeric	numeric code for the intervention type (see ‘Details’)

Details

Asthma management guidelines typically recommend for patients to receive education and regular medical review. While self-management programs have been shown to increase patient knowledge, it is less clear to what extent they actually impact health outcomes. The systematic review by Gibson et al. (2002) examined the effectiveness of self-management education and regular medical review for adults with asthma. In each study, participants receiving a certain management intervention were compared against those in a control/comparison group with respect to a variety of health outcomes. One of the outcomes examined in a number of studies was the number of days off work/school.

The majority of studies reporting this outcome provided means and standard deviations allowing a meta-analysis of standardized mean differences. Seven studies also reported the number of participants who had one or more days off work/school in each group. These studies could be meta-analyzed using, for example, (log) risk ratios. Finally, one could also consider a combined analysis based on standardized mean differences computed from the means and standard deviations where available and using probit transformed risk differences (which also provide estimates of the standardized mean difference) for the remaining studies.

Some degree of patient education was provided in all studies. In addition, the type variable indicates what additional intervention components were included in each study:

optimal self-management (writing action plan, self-monitoring, regular medical review),

self-monitoring and regular medical review,

self-monitoring only,

regular medical review only,

written action plan only.

Source

Gibson, P. G., Powell, H., Wilson, A., Abramson, M. J., Haywood, P., Bauman, A., Hensley, M. J., Walters, E. H., & Roberts, J. J. L. (2002). Self-management education and regular practitioner review for adults with asthma. Cochrane Database of Systematic Reviews, 3, CD001117. https://doi.org/10.1002/14651858.CD001117

Examples

### copy data into 'dat' and examine data
dat <- dat.gibson2002
dat

#>            author year n1i   m1i  sd1i n2i   m2i  sd2i ai bi ci di type
#> 1            Cote 1997  50  2.20 12.73  54  5.20 12.50 NA NA NA NA    1
#> 2           Ghosh 1998 140 17.60 24.20 136 34.10 38.80 NA NA NA NA    1
#> 3         Hayward 1996  23  0.38  0.56  19  0.23  0.29 NA NA NA NA    1
#> 4           Heard 1999  97  2.09  5.93  94  2.66  4.95 34 63 36 58    1
#> 5  Ignacio-Garcia 1995  35  4.92  6.05  35 20.00 26.34 24 11 29  6    1
#> 6          Knoell 1998  45  0.85  4.75  55  2.31  9.16 NA NA NA NA    1
#> 7       Lahdensuo 1996  56  2.80  9.00  59  4.80  7.20 13 43 25 34    1
#> 8       Sommaruga 1995  20 24.10 11.80  20 31.80 17.90 NA NA NA NA    1
#> 9          Zeiger 1991 128  1.40  3.30 143  2.30  7.60 NA NA NA NA    1
#> 10         Garret 1994 119  6.23 12.20 100  5.71  8.57 58 42 57 33    2
#> 11           Neri 1996  32  2.10  8.00  33  5.10 14.00  7 25 13 20    3
#> 12         Hilton 1986  86  0.73  1.48 100  0.47  1.20 NA NA NA NA    4
#> 13      Gallefoss 1999  25  8.00 32.00  24 26.00 70.00 NA NA NA NA    5
#> 14           Yoon 1993  28    NA    NA  28    NA    NA  5 23  4 24    1
#> 15         Brewin 1995  12    NA    NA  33    NA    NA  0 12 16 17    3

### fixed-effects model analysis of the standardized mean differences
dat <- escalc(measure="SMD", m1i=m1i, sd1i=sd1i, n1i=n1i, m2i=m2i, sd2i=sd2i, n2i=n2i, data=dat)
res <- rma(yi, vi, data=dat, method="FE")

#> Warning: Studies with NAs omitted from model fitting.

print(res, digits=2)

#> 
#> Fixed-Effects Model (k = 13)
#> 
#> I^2 (total heterogeneity / total variability):   55.36%
#> H^2 (total variability / sampling variability):  2.24
#> 
#> Test for Heterogeneity:
#> Q(df = 12) = 26.88, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.18  0.05  -3.77  <.01  -0.28  -0.09  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### fixed-effects model analysis of the (log) risk ratios
dat <- escalc(measure="RR", ai=ai, bi=bi, ci=ci, di=di, data=dat)
res <- rma(yi, vi, data=dat, method="FE")

#> Warning: Studies with NAs omitted from model fitting.

print(res, digits=2)

#> 
#> Fixed-Effects Model (k = 7)
#> 
#> I^2 (total heterogeneity / total variability):   18.22%
#> H^2 (total variability / sampling variability):  1.22
#> 
#> Test for Heterogeneity:
#> Q(df = 6) = 7.34, p-val = 0.29
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.17  0.08  -2.31  0.02  -0.32  -0.03  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub 
#>  0.84  0.72  0.97 
#> 

### fixed-effects model analysis of the standardized mean differences and the probit transformed
### risk differences (which also provide estimates of the standardized mean difference)
dat <- escalc(measure="SMD", m1i=m1i, sd1i=sd1i, n1i=n1i, m2i=m2i, sd2i=sd2i, n2i=n2i, data=dat)
dat <- escalc(measure="PBIT", ai=ai, bi=bi, ci=ci, di=di, data=dat, replace=FALSE)
dat

#>            author year n1i   m1i  sd1i n2i   m2i  sd2i ai bi ci di type      yi 
#> 1            Cote 1997  50  2.20 12.73  54  5.20 12.50 NA NA NA NA    1 -0.2361 
#> 2           Ghosh 1998 140 17.60 24.20 136 34.10 38.80 NA NA NA NA    1 -0.5105 
#> 3         Hayward 1996  23  0.38  0.56  19  0.23  0.29 NA NA NA NA    1  0.3209 
#> 4           Heard 1999  97  2.09  5.93  94  2.66  4.95 34 63 36 58    1 -0.1038 
#> 5  Ignacio-Garcia 1995  35  4.92  6.05  35 20.00 26.34 24 11 29  6    1 -0.7804 
#> 6          Knoell 1998  45  0.85  4.75  55  2.31  9.16 NA NA NA NA    1 -0.1930 
#> 7       Lahdensuo 1996  56  2.80  9.00  59  4.80  7.20 13 43 25 34    1 -0.2445 
#> 8       Sommaruga 1995  20 24.10 11.80  20 31.80 17.90 NA NA NA NA    1 -0.4978 
#> 9          Zeiger 1991 128  1.40  3.30 143  2.30  7.60 NA NA NA NA    1 -0.1504 
#> 10         Garret 1994 119  6.23 12.20 100  5.71  8.57 58 42 57 33    2  0.0484 
#> 11           Neri 1996  32  2.10  8.00  33  5.10 14.00  7 25 13 20    3 -0.2589 
#> 12         Hilton 1986  86  0.73  1.48 100  0.47  1.20 NA NA NA NA    4  0.1937 
#> 13      Gallefoss 1999  25  8.00 32.00  24 26.00 70.00 NA NA NA NA    5 -0.3277 
#> 14           Yoon 1993  28    NA    NA  28    NA    NA  5 23  4 24    1  0.1467 
#> 15         Brewin 1995  12    NA    NA  33    NA    NA  0 12 16 17    3 -1.7320 
#>        vi 
#> 1  0.0388 
#> 2  0.0150 
#> 3  0.0973 
#> 4  0.0210 
#> 5  0.0615 
#> 6  0.0406 
#> 7  0.0351 
#> 8  0.1031 
#> 9  0.0148 
#> 10 0.0184 
#> 11 0.0621 
#> 12 0.0217 
#> 13 0.0828 
#> 14 0.1627 
#> 15 0.4546 

res <- rma(yi, vi, data=dat, method="FE")
print(res, digits=2)

#> 
#> Fixed-Effects Model (k = 15)
#> 
#> I^2 (total heterogeneity / total variability):   57.35%
#> H^2 (total variability / sampling variability):  2.34
#> 
#> Test for Heterogeneity:
#> Q(df = 14) = 32.82, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.19  0.05  -3.87  <.01  -0.28  -0.09  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>