Studies on the Effectiveness of Injected Cholera Vaccines

Results from 17 studies on the effectiveness of injected vaccines against cholera.

dat.graves2010

Format

The data frame contains the following columns:

study	character	author/study name and publication year

ai	numeric	number of cholera cases in the vaccinated group

n1i	numeric	number of individuals in the vaccinated group

ci	numeric	number of cholera cases in the placebo group

n2i	numeric	number of individuals in the placebo group

Details

Cholera is an infection caused by certain strains of the bacterium Vibrio cholerae. When untreated, mortality rates can be as high as 50-60%. Proper sanitation practices are usually effective in preventing outbreaks, but a number of oral and injectable vaccines have also been developed. The Cochrane review by Graves et al. (2010) examined the effectiveness of injectable vaccines for preventing cholera cases and death. The present dataset includes results from 17 studies that reported the number of cholera cases in vaccinated and placebo/comparison groups up to 7 months after the treatment.

Source

Graves, P. M., Deeks, J. J., Demicheli, V., & Jefferson, T. (2010). Vaccines for preventing cholera: Killed whole cell or other subunit vaccines (injected). Cochrane Database of Systematic Reviews, 8, CD000974. https://doi.org/10.1002/14651858.CD000974.pub2

Examples

### copy data into 'dat' and examine data
dat <- dat.graves2010
dat

#>              study  ai    n1i ci    n2i
#> 1    Oseasohn 1965   8   6956 33   7103
#> 2   McCormack 1969  44  29939 35   9923
#> 3    Mosley 1970-i  32  11491 13   3810
#> 4   Mosley 1970-ii   2  11435 12   3810
#> 5      Taneja 1965  73  40326 27  10789
#> 6        PCC 1973b  72  82220 93  40620
#> 7      PCC 1973a-i  38  44500 23  11050
#> 8     PCC 1973a-ii  33  45750 22  11050
#> 9    PCC 1973a-iii  36  44450 23  11050
#> 10    PCC 1973a-iv  26  44700 22  11050
#> 11   Azurin 1965-i 110 145500 52  48934
#> 12  Azurin 1965-ii  92 148100 52  48933
#> 13 Azurin 1965-iii  71 143600 52  48933
#> 14        Pal 1980  13 101096 33 101030
#> 15        PCC 1968  54 268700 41  90900
#> 16   Saroso 1978-i  18 156300 19  79250
#> 17  Saroso 1978-ii  10 155600 18  79250

### analysis using the Mantel-Haenszel method
rma.mh(measure="RR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat, digits=2)

#> 
#> Fixed-Effects Model (k = 17)
#> 
#> I^2 (total heterogeneity / total variability):  51.53%
#> H^2 (total variability / sampling variability): 2.06
#> 
#> Test for Heterogeneity: 
#> Q(df = 16) = 33.01, p-val < .01
#> 
#> Model Results (log scale):
#> 
#> estimate    se    zval  pval  ci.lb  ci.ub 
#>    -0.77  0.06  -13.51  <.01  -0.89  -0.66 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.46   0.41   0.52 
#>