Studies on Lung Cancer Risk from ETS Exposure

Results from 37 studies on the risk of lung cancer from environmental tobacco smoke (ETS) exposure.

dat.hackshaw1998

Format

The data frame contains the following columns:

study	numeric	study number

author	character	first author of study

year	numeric	publication year

country	character	country where study was conducted

design	character	study design (either cohort or case-control)

cases	numeric	number of lung cancer cases

yi	numeric	log odds ratio

vi	numeric	corresponding sampling variance

Details

The dataset includes the results from 37 studies (4 cohort, 33 case-control) on the risk of lung cancer from environmental tobacco smoke (ETS) exposure from the spouse in women who are lifelong nonsmokers. Values of the log odds ratio greater than 0 indicate an increased risk of cancer in exposed women compared to women not exposed to ETS from their spouse.

Source

Hackshaw, A. K., Law, M. R., & Wald, N. J. (1997). The accumulated evidence on lung cancer and environmental tobacco smoke. British Medical Journal, 315, 980--988.

Hackshaw, A. K. (1998). Lung cancer and passive smoking. Statistical Methods in Medical Research, 7, 119--136.

Examples

### copy data into 'dat' and examine data
dat <- dat.hackshaw1998
dat

#>    study       author year   country       design cases      yi     vi 
#> 1      1    Garfinkel 1981       USA       cohort   153  0.1660 0.0188 
#> 2      2     Hirayama 1984     Japan       cohort   200  0.3720 0.0330 
#> 3      3       Butler 1988       USA       cohort     8  0.7030 0.5402 
#> 4      4     Cardenas 1997       USA       cohort   150  0.1820 0.0313 
#> 5      5         Chan 1982 Hong Kong case-control    84 -0.2880 0.0797 
#> 6      6       Correa 1983       USA case-control    22  0.7280 0.2273 
#> 7      7 Trichopolous 1983    Greece case-control    62  0.7560 0.0889 
#> 8      8      Buffler 1984       USA case-control    41 -0.2230 0.1927 
#> 9      9        Kabat 1984       USA case-control    24 -0.2360 0.3390 
#> 10    10          Lam 1985 Hong Kong case-control    60  0.6980 0.0981 
#> 11    11    Garfinkel 1985       USA case-control   134  0.2070 0.0456 
#> 12    12           Wu 1985       USA case-control    29  0.1820 0.2317 
#> 13    13        Akiba 1986     Japan case-control    94  0.4190 0.0796 
#> 14    14          Lee 1986        UK case-control    32  0.0300 0.2174 
#> 15    15          Koo 1987 Hong Kong case-control    86  0.4380 0.0770 
#> 16    16    Pershagen 1987    Sweden case-control    70  0.0300 0.0715 
#> 17    17       Humble 1987       USA case-control    20  0.8500 0.2926 
#> 18    18          Lam 1987 Hong Kong case-control   199  0.5010 0.0324 
#> 19    19          Gao 1987     China case-control   246  0.1740 0.0363 
#> 20    20     Brownson 1987       USA case-control    19  0.4190 0.4838 
#> 21    21         Geng 1988     China case-control    54  0.7700 0.1238 
#> 22    22      Shimizu 1988     Japan case-control    90  0.0770 0.0711 
#> 23    23        Inoue 1988     Japan case-control    22  0.9360 0.3982 
#> 24    24    Kalandidi 1990    Greece case-control    90  0.4820 0.0896 
#> 25    25        Sobue 1990     Japan case-control   144  0.0580 0.0337 
#> 26    26  Wu-Williams 1990     China case-control   417 -0.2360 0.0161 
#> 27    27          Liu 1991     China case-control    54 -0.3010 0.1802 
#> 28    28       Jockel 1991   Germany case-control    23  0.8200 0.3171 
#> 29    29     Brownson 1992       USA case-control   431 -0.0300 0.0125 
#> 30    30    Stockwell 1992       USA case-control   210  0.4700 0.1137 
#> 31    31           Du 1993     China case-control    75  0.1740 0.0893 
#> 32    32          Liu 1993     China case-control    38  0.5070 0.1760 
#> 33    33      Fontham 1994       USA case-control   651  0.2310 0.0100 
#> 34    34        Kabat 1995       USA case-control    67  0.0950 0.0862 
#> 35    35      Zaridze 1995    Russia case-control   162  0.5070 0.0399 
#> 36    36          Sun 1996     China case-control   230  0.1480 0.0364 
#> 37    37         Wang 1996     China case-control   135  0.1040 0.0664 

### random-effects model using the log odds ratios
res <- rma(yi, vi, data=dat, method="DL")
res

#> 
#> Random-Effects Model (k = 37; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0171 (SE = 0.0172)
#> tau (square root of estimated tau^2 value):      0.1307
#> I^2 (total heterogeneity / total variability):   24.28%
#> H^2 (total variability / sampling variability):  1.32
#> 
#> Test for Heterogeneity:
#> Q(df = 36) = 47.5453, p-val = 0.0944
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.2139  0.0471  4.5378  <.0001  0.1215  0.3063  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### estimated average odds ratio with CI (and credibility/prediction interval)
predict(res, transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub cr.lb cr.ub 
#>  1.24  1.13  1.36  0.94  1.63 
#>