Studies on the Effectiveness of Counseling for Smoking Cessation

Results from 24 studies on the effectiveness of various counseling types for smoking cessation.

dat.hasselblad1998

Format

The data frame contains the following columns:

id	numeric	id number for each treatment arm

study	numeric	study id number

authors	character	study author(s)

year	numeric	publication year

trt	character	intervention group

xi	numeric	number of individuals abstinent

ni	numeric	number of individuals in group

Details

The dataset includes the results from 24 studies on the effectiveness of various counseling types for smoking cessation (i.e., self-help, individual counseling, group counseling, and no contact). The dataset indicates the total number of individuals within each study arm and the number that were abstinent from 6 to 12 months. The majority of the studies compared two interventions types against each other, while 2 studies compared three types against each other simultaneously.

The data can be used for a ‘network meta-analysis’ (also called ‘mixed treatment comparison meta-analysis’). The code below shows how such an analysis can be conducted using an arm-based and a contrast-based model (see Salanti et al., 2008, for more details).

Source

Hasselblad, V. (1998). Meta-analysis of multitreatment studies. Medical Decision Making, 18, 37--43.

References

Gleser, L. J., & Olkin, I. (2009). Stochastically dependent effect sizes. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (2nd ed., pp. 357--376). New York: Russell Sage Foundation.

Salanti, G., Higgins, J. P. T., Ades, A. E., & Ioannidis, J. P. A. (2008). Evaluation of networks of randomized trials. Statistical Methods in Medical Research, 17, 279--301.

Examples

### load data
dat <- get(data(dat.hasselblad1998))

### calculate log odds for each study arm
dat <- escalc(measure="PLO", xi=xi, ni=ni, add=1/2, to="all", data=dat)

### create network graph ('plyr' and 'igraph' packages must be installed)
# NOT RUN {
require(plyr)
require(igraph)
pairs <- do.call(rbind, sapply(split(dat$trt, dat$study), function(x) t(combn(x,2))))
pairs <- ddply(data.frame(pairs), .(X1, X2), count)
g <- graph.edgelist(as.matrix(pairs[,1:2]), directed=FALSE)
plot(g, edge.curved=FALSE, edge.width=pairs$freq, vertex.label.dist=.7,
vertex.label=c("Individual\nCounseling", "Group\nCounseling", "No Contact", "Self-Help"))
# }

### convert trt variable to factor with desired ordering of levels
dat$trt <- factor(dat$trt, levels=c("no_contact", "self_help", "ind_counseling", "grp_counseling"))

### add a space before each level (this makes the output a bit more legible)
levels(dat$trt) <- paste0(" ", levels(dat$trt))

### network meta-analysis using an arm-based model with fixed study effects
### by setting rho=1/2, tau^2 reflects the amount of heterogeneity for all treatment comparisons
res <- rma.mv(yi, vi, mods = ~ factor(study) + trt - 1,
              random = ~ trt | study, rho=1/2, data=dat, btt=25:27)
res

#> 
#> Multivariate Meta-Analysis Model (k = 50; method: REML)
#> 
#> Variance Components:
#> 
#> outer factor: study (nlvls = 24)
#> inner factor: trt   (nlvls = 4)
#> 
#>             estim    sqrt  fixed 
#> tau^2      0.4324  0.6575     no 
#> rho        0.5000            yes 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 23) = 202.3334, p-val < .0001
#> 
#> Test of Moderators (coefficients 25:27):
#> QM(df = 3) = 14.2278, p-val = 0.0026
#> 
#> Model Results:
#> 
#>                     estimate      se     zval    pval    ci.lb    ci.ub 
#> factor(study)1       -1.3925  0.5820  -2.3926  0.0167  -2.5333  -0.2518    * 
#> factor(study)2       -2.7302  0.5842  -4.6731  <.0001  -3.8753  -1.5851  *** 
#> factor(study)3       -3.7217  0.6681  -5.5703  <.0001  -5.0312  -2.4121  *** 
#> factor(study)4       -2.2710  0.5829  -3.8958  <.0001  -3.4136  -1.1285  *** 
#> factor(study)5       -2.3914  0.7374  -3.2430  0.0012  -3.8367  -0.9461   ** 
#> factor(study)6       -0.9698  0.6421  -1.5104  0.1309  -2.2284   0.2887      
#> factor(study)7       -2.0855  0.6369  -3.2744  0.0011  -3.3339  -0.8372   ** 
#> factor(study)8       -1.9592  0.6674  -2.9357  0.0033  -3.2672  -0.6512   ** 
#> factor(study)9       -2.3487  0.6032  -3.8936  <.0001  -3.5310  -1.1664  *** 
#> factor(study)10      -1.8062  0.6296  -2.8688  0.0041  -3.0402  -0.5722   ** 
#> factor(study)11      -2.2618  0.5978  -3.7835  0.0002  -3.4334  -1.0901  *** 
#> factor(study)12      -3.5643  0.6139  -5.8058  <.0001  -4.7675  -2.3610  *** 
#> factor(study)13      -2.3454  0.5836  -4.0188  <.0001  -3.4892  -1.2015  *** 
#> factor(study)14      -2.0624  0.6492  -3.1767  0.0015  -3.3348  -0.7899   ** 
#> factor(study)15      -2.4575  0.6868  -3.5781  0.0003  -3.8037  -1.1114  *** 
#> factor(study)16      -2.1438  0.6707  -3.1964  0.0014  -3.4583  -0.8293   ** 
#> factor(study)17      -2.6810  0.6448  -4.1581  <.0001  -3.9447  -1.4173  *** 
#> factor(study)18      -2.4066  0.5964  -4.0353  <.0001  -3.5755  -1.2377  *** 
#> factor(study)19      -1.4441  0.8108  -1.7811  0.0749  -3.0331   0.1450    . 
#> factor(study)20      -2.4041  0.6331  -3.7974  0.0001  -3.6450  -1.1633  *** 
#> factor(study)21      -2.4359  0.5817  -4.1877  <.0001  -3.5760  -1.2958  *** 
#> factor(study)22      -2.8559  0.5991  -4.7667  <.0001  -4.0301  -1.6816  *** 
#> factor(study)23      -2.3268  0.5838  -3.9857  <.0001  -3.4710  -1.1826  *** 
#> factor(study)24      -3.0893  0.5847  -5.2836  <.0001  -4.2353  -1.9433  *** 
#> trt self_help         0.3888  0.3221   1.2070  0.2274  -0.2426   1.0202      
#> trt ind_counseling    0.6864  0.1904   3.6055  0.0003   0.3133   1.0596  *** 
#> trt grp_counseling    0.8438  0.3641   2.3176  0.0205   0.1302   1.5574    * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### all pairwise odds ratios of interventions versus no contact
predict(res, newmods=cbind(matrix(0, nrow=3, ncol=24), diag(3)),
        intercept=FALSE, transf=exp, digits=2)

#> 
#>   pred ci.lb ci.ub cr.lb cr.ub 
#> 1 1.48  0.78  2.77  0.35  6.20 
#> 2 1.99  1.37  2.89  0.52  7.60 
#> 3 2.33  1.14  4.75  0.53 10.14 
#> 

### all pairwise odds ratios comparing interventions (ic vs sh, gc vs sh, and gc vs ic)
predict(res, newmods=cbind(matrix(0, nrow=3, ncol=24), rbind(c(-1,1,0), c(-1,0,1), c(0,-1,1))),
        intercept=FALSE, transf=exp, digits=2)

#> 
#>   pred ci.lb ci.ub cr.lb cr.ub 
#> 1 1.35  0.70  2.58  0.32  5.71 
#> 2 1.58  0.72  3.43  0.35  7.10 
#> 3 1.17  0.59  2.30  0.27  5.02 
#> 

### forest plot of ORs of interventions versus no contact
dev.new(width=7, height=4)
par(mar=c(5,4,1,2))
forest(c(0,res$beta[25:27]), sei=c(0,res$se[25:27]), psize=1, xlim=c(-3,4), digits=c(2,1), efac=2,
       slab=c("No Contact", "Self-Help", "Individual Counseling", "Group Counseling"),
       atransf=exp, at=log(c(.5, 1, 2, 4, 8)), xlab="Odds Ratio for Intervention vs. No Contact")
text(-3, 6, "Intervention", pos=4)
text( 4, 6, "Odds Ratio [95% CI]",  pos=2)

### restructure dataset to a contrast-based format
dat.c <- lapply(split(dat.hasselblad1998, dat.hasselblad1998$study),
                function(x) cbind(x[rep(1,nrow(x)-1),], x[-1,5:7]))
dat.c <- do.call(rbind, dat.c)
dat.c <- dat.c[,c(1:5,8,6:7,9:10)]
names(dat.c)[5:10] <- c("trt1", "trt2", "ai", "n1i", "ci", "n2i")
rownames(dat.c) <- 1:nrow(dat.c)
dat.c$id <- 1:nrow(dat.c)
dat.c

#>    id study            authors year           trt1           trt2 ai  n1i  ci
#> 1   1     1        Reid et al. 1974     no_contact ind_counseling 75  731 363
#> 2   2     2    Cottraux et al. 1983     no_contact ind_counseling  9  140  23
#> 3   3     2    Cottraux et al. 1983     no_contact grp_counseling  9  140  10
#> 4   4     3       Slama et al. 1990     no_contact ind_counseling  2  106   9
#> 5   5     4    Jamrozik et al. 1984     no_contact ind_counseling 58  549 237
#> 6   6     5      Rabkin et al. 1984     no_contact ind_counseling  0   33   9
#> 7   7     6   Decker and Evans 1989      self_help ind_counseling 20   49  16
#> 8   8     7    Richmond et al. 1986     no_contact ind_counseling  3  100  31
#> 9   9     8              Leung 1991     no_contact ind_counseling  1   31  26
#> 10 10     9  Mothersill et al. 1988      self_help ind_counseling 11   78  12
#> 11 11     9  Mothersill et al. 1988      self_help grp_counseling 11   78  29
#> 12 12    10    Langford et al. 1983     no_contact ind_counseling  6   39  17
#> 13 13    11       Gritz et al. 1992     no_contact      self_help 79  702  77
#> 14 14    12    Campbell et al. 1986     no_contact      self_help 18  671  21
#> 15 15    13     Sanders et al. 1989     no_contact ind_counseling 64  642 107
#> 16 16    14    Hilleman et al. 1993 ind_counseling grp_counseling 12   76  20
#> 17 17    15     Gillams et al. 1984 ind_counseling grp_counseling  9   55   3
#> 18 18    16 Mogielnicki et al. 1986      self_help grp_counseling  7   66  32
#> 19 19    17        Page et al. 1986     no_contact ind_counseling  5   62   8
#> 20 20    18    Vetter and Ford 1990     no_contact ind_counseling 20  234  34
#> 21 21    19  Williams and Hall 1988     no_contact grp_counseling  0   20   9
#> 22 22    20    Pallonen et al. 1994     no_contact      self_help  8  116  19
#> 23 23    21     Russell et al. 1983     no_contact ind_counseling 95 1107 143
#> 24 24    22 Stewart and Rosser 1982     no_contact ind_counseling 15  187  36
#> 25 25    23     Russell et al. 1979     no_contact ind_counseling 78  584  73
#> 26 26    24    Kendrick et al. 1995     no_contact ind_counseling 69 1177  54
#>     n2i
#> 1   714
#> 2   140
#> 3   138
#> 4   205
#> 5  1561
#> 6    48
#> 7    43
#> 8    98
#> 9    95
#> 10   85
#> 11  170
#> 12   77
#> 13  694
#> 14  535
#> 15  761
#> 16   74
#> 17   26
#> 18  127
#> 19   90
#> 20  237
#> 21   20
#> 22  149
#> 23 1031
#> 24  504
#> 25  675
#> 26  888

### calculate log odds ratios for each treatment comparison
dat.c <- escalc(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, add=1/2, to="all", data=dat.c)
dat.c

#>    id study            authors year           trt1           trt2 ai  n1i  ci 
#> 1   1     1        Reid et al. 1974     no_contact ind_counseling 75  731 363 
#> 2   2     2    Cottraux et al. 1983     no_contact ind_counseling  9  140  23 
#> 3   3     2    Cottraux et al. 1983     no_contact grp_counseling  9  140  10 
#> 4   4     3       Slama et al. 1990     no_contact ind_counseling  2  106   9 
#> 5   5     4    Jamrozik et al. 1984     no_contact ind_counseling 58  549 237 
#> 6   6     5      Rabkin et al. 1984     no_contact ind_counseling  0   33   9 
#> 7   7     6   Decker and Evans 1989      self_help ind_counseling 20   49  16 
#> 8   8     7    Richmond et al. 1986     no_contact ind_counseling  3  100  31 
#> 9   9     8              Leung 1991     no_contact ind_counseling  1   31  26 
#> 10 10     9  Mothersill et al. 1988      self_help ind_counseling 11   78  12 
#> 11 11     9  Mothersill et al. 1988      self_help grp_counseling 11   78  29 
#> 12 12    10    Langford et al. 1983     no_contact ind_counseling  6   39  17 
#> 13 13    11       Gritz et al. 1992     no_contact      self_help 79  702  77 
#> 14 14    12    Campbell et al. 1986     no_contact      self_help 18  671  21 
#> 15 15    13     Sanders et al. 1989     no_contact ind_counseling 64  642 107 
#> 16 16    14    Hilleman et al. 1993 ind_counseling grp_counseling 12   76  20 
#> 17 17    15     Gillams et al. 1984 ind_counseling grp_counseling  9   55   3 
#> 18 18    16 Mogielnicki et al. 1986      self_help grp_counseling  7   66  32 
#> 19 19    17        Page et al. 1986     no_contact ind_counseling  5   62   8 
#> 20 20    18    Vetter and Ford 1990     no_contact ind_counseling 20  234  34 
#> 21 21    19  Williams and Hall 1988     no_contact grp_counseling  0   20   9 
#> 22 22    20    Pallonen et al. 1994     no_contact      self_help  8  116  19 
#> 23 23    21     Russell et al. 1983     no_contact ind_counseling 95 1107 143 
#> 24 24    22 Stewart and Rosser 1982     no_contact ind_counseling 15  187  36 
#> 25 25    23     Russell et al. 1979     no_contact ind_counseling 78  584  73 
#> 26 26    24    Kendrick et al. 1995     no_contact ind_counseling 69 1177  54 
#>     n2i      yi     vi 
#> 1   714 -2.1964 0.0204 
#> 2   140 -1.0183 0.1639 
#> 3   138 -0.1232 0.2159 
#> 4   205 -0.7035 0.5199 
#> 5  1561 -0.4098 0.0241 
#> 6    48 -2.7797 2.1604 
#> 7    43  0.1469 0.1796 
#> 8    98 -2.5649 0.3425 
#> 9    95 -2.0481 0.7516 
#> 10   85  0.0018 0.1954 
#> 11  170 -0.2019 0.1427 
#> 12   77 -0.3993 0.2574 
#> 13  694  0.0158 0.0287 
#> 14  535 -0.3894 0.1040 
#> 15  761 -0.3874 0.0281 
#> 16   74 -0.6632 0.1626 
#> 17   26  0.3161 0.4550 
#> 18  127 -0.9932 0.1914 
#> 19   90 -0.0743 0.3290 
#> 20  237 -0.5732 0.0873 
#> 21   20 -3.5225 2.2410 
#> 22  149 -0.6457 0.1858 
#> 23 1031 -0.5379 0.0196 
#> 24  504  0.1427 0.0998 
#> 25  675  0.2394 0.0300 
#> 26  888 -0.0408 0.0348 

### calculate the variance-covariance matrix of the log odds ratios for multitreatment studies
### see Gleser & Olkin (2009), equation (19.11), for the covariance equation
calc.v <- function(x) {
   v <- matrix(1/(x$ai[1]+1/2) + 1/(x$n1i[1] - x$ai[1] + 1/2), nrow=nrow(x), ncol=nrow(x))
   diag(v) <- x$vi
   v
}
V <- bldiag(lapply(split(dat.c, dat.c$study), calc.v))

### convert trt1 and trt2 variables to factors with desired ordering of levels
lvls <- c("no_contact", "self_help", "ind_counseling", "grp_counseling")
dat.c$trt1 <- factor(dat.c$trt1, levels=lvls)
dat.c$trt2 <- factor(dat.c$trt2, levels=lvls)

### create variables to indicate the contrast examined
dat.c <- cbind(dat.c, model.matrix(~ dat.c$trt1 - 1) - model.matrix(~ dat.c$trt2 - 1))
names(dat.c)[(ncol(dat.c)-3):ncol(dat.c)] <- lvls

### network meta-analysis using a contrast-based random-effects model
### by setting rho=1/2, tau^2 reflects the amount of heterogeneity for all treatment comparisons
res <- rma.mv(yi, V, mods = ~ self_help + ind_counseling + grp_counseling - 1,
              random = ~ factor(id) | study, rho=1/2, data=dat.c)
res

#> 
#> Multivariate Meta-Analysis Model (k = 26; method: REML)
#> 
#> Variance Components:
#> 
#> outer factor: study      (nlvls = 24)
#> inner factor: factor(id) (nlvls = 26)
#> 
#>             estim    sqrt  fixed 
#> tau^2      0.4324  0.6575     no 
#> rho        0.5000            yes 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 23) = 202.3334, p-val < .0001
#> 
#> Test of Moderators (coefficients 1:3):
#> QM(df = 3) = 14.2278, p-val = 0.0026
#> 
#> Model Results:
#> 
#>                 estimate      se    zval    pval    ci.lb   ci.ub 
#> self_help         0.3888  0.3221  1.2070  0.2274  -0.2426  1.0202      
#> ind_counseling    0.6864  0.1904  3.6055  0.0003   0.3133  1.0596  *** 
#> grp_counseling    0.8438  0.3641  2.3176  0.0205   0.1302  1.5574    * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>