Studies on the Effectiveness of Azithromycin for Treating Lower Respiratory Tract Infections

Results from 15 studies the effectiveness of azithromycin versus amoxycillin or amoxycillin/clavulanic acid (amoxyclav) in the treatment of acute lower respiratory tract infections.

dat.laopaiboon2015

Format

The data frame contains the following columns:

author	character	author(s)

year	numeric	publication year

ai	numeric	number of clinical failures in the group treated with azithromycin

n1i	numeric	number of patients in the group treated with azithromycin

ci	numeric	number of clinical failures in the group treated with amoxycillin or amoxyclav

n2i	numeric	number of patients in the group treated with amoxycillin or amoxyclav

age	character	whether the trial included adults or children

diag.ab	numeric	trial included patients with a diagnosis of acute bacterial bronchitis

diag.cb	numeric	trial included patients with a diagnosis of chronic bronchitis with acute exacerbation

diag.pn	numeric	trial included patients with a diagnosis of pneumonia

ctrl	character	antibiotic in control group (amoxycillin or amoxyclav)

Details

Azithromycin is an antibiotic useful for the treatment of a number of bacterial infections. Laopaiboon et al. (2015) conducted a meta-analysis of trials comparing the effectiveness of azithromycin versus amoxycillin or amoxycillin/clavulanic acid (amoxyclav) in the treatment of acute lower respiratory tract infections, including acute bacterial bronchitis, acute exacerbations of chronic bronchitis, and pneumonia. The results from 15 trials are included in this dataset.

Source

Laopaiboon, M., Panpanich, R., & Swa Mya, K. (2015). Azithromycin for acute lower respiratory tract infections. Cochrane Database of Systematic Reviews, 3, CD001954.

Examples

### load data
dat <- get(data(dat.laopaiboon2015))

### analysis using the Mantel-Haenszel method
rma.mh(measure="RR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat, digits=3)

#> 
#> Fixed-Effects Model (k = 15)
#> 
#> Test for Heterogeneity: 
#> Q(df = 14) = 40.348, p-val < .001
#> 
#> Model Results (log scale):
#> 
#> estimate     se    zval   pval   ci.lb  ci.ub 
#>   -0.083  0.117  -0.709  0.479  -0.311  0.146 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>    0.921  0.733  1.157 
#> 

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat)

### random-effects model
res <- rma(yi, vi, data=dat)
res

#> 
#> Random-Effects Model (k = 15; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.5874 (SE = 0.3897)
#> tau (square root of estimated tau^2 value):      0.7664
#> I^2 (total heterogeneity / total variability):   64.85%
#> H^2 (total variability / sampling variability):  2.84
#> 
#> Test for Heterogeneity:
#> Q(df = 14) = 38.4438, p-val = 0.0004
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval    ci.lb   ci.ub 
#>   0.0852  0.2673  0.3186  0.7501  -0.4388  0.6092    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### average risk ratio with 95% CI
predict(res, transf=exp)

#> 
#>    pred  ci.lb  ci.ub  cr.lb  cr.ub 
#>  1.0889 0.6448 1.8389 0.2218 5.3447 
#>