Studies on the association between maternal size, offspring size, and number of offsprings

Results from studies examining the association between maternal size, offspring size, and number of offsprings.

dat.lim2014

Format

The object is a list containing data frames m_o_size, m_o_fecundity, o_o_unadj, and o_o_adj that contain the following columns and the corresponding phylogenetic trees called m_o_size_tree, m_o_fecundity_tree, o_o_unadj_tree, and o_o_adj_tree:

article	numeric	article id

author	character	study author(s)

year	numeric	publication year

species	character	species

amniotes	character	whether the species was amniotic

environment	character	whether the species were wild or captive

reprounit	character	whether the data were based on lifetime reproductive output or a single reproductive event (only in m_o_size and m_o_fecundity)

ri	numeric	correlation coefficient

ni	numeric	sample size

Details

The object dat.lim2014 includes 4 datasets:

m_o_size	on the correlation between maternal size and offspring size

m_o_fecundity	on the correlation between maternal size and number of offsprings

o_o_unadj	on the correlation between offspring size and number of offsprings

o_o_adj	on the correlation between offspring size and number of offsprings adjusted for maternal size

Objects m_o_size_tree, m_o_fecundity_tree, o_o_unadj_tree, and o_o_adj_tree are the corresponding phylogenetic trees for the species included in each of these datasets.

Source

Lim, J. N., Senior, A. M., & Nakagawa, S. (2014). Heterogeneity in individual quality and reproductive trade-offs within species. Evolution, 68(8), 2306--2318. https://doi.org/10.1111/evo.12446

References

Hadfield, J. D., & Nakagawa, S. (2010). General quantitative genetic methods for comparative biology: Phylogenies, taxonomies and multi-trait models for continuous and categorical characters. Journal of Evolutionary Biology, 23(3), 494--508. https://doi.org/10.1111/j.1420-9101.2009.01915.x

Nakagawa, S., & Santos, E. S. A. (2012). Methodological issues and advances in biological meta-analysis. Evolutionary Ecology, 26(5), 1253--1274. https://doi.org/10.1007/s10682-012-9555-5

Examples

### copy data into 'dat' and examine data
dat <- dat.lim2014$o_o_unadj
head(dat)

#>     article                                     author year             species
#> 21        1                                Abell, A.J. 1999 Sceloporus_virgatus
#> 22        1                                Abell, A.J. 1999 Sceloporus_virgatus
#> 132       1                                Abell, A.J. 1999 Sceloporus_virgatus
#> 133       1                                Abell, A.J. 1999 Sceloporus_virgatus
#> 134       2     Allaine, D., Graziani, L. & Coulon, J. 1998     Marmota_marmota
#> 135       3 Baron, J.P., Tully, T. & Le Galliard, J.F. 2010      Vipera_ursinii
#>     amniotes environment     ri  ni
#> 21       yes        wild  0.100  21
#> 22       yes        wild -0.170  14
#> 132      yes        wild -0.070  21
#> 133      yes        wild -0.140  14
#> 134      yes        wild -0.540  74
#> 135      yes        wild  0.487 105

### calculate r-to-z transformed correlations and corresponding sampling variances
dat <- escalc(measure="ZCOR", ri=ri, ni=ni, data=dat)

# \dontrun{
### load 'ape' package
require(ape)

#> Loading required package: ape

#> 
#> Attaching package: ‘ape’

#> The following objects are masked from ‘package:igraph’:
#> 
#>     edges, mst, ring

### copy tree to 'tree'
tree <- dat.lim2014$o_o_unadj_tree

### compute branch lengths
tree <- compute.brlen(tree)

### compute phylogenetic correlation matrix
A <- vcv(tree, corr=TRUE)

### make copy of the species variable
dat$species.phy <- dat$species

### create effect size id variable
dat$esid <- 1:nrow(dat)

### fit multilevel phylogenetic meta-analytic model
res <- rma.mv(yi, vi,
   random = list(~ 1 | article, ~ 1 | esid, ~ 1 | species, ~ 1 | species.phy),
   R=list(species.phy=A), data=dat)
res

#> 
#> Multivariate Meta-Analysis Model (k = 170; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed       factor    R 
#> sigma^2.1  0.1387  0.3725    125     no      article   no 
#> sigma^2.2  0.0093  0.0962    170     no         esid   no 
#> sigma^2.3  0.0000  0.0000    120     no      species   no 
#> sigma^2.4  0.0572  0.2392    120     no  species.phy  yes 
#> 
#> Test for Heterogeneity:
#> Q(df = 169) = 1823.4239, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se     zval    pval    ci.lb   ci.ub 
#>  -0.1564  0.1277  -1.2242  0.2209  -0.4067  0.0940    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

# }