Studies on the Effectiveness of St. John's Wort for Treating Depression

Results from 26 studies on the effectiveness of Hypericum perforatum extracts (St. John's wort) for treating depression.

dat.linde2005

Format

The data frame contains the following columns:

id	numeric	study number

study	character	study author(s)

year	numeric	publication year

country	character	study location

ni	numeric	total sample size

major	numeric	sample restricted to patients who met criteria for major depression

baseline	numeric	HRSD baseline score

version	numeric	HRSD version (17 or 21 items)

duration	numeric	study duration (in weeks)

prep	character	Hypericum extract preparation

dosage	numeric	dosage (in mg)

response	numeric	definition of response (1 = HRSD score reduction of at least 50% or HRSDscore after therapy <10; 2 = HRSDreduction of at least 50%; 3 = based on HRSD scale but exact definition not reported; 4 = global patient assessment of efficacy; 5 = at least 'much improved' on the Clinical Global Impression sub-scale global improvement)

ai	numeric	number of responses in treatment group

n1i	numeric	number of patients in treatment group

ci	numeric	number of responses in placebo group

n2i	numeric	number of patients in placebo group

group	numeric	stratification variable used by the authors (1 = smaller trials restricted to major depression; 2 = larger trials restricted to major depression; 3 = smaller trials not restricted to major depression; 4 = larger trials not restricted to major depression)

Details

The dataset includes the results from 26 double-blind placebo-controlled trials on the effectiveness of Hypericum perforatum extracts (St. John's wort) for treating depression (note that 2 studies did not provide sufficient response information).

Data were extracted from Table 1 and Figure 3. For study duration, the assessment week (instead of the total study duration) was coded for Philipp et al. (1999) and Montgomery et al. (2000). For dosage, the midpoint was coded when a range of values was given.

Source

Linde, K., Berner, M., Egger, M., & Mulrow, C. (2005). St John's wort for depression: Meta-analysis of randomised controlled trials. British Journal of Psychiatry, 186, 99--107.

References

Viechtbauer, W. (2007). Accounting for heterogeneity via random-effects models and moderator analyses in meta-analysis. Zeitschrift für Psychologie / Journal of Psychology, 215, 104--121.

Examples

### load data
dat <- get(data(dat.linde2005))

### remove studies with no response information and study with no responses in either group
dat <- dat[-c(5,6,26),]

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=ai, ci=ci, n1i=n1i, n2i=n2i, data=dat)

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(yi, vi, data=dat, method="DL")
res

#> 
#> Random-Effects Model (k = 23; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.1413 (SE = 0.0697)
#> tau (square root of estimated tau^2 value):      0.3759
#> I^2 (total heterogeneity / total variability):   73.94%
#> H^2 (total variability / sampling variability):  3.84
#> 
#> Test for Heterogeneity:
#> Q(df = 22) = 84.4155, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.5574  0.1002  5.5631  <.0001  0.3610  0.7537  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### mixed-effects meta-regression model with stratification variable
res <- rma(yi, vi, mods = ~ factor(group) - 1, data=dat, method="DL")
res

#> 
#> Mixed-Effects Model (k = 23; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0555 (SE = 0.0377)
#> tau (square root of estimated tau^2 value):             0.2357
#> I^2 (residual heterogeneity / unaccounted variability): 53.18%
#> H^2 (unaccounted variability / sampling variability):   2.14
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 19) = 40.5792, p-val = 0.0027
#> 
#> Test of Moderators (coefficients 1:4):
#> QM(df = 4) = 68.4988, p-val < .0001
#> 
#> Model Results:
#> 
#>                 estimate      se    zval    pval    ci.lb   ci.ub 
#> factor(group)1    0.6826  0.1546  4.4138  <.0001   0.3795  0.9857  *** 
#> factor(group)2    0.1548  0.1136  1.3622  0.1731  -0.0679  0.3774      
#> factor(group)3    1.6681  0.2918  5.7166  <.0001   1.0962  2.2401  *** 
#> factor(group)4    0.5442  0.1430  3.8056  0.0001   0.2639  0.8245  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### predicted average risk ratio for each level of the stratification variable
predict(res, newmods=diag(4), transf=exp, digits=2)

#> 
#>   pred ci.lb ci.ub cr.lb cr.ub 
#> 1 1.98  1.46  2.68  1.14  3.44 
#> 2 1.17  0.93  1.46  0.70  1.95 
#> 3 5.30  2.99  9.39  2.54 11.06 
#> 4 1.72  1.30  2.28  1.00  2.96 
#>