Studies on Assortative Mating

Results from 457 studies on assortative mating in various species.

dat.moura2021

Format

The object is a list containing a data frame called dat that contains the following columns and a phylogenetic tree called tree:

study.id	character	study id

effect.size.id	numeric	effect size id

species	character	species

species.id	character	species id (as in the Open Tree of Life reference taxonomy)

subphylum	character	the subphyla of the species

phylum	character	the phyla of the species

assortment.trait	character	the measure of body size

trait.dimensions	character	dimensionality of the measure

field.collection	character	whether data were collected in the field

publication.year	numeric	publication year of the study

pooled.data	character	whether data were pooled either spatially and/or temporally

spatially.pooled.data	character	whether data were pooled spatially

temporally.pooled.data	character	whether data were pooled temporally

ri	numeric	correlation coefficient

ni	numeric	sample size

Blah.

Details

The 457 studies included in this dataset provide 1828 correlation coefficients describing the similarity in some measure of body size in mating couples in 341 different species.

Source

Rios Moura, R., Oliveira Gonzaga, M., Silva Pinto, N., Vasconcellos-Neto, J., & Requena, G. S. (2021). Assortative mating in space and time: Patterns and biases. Ecology Letters, 24(5), 1089--1102. https://doi.org/10.1111/ele.13690

References

Hadfield, J. D., & Nakagawa, S. (2010). General quantitative genetic methods for comparative biology: Phylogenies, taxonomies and multi-trait models for continuous and categorical characters. Journal of Evolutionary Biology, 23(3), 494--508. https://doi.org/10.1111/j.1420-9101.2009.01915.x

Nakagawa, S., & Santos, E. S. A. (2012). Methodological issues and advances in biological meta-analysis. Evolutionary Ecology, 26(5), 1253--1274. https://doi.org/10.1007/s10682-012-9555-5

Examples

### copy data into 'dat' and examine data
dat <- dat.moura2021$dat
head(dat)

#>                   study.id effect.size.id             species
#> 1 Adams and Greenwood 1983              1      Gammarus pulex
#> 2       Adams et al. 1985a              2 Trapezia ferruginea
#> 3       Adams et al. 1985a              3 Trapezia ferruginea
#> 4       Adams et al. 1985a              4 Trapezia ferruginea
#> 5       Adams et al. 1985b              5   Asellus aquaticus
#> 6       Adams et al. 1985b              6   Asellus aquaticus
#>                           species.id subphylum     phylum assortment.trait
#> 1 Gammarus_pulex_gallicus_ott1024016 Crustacea Arthropoda      body length
#> 2       Trapezia_bidentata_ott787242 Crustacea Arthropoda  carapace length
#> 3       Trapezia_bidentata_ott787242 Crustacea Arthropoda   carapace width
#> 4       Trapezia_bidentata_ott787242 Crustacea Arthropoda  cheliped length
#> 5        Asellus_aquaticus_ott335971 Crustacea Arthropoda           weight
#> 6        Asellus_aquaticus_ott335971 Crustacea Arthropoda           weight
#>   trait.dimensions field.collection publication.year pooled.data.
#> 1              uni              yes             1983           no
#> 2              uni              yes             1985          yes
#> 3              uni              yes             1985          yes
#> 4              uni              yes             1985          yes
#> 5              uni              yes             1985           no
#> 6              uni              yes             1985           no
#>   spatially.pooled.data. temporally.pooled.data. study.id   ri ni
#> 1                     no                      no     <NA> 0.23 53
#> 2                    yes                      no     <NA> 0.57 23
#> 3                    yes                      no     <NA> 0.54 23
#> 4                    yes                      no     <NA> 0.53 23
#> 5                     no                      no     <NA> 0.85 12
#> 6                     no                      no     <NA> 0.82 15

### calculate r-to-z transformed correlations and corresponding sampling variances
dat <- escalc(measure="ZCOR", ri=ri, ni=ni, data=dat)

# \dontrun{
### load 'ape' package
require(ape)

### copy tree to 'tree'
tree <- dat.moura2021$tree

### turn tree into an ultrametric one
tree <- compute.brlen(tree)

### compute phylogenetic correlation matrix
A <- vcv(tree, corr=TRUE)

### make copy of the species.id variable
dat$species.id.phy <- dat$species.id

### fit multilevel phylogenetic meta-analytic model
res <- rma.mv(yi, vi,
   random = list(~ 1 | study.id, ~ 1 | effect.size.id, ~ 1 | species.id, ~ 1 | species.id.phy),
   R=list(species.id.phy=A), data=dat)
res

#> 
#> Multivariate Meta-Analysis Model (k = 1828; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed          factor    R 
#> sigma^2.1  0.0192  0.1384    457     no        study.id   no 
#> sigma^2.2  0.0145  0.1202   1828     no  effect.size.id   no 
#> sigma^2.3  0.0557  0.2359    341     no      species.id   no 
#> sigma^2.4  0.0512  0.2263    341     no  species.id.phy  yes 
#> 
#> Test for Heterogeneity:
#> Q(df = 1827) = 10743.8076, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.3682  0.1300  2.8311  0.0046  0.1133  0.6230  ** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### examine if spatial and/or temporal pooling of data tends to yield larger correlations
res <- rma.mv(yi, vi,
   mods = ~ spatially.pooled.data * temporally.pooled.data,
   random = list(~ 1 | study.id, ~ 1 | effect.size.id, ~ 1 | species.id, ~ 1 | species.id.phy),
   R=list(species.id.phy=A), data=dat)

#> Error in eval(predvars, data, env): object 'spatially.pooled.data' not found

res

#> 
#> Multivariate Meta-Analysis Model (k = 1828; method: REML)
#> 
#> Variance Components:
#> 
#>             estim    sqrt  nlvls  fixed          factor    R 
#> sigma^2.1  0.0192  0.1384    457     no        study.id   no 
#> sigma^2.2  0.0145  0.1202   1828     no  effect.size.id   no 
#> sigma^2.3  0.0557  0.2359    341     no      species.id   no 
#> sigma^2.4  0.0512  0.2263    341     no  species.id.phy  yes 
#> 
#> Test for Heterogeneity:
#> Q(df = 1827) = 10743.8076, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.3682  0.1300  2.8311  0.0046  0.1133  0.6230  ** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### estimated average correlation without pooling, when pooling spatially,
### when pooling temporally, and when pooling spatially and temporally
predict(res, newmods = rbind(c(0,0,0),c(1,0,0),c(0,1,0),c(1,1,1)), transf=transf.ztor, digits=2)

#> Error in predict.rma(res, newmods = rbind(c(0, 0, 0), c(1, 0, 0), c(0,     1, 0), c(1, 1, 1)), transf = transf.ztor, digits = 2): Cannot specify new moderator values for models without moderators.

# }