Studies on the Effectiveness of Nonsurgical Treatments in Cirrhosis

Results from 26 trials examining the effectiveness of beta-blockers and sclerotherapy for the prevention of first bleeding in patients with cirrhosis

dat.pagliaro1992

Format

The data frame contains the following columns:

study	numeric	study id

trt	character	either beta-blockers, sclerotherapy, or control

xi	numeric	number of patients with first bleeding

ni	numeric	number of patients treated

Details

The dataset includes the results from 26 randomized controlled trials examining the effectiveness of nonsurgical treatments for the prevention of first bleeding in patients with cirrhosis. Patients were either treated with beta-blockers, endoscopic sclerotherapy, or with a nonactive treatment (control). Two trials included all three treatment conditions, 7 trials compared beta-blockers against control, and 17 trials compared sclerotherapy against control. The dataset has been used in various papers to illustrate methods for conducting a network meta-analysis / mixed treatment comparison.

Source

Pagliaro, L., D'Amico, G., Sörensen, T. I. A., Lebrec, D., Burroughs, A. K., Morabito, A., Tiné, F., Politi, F., & Traina, M. (1992). Prevention of first bleeding in cirrhosis: A meta-analysis of randomized trials of nonsurgical treatment. Annals of Internal Medicine, 117, 59--70.

Examples

### load data
dat <- get(data(dat.pagliaro1992))

### restructure dataset to a contrast-based format
dat.c <- lapply(split(dat.pagliaro1992, dat.pagliaro1992$study),
                function(x) cbind(x[-nrow(x),2:4], x[rep(nrow(x),nrow(x)-1),]))
dat.c <- do.call(rbind, dat.c)
dat.c <- dat.c[,c(4,1,5,2,3,6,7)]
names(dat.c) <- c("study", "trt1", "trt2", "ai", "n1i", "ci", "n2i")
rownames(dat.c) <- 1:nrow(dat.c)
dat.c

#>    study          trt1    trt2 ai n1i ci n2i
#> 1      1 beta-blockers control  2  43 13  41
#> 2      1 sclerotherapy control  9  42 13  41
#> 3      2 beta-blockers control 12  68 13  72
#> 4      2 sclerotherapy control 13  73 13  72
#> 5      3 beta-blockers control  4  20  4  16
#> 6      4 beta-blockers control 20 116 30 111
#> 7      5 beta-blockers control  1  30 11  49
#> 8      6 beta-blockers control  7  53 10  53
#> 9      7 beta-blockers control 18  85 31  89
#> 10     8 beta-blockers control  2  51 11  51
#> 11     9 beta-blockers control  8  23  2  25
#> 12    10 sclerotherapy control  4  18  0  19
#> 13    11 sclerotherapy control  3  35 22  36
#> 14    12 sclerotherapy control  5  56 30  53
#> 15    13 sclerotherapy control  5  16  6  18
#> 16    14 sclerotherapy control  3  23  9  22
#> 17    15 sclerotherapy control 11  49 31  46
#> 18    16 sclerotherapy control 19  53  9  60
#> 19    17 sclerotherapy control 17  53 26  60
#> 20    18 sclerotherapy control 10  71 29  69
#> 21    19 sclerotherapy control 12  41 14  41
#> 22    20 sclerotherapy control  0  21  3  20
#> 23    21 sclerotherapy control 13  33 14  35
#> 24    22 sclerotherapy control 31 143 23 138
#> 25    23 sclerotherapy control 20  55 19  51
#> 26    24 sclerotherapy control  3  13 12  16
#> 27    25 sclerotherapy control  3  21  5  28
#> 28    26 sclerotherapy control  6  22  2  24

### Mantel-Haenszel results for beta-blockers and sclerotherapy versus control, respectively
rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i,
       data=dat.c, subset=(trt1=="beta-blockers"), digits=2)

#> 
#> Fixed-Effects Model (k = 9)
#> 
#> Test for Heterogeneity: 
#> Q(df = 8) = 18.98, p-val = 0.01
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.62  0.16  -3.78  <.01  -0.94  -0.30 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.54   0.39   0.74 
#> 
#> Cochran-Mantel-Haenszel Test:    CMH = 14.14, df = 1, p-val < 0.01
#> Tarone's Test for Heterogeneity: X^2 = 22.76, df = 8, p-val < 0.01
#> 

rma.mh(measure="OR", ai=ai, n1i=n1i, ci=ci, n2i=n2i,
       data=dat.c, subset=(trt1=="sclerotherapy"), digits=2)

#> 
#> Fixed-Effects Model (k = 19)
#> 
#> Test for Heterogeneity: 
#> Q(df = 18) = 81.60, p-val < .01
#> 
#> Model Results (log scale):
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.53  0.11  -4.84  <.01  -0.74  -0.31 
#> 
#> Model Results (OR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>     0.59   0.48   0.73 
#> 
#> Cochran-Mantel-Haenszel Test:    CMH = 24.20, df = 1,  p-val < 0.01
#> Tarone's Test for Heterogeneity: X^2 = 94.46, df = 18, p-val < 0.01
#> 

### calculate log odds for each study arm
dat <- escalc(measure="PLO", xi=xi, ni=ni, data=dat)

### turn treatment variable into factor and set reference level
dat$trt <- relevel(factor(dat$trt), ref="control")

### add a space before each level (this makes the output a bit more legible)
levels(dat$trt) <- paste0(" ", levels(dat$trt))

### network meta-analysis using an arm-based random-effects model with fixed study effects
### by setting rho=1/2, tau^2 reflects the amount of heterogeneity for all treatment comparisons
res <- rma.mv(yi, vi, mods = ~ factor(study) + trt - 1, random = ~ trt | study, rho=1/2, data=dat)
res

#> 
#> Multivariate Meta-Analysis Model (k = 54; method: REML)
#> 
#> Variance Components:
#> 
#> outer factor: study (nlvls = 26)
#> inner factor: trt   (nlvls = 3)
#> 
#>             estim    sqrt  fixed 
#> tau^2      0.9913  0.9957     no 
#> rho        0.5000            yes 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 26) = 99.3092, p-val < .0001
#> 
#> Test of Moderators (coefficients 1:28):
#> QM(df = 28) = 56.4980, p-val = 0.0011
#> 
#> Model Results:
#> 
#>                    estimate      se     zval    pval    ci.lb    ci.ub 
#> factor(study)1      -1.1071  0.8684  -1.2748  0.2024  -2.8091   0.5950    
#> factor(study)2      -1.0951  0.8483  -1.2910  0.1967  -2.7577   0.5675    
#> factor(study)3      -0.8816  0.9712  -0.9077  0.3641  -2.7851   1.0220    
#> factor(study)4      -0.9237  0.8980  -1.0286  0.3037  -2.6838   0.8364    
#> factor(study)5      -1.6433  0.9724  -1.6900  0.0910  -3.5491   0.2625  . 
#> factor(study)6      -1.3172  0.9221  -1.4286  0.1531  -3.1244   0.4900    
#> factor(study)7      -0.6125  0.8997  -0.6808  0.4960  -2.3758   1.1509    
#> factor(study)8      -1.7386  0.9479  -1.8341  0.0666  -3.5966   0.1193  . 
#> factor(study)9      -0.9195  0.9816  -0.9367  0.3489  -2.8435   1.0045    
#> factor(study)10     -1.3962  1.0750  -1.2988  0.1940  -3.5032   0.7107    
#> factor(study)11     -0.4776  0.9305  -0.5133  0.6077  -2.3015   1.3462    
#> factor(study)12     -0.6253  0.9099  -0.6872  0.4920  -2.4086   1.1581    
#> factor(study)13     -0.4554  0.9463  -0.4812  0.6304  -2.3101   1.3994    
#> factor(study)14     -0.7814  0.9455  -0.8264  0.4086  -2.6346   1.0718    
#> factor(study)15      0.0456  0.9027   0.0505  0.9597  -1.7238   1.8149    
#> factor(study)16     -0.8312  0.9032  -0.9203  0.3574  -2.6014   0.9390    
#> factor(study)17     -0.2181  0.8941  -0.2440  0.8072  -1.9706   1.5343    
#> factor(study)18     -0.7510  0.8961  -0.8380  0.4020  -2.5074   1.0055    
#> factor(study)19     -0.4790  0.9043  -0.5297  0.5963  -2.2514   1.2934    
#> factor(study)20     -2.1080  1.0763  -1.9585  0.0502  -4.2176   0.0016  . 
#> factor(study)21     -0.1279  0.9071  -0.1410  0.8878  -1.9057   1.6499    
#> factor(study)22     -1.1504  0.8861  -1.2983  0.1942  -2.8871   0.5863    
#> factor(study)23     -0.2472  0.8957  -0.2760  0.7825  -2.0028   1.5083    
#> factor(study)24      0.2882  0.9746   0.2957  0.7675  -1.6219   2.1983    
#> factor(study)25     -1.3812  0.9557  -1.4453  0.1484  -3.2542   0.4919    
#> factor(study)26     -1.2182  0.9737  -1.2511  0.2109  -3.1266   0.6902    
#> trt beta-blockers   -0.7163  0.3871  -1.8508  0.0642  -1.4750   0.0423  . 
#> trt sclerotherapy   -0.5839  0.2684  -2.1753  0.0296  -1.1101  -0.0578  * 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### average odds ratio comparing beta-blockers and sclerotherapy versus control, respectively
predict(res, newmods=c(rep(0,26), 1, 0), transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub cr.lb cr.ub 
#>  0.49  0.23  1.04  0.06  3.96 
#> 

predict(res, newmods=c(rep(0,26), 0, 1), transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub cr.lb cr.ub 
#>  0.56  0.33  0.94  0.07  4.21 
#> 

### average odds ratio comparing beta-blockers versus sclerotherapy
predict(res, newmods=c(rep(0,26), 1, -1), transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub cr.lb cr.ub 
#>  0.88  0.36  2.11  0.10  7.45 
#>