Calculate Effect Sizes and Outcome Measures

The function can be used to calculate various effect sizes or outcome measures (and the corresponding sampling variances) that are commonly used in meta-analyses.

escalc(measure, ai, bi, ci, di, n1i, n2i, x1i, x2i, t1i, t2i,
       m1i, m2i, sd1i, sd2i, xi, mi, ri, ti, sdi, r2i, ni, yi, vi, sei,
       data, slab, subset,
       add=1/2, to="only0", drop00=FALSE, vtype="LS",
       var.names=c("yi","vi"), add.measure=FALSE,
       append=TRUE, replace=TRUE, digits=4, …)

Arguments

measure
a character string indicating which effect size or outcome measure should be calculated. See ‘Details’ for possible options and how the data should be specified.

ai
vector to specify the \(2 \times 2\) table frequencies (upper left cell).

bi
vector to specify the \(2 \times 2\) table frequencies (upper right cell).

ci
vector to specify the \(2 \times 2\) table frequencies (lower left cell).

di
vector to specify the \(2 \times 2\) table frequencies (lower right cell).

n1i
vector to specify the group sizes or row totals (first group/row).

n2i
vector to specify the group sizes or row totals (second group/row).

x1i
vector to specify the number of events (first group).

x2i
vector to specify the number of events (second group).

t1i
vector to specify the total person-times (first group).

t2i
vector to specify the total person-times (second group).

m1i
vector to specify the means (first group or time point).

m2i
vector to specify the means (second group or time point).

sd1i
vector to specify the standard deviations (first group or time point).

sd2i
vector to specify the standard deviations (second group or time point).

xi
vector to specify the frequencies of the event of interest.

mi
vector to specify the frequencies of the complement of the event of interest or the group means.

ri
vector to specify the raw correlation coefficients.

ti
vector to specify the total person-times.

sdi
vector to specify the standard deviations.

r2i
vector to specify the \(R^2\) values.

ni
vector to specify the sample/group sizes.

yi
vector to specify the observed effect size or outcomes.

vi
vector to specify the corresponding sampling variances.

sei
vector to specify the corresponding standard errors.

data
optional data frame containing the variables given to the arguments above.

slab
optional vector with labels for the studies.

subset
optional vector indicating the subset of studies that should be used. This can be a logical vector or a numeric vector indicating the indices of the studies to include.

add
a non-negative number indicating the amount to add to zero cells, counts, or frequencies. See ‘Details’.

to
a character string indicating when the values under add should be added (either "all", "only0", "if0all", or "none"). See ‘Details’.

drop00
logical indicating whether studies with no cases/events (or only cases) in both groups should be dropped when calculating the observed outcomes of the individual studies. See ‘Details’.

vtype
a character string indicating the type of sampling variances to calculate (either "LS", "UB", "HO", "ST", or vtype="CS"). See ‘Details’.

var.names
a character string with two elements, specifying the name of the variable for the observed outcomes and the name of the variable for the corresponding sampling variances (the default is "yi" and "vi").

add.measure
logical indicating whether a variable should be added to the data frame (with default name "measure") that indicates the type of outcome measure computed. When using this option, var.names can have a third element to change this variable name.

append
logical indicating whether the data frame specified via the data argument (if one has been specified) should be returned together with the observed outcomes and corresponding sampling variances (the default is TRUE).

replace
logical indicating whether existing values for yi and vi in the data frame should be replaced or not. Only relevant when append=TRUE and the data frame already contains the yi and vi variables. If replace=TRUE (the default), all of the existing values will be overwritten. If replace=FALSE, only NA values will be replaced. See ‘Value’ section below for more details.

digits
integer specifying the number of decimal places to which the printed results should be rounded (the default is 4). Note that the values are stored without rounding in the returned object.

…
other arguments.

Details

Before a meta-analysis can be conducted, the relevant results from each study must be quantified in such a way that the resulting values can be further aggregated and compared. Depending on (a) the goals of the meta-analysis, (b) the design and types of studies included, and (c) the information provided therein, one of the various effect size or outcome measures described below may be appropriate for the meta-analysis and can be computed with the escalc function.

The measure argument is a character string specifying which outcome measure should be calculated (see below for the various options), arguments ai through ni are then used to specify the information needed to calculate the various measures (depending on the chosen outcome measure, different arguments need to be specified), and data can be used to specify a data frame containing the variables given to the previous arguments. The add, to, and drop00 arguments may be needed when dealing with frequency or count data that may need special handling when some of the frequencies or counts are equal to zero (see below for details). Finally, the vtype argument is used to specify how to estimate the sampling variances (again, see below for details).

To provide a structure to the various effect size or outcome measures that can be calculated with the escalc function, we can distinguish between measures that are used to:

contrast two independent (either experimentally created or naturally occuring) groups,

describe the direction and strength of the association between two variables,

summarize some characteristic or attribute of individual groups, or

quantify change within a single group or the difference between two matched pairs samples.

Furthermore, where appropriate, we can further distinguish between measures that are applicable when the characteristic, response, or dependent variable assessed in the individual studies is:

a dichotomous (binary) variable (e.g., remission versus no remission),

a count of events per time unit (e.g., number of migraines per year),

a quantitative variable (e.g., amount of depression as assessed by a rating scale).

Outcome Measures for Two-Group Comparisons
In many meta-analyses, the goal is to synthesize the results from studies that compare or contrast two groups. The groups may be experimentally defined (e.g., a treatment and a control group created via random assignment) or may occur naturally (e.g., men and women, employees working under high- versus low-stress conditions, people exposed to some environmental risk factor versus those not exposed).
Measures for Dichotomous Variables
In various fields (such as the health and medical sciences), the response or outcome variable measured is often dichotomous (binary), so that the data from a study comparing two different groups can be expressed in terms of a \(2 \times 2\) table, such as:

outcome 1	outcome 2	total

group 1	ai	bi	n1i

group 2	ci	di	n2i

where ai, bi, ci, and di denote the cell frequencies (i.e., the number of people falling into a particular category) and n1i and n2i are the row totals (i.e., the group sizes). For example, in a set of randomized clinical trials, group 1 and group 2 may refer to the treatment and placebo/control group, respectively, with outcome 1 denoting some event of interest (e.g., death, complications, failure to improve under the treatment) and outcome 2 its complement. Similarly, in a set of cohort studies, group 1 and group 2 may denote those who engage in and those who do not engage in a potentially harmful behavior (e.g., smoking), with outcome 1 denoting the development of a particular disease (e.g., lung cancer) during the follow-up period. Finally, in a set of case-control studies, group 1 and group 2 may refer to those with the disease (i.e., cases) and those free of the disease (i.e., controls), with outcome 1 denoting, for example, exposure to some risk environmental risk factor in the past and outcome 2 non-exposure. Note that in all of these examples, the stratified sampling scheme fixes the row totals (i.e., the group sizes) by design. A meta-analysis of studies reporting results in terms of \(2 \times 2\) tables can be based on one of several different outcome measures, including the risk ratio (also called the relative risk), the odds ratio, the risk difference, and the arcsine square root transformed risk difference (e.g., Fleiss & Berlin, 2009, Rücker et al., 2009). For any of these outcome measures, one needs to specify the cell frequencies via the ai, bi, ci, and di arguments (or alternatively, one can use the ai, ci, n1i, and n2i arguments). The options for the measure argument are then:

"RR" for the log risk ratio.

"OR" for the log odds ratio.

"RD" for the risk difference.

"AS" for the arcsine square root transformed risk difference (Rücker et al., 2009).

"PETO" for the log odds ratio estimated with Peto's method (Yusuf et al., 1985).

Note that the log is taken of the risk ratio and the odds ratio, which makes these outcome measures symmetric around 0 and yields corresponding sampling distributions that are closer to normality. Cell entries with a zero count can be problematic, especially for the risk ratio and the odds ratio. Adding a small constant to the cells of the \(2 \times 2\) tables is a common solution to this problem. When to="only0" (the default), the value of add (the default is 1/2; but see ‘Note’) is added to each cell of those \(2 \times 2\) tables with at least one cell equal to 0. When to="all", the value of add is added to each cell of all \(2 \times 2\) tables. When to="if0all", the value of add is added to each cell of all \(2 \times 2\) tables, but only when there is at least one \(2 \times 2\) table with a zero cell. Setting to="none" or add=0 has the same effect: No adjustment to the observed table frequencies is made. Depending on the outcome measure and the data, this may lead to division by zero inside of the function (when this occurs, the resulting value is recoded to NA). Also, studies where ai=ci=0 or bi=di=0 may be considered to be uninformative about the size of the effect and dropping such studies has sometimes been recommended (Higgins & Green, 2008). This can be done by setting drop00=TRUE. The values for such studies will then be set to NA. Datasets corresponding to data of this type are provided in dat.bcg, dat.collins1985a, dat.collins1985b, dat.egger2001, dat.hine1989, dat.laopaiboon2015, dat.lee2004, dat.li2007, dat.linde2005, dat.nielweise2007, and dat.yusuf1985. Assuming that the dichotomous outcome is actually a dichotomized version of the responses on an underlying quantitative scale, it is also possible to estimate the standardized mean difference based on \(2 \times 2\) table data, using either the probit transformed risk difference or a transformation of the odds ratio (e.g., Cox & Snell, 1989; Chinn, 2000; Hasselblad & Hedges, 1995; Sánchez-Meca et al., 2003). The options for the measure argument are then:

"PBIT" for the probit transformed risk difference as an estimate of the standardized mean difference.

"OR2DN" for the transformed odds ratio as an estimate of the standardized mean difference (normal distributions).

"OR2DL" for the transformed odds ratio as an estimate of the standardized mean difference (logistic distributions).

The probit transformation assumes that the responses on the underlying quantitative scale are normally distributed. There are two versions of the odds ratio transformation, the first also assuming normal distributions within the two groups, while the second assumes that the responses follow logistic distributions. A dataset corresponding to data of this type is provided in dat.gibson2002.
Measures for Event Counts
In medical and epidemiological studies comparing two different groups (e.g., treated versus untreated patients, exposed versus unexposed individuals), results are sometimes reported in terms of event counts (i.e., the number of events, such as strokes or myocardial infarctions) over a certain period of time. Data of this type are also referred to as ‘person-time data’. In particular, assume that the studies report data in the form:

number of events	total person-time

group 1	x1i	t1i

group 2	x2i	t2i

where x1i and x2i denote the number of events in the first and the second group, respectively, and t1i and t2i the corresponding total person-times at risk. Often, the person-time is measured in years, so that t1i and t2i denote the total number of follow-up years in the two groups. This form of data is fundamentally different from what was described in the previous section, since the total follow-up time may differ even for groups of the same size and the individuals studied may experience the event of interest multiple times. Hence, different outcome measures than the ones described in the previous section must be considered when data are reported in this format. These include the incidence rate ratio, the incidence rate difference, and the square root transformed incidence rate difference (Bagos & Nikolopoulos, 2009; Rothman et al., 2008). For any of these outcome measures, one needs to specify the total number of events via the x1i and x2i arguments and the corresponding total person-time values via the t1i and t2i arguments. The options for the measure argument are then:

"IRR" for the log incidence rate ratio.

"IRD" for the incidence rate difference.

"IRSD" for the square root transformed incidence rate difference.

Note that the log is taken of the incidence rate ratio, which makes this outcome measure symmetric around 0 and yields a corresponding sampling distribution that is closer to normality. Studies with zero events in one or both groups can be problematic, especially for the incidence rate ratio. Adding a small constant to the number of events is a common solution to this problem. When to="only0" (the default), the value of add (the default is 1/2; but see ‘Note’) is added to x1i and x2i only in the studies that have zero events in one or both groups. When to="all", the value of add is added to x1i and x2i in all studies. When to="if0all", the value of add is added to x1i and x2i in all studies, but only when there is at least one study with zero events in one or both groups. Setting to="none" or add=0 has the same effect: No adjustment to the observed number of events is made. Depending on the outcome measure and the data, this may lead to division by zero inside of the function (when this occurs, the resulting value is recoded to NA). Like for \(2 \times 2\) table data, studies where x1i=x2i=0 may be considered to be uninformative about the size of the effect and dropping such studies has sometimes been recommended. This can be done by setting drop00=TRUE. The values for such studies will then be set to NA. Datasets corresponding to data of this type are provided in dat.hart1999 and dat.nielweise2008.
Measures for Quantitative Variables
When the response or dependent variable assessed in the individual studies is measured on some quantitative scale, it is customary to report certain summary statistics, such as the mean and standard deviation of the observations. The data layout for a study comparing two groups with respect to such a variable is then of the form:

mean	standard deviation	group size

group 1	m1i	sd1i	n1i

group 2	m2i	sd2i	n2i

where m1i and m2i are the observed means of the two groups, sd1i and sd2i are the observed standard deviations, and n1i and n2i denote the number of individuals in each group. Again, the two groups may be experimentally created (e.g., a treatment and control group based on random assignment) or naturally occurring (e.g., men and women). In either case, the raw mean difference, the standardized mean difference, and the (log transformed) ratio of means (also called log response ratio) are useful outcome measures when meta-analyzing studies of this type (e.g., Borenstein, 2009). The options for the measure argument are then:

"MD" for the raw mean difference.

"SMD" for the standardized mean difference.

"SMDH" for the standardized mean difference with heteroscedastic population variances in the two groups (Bonett, 2008, 2009).

"ROM" for the log transformed ratio of means (Hedges et al., 1999; Lajeunesse, 2011).

Note that the log is taken of the ratio of means, which makes this outcome measures symmetric around 0 and yields a corresponding sampling distribution that is closer to normality. However, note that if m1i and m2i have opposite signs, this outcome measure cannot be computed. The positive bias in the standardized mean difference is automatically corrected for within the function, yielding Hedges' g for measure="SMD" (Hedges, 1981). Similarly, the same bias correction is applied for measure="SMDH" (Bonett, 2009). For measure="SMD", one can choose between vtype="LS" (the default) and vtype="UB". The former uses the usual large-sample approximation to compute the sampling variances. The latter provides unbiased estimates of the sampling variances. Finally, for measure="MD" and measure="ROM", one can choose between vtype="LS" (the default) and vtype="HO". The former computes the sampling variances without assuming homoscedasticity (i.e., that the true variances of the measurements are the same in group 1 and group 2 within each study), while the latter assumes homoscedasticity. A dataset corresponding to data of this type is provided in dat.normand1999 (for mean differences and standardized mean differences). A dataset showing the use of the ratio of means measure is provided in dat.curtis1998. It is also possible to transform standardized mean differences into log odds ratios (e.g., Cox & Snell, 1989; Chinn, 2000; Hasselblad & Hedges, 1995; Sánchez-Meca et al., 2003). The options for the measure argument are then:

"D2ORN" for the transformed standardized mean difference as an estimate of the log odds ratio (normal distributions).

"D2ORL" for the transformed standardized mean difference as an estimate of the log odds ratio (logistic distributions).

Both of these transformations provide an estimate of the log odds ratio, the first assuming that the responses within the two groups are normally distributed, while the second assumes that the responses follow logistic distributions. A dataset illustrating the combined analysis of standardized mean differences and probit transformed risk differences is provided in dat.gibson2002. Finally, interest may also be focused on differences between the two groups with respect to their variability. Here, the (log transformed) ratio of the coefficient of variation of the two groups (also called the coefficient of variation ratio) can be a useful measure (Nakagawa et al., 2015). If focus is solely on the variability of the measurements within the two groups, then the (log transformed) ratio of the standard deviations (also called the variability ratio) can be used (Nakagawa et al., 2015). For the latter, one only needs to specify sd1i, sd2i, n1i, and n2i. The options for the measure argument are:

"CVR" for the log transformed coefficient of variation ratio.

"VR" for the log transformed variability ratio.

Note that a slight bias correction is applied for both of these measures (Nakagawa et al., 2015). Also, the sampling variance for measure="CVR" is computed as given by equation 12 in Nakagawa et al. (2015), but without the ‘\(-2 \rho \ldots\)’ terms, since for normally distributed data (which we assume here) the mean and variance (and transformations thereof) are independent.

Outcome Measures for Variable Association
Meta-analyses are often used to synthesize studies that examine the direction and strength of the association between two variables measured concurrently and/or without manipulation by experimenters. In this section, a variety of outcome measures will be discussed that may be suitable for a meta-analyses with this purpose. We can distinguish between measures that are applicable when both variables are measured on quantitative scales, when both variables measured are dichotomous, and when the two variables are of mixed types.
Measures for Two Quantitative Variables
The (Pearson or product moment) correlation coefficient quantifies the direction and strength of the (linear) relationship between two quantitative variables and is therefore frequently used as the outcome measure for meta-analyses (e.g., Borenstein, 2009). Two alternative measures are a bias-corrected version of the correlation coefficient and Fisher's r-to-z transformed correlation coefficient. For these measures, one needs to specify ri, the vector with the raw correlation coefficients, and ni, the corresponding sample sizes. The options for the measure argument are then:

"COR" for the raw correlation coefficient.

"UCOR" for the raw correlation coefficient corrected for its slight negative bias (based on equation 2.3 in Olkin & Pratt, 1958).

"ZCOR" for Fisher's r-to-z transformed correlation coefficient (Fisher, 1921).

For measure="UCOR", one can choose between vtype="LS" (the default) and vtype="UB". The former uses the usual large-sample approximation to compute the sampling variances. The latter provides unbiased estimates of the sampling variances (see Hedges, 1989, but using the exact equation instead of the approximation). Datasets corresponding to data of this type are provided in dat.mcdaniel1994 and dat.molloy2014.
Measures for Two Dichotomous Variables
When the goal of a meta-analysis is to examine the relationship between two dichotomous variables, the data for each study can again be presented in the form of a \(2 \times 2\) table, except that there may not be a clear distinction between the group (i.e., the row) and the outcome (i.e., the column) variable. Moreover, the table may be a result of cross-sectional (i.e., multinomial) sampling, where none of the table margins (except the total sample size) are fixed by the study design. The phi coefficient and the odds ratio are commonly used measures of association for \(2 \times 2\) table data (e.g., Fleiss & Berlin, 2009). The latter is particularly advantageous, as it is directly comparable to values obtained from stratified sampling (as described earlier). Yule's Q and Yule's Y (Yule, 1912) are additional measures of association for \(2 \times 2\) table data (although they are not typically used in meta-analyses). Finally, assuming that the two dichotomous variables are actually dichotomized versions of the responses on two underlying quantitative scales (and assuming that the two variables follow a bivariate normal distribution), it is also possible to estimate the correlation between the two variables using the tetrachoric correlation coefficient (Pearson, 1900; Kirk, 1973). For any of these outcome measures, one needs to specify the cell frequencies via the ai, bi, ci, and di arguments (or alternatively, one can use the ai, ci, n1i, and n2i arguments). The options for the measure argument are then:

"OR" for the log odds ratio.

"PHI" for the phi coefficient.

"YUQ" for Yule's Q (Yule, 1912).

"YUY" for Yule's Y (Yule, 1912).

"RTET" for the tetrachoric correlation.

Tables with one or two zero counts are handled as described earlier. For measure="PHI", one must indicate via vtype="ST" or vtype="CS" whether the data for the studies were obtained using stratified or cross-sectional (i.e., multinomial) sampling, respectively (it is also possible to specify an entire vector for the vtype argument in case the sampling schemes differed for the various studies). A dataset corresponding to data of this type is provided in dat.bourassa1996.
Measures for Mixed Variable Types
Finally, we can consider outcome measures that can be used to describe the relationship between two variables, where one variable is dichotomous and the other variable measures some quantitative characteristic. In that case, it is likely that study authors again report summary statistics, such as the mean and standard deviation of the measurements within the two groups (defined by the dichotomous variable). In that case, one can compute the point-biserial correlation (Tate, 1954) as a measure of association between the two variables. If the dichotomous variable is actually a dichotomized version of the responses on an underlying quantitative scale (and assuming that the two variables follow a bivariate normal distribution), it is also possible to estimate the correlation between the two variables using the biserial correlation coefficient (Pearson, 1909; Soper, 1914; Jacobs & Viechtbauer, 2017). Here, one again needs to specify m1i and m2i for the observed means of the two groups, sd1i and sd2i for the observed standard deviations, and n1i and n2i for the number of individuals in each group. The options for the measure argument are then:

"RPB" for the point-biserial correlation.

"RBIS" for the biserial correlation.

For measure="RPB", one must indicate via vtype="ST" or vtype="CS" whether the data for the studies were obtained using stratified or cross-sectional (i.e., multinomial) sampling, respectively (it is also possible to specify an entire vector for the vtype argument in case the sampling schemes differed for the various studies).

Outcome Measures for Individual Groups
In this section, outcome measures will be described which may be useful when the goal of a meta-analysis is to synthesize studies that characterize some property of individual groups. We will again distinguish between measures that are applicable when the characteristic of interest is a dichotomous variable, when the characteristic represents an event count, or when the characteristic assessed is a quantitative variable.
Measures for Dichotomous Variables
A meta-analysis may be conducted to aggregate studies that provide data for individual groups with respect to a dichotomous dependent variable. Here, one needs to specify xi and ni, denoting the number of individuals experiencing the event of interest and the total number of individuals, respectively. Instead of specifying ni, one can use mi to specify the number of individuals that do not experience the event of interest. The options for the measure argument are then:

"PR" for the raw proportion.

"PLN" for the log transformed proportion.

"PLO" for the logit transformed proportion (i.e., log odds).

"PAS" for the arcsine square root transformed proportion (i.e., the angular transformation).

"PFT" for the Freeman-Tukey double arcsine transformed proportion (Freeman & Tukey, 1950).

Zero cell entries can be problematic for certain outcome measures. When to="only0" (the default), the value of add (the default is 1/2; but see ‘Note’) is added to xi and mi only for studies where xi or mi is equal to 0. When to="all", the value of add is added to xi and mi in all studies. When to="if0all", the value of add is added in all studies, but only when there is at least one study with a zero value for xi or mi. Setting to="none" or add=0 has the same effect: No adjustment to the observed values is made. Depending on the outcome measure and the data, this may lead to division by zero inside of the function (when this occurs, the resulting value is recoded to NA). Datasets corresponding to data of this type are provided in dat.pritz1997 and dat.debruin2009.
Measures for Event Counts
Various measures can be used to characterize individual groups when the dependent variable assessed is an event count. Here, one needs to specify xi and ti, denoting the number of events that occurred and the total person-times at risk, respectively. The options for the measure argument are then:

"IR" for the raw incidence rate.

"IRLN" for the log transformed incidence rate.

"IRS" for the square root transformed incidence rate.

"IRFT" for the Freeman-Tukey transformed incidence rate (Freeman & Tukey, 1950).

Studies with zero events can be problematic, especially for the log transformed incidence rate. Adding a small constant to the number of events is a common solution to this problem. When to="only0" (the default), the value of add (the default is 1/2; but see ‘Note’) is added to xi only in the studies that have zero events. When to="all", the value of add is added to xi in all studies. When to="if0all", the value of add is added to xi in all studies, but only when there is at least one study with zero events. Setting to="none" or add=0 has the same effect: No adjustment to the observed number of events is made. Depending on the outcome measure and the data, this may lead to division by zero inside of the function (when this occurs, the resulting value is recoded to NA).
Measures for Quantitative Variables
The goal of a meta-analysis may also be to characterize individual groups, where the response, characteristic, or dependent variable assessed in the individual studies is measured on some quantitative scale. In the simplest case, the raw mean for the quantitative variable is reported for each group, which then becomes the observed outcome for the meta-analysis. Here, one needs to specify mi, sdi, and ni for the observed means, the observed standard deviations, and the sample sizes, respectively. For ratio scale measurements, the log transformed mean or the log transformed coefficient of variation (with bias correction) may also be of interest (Nakagawa et al., 2015). If focus is solely on the variability of the measurements, then the log transformed standard deviation (with bias correction) is a useful measure (Nakagawa et al., 2015; Raudenbush & Bryk, 1987). Here, one only needs to specify sdi and ni. The options for the measure argument are:

"MN" for the raw mean.

"MNLN" for the log transformed mean.

"CVLN" for the log transformed coefficient of variation.

"SDLN" for the log transformed standard deviation.

Note that sdi is used to specify the standard deviations of the observed values of the response, characteristic, or dependent variable and not the standard errors of the means. Also, the sampling variance for measure="CVLN" is computed as given by equation 27 in Nakagawa et al. (2015), but without the ‘\(-2 \rho \ldots\)’ term, since for normally distributed data (which we assume here) the mean and variance (and transformations thereof) are independent.

Outcome Measures for Change or Matched Pairs
A more complicated situation arises when the purpose of the meta-analysis is to assess the amount of change within individual groups (e.g., before and after a treatment or under two different treatments) or when dealing with matched pairs designs.
Measures for Dichotomous Variables
For dichotomous variables, the data for a study of this type gives rise to a paired \(2 \times 2\) table, which is of the form:

trt 2 outcome 1	trt 2 outcome 2

trt 1 outcome 1	ai	bi

trt 1 outcome 2	ci	di

where ai, bi, ci, and di denote the cell frequencies. Note that ‘trt1’ and ‘trt2’ may be applied to a single group of subjects or to matched pairs of subjects. The data from such a study can be rearranged into a marginal table of the form:

outcome 1	outcome 2

trt 1	ai+bi	ci+di

trt 2	ai+ci	bi+di

which is of the same form as a \(2 \times 2\) table that would arise in a study comparing/contrasting two independent groups. The options for the measure argument that will compute outcome measures based on the marginal table are:

"MPRR" for the matched pairs marginal log risk ratio.

"MPOR" for the matched pairs marginal log odds ratio.

"MPRD" for the matched pairs marginal risk difference.

See Becker and Balagtas (1993), Curtin et al. (2002), Elbourne et al. (2002), Fagerland et al. (2014), May and Johnson (1997), Newcombe (1998), Stedman et al. (2011), and Zou (2007) for discussions of these measures. The options for the measure argument that will compute outcome measures based on the paired table are:

"MPORC" for the conditional log odds ratio.

"MPPETO" for the conditional log odds ratio estimated with Peto's method.

See Curtin et al. (2002) and Zou (2007) for discussions of these measures.
Measures for Quantitative Variables
When the response or dependent variable assessed in the individual studies is measured on some quantitative scale, the raw mean change, standardized versions thereof, or the (log transformed) ratio of means (log response ratio) can be used as outcome measures (Becker, 1988; Gibbons et al., 1993; Lajeunesse, 2011; Morris, 2000). Here, one needs to specify m1i and m2i, the observed means at the two measurement occasions, sd1i and sd2i for the corresponding observed standard deviations, ri for the correlation between the scores observed at the two measurement occasions, and ni for the sample size. The options for the measure argument are then:

"MC" for the raw mean change.

"SMCC" for the standardized mean change using change score standardization.

"SMCR" for the standardized mean change using raw score standardization.

"SMCRH" for the standardized mean change using raw score standardization with heteroscedastic population variances at the two measurement occasions (Bonett, 2008).

"ROMC" for the log transformed ratio of means (Lajeunesse, 2011).

See also Morris and DeShon (2002) for a thorough discussion of the difference between the change score measures. A few notes about the change score measures. In practice, one often has a mix of information available from the individual studies to compute these measures. In particular, if m1i and m2i are unknown, but the raw mean change is directly reported in a particular study, then you can set m1i to that value and m2i to 0 (making sure that the raw mean change was computed as m1i-m2i within that study and not the other way around). Also, for the raw mean change ("MC") or the standardized mean change using change score standardization ("SMCC"), if sd1i, sd2i, and ri are unknown, but the standard deviation of the change scores is directly reported, then you can set sd1i to that value and both sd2i and ri to 0. Finally, for the standardized mean change using raw score standardization ("SMCR"), argument sd2i is actually not needed, as the standardization is only based on sd1i (Becker, 1988; Morris, 2000), which is usually the pre-test standard deviation (if the post-test standard deviation should be used, then set sd1i to that). Finally, all of these measures are also applicable for matched-pairs designs (subscripts 1 and 2 then simply denote the first and second group that are formed by the matching).

Other Outcome Measures for Meta-Analyses
Other outcome measures are sometimes used for meta-analyses that do not directly fall into the categories above. These are described in this section.
Cronbach's alpha and Transformations Thereof
Meta-analytic methods can also be used to aggregate Cronbach's alpha values. This is usually referred to as a ‘reliability generalization meta-analysis’ (Vacha-Haase, 1998). Here, one needs to specify ai, mi, and ni for the observed alpha values, the number of items/replications/parts of the measurement instrument, and the sample sizes, respectively. One can either directly analyze the raw Cronbach's alpha values or transformations thereof (Bonett, 2002, 2010; Hakstian & Whalen, 1976). The options for the measure argument are then:

"ARAW" for raw alpha values.

"AHW" for transformed alpha values (Hakstian & Whalen, 1976).

"ABT" for transformed alpha values (Bonett, 2002).

Note that the transformations implemented here are slightly different from the ones described by Hakstian and Whalen (1976) and Bonett (2002). In particular, for "AHW", the transformation \(1-(1-\alpha)^{1/3}\) is used, while for "ABT", the transformation \(-ln(1-\alpha)\) is used. This ensures that the transformed values are monotonically increasing functions of \(\alpha\). A dataset corresponding to data of this type is provided in dat.bonett2010.
Partial and Semi-Partial Correlations
Aloe and Becker (2012), Aloe and Thompson (2013), and Aloe (2014) describe the use of partial and semi-partial correlation coefficients as a method for meta-analyzing the results from regression models (when the focus is on a common regression coefficient of interest across studies). To compute these measures, one needs to specify ti for the test statistics (i.e., t-tests) of the regression coefficient of interest, ni for the sample sizes of the studies, mi for the number of predictors in the regression models, and r2i for the \(R^2\) value of the regression models (the latter is only needed when measure="SPCOR"). The options for the measure argument are then:

"PCOR" for the partial correlation coefficient.

"ZPCOR" for Fisher's r-to-z transformed partial correlation coefficient.

"SPCOR" for the semi-partial correlation coefficient.

Note that the sign of the (semi-)partial correlation coefficients is determined based on the signs of the values specified via the ti argument. Also, Fisher's variance stabilizing transformation can only be applied to partial correlation coefficient, not semi-partial coefficients.

Converting a Data Frame to an 'escalc' Object
The function can also be used to convert a regular data frame to an 'escalc' object. One simply sets the measure argument to one of the options described above (or to measure="GEN" for a generic outcome measure not further specified) and passes the observed effect sizes or outcomes via the yi argument and the corresponding sampling variances via the vi argument (or the standard errors via the sei argument).
Value

An object of class c("escalc","data.frame"). The object is a data frame containing the following components:

yi

observed outcomes or effect size estimates.

vi

corresponding (estimated) sampling variances.

If append=TRUE and a data frame was specified via the data argument, then yi and vi are append to this data frame. Note that the var.names argument actually specifies the names of these two variables. If the data frame already contains two variables with names as specified by the var.names argument, the values for these two variables will be overwritten when replace=TRUE (which is the default). By setting replace=FALSE, only values that are NA will be replaced. The object is formatted and printed with the print.escalc function. The summary.escalc function can be used to obtain confidence intervals for the individual outcomes.
Note

The variable names specified under var.names should be syntactically valid variable names. If necessary, they are adjusted so that they are.

Although the default value for add is 1/2, for certain measures the use of such a bias correction makes little sense and for these measures, the function internally sets add = 0. This applies to the following measures: "AS", "PHI", "RTET", "IRSD", "PAS", "PFT", "IRS", and "IRFT". One can still force the use of the bias correction by explicitly setting the add argument to some non-zero value.

References

Aloe, A. M. (2014). An empirical investigation of partial effect sizes in meta-analysis of correlational data. Journal of General Psychology, 141, 47--64.

Aloe, A. M., & Becker, B. J. (2012). An effect size for regression predictors in meta-analysis. Journal of Educational and Behavioral Statistics, 37, 278--297.

Aloe, A. M., & Thompson, C. G. (2013). The synthesis of partial effect sizes. Journal of the Society for Social Work and Research, 4, 390--405.

Bagos, P. G., & Nikolopoulos, G. K. (2009). Mixed-effects Poisson regression models for meta-analysis of follow-up studies with constant or varying durations. The International Journal of Biostatistics, 5(1), article 21.

Becker, B. J. (1988). Synthesizing standardized mean-change measures. British Journal of Mathematical and Statistical Psychology, 41, 257--278.

Becker, M. P., & Balagtas, C. C. (1993). Marginal modeling of binary cross-over data. Biometrics, 49, 997--1009.

Bonett, D. G. (2002). Sample size requirements for testing and estimating coefficient alpha. Journal of Educational and Behavioral Statistics, 27, 335--340.

Bonett, D. G. (2008). Confidence intervals for standardized linear contrasts of means. Psychological Methods, 13, 99--109.

Bonett, D. G. (2009). Meta-analytic interval estimation for standardized and unstandardized mean differences. Psychological Methods, 14, 225--238.

Bonett, D. G. (2010). Varying coefficient meta-analytic methods for alpha reliability. Psychological Methods, 15, 368--385.

Borenstein, M. (2009). Effect sizes for continuous data. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (2nd ed., pp. 221--235). New York: Russell Sage Foundation.

Chinn, S. (2000). A simple method for converting an odds ratio to effect size for use in meta-analysis. Statistics in Medicine, 19, 3127--3131.

Cox, D. R., & Snell, E. J. (1989). Analysis of binary data (2nd ed.). London: Chapman & Hall.

Curtin, F., Elbourne, D., & Altman, D. G. (2002). Meta-analysis combining parallel and cross-over clinical trials. II: Binary outcomes. Statistics in Medicine, 21, 2145--2159.

Elbourne, D. R., Altman, D. G., Higgins, J. P. T., Curtin, F., Worthington, H. V., & Vail, A. (2002). Meta-analyses involving cross-over trials: Methodological issues. International Journal of Epidemiology, 31, 140--149.

Fagerland, M. W., Lydersen, S., & Laake, P. (2014). Recommended tests and confidence intervals for paired binomial proportions. Statistics in Medicine, 33, 2850--2875.

Fisher, R. A. (1921). On the “probable error” of a coefficient of correlation deduced from a small sample. Metron, 1, 1--32.

Fleiss, J. L., & Berlin, J. (2009). Effect sizes for dichotomous data. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (2nd ed., pp. 237--253). New York: Russell Sage Foundation.

Freeman, M. F., & Tukey, J. W. (1950). Transformations related to the angular and the square root. Annals of Mathematical Statistics, 21, 607--611.

Gibbons, R. D., Hedeker, D. R., & Davis, J. M. (1993). Estimation of effect size from a series of experiments involving paired comparisons. Journal of Educational Statistics, 18, 271--279.

Hakstian, A. R., & Whalen, T. E. (1976). A k-sample significance test for independent alpha coefficients. Psychometrika, 41, 219--231.

Hasselblad, V., & Hedges, L. V. (1995). Meta-analysis of screening and diagnostic tests. Psychological Bulletin, 117(1), 167-178.

Hedges, L. V. (1981). Distribution theory for Glass's estimator of effect size and related estimators. Journal of Educational Statistics, 6, 107--128.

Hedges, L. V. (1989). An unbiased correction for sampling error in validity generalization studies. Journal of Applied Psychology, 74, 469--477.

Hedges, L. V., Gurevitch, J., & Curtis, P. S. (1999). The meta-analysis of response ratios in experimental ecology. Ecology, 80, 1150--1156.

Higgins, J. P. T., & Green, S. (Eds.) (2008). Cochrane handbook for systematic reviews of interventions. Chichester, England: Wiley.

Jacobs, P., & Viechtbauer, W. (2017). Estimation of the biserial correlation and its sampling variance for use in meta-analysis. Research Synthesis Methods, 8, 161--180.

Kirk, D. B. (1973). On the numerical approximation of the bivariate normal (tetrachoric) correlation coefficient. Psychometrika, 38, 259--268.

Lajeunesse, M. J. (2011). On the meta-analysis of response ratios for studies with correlated and multi-group designs. Ecology, 92, 2049--2055.

May, W. L., & Johnson, W. D. (1997). Confidence intervals for differences in correlated binary proportions. Statistics in Medicine, 16, 2127--2136.

Morris, S. B. (2000). Distribution of the standardized mean change effect size for meta-analysis on repeated measures. British Journal of Mathematical and Statistical Psychology, 53, 17--29.

Morris, S. B., & DeShon, R. P. (2002). Combining effect size estimates in meta-analysis with repeated measures and independent-groups designs. Psychological Methods, 7, 105--125.

Nakagawa, S., Poulin, R., Mengersen, K., Reinhold, K., Engqvist, L., Lagisz, M., & Senior, A. M. (2015). Meta-analysis of variation: Ecological and evolutionary applications and beyond. Methods in Ecology and Evolution, 6, 143--152.

Newcombe, R. G. (1998). Improved confidence intervals for the difference between binomial proportions based on paired data. Statistics in Medicine, 17, 2635--2650.

Olkin, I., & Pratt, J. W. (1958). Unbiased estimation of certain correlation coefficients. Annals of Mathematical Statistics, 29, 201--211.

Pearson, K. (1900). Mathematical contribution to the theory of evolution. VII. On the correlation of characters not quantitatively measurable. Philosophical Transactions of the Royal Society of London, Series A, 195, 1--47.

Pearson, K. (1909). On a new method of determining correlation between a measured character A, and a character B, of which only the percentage of cases wherein B exceeds (or falls short of) a given intensity is recorded for each grade of A. Biometrika, 7, 96--105.

Raudenbush, S. W., & Bryk, A. S. (1987). Examining correlates of diversity. Journal of Educational Statistics, 12, 241--269.

Rothman, K. J., Greenland, S., & Lash, T. L. (2008). Modern epidemiology (3rd ed.). Philadelphia: Lippincott Williams & Wilkins.

Rücker, G., Schwarzer, G., Carpenter, J., & Olkin, I. (2009). Why add anything to nothing? The arcsine difference as a measure of treatment effect in meta-analysis with zero cells. Statistics in Medicine, 28, 721--738.

Sánchez-Meca, J., Marín-Martínez, F., & Chacón-Moscoso, S. (2003). Effect-size indices for dichotomized outcomes in meta-analysis. Psychological Methods, 8, 448--467.

Soper, H. E. (1914). On the probable error of the bi-serial expression for the correlation coefficient. Biometrika, 10, 384--390.

Stedman, M. R., Curtin, F., Elbourne, D. R., Kesselheim, A. S., & Brookhart, M. A. (2011). Meta-analyses involving cross-over trials: Methodological issues. International Journal of Epidemiology, 40, 1732--1734.

Tate, R. F. (1954). Correlation between a discrete and a continuous variable: Point-biserial correlation. Annals of Mathematical Statistics, 25, 603--607.

Vacha-Haase, T. (1998). Reliability generalization: Exploring variance in measurement error affecting score reliability across studies. Educational and Psychological Measurement, 58, 6--20.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. http://www.jstatsoft.org/v36/i03/.

Yule, G. U. (1912). On the methods of measuring association between two attributes. Journal of the Royal Statistical Society, 75, 579--652.

Yusuf, S., Peto, R., Lewis, J., Collins, R., & Sleight, P. (1985). Beta blockade during and after myocardial infarction: An overview of the randomized trials. Progress in Cardiovascular Disease, 27, 335--371.

Zou, G. Y. (2007). One relative risk versus two odds ratios: Implications for meta-analyses involving paired and unpaired binary data. Clinical Trials, 4, 25--31.

See also

Examples

### load BCG vaccine data
dat <- get(data(dat.bcg))

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat)
dat

#>    trial               author year tpos  tneg cpos  cneg ablat      alloc 
#> 1      1              Aronson 1948    4   119   11   128    44     random 
#> 2      2     Ferguson & Simes 1949    6   300   29   274    55     random 
#> 3      3      Rosenthal et al 1960    3   228   11   209    42     random 
#> 4      4    Hart & Sutherland 1977   62 13536  248 12619    52     random 
#> 5      5 Frimodt-Moller et al 1973   33  5036   47  5761    13  alternate 
#> 6      6      Stein & Aronson 1953  180  1361  372  1079    44  alternate 
#> 7      7     Vandiviere et al 1973    8  2537   10   619    19     random 
#> 8      8           TPT Madras 1980  505 87886  499 87892    13     random 
#> 9      9     Coetzee & Berjak 1968   29  7470   45  7232    27     random 
#> 10    10      Rosenthal et al 1961   17  1699   65  1600    42 systematic 
#> 11    11       Comstock et al 1974  186 50448  141 27197    18 systematic 
#> 12    12   Comstock & Webster 1969    5  2493    3  2338    33 systematic 
#> 13    13       Comstock et al 1976   27 16886   29 17825    33 systematic 
#>         yi     vi 
#> 1  -0.8893 0.3256 
#> 2  -1.5854 0.1946 
#> 3  -1.3481 0.4154 
#> 4  -1.4416 0.0200 
#> 5  -0.2175 0.0512 
#> 6  -0.7861 0.0069 
#> 7  -1.6209 0.2230 
#> 8   0.0120 0.0040 
#> 9  -0.4694 0.0564 
#> 10 -1.3713 0.0730 
#> 11 -0.3394 0.0124 
#> 12  0.4459 0.5325 
#> 13 -0.0173 0.0714 

### suppose that for a particular study, yi and vi are known (i.e., have
### already been calculated) but the 2x2 table counts are not known; with
### replace=FALSE, the yi and vi values for that study are not replaced
dat[1:12,10:11] <- NA
dat[13,4:7] <- NA
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat, replace=FALSE)
dat

#>    trial               author year tpos  tneg cpos  cneg ablat      alloc 
#> 1      1              Aronson 1948    4   119   11   128    44     random 
#> 2      2     Ferguson & Simes 1949    6   300   29   274    55     random 
#> 3      3      Rosenthal et al 1960    3   228   11   209    42     random 
#> 4      4    Hart & Sutherland 1977   62 13536  248 12619    52     random 
#> 5      5 Frimodt-Moller et al 1973   33  5036   47  5761    13  alternate 
#> 6      6      Stein & Aronson 1953  180  1361  372  1079    44  alternate 
#> 7      7     Vandiviere et al 1973    8  2537   10   619    19     random 
#> 8      8           TPT Madras 1980  505 87886  499 87892    13     random 
#> 9      9     Coetzee & Berjak 1968   29  7470   45  7232    27     random 
#> 10    10      Rosenthal et al 1961   17  1699   65  1600    42 systematic 
#> 11    11       Comstock et al 1974  186 50448  141 27197    18 systematic 
#> 12    12   Comstock & Webster 1969    5  2493    3  2338    33 systematic 
#> 13    13       Comstock et al 1976   NA    NA   NA    NA    33 systematic 
#>         yi     vi 
#> 1  -0.8893 0.3256 
#> 2  -1.5854 0.1946 
#> 3  -1.3481 0.4154 
#> 4  -1.4416 0.0200 
#> 5  -0.2175 0.0512 
#> 6  -0.7861 0.0069 
#> 7  -1.6209 0.2230 
#> 8   0.0120 0.0040 
#> 9  -0.4694 0.0564 
#> 10 -1.3713 0.0730 
#> 11 -0.3394 0.0124 
#> 12  0.4459 0.5325 
#> 13 -0.0173 0.0714 

### convert a regular data frame to an 'escalc' object
### dataset from Lipsey & Wilson (2001), Table 7.1, page 130
dat <- data.frame(id = c(100, 308, 1596, 2479, 9021, 9028, 161, 172, 537, 7049),
                  yi = c(-0.33, 0.32, 0.39, 0.31, 0.17, 0.64, -0.33, 0.15, -0.02, 0.00),
                  vi = c(0.084, 0.035, 0.017, 0.034, 0.072, 0.117, 0.102, 0.093, 0.012, 0.067),
                  random = c(0, 0, 0, 0, 0, 0, 1, 1, 1, 1),
                  intensity = c(7, 3, 7, 5, 7, 7, 4, 4, 5, 6))
dat <- escalc(measure="SMD", yi=yi, vi=vi, data=dat, slab=paste("Study ID:", id), digits=3)
dat

#>      id     yi    vi random intensity 
#> 1   100 -0.330 0.084      0         7 
#> 2   308  0.320 0.035      0         3 
#> 3  1596  0.390 0.017      0         7 
#> 4  2479  0.310 0.034      0         5 
#> 5  9021  0.170 0.072      0         7 
#> 6  9028  0.640 0.117      0         7 
#> 7   161 -0.330 0.102      1         4 
#> 8   172  0.150 0.093      1         4 
#> 9   537 -0.020 0.012      1         5 
#> 10 7049  0.000 0.067      1         6