Fitted Values for 'rma' Objects

The function computes the fitted values for objects of class "rma".

# S3 method for rma
fitted(object, ...)

Arguments

object
an object of class "rma".

...
other arguments.

Value

A vector with the fitted values.

Note

The predict.rma function also provides standard errors and confidence intervals for the fitted values. Best linear unbiased predictions (BLUPs) that combine the fitted values based on the fixed effects and the estimated contributions of the random effects can be obtained with blup.rma.uni (only for objects of class "rma.uni").

For objects not involving moderators, the fitted values are all identical to the estimated value of the model intercept.

Author

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://doi.org/10.18637/jss.v036.i03

See also

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### fit mixed-effects model with absolute latitude and publication year as moderators
res <- rma(yi, vi, mods = ~ ablat + year, data=dat)

### compute the fitted values
fitted(res)

#>          1          2          3          4          5          6          7 
#> -1.0620809 -1.3682974 -0.9831677 -1.2308520 -0.1460425 -1.0525432 -0.3141101 
#>          8          9         10         11         12         13 
#> -0.1326896 -0.5477381 -0.9812602 -0.2841913 -0.7138982 -0.7005453