influence.rma.uni {metafor}	R Documentation

Outlier and Influential Case Diagnostics for 'rma.uni' Objects

Description

The functions compute various outlier and influential case diagnostics (some of which indicate the influence of deleting one case/study at a time on the model fit and the fitted/residual values) for objects of class "rma.uni".

Usage

## S3 method for class 'rma.uni'
influence(model, digits, progbar=FALSE, ...)

## S3 method for class 'infl.rma.uni'
print(x, digits=x$digits, infonly=FALSE, ...)

## S3 method for class 'rma.uni'
cooks.distance(model, progbar=FALSE, ...)
## S3 method for class 'rma.uni'
dfbetas(model, progbar=FALSE, ...)
## S3 method for class 'rma.uni'
hatvalues(model, type="diagonal", ...)

Arguments

model
an object of class "rma.uni".

x
an object of class "infl.rma.uni" (for print).

digits
integer specifying the number of decimal places to which the printed results should be rounded (if unspecified, the default is to take the value from the object).

progbar
logical indicating whether a progress bar should be shown (the default is FALSE).

infonly
logical indicating whether only the influential cases should be printed (the default is FALSE).

type
character string indicating whether to return only the diagonal of the hat matrix ("diagonal") or the entire hat matrix ("matrix").

...
other arguments.

Details

The influence function calculates the following leave-one-out diagnostics for each study:

externally standardized residual,

DFFITS value,

Cook's distance,

covariance ratio,

the leave-one-out amount of (residual) heterogeneity,

the leave-one-out test statistic for the test of (residual) heterogeneity,

DFBETAS value(s).

The diagonal elements of the hat matrix and the weights (in %) given to the observed effects or outcomes during the model fitting are also provided (except for their scaling, the hat values and weights are the same for models without moderators, but will differ when moderators are included).

For details on externally standardized residuals, see rstudent.rma.uni.

The DFFITS value essentially indicates how many standard deviations the predicted (average) effect for the ith study changes after excluding the ith study from the model fitting.

Cook's distance can be interpreted as the Mahalanobis distance between the entire set of predicted values once with the ith study included and once with the ith study excluded from the model fitting.

The covariance ratio is defined as the determinant of the variance-covariance matrix of the parameter estimates based on the dataset with the ith study removed divided by the determinant of the variance-covariance matrix of the parameter estimates based on the complete dataset. A value below 1 therefore indicates that removal of the ith study yields more precise estimates of the model coefficients.

The leave-one-out amount of (residual) heterogeneity is the estimated value of τ² based on the dataset with the ith study removed. Note that this is always equal to 0 for fixed-effects models.

Similarly, the leave-one-out test statistic for the test of (residual) heterogeneity is the value of the test statistic of the test for (residual) heterogeneity calculated based on the dataset with the ith study removed.

Finally, the DFBETAS value(s) essentially indicate(s) how many standard deviations the estimated coefficient(s) change(s) after excluding the ith study from the model fitting.

A study may be considered to be ‘influential’ if at least one of the following is true:

The absolute DFFITS value is larger than 3√(p/(k-p)), where p is the number of model coefficients and k the number of studies.

The lower tail area of a chi-square distribution with p degrees of freedom cut off by the Cook's distance is larger than 50%.

The hat value is larger than 3(p/k).

Any DFBETAS value is larger than 1.

Studies which are considered influential with respect to any of these measures are marked with an asterisk. Note that the chosen cut-offs are (somewhat) arbitrary. Substantively informed judgment should always be used when examining the influence of each study on the results.

Value

An object of class "infl.rma.uni", which is a list containing the following components:

inf
an element of class "list.rma" with the externally standardized residuals, DFFITS values, Cook's distances, covariance ratios, leave-one-out τ² estimates, leave-one-out (residual) heterogeneity test statistics, hat values, weights, and an indicator whether a study is influential or not.

dfbs
an element of class "list.rma" with the the DFBETAS values.

...
some additional elements/values.

The results are printed with print.infl.rma.uni and plotted with plot.infl.rma.uni.

Note

Right now, leave-one-out diagnostics are calculated by refitting the model k times. Depending on how large k is, it may take a few moments to finish the calculations. There are shortcuts for calculating at least some of these values without refitting the model each time, but these are currently not implemented (and may not exist for all of the leave-one-out diagnostics calculated by the function).

It may not be possible to fit the model after deletion of the ith study from the dataset. This will result in NA values for that study.

Certain relationships between the leave-one-out diagnostics and the (internally or externally) standardized residuals (Belsley, Kuh, & Welsch, 1980; Cook & Weisberg, 1982) no longer hold for the meta-analytic models. Maybe there are other relationships. These remain to be determined.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Belsley, D. A., Kuh, E., & Welsch, R. E. (1980). Regression diagnostics. New York: Wiley.

Cook, R. D., & Weisberg, S. (1982). Residuals and influence in regression. London: Chapman and Hall.

Hedges, L. V., & Olkin, I. (1985). Statistical methods for meta-analysis. San Diego, CA: Academic Press.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

Viechtbauer, W., & Cheung, M. W.-L. (2010). Outlier and influence diagnostics for meta-analysis. Research Synthesis Methods, 1, 112–125.

See Also

plot.infl.rma.uni, rstudent.rma.uni, weights.rma.uni

Examples

### meta-analysis of the log risk ratios using a mixed-effects model
### with two moderators (absolute latitude and publication year)
res <- rma(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg,
           mods = ~ ablat + year, data=dat.bcg)
influence(res)
plot(influence(res))

cooks.distance(res)
dfbetas(res)
hatvalues(res)

[Package metafor version 2.4-0 Index]