leave1out.rma.uni {metafor}	R Documentation

Leave-One-Out Diagnostics for rma.uni Objects

Description

The function leave1out.rma.uni repeatedly fits the specified model, leaving out one observation at a time.

Usage

## S3 method for class 'rma.uni':
leave1out(x, digits=x$digits, transf=FALSE, targs=NULL, ...)

Arguments

x 	 an object of class "rma.uni".

digits 	 an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

transf 	 an optional argument specifying the name of a function that should be used to transform the model coefficients and interval bounds (e.g., transf=exp). Defaults to FALSE, which means that no transformation is used.

targs 	 optional arguments needed by the function specified under transf.

... 	 other arguments.

Details

The model specified by x must be a model without moderators (i.e., either a fixed- or a random-effects model).

Value

An object of class "list.rma". The object is a list containing the following components:

estimate 	 estimated coefficients of the model.

se 	 standard errors of the coefficients. NA if transf is used to transform the coefficients.

zval 	 test statistics of the coefficients.

pval 	 p-values for the test statistics.

ci.lb 	 lower bounds of the confidence intervals for the coefficients.

ci.ub 	 upper bounds of the confidence intervals for the coefficients.

Q 	 test statistics for the tests of heterogeneity.

Qp 	 p-values for the tests of heterogeneity.

tau2 	 estimated amounts of (residual) heterogeneity (only for random-effects models).

I2 	 values of I^2 (only for random-effects models).

H2 	 values of H^2 (only for random-effects models).

The "list.rma" object is formated and printed with print.list.rma.

Note

Various case diagnostics for objects of class "rma.uni" can also be obtained with the influence.rma.uni function.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

leave1out, influence.rma.uni

Examples

### load BCG vaccine data
data(dat.bcg)

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)
dat <- cbind(dat.bcg, dat)

### random-effects model
res <- rma(yi, vi, data=dat, method="REML")

leave1out(res)
leave1out(res, transf=exp)

[Package metafor version 0.5-7 Index]