plot.rma.uni {metafor}	R Documentation

Plot Method for rma Objects

Description

Plot method for objects of class "rma.uni", "rma.mh", and "rma.peto".

Usage

## S3 method for class 'rma.uni':
plot(x, qqplot=FALSE, ...)
## S3 method for class 'rma.mh':
plot(x, qqplot=FALSE, ...)
## S3 method for class 'rma.peto':
plot(x, qqplot=FALSE, ...)

Arguments

x 	 an object of class "rma.uni", "rma.mh", or "rma.peto".

qqplot 	 logical indicating whether a normal QQ plot should be drawn (default is FALSE).

... 	 other arguments.

Details

Four plots are produced. If the model does not contain any moderators, then a forest plot, funnel plot, radial plot, and a plot of the standardized residuals is provided. If qqplot=TRUE, the last plot is replaced by a normal QQ plot of the standardized residuals.

If the model contains moderators, then a forest plot, funnel plot, plot of the standardized residuals against the fitted values, and a plot of the standardized residuals is provided. If qqplot=TRUE, the last plot is replaced by a normal QQ plot of the standardized residuals.

Note

If the number of studies is large, the forest plot may become quite impossible to read.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

forest, funnel, radial, qqnorm.rma.uni

Examples

### load BCG vaccine data
data(dat.bcg)

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg, 
           measure="RR", method="REML")
plot(res, qqplot=TRUE)

### mixed-effects model with two moderators (absolute latitude and publication year)
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, mods=cbind(ablat, year),
           data=dat.bcg, measure="RR", method="REML")
plot(res, qqplot=TRUE)

[Package metafor version 0.5-7 Index]