print.anova.rma.uni {metafor}	R Documentation

Print Method for anova.rma.uni Objects

Description

Print method for objects of class "anova.rma.uni".

Usage

## S3 method for class 'anova.rma.uni':
print(x, digits=x$digits, ...)

Arguments

x 	 an object of class "anova.rma.uni".

digits 	 an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

... 	 other arguments.

Details

The output includes:

the number of parameters in the full and the reduced model.
the AIC, BIC, and log likelihood of the full and the reduced model.
the value of the likelihood ratio test statistic.
the p-value for the likelihood ratio test.
the test statistic for the test of (residual) heterogeneity for the full and the reduced model.
the estimate of tau^2 from the full and the reduced model. Suppressed for fixed-effects models.
the amount of (residual) heterogeneity in the reduced model that is accounted for in the full model (in percent). NA for fixed-effects models or if the amount of heterogeneity in the reduced model is equal to zero.

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

anova.rma.uni

[Package metafor version 0.5-7 Index]