print.confint.rma {metafor}	R Documentation

Print Methods for 'confint.rma' and 'list.confint.rma' Objects

Description

Print methods for objects of class "confint.rma" and "list.confint.rma".

Usage

## S3 method for class 'confint.rma'
print(x, digits=x$digits, ...)
## S3 method for class 'list.confint.rma'
print(x, digits=x$digits, ...)

Arguments

x
an object of class "confint.rma" or "list.confint.rma" obtained with confint.

digits
integer to specify the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

...
other arguments.

Details

The output includes:

estimate of the model coefficient or variance/correlation parameter

lower bound of the confidence interval

upper bound of the confidence interval

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

See Also

confint.rma.uni and confint.rma.mv for the functions to create confint.rma and list.confint.rma objects.

[Package metafor version 3.4-0 Index]