print.gosh.rma {metafor}	R Documentation

Print Method for 'gosh.rma' Objects

Description

Print method for objects of class "gosh.rma".

Usage

## S3 method for class 'gosh.rma'
print(x, digits=x$digits, ...)

Arguments

x
an object of class "gosh.rma" obtained with gosh.

digits
integer to specify the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

...
other arguments.

Details

The output shows how many model fits were attempted, how many succeeded, and summary statistics (i.e., the mean, minimum, first quartile, median, third quartile, and maximum) for the various measures of (residual) heterogeneity and the model coefficient(s) computed across all of the subsets.

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

See Also

gosh.rma for the function to create gosh.rma objects.

[Package metafor version 3.4-0 Index]