Print Method for 'hc.rma.uni' Objects

Function to print objects of class "hc.rma.uni".

# S3 method for class 'hc.rma.uni'
print(x, digits=x$digits, ...)

Arguments

x

an object of class "hc.rma.uni" obtained with hc.

digits

integer to specify the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

...

other arguments.

Details

The output is a data frame with two rows, the first (labeled rma) corresponding to the results based on the usual estimation method, the second (labeled hc) corresponding to the results based on the method by Henmi and Copas (2010). The data frame includes the following variables:

the method used to estimate \(\tau^2\) (always DL for hc)

the estimated amount of heterogeneity

the estimated average true outcome

the corresponding standard error (NA when transf argument has been used)

the lower and upper confidence interval bounds

Value

The function returns the data frame invisibly.

Author

Wolfgang Viechtbauer (wvb@metafor-project.org, https://www.metafor-project.org).

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

See also