print.infl.rma.uni {metafor}	R Documentation

Print Method for infl.rma.uni Objects

Description

Print method for objects of class "infl.rma.uni".

Usage

## S3 method for class 'infl.rma.uni':
print(x, digits=x$digits, ...)

Arguments

x 	 an object of class "infl.rma.uni".

digits 	 an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

... 	 other arguments.

Value

See the documentation of the influence.rma.uni function for details on what is printed.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

influence.rma.uni

[Package metafor version 0.5-7 Index]