Print method for 'list.rma' Objects

Print method for objects of class "list.rma".

# S3 method for list.rma
print(x, digits=x$digits, ...)

Arguments

x
an object of class "list.rma".

digits
integer to specify the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

...
other arguments.

Value

See the documentation of the function that creates the "list.rma" object for details on what is printed. Regardless of what is printed, a data frame with the results is also returned invisibly.

Author

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://doi.org/10.18637/jss.v036.i03