print.matreg {metafor}	R Documentation

Print Method for 'matreg' Objects

Description

Print method for objects of class "matreg".

Usage

## S3 method for class 'matreg'
print(x, digits=x$digits,
      signif.stars=getOption("show.signif.stars"), signif.legend=signif.stars, ...)

Arguments

x
an object of class "matreg" obtained with matreg.

digits
integer to specify the number of decimal places to which the printed results should be rounded. If unspecified, the default is to take the value from the object.

signif.stars
logical to specify whether p-values should be encoded visually with ‘significance stars’. Defaults to the show.signif.stars slot of options.

signif.legend
logical to specify whether the legend for the ‘significance stars’ should be printed. Defaults to the value for signif.stars.

...
other arguments.

Details

The output is a table with the estimated coefficients, corresponding standard errors, test statistics, p-values, and confidence interval bounds.

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

See Also

matreg for the function to create matreg objects.

[Package metafor version 3.4-0 Index]