Print Method for 'permutest.rma.uni' Objects

Print method for objects of class "permutest.rma.uni".

# S3 method for permutest.rma.uni
print(x, digits=x$digits, signif.stars=getOption("show.signif.stars"),
      signif.legend=signif.stars, ...)

Arguments

x

an object of class "permutest.rma.uni".

digits

integer to specify the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

signif.stars

logical to specify whether p-values should be encoded visually with ‘significance stars’. Defaults to the show.signif.stars slot of options.

signif.legend

logical to specify whether the legend for the ‘significance stars’ should be printed. Defaults to the value for signif.stars.

...

other arguments.

Details

The output includes:

the results of the omnibus test of moderators. Suppressed if the model includes only one coefficient (e.g., only an intercept, like in the fixed- and random-effects model). The p-value is based on the permutation test.

a table with the estimated coefficients, corresponding standard errors, test statistics, p-values, and confidence interval bounds. The p-values are based on permutation tests. If permci was set to TRUE, then the permutation-based CI bounds are shown.

Value

The function does not return an object.

Author

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://doi.org/10.18637/jss.v036.i03

See also