print.ranktest.rma {metafor}	R Documentation

Print Method for ranktest.rma Objects

Description

Print method for objects of class "ranktest.rma".

Usage

## S3 method for class 'ranktest.rma':
print(x, digits=x$digits, ...)

Arguments

x 	 an object of class "ranktest.rma".

digits 	 an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

... 	 other arguments.

Details

The output includes:

the estimated value of Kendall's tau rank correlation coefficient
the corresponding p-value for the test that the true tau is equal to zero

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

ranktest.rma

[Package metafor version 0.5-7 Index]