print.rma.uni {metafor}	R Documentation

Print Methods for rma Objects

Description

Print methods for objects of class "rma.uni", "rma.mh", and "rma.peto".

Usage

## S3 method for class 'rma.uni':
print(x, digits=x$digits, showfit=FALSE, signif.legend=TRUE, ...)
## S3 method for class 'rma.mh':
print(x, digits=x$digits, showfit=FALSE, ...)
## S3 method for class 'rma.peto':
print(x, digits=x$digits, showfit=FALSE, ...)
## S3 method for class 'rma':
summary(object, digits=object$digits, showfit=TRUE, signif.legend=TRUE, ...)

Arguments

x 	 an object of class "rma.uni", "rma.mh", or "rma.peto" (for print).

object 	 an object of class "rma" (for summary).

digits 	 an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

showfit 	 logical indicating whether the fit statistics and information criteria should be printed.

signif.legend 	 logical indicating whether the legend for the ‘significance stars’ should be printed.

... 	 other arguments.

Details

The output includes:

the log likelihood, deviance, AIC, and BIC value (when showfit=TRUE).
the amount of (residual) heterogeneity in the random/mixed-effects model (i.e., the estimate of tau^2 and its square root). Suppressed for fixed-effects models. For ML and REML estimation, the asymptotic standard error of the estimate of tau^2 is also provided.
the I^2 statistic estimates (in percent) how much of the total variability in the effect size estimates (which is composed of heterogeneity and sampling variability) can be attributed to heterogeneity among the true effects. Only supplied when fitting a random-effects model.
the H^2 statistic is the ratio of the total amount of variability in the observed outcomes to the amount of sampling variability. Only supplied when fitting a random-effects model.
the results of the test for (residual) heterogeneity. This is the usual Q-test for heterogeneity when not including moderators in the model and the QE-test for residual heterogeneity when moderators are included.
the results of the omnibus test of the coefficients in the model (the indices of the coefficients tested are also indicated). Suppressed if the model includes only one coefficient (e.g., only an intercept, like in the fixed- and random-effects model).
a table with the estimated coefficients, corresponding standard errors, test statistics, p-values, and confidence interval bounds.

When analyzing odds ratios using the Mantel-Haenszel method, the Cochran-Mantel-Haenszel test and Tarone's test for heterogeneity are also provided.

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

rma.uni, rma.mh, rma.peto

[Package metafor version 0.5-7 Index]