print.robust.rma {metafor}	R Documentation

Print Method for 'robust.rma' Objects

Description

Print method for objects of class "robust.rma".

Usage

## S3 method for class 'robust.rma'
print(x, digits=x$digits, signif.stars=getOption("show.signif.stars"),
      signif.legend=signif.stars, ...)

Arguments

x
an object of class "robust.rma".

digits
integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

signif.stars
logical indicating whether p-values should be encoded visually with ‘significance stars’. Defaults to the show.signif.stars slot of options.

signif.legend
logical indicating whether the legend for the ‘significance stars’ should be printed. Defaults to the value for signif.stars.

...
other arguments.

Details

The output includes:

information on the number of outcomes, number of clusters, and the number of outcomes per cluster.

the results of the omnibus (Wald-type) test of the coefficients in the model (the indices of the coefficients tested are also indicated). Suppressed if the model includes only one coefficient (e.g., only an intercept, like in the fixed- and random-effects model).

a table with the estimated coefficients, corresponding standard errors, test statistics, p-values, and confidence interval bounds.

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

robust.rma.uni, robust.rma.mv

[Package metafor version 2.4-0 Index]