profile.rma.uni {metafor}	R Documentation

Profile Plots for 'rma' Objects

Description

Function to profile the (restricted) log-likelihood for objects of class "rma.uni" and "rma.mv".

Usage

## S3 method for class 'rma.uni'
profile(fitted, xlim, ylim, steps=20,
        progbar=TRUE, parallel="no", ncpus=1, cl=NULL,
        plot=TRUE, pch=19, cline=FALSE, ...)

## S3 method for class 'rma.mv'
profile(fitted, sigma2, tau2, rho, gamma2, phi,
        xlim, ylim, steps=20, lltol=1e-06, startmethod="init",
        progbar=TRUE, parallel="no", ncpus=1, cl=NULL,
        plot=TRUE, pch=19, cline=FALSE, ...)

## S3 method for class 'profile.rma'
print(x, ...)
## S3 method for class 'profile.rma'
plot(x, xlim, ylim, pch=19,
     xlab, ylab, main, cline=FALSE, ...)

Arguments

fitted
an object of class "rma.uni" or "rma.mv".

x
an object of class "profile.rma" (for plot and print).

sigma2
integer specifying for which σ² value the likelihood should be profiled (only relevant for "rma.mv" objects).

tau2
integer specifying for which τ² value the likelihood should be profiled (only relevant for "rma.mv" objects).

rho
integer specifying for which ρ value the likelihood should be profiled (only relevant for "rma.mv" objects).

gamma2
integer specifying for which γ² value the likelihood should be profiled (only relevant for "rma.mv" objects).

phi
integer specifying for which φ value the likelihood should be profiled (only relevant for "rma.mv" objects).

xlim
optional vector specifying the lower and upper limit of the parameter over which the profiling should be done. If unspecified, the function tries to set these limits automatically.

ylim
optional vector specifying the y-axis limits when plotting the profiled likelihood. If unspecified, the function tries to set these limits automatically.

steps
number of points between xlim[1] and xlim[2] (inclusive) for which the likelihood should be evaluated (the default is 20).

lltol
numerical tolerance used when comparing values of the profiled log-likelihood with the log-likelihood of the fitted model (the default is 1e-06).

startmethod
method for picking starting values for the optimization. Default is "init", which starts each model fit at the default initial values. The alternative is "prev", which starts at the estimates from the previous model fit (usually faster, but also less stable for ill-defined models).

progbar
logical indicating whether a progress bar should be shown (the default is TRUE). Ignored when using parallel processing.

parallel
character string indicating whether parallel processing should be used (the default is "no"). For parallel processing, set to either "snow" or "multicore". See ‘Details’.

ncpus
integer specifying the number of processes to use in the parallel processing.

cl
optional snow cluster to use if parallel="snow". If not supplied, a cluster on the local machine is created for the duration of the call.

plot
logical indicating whether the profile plot should be drawn after profiling is finished (the default is TRUE).

pch
plotting symbol to use. By default, a filled circle is used. See points for other options.

cline
logical indicating whether a horizontal reference line should be added to the plot that indicates the log-likelihood value corresponding to the 95% profile confidence interval (the default is FALSE).

xlab
title for the x-axis. If unspecified, the function tries to set an appropriate axis title.

ylab
title for the y-axis. If unspecified, the function tries to set an appropriate axis title.

main
title for the plot. If unspecified, the function tries to set an appropriate title.

...
other arguments.

Details

The function fixes a particular variance component or correlation parameter of the model and then computes the maximized (restricted) log-likelihood over the remaining parameters of the model. By doing this for a range of values for the parameter that was fixed, a profile of the (restricted) log-likelihood is constructed.

For objects of class "rma.uni" obtained with the rma.uni function, the function profiles over parameter τ². If the model was fitted with method="ML" or method="REML", the profiled (restricted) log-likelihood should be maximized at the ML/REML estimate of τ².

For objects of class "rma.mv" obtained with the rma.mv function, profiling is done by default over all (non-fixed) variance and correlation components of the model. Alternatively, one can use the sigma2, tau2, rho, gamma2, or phi arguments to specify over which parameter the profiling should be done. Only one of these arguments can be used at a time. A single integer is used to specify the number of the parameter. Each profile plot should show a peak at the corresponding ML/REML estimate (the function checks whether any of the profiled log-likelihood values is larger than the log-likelihood of the fitted model, using a numerical tolerance of lltol; if so, a warning is issued as this might indicate that the optimizer did not identify the actual ML/REML estimates). If the profiled likelihood is flat (over the entire parameter space or large portions of it), then this suggests that at least some of the parameters of the model are not identifiable (and the parameter estimates obtained are to some extent arbitrary).

Profiling requires repeatedly refitting the same model, which can be slow when k is large and/or the model is complex (the latter especially applies to "rma.mv" objects). On machines with multiple cores, one can usually speed things up by delegating the model fitting to separate worker processes, that is, by setting parallel="snow" or parallel="multicore" and ncpus to some value larger than 1. Parallel processing makes use of the parallel package, using the makePSOCKcluster and parLapply functions when parallel="snow" or using mclapply when parallel="multicore" (the latter only works on Unix/Linux-alikes). With parallel::detectCores(), one can check on the number of available cores on the local machine. Note that argument startmethod is ignored when using parallel processing and no progress bar will be shown.

Value

An object of class "profile.rma". The object is a list (or list of lists) containing the following components:

sigma2
values of σ² over which the likelihood was profiled (only when profiling was actually done over σ²).

tau2
values of τ² over which the likelihood was profiled (only when profiling was actually done over τ²).

rho
values of ρ over which the likelihood was profiled (only when profiling was actually done over ρ).

gamma2
values of γ² over which the likelihood was profiled (only when profiling was actually done over γ²).

phi
values of φ over which the likelihood was profiled (only when profiling was actually done over φ).

ll
(restricted) log-likelihood at the corresponding parameter value.

beta
a matrix with the estimated model coefficients at the corresponding parameter value.

ci.lb
a matrix with the lower confidence interval bounds for the model coefficients at the corresponding parameter value.

ci.ub
a matrix with the upper confidence interval bounds for the model coefficients at the corresponding parameter value.

...
some additional elements/values.

Note that the list is returned invisibly.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

rma.uni, rma.mv

Examples

### calculate log odds ratios and corresponding sampling variances
dat <- escalc(measure="OR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### random-effects model using rma.uni()
res <- rma(yi, vi, data=dat)

### profile over tau^2
profile(res, progbar=FALSE)

### change data into long format
dat.long <- to.long(measure="OR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### set levels of group variable ("exp" = experimental/vaccinated; "con" = control/non-vaccinated)
levels(dat.long$group) <- c("exp", "con")

### set "con" to reference level
dat.long$group <- relevel(dat.long$group, ref="con")

### calculate log odds and corresponding sampling variances
dat.long <- escalc(measure="PLO", xi=out1, mi=out2, data=dat.long)

### bivariate random-effects model using rma.mv()
res <- rma.mv(yi, vi, mods = ~ group, random = ~ group | study, struct="UN", data=dat.long)
res

### profile over tau^2_1, tau^2_2, and rho
### note: for rho, adjust region over which profiling is done ('zoom in' on area around estimate)
## Not run: 
par(mfrow=c(3,1))
profile(res, tau2=1)
profile(res, tau2=2)
profile(res, rho=1, xlim=c(.90, .98))
## End(Not run)

[Package metafor version 2.4-0 Index]