Regression Test for Funnel Plot Asymmetry

The function can be used to carry out (various versions of) Egger's regression test for funnel plot asymmetry.

regtest(x, ...)

# S3 method for rma
regtest(x, model="rma", predictor="sei", ret.fit=FALSE, digits, ...)

# S3 method for default
regtest(x, vi, sei, ni, subset, model="rma", predictor="sei", ret.fit=FALSE, digits, ...)

Arguments

x
an object of class "rma" or a vector with the observed effect sizes or outcomes.

vi
vector with the corresponding sampling variances.

sei
vector with the corresponding standard errors.

ni
vector with the corresponding sample sizes.

subset
optional vector indicating the subset of studies that should be included in the test. This can be a logical vector of the same length as x or a numeric vector indicating the indices of the observations to include. Only relevant when passing a vector via x.

model
either "rma" or "lm" to indicate the type of model to use for the regression test. See ‘Details’.

predictor
either "sei" "vi", "ni", "ninv", "sqrtni", or "sqrtninv" to indicate the predictor to use for the regression test. See ‘Details’.

ret.fit
logical indicating whether the full results from the fitted model should also be returned.

digits
integer specifying the number of decimal places to which the printed results should be rounded (the default is 4).

...
other arguments.

Details

Various tests for funnel plot asymmetry have been suggested in the literature, including the rank correlation test by Begg and Mazumdar (1994) and the regression test by Egger et al. (1997). Extensions, modifications, and further developments of the regression test are described (among others) by Macaskill, Walter, and Irwig (2001), Sterne and Egger (2005), Harbord, Egger, and Sterne (2006), Peters et al. (2006), Rücker et al. (2008), and Moreno et al. (2009). The various versions of the regression test differ in terms of the model (either a weighted regression model with a multiplicative dispersion term or a fixed/random/mixed-effects meta-regression model is used), in terms of the predictor variable that the observed outcomes are hypothesized to be related to when publication bias is present (suggested predictors include the standard error, the sampling variance, and the sample size -- or transformations thereof), and in terms of the outcome measure used (e.g., for \(2 \times 2\) table data, one has the choice between various outcome measures). The idea behind the various tests is the same though: If there is a relationship between the observed outcomes and the chosen predictor, then this usually implies asymmetry in the funnel plot, which in turn may be an indication of publication bias.

The regtest.rma function can be used to carry out various versions of the regression test. The model is chosen via the model argument, with model="lm" for weighted regression with a multiplicative dispersion term or model="rma" for the meta-analytic models. In the latter case, arguments such as method, weighted, and test used during the initial model fitting are also used for the regression test. Therefore, if one wants to conduct the regression test with a random/mixed-effects model, one should first fit a random-effects model with the rma function and then use the regtest.rma function on the fitted model object.

The predictor is chosen via the predictor argument:

predictor="sei" for the standard error,

predictor="vi" for the sampling variance,

predictor="ni" for the sample size,

predictor="ninv" for the inverse of the sample size,

predictor="sqrtni" for the square root transformed sample size, and

predictor="sqrtninv" for the inverse of the square root transformed sample size.

For predictors based on the sample size, the object x obviously must contain the information about the sample sizes. This will automatically be the case when measure was not equal to "GEN" or the ni values were explicitly specified during the initial model fitting.

If the model passed to the regtest.rma function already included one or more moderators, then regtest.rma will add the chosen predictor to the moderator(s) already included in the model. This way, one can test for funnel plot asymmetry after accounting first for the influence of the moderator(s) already included.

One can also pass a vector of observed effect sizes or outcomes (via x) and the corresponding sampling variances via vi (or the standard errors via sei) directly to the function (in this case, the regtest.default function is used). When the predictor is the sample size or a transformation thereof, then ni needs to be specified here as well.

The outcome measure used for the regression test is simply determined by what measure was used in fitting the original model (or what values are passed to regtest.default).

Value

An object of class "regtest.rma". The object is a list containing the following components:

model

the model used for the regression test.

predictor

the predictor used for the regression test.

zval

the value of the test statistic.

pval

the corresponding p-value

dfs

the degrees of freedom of the test stastistic (if the test is based on a t-distribution).

fit

the full results from the fitted model.

The results are formatted and printed with the print.regtest.rma function.
Note

The classical ``Egger test'' is obtained by setting model="lm" and predictor="sei". For the random/mixed-effects version of the Egger test, one should first fit a random-effects model to the data and then set model="rma" and predictor="sei" when using the regtest.rma function. See Sterne and Egger (2005) for details on these two types of models/tests.

All of the tests do not directly test for publication bias, but for a relationship between the observed outcomes and the chosen predictor. If such a relationship is present, then this usually implies asymmetry in the funnel plot, which in turn may be an indication of publication bias. However, it is important to keep in mind that there can be other reasons besides publication bias that could lead to asymmetry in the funnel plot.

References

Begg, C. B., & Mazumdar, M. (1994). Operating characteristics of a rank correlation test for publication bias. Biometrics, 50, 1088--1101.

Egger, M., Davey Smith, G., Schneider, M., & Minder, C. (1997). Bias in meta-analysis detected by a simple, graphical test. British Medical Journal, 315, 629--634.

Harbord, R. M., Egger, M., & Sterne, J. A. C. (2006). A modified test for small-study effects in meta-analyses of controlled trials with binary endpoints. Statistics in Medicine, 25, 3443--3457.

Macaskill, P., Walter, S. D., & Irwig, L. (2001). A comparison of methods to detect publication bias in meta-analysis. Statistics in Medicine, 20, 641--654.

Moreno, S. G., Sutton, A. J., Ades, A. E., Stanley, T. D., Abrams, K. R., Peters, J. L., & Cooper, N. J. (2009). Assessment of regression-based methods to adjust for publication bias through a comprehensive simulation study. BMC Medical Research Methodology, 9:2.

Peters, J. L., Sutton, A. J., Jones, D. R., Abrams, K. R., & Rushton, L. (2006). Comparison of two methods to detect publication bias in meta-analysis. Journal of the American Medical Association, 295, 676--680.

Rücker, G., Schwarzer, G., & Carpenter, J. (2008). Arcsine test for publication bias in meta-analyses with binary outcomes. Statistics in Medicine, 27, 746--763.

Sterne, J. A. C., & Egger, M. (2005). Regression methods to detect publication and other bias in meta-analysis. In H. R. Rothstein, A. J. Sutton, & M. Borenstein (Eds.) Publication bias in meta-analysis: Prevention, assessment, and adjustments (pp. 99--110). Chichester, England: Wiley.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. http://www.jstatsoft.org/v36/i03/.

See also

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### fit random-effects model
res <- rma(yi, vi, data=dat)

### classical Egger test
regtest(res, model="lm")

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     weighted regression with multiplicative dispersion
#> predictor: standard error
#> 
#> test for funnel plot asymmetry: t = -1.4013, df = 11, p = 0.1887
#> 

### random/mixed-effects version of the Egger test
regtest(res)

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     mixed-effects meta-regression model
#> predictor: standard error
#> 
#> test for funnel plot asymmetry: z = -0.8033, p = 0.4218
#> 

### same tests, but passing outcomes directly
regtest(dat$yi, dat$vi, model="lm")

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     weighted regression with multiplicative dispersion
#> predictor: standard error
#> 
#> test for funnel plot asymmetry: t = -1.4013, df = 11, p = 0.1887
#> 

regtest(dat$yi, dat$vi)

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     mixed-effects meta-regression model
#> predictor: standard error
#> 
#> test for funnel plot asymmetry: z = -0.8033, p = 0.4218
#> 

### examples using the sample size (or a transformation thereof) as predictor
regtest(res, model="lm", predictor="ni")

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     weighted regression with multiplicative dispersion
#> predictor: sample size
#> 
#> test for funnel plot asymmetry: t = 4.4112, df = 11, p = 0.0010
#> 

regtest(res, model="lm", predictor="ninv")

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     weighted regression with multiplicative dispersion
#> predictor: inverse of the sample size
#> 
#> test for funnel plot asymmetry: t = -1.4485, df = 11, p = 0.1754
#> 

regtest(res, model="rma", predictor="ni")

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     mixed-effects meta-regression model
#> predictor: sample size
#> 
#> test for funnel plot asymmetry: z = 2.0488, p = 0.0405
#> 

regtest(res, model="rma", predictor="ninv")

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     mixed-effects meta-regression model
#> predictor: inverse of the sample size
#> 
#> test for funnel plot asymmetry: z = -1.2566, p = 0.2089
#> 

### if dat$yi is computed with escalc(), sample size information is stored in attributes
dat$yi

#>  [1] -0.88931133 -1.58538866 -1.34807315 -1.44155119 -0.21754732 -0.78611559
#>  [7] -1.62089822  0.01195233 -0.46941765 -1.37134480 -0.33935883  0.44591340
#> [13] -0.01731395
#> attr(,"measure")
#> [1] "RR"
#> attr(,"ni")
#>  [1]    262    609    451  26465  10877   2992   3174 176782  14776   3381
#> [11]  77972   4839  34767

### then this will work
regtest(dat$yi, dat$vi, predictor="ni")

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     mixed-effects meta-regression model
#> predictor: sample size
#> 
#> test for funnel plot asymmetry: z = 2.0488, p = 0.0405
#> 

### otherwise have to supply sample sizes manually
dat$ni <- with(dat, tpos + tneg + cpos + cneg)
regtest(dat$yi, dat$vi, dat$ni, predictor="ni")

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     mixed-effects meta-regression model
#> predictor: sample size
#> 
#> test for funnel plot asymmetry: z = 2.0488, p = 0.0405
#> 

### testing for asymmetry after accounting for the influence of a moderator
res <- rma(yi, vi, mods = ~ ablat, data=dat)
regtest(res, model="lm")

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     weighted regression with multiplicative dispersion
#> predictor: standard error
#> 
#> test for funnel plot asymmetry: t = -1.0876, df = 10, p = 0.3023
#> 

regtest(res)

#> 
#> Regression Test for Funnel Plot Asymmetry
#> 
#> model:     mixed-effects meta-regression model
#> predictor: standard error
#> 
#> test for funnel plot asymmetry: z = -0.5990, p = 0.5492
#>