replmiss {metafor}	R Documentation

Replace Missing Values in a Vector

Description

Function to replace missing (NA) values in a vector.

Usage

replmiss(x, y)

Arguments

x
vector that may include one or more missing values.

y
either a scalar or a vector of the same length as x with the value(s) to replace missing values with.

Value

Vector x with the missing values replaced based on the scalar or vector y.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

Examples

x <- c(4,2,7,NA,1,NA,5)
x <- replmiss(x,0)
x

x <- c(4,2,7,NA,1,NA,5)
y <- c(2,3,6,5,8,1,2)
x <- replmiss(x,y)
x

[Package metafor version 3.4-0 Index]