Meta-Analysis via the Mantel-Haenszel Method

Function to fit fixed-effects models to \(2 \times 2\) table and person-time data via the Mantel-Haenszel method. See below and the documentation of the metafor-package for more details on these models.

rma.mh(ai, bi, ci, di, n1i, n2i, x1i, x2i, t1i, t2i,
       measure="OR", data, slab, subset,
       add=1/2, to="only0", drop00=TRUE,
       correct=TRUE, level=95, digits, verbose=FALSE, ...)

Arguments

ai
vector to specify the \(2 \times 2\) table frequencies (upper left cell). See below and the documentation of the escalc function for more details.

bi
vector to specify the \(2 \times 2\) table frequencies (upper right cell). See below and the documentation of the escalc function for more details.

ci
vector to specify the \(2 \times 2\) table frequencies (lower left cell). See below and the documentation of the escalc function for more details.

di
vector to specify the \(2 \times 2\) table frequencies (lower right cell). See below and the documentation of the escalc function for more details.

n1i
vector to specify the group sizes or row totals (first group). See below and the documentation of the escalc function for more details.

n2i
vector to specify the group sizes or row totals (second group). See below and the documentation of the escalc function for more details.

x1i
vector to specify the number of events (first group). See below and the documentation of the escalc function for more details.

x2i
vector to specify the number of events (second group). See below and the documentation of the escalc function for more details.

t1i
vector to specify the total person-times (first group). See below and the documentation of the escalc function for more details.

t2i
vector to specify the total person-times (second group). See below and the documentation of the escalc function for more details.

measure
character string indicating the outcome measure to use for the meta-analysis. Possible options are the risk ratio ("RR"), the odds ratio ("OR"), the risk difference ("RD"), the incidence rate ratio ("IRR"), or the incidence rate difference ("IRD").

data
optional data frame containing the data supplied to the function.

slab
optional vector with labels for the \(k\) studies.

subset
optional (logical or numeric) vector indicating the subset of studies that should be used for the analysis.

add
non-negative number indicating the amount to add to zero cells, counts, or frequencies when calculating the observed outcomes of the individual studies. Can also be a vector of two numbers, where the first number is used in the calculation of the observed outcomes and the second number is used when applying the Mantel-Haenszel method. See below and the documentation of the escalc function for more details.

to
character string indicating when the values under add should be added (either "only0", "all", "if0all", or "none"). Can also be a character vector, where the first string again applies when calculating the observed outcomes and the second string when applying the Mantel-Haenszel method. See below and the documentation of the escalc function for more details.

drop00
logical indicating whether studies with no cases/events (or only cases) in both groups should be dropped when calculating the observed outcomes (the outcomes for such studies are set to NA). Can also be a vector of two logicals, where the first applies to the calculation of the observed outcomes and the second when applying the Mantel-Haenszel method. See below and the documentation of the escalc function for more details.

correct
logical indicating whether to apply a continuity correction when computing the Cochran-Mantel-Haenszel test statistic.

level
numerical value between 0 and 100 specifying the confidence interval level (the default is 95).

digits
integer specifying the number of decimal places to which the printed results should be rounded (if unspecified, the default is 4).

verbose
logical indicating whether output should be generated on the progress of the model fitting (the default is FALSE).

...
additional arguments.

Details

Specifying the Data

When the outcome measure is either the risk ratio (measure="RR"), odds ratio (measure="OR"), or risk difference (measure="RD"), the studies are assumed to provide data in terms of \(2 \times 2\) tables of the form:

outcome 1	outcome 2	total

group 1	ai	bi	n1i

group 2	ci	di	n2i

where ai, bi, ci, and di denote the cell frequencies and n1i and n2i the row totals. For example, in a set of randomized clinical trials (RCTs) or cohort studies, group 1 and group 2 may refer to the treatment (exposed) and placebo/control (not exposed) group, with outcome 1 denoting some event of interest (e.g., death) and outcome 2 its complement. In a set of case-control studies, group 1 and group 2 may refer to the group of cases and the group of controls, with outcome 1 denoting, for example, exposure to some risk factor and outcome 2 non-exposure. For these outcome measures, one needs to specify either ai, bi, ci, and di or alternatively ai, ci, n1i, and n2i.

Alternatively, when the outcome measure is the incidence rate ratio (measure="IRR") or the incidence rate difference (measure="IRD"), the studies are assumed to provide data in terms of tables of the form:

events	person-time

group 1	x1i	t1i

group 2	x2i	t2i

where x1i and x2i denote the number of events in the first and the second group, respectively, and t1i and t2i the corresponding total person-times at risk.

Mantel-Haenszel Method

An approach for aggregating table data of these types was suggested by Mantel and Haenszel (1959) and later extended by various authors (see references). The Mantel-Haenszel method provides a weighted estimate under a fixed-effects model. The method is particularly advantageous when aggregating a large number of studies with small sample sizes (the so-called sparse data or increasing strata case).

When analyzing odds ratios, the Cochran-Mantel-Haenszel (CMH) test (Cochran, 1954; Mantel & Haenszel, 1959) and Tarone's test for heterogeneity (Tarone, 1985) are also provided (by default, the CMH test statistic is computed with the continuity correction; this can be switched off with correct=FALSE). When analyzing incidence rate ratios, the Mantel-Haenszel (MH) test (Rothman et al., 2008) for person-time data is also provided (again, the correct argument controls whether the continuity correction is applied). When analyzing risk ratios, odds ratios, or incidence rate ratios, the printed results are given both in terms of the log and the raw units (for easier interpretation).

Observed Outcomes of the Individual Studies

The Mantel-Haenszel method itself does not require the calculation of the observed outcomes of the individual studies (e.g., the observed odds or incidence rate ratios of the \(k\) studies) and directly makes use of the table/event counts. Zero cells/events are not a problem (except in extreme cases, such as when one of the two outcomes never occurs in any of the \(2 \times 2\) tables or when there are no events for one of the two groups in any of the tables). Therefore, it is unnecessary to add some constant to the cell/event counts when there are zero cells/events.

However, for plotting and various other functions, it is necessary to calculate the observed outcomes for the \(k\) studies. Here, zero cells/events can be problematic, so adding a constant value to the cell/event counts ensures that all \(k\) values can be calculated. The add and to arguments are used to specify what value should be added to the cell/event counts and under what circumstances when calculating the observed outcomes and when applying the Mantel-Haenszel method. Similarly, the drop00 argument is used to specify how studies with no cases/events (or only cases) in both groups should be handled. The documentation of the escalc function explains how the add, to, and drop00 arguments work. If only a single value for these arguments is specified (as per default), then these values are used when calculating the observed outcomes and no adjustment to the cell/event counts is made when applying the Mantel-Haenszel method. Alternatively, when specifying two values for these arguments, the first value applies when calculating the observed outcomes and the second value when applying the Mantel-Haenszel method.

Note that drop00 is set to TRUE by default. Therefore, the observed outcomes for studies where ai=ci=0 or bi=di=0 or studies where x1i=x2i=0 are set to NA. When applying the Mantel-Haenszel method, such studies are not explicitly dropped (unless the second value of drop00 argument is also set to TRUE), but this is practically not necessary, as they do not actually influence the results (assuming no adjustment to the cell/event counts are made when applying the Mantel-Haenszel method).

Value

An object of class c("rma.mh","rma"). The object is a list containing the following components:

beta

aggregated log risk ratio, log odds ratio, risk difference, log rate ratio, or rate difference.

se

standard error of the aggregated value.

zval

test statistics of the aggregated value.

pval

corresponding p-value.

ci.lb

lower bound of the confidence interval.

ci.ub

upper bound of the confidence interval.

QE

test statistic of the test for heterogeneity.

QEp

correspinding p-value.

MH

Cochran-Mantel-Haenszel test statistic (measure="OR") or Mantel-Haenszel test statistic (measure="IRR").

MHp

corresponding p-value.

TA

test statistic of Tarone's test for heterogeneity (only when measure="OR").

TAp

corresponding p-value (only when measure="OR").

k

number of studies included in the analysis.

yi, vi

the vector of outcomes and corresponding sampling variances.

fit.stats

a list with the log-likelihood, deviance, AIC, BIC, and AICc values under the unrestricted and restricted likelihood.

...

some additional elements/values.

Methods

The results of the fitted model are formatted and printed with the print.rma.mh function. If fit statistics should also be given, use summary.rma (or use the fitstats.rma function to extract them).

The residuals.rma, rstandard.rma.mh, and rstudent.rma.mh functions extract raw and standardized residuals. Leave-one-out diagnostics can be obtained with leave1out.rma.mh.

Forest, funnel, radial, L'Abbé, and Baujat plots can be obtained with forest.rma, funnel.rma, radial.rma, labbe.rma, and baujat. The qqnorm.rma.mh function provides normal QQ plots of the standardized residuals. One can also just call plot.rma.mh on the fitted model object to obtain various plots at once.

A cumulative meta-analysis (i.e., adding one observation at a time) can be obtained with cumul.rma.mh.

Other extractor functions include coef.rma, vcov.rma, logLik.rma, deviance.rma, AIC.rma, and BIC.rma.

References

Cochran, W. G. (1954). Some methods for strengthening the common \(\chi^2\) tests. Biometrics, 10(4), 417--451. https://doi.org/10.2307/3001616

Greenland, S., & Robins, J. M. (1985). Estimation of a common effect parameter from sparse follow-up data. Biometrics, 41(1), 55--68. https://doi.org/10.2307/2530643

Mantel, N., & Haenszel, W. (1959). Statistical aspects of the analysis of data from retrospective studies of disease. Journal of the National Cancer Institute, 22(4), 719--748. https://doi.org/10.1093/jnci/22.4.719

Nurminen, M. (1981). Asymptotic efficiency of general noniterative estimators of common relative risk. Biometrika, 68(2), 525--530. https://doi.org/10.1093/biomet/68.2.525

Robins, J., Breslow, N., & Greenland, S. (1986). Estimators of the Mantel-Haenszel variance consistent in both sparse data and large-strata limiting models. Biometrics, 42(2), 311--323. https://doi.org/10.2307/2531052

Rothman, K. J., Greenland, S., & Lash, T. L. (2008). Modern epidemiology (3rd ed.). Philadelphia: Lippincott Williams & Wilkins.

Sato, T., Greenland, S., & Robins, J. M. (1989). On the variance estimator for the Mantel-Haenszel risk difference. Biometrics, 45(4), 1323--1324. https://www.jstor.org/stable/2531784

Tarone, R. E. (1981). On summary estimators of relative risk. Journal of Chronic Diseases, 34(9-10), 463--468. https://doi.org/10.1016/0021-9681(81)90006-0

Tarone, R. E. (1985). On heterogeneity tests based on efficient scores. Biometrika, 72(1), 91--95. https://doi.org/10.1093/biomet/72.1.91

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://doi.org/10.18637/jss.v036.i03

See also

Examples

### meta-analysis of the (log) odds ratios using the Mantel-Haenszel method
rma.mh(measure="OR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

#> 
#> Fixed-Effects Model (k = 13)
#> 
#> I^2 (total heterogeneity / total variability):  92.68%
#> H^2 (total variability / sampling variability): 13.66
#> 
#> Test for Heterogeneity: 
#> Q(df = 12) = 163.9426, p-val < .0001
#> 
#> Model Results (log scale):
#> 
#> estimate      se      zval    pval    ci.lb    ci.ub 
#>  -0.4734  0.0410  -11.5444  <.0001  -0.5538  -0.3930 
#> 
#> Model Results (OR scale):
#> 
#> estimate   ci.lb   ci.ub 
#>   0.6229  0.5748  0.6750 
#> 
#> Cochran-Mantel-Haenszel Test:    CMH = 135.6889, df = 1,  p-val < 0.0001
#> Tarone's Test for Heterogeneity: X^2 = 171.7567, df = 12, p-val < 0.0001
#> 

### meta-analysis of the (log) risk ratios using the Mantel-Haenszel method
rma.mh(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

#> 
#> Fixed-Effects Model (k = 13)
#> 
#> I^2 (total heterogeneity / total variability):  92.13%
#> H^2 (total variability / sampling variability): 12.71
#> 
#> Test for Heterogeneity: 
#> Q(df = 12) = 152.5676, p-val < .0001
#> 
#> Model Results (log scale):
#> 
#> estimate      se      zval    pval    ci.lb    ci.ub 
#>  -0.4537  0.0393  -11.5338  <.0001  -0.5308  -0.3766 
#> 
#> Model Results (RR scale):
#> 
#> estimate   ci.lb   ci.ub 
#>   0.6353  0.5881  0.6862 
#>