Transformation Functions

A set of transformation functions useful for meta-analyses.

Usage

transf.rtoz(xi, ...)
transf.ztor(xi, ...)
transf.logit(xi, ...)
transf.ilogit(xi, ...)
transf.arcsin(xi, ...)
transf.iarcsin(xi, ...)
transf.pft(xi, ni, ...)
transf.ipft(xi, ni, ...)
transf.ipft.hm(xi, targs, ...)
transf.isqrt(xi, ...)
transf.irft(xi, ti, ...)
transf.iirft(xi, ti, ...)
transf.ahw(xi, ...)
transf.iahw(xi, ...)
transf.abt(xi, ...)
transf.iabt(xi, ...)
transf.ztor.int(xi, targs, ...)
transf.exp.int(xi, targs, ...)
transf.ilogit.int(xi, targs, ...)
transf.dtou1(xi, ...)
transf.dtou2(xi, ...)
transf.dtou3(xi, ...)
transf.dtorpb(xi, n1i, n2i, ...)
transf.dtobesd(xi, ...)
transf.dtomd(xi, targs, ...)
transf.logortord(xi, pc, ...)
transf.logortorr(xi, pc, ...)

Arguments

xi

vector of values to be transformed.

ni

vector of sample sizes.

n1i

vector of sample sizes for the first group.

n2i

vector of sample sizes for the second group.

ti

vector of person-times at risk.

pc

control group risk (either a single value or a vector).

targs

list with additional arguments for the transformation function. See ‘Details’.

...

other arguments.

Details

The following transformation functions are currently implemented:

transf.rtoz: Fisher's r-to-z transformation for correlations.

transf.ztor: inverse of Fisher's r-to-z transformation.

transf.logit: logit (log odds) transformation for proportions.

transf.ilogit: inverse of the logit transformation.

transf.arcsin: arcsine square root transformation for proportions.

transf.iarcsin: inverse of the arcsine transformation.

transf.pft: Freeman-Tukey (double arcsine) transformation for proportions. See Freeman & Tukey (1950). The xi argument is used to specify the proportions and the ni argument the corresponding sample sizes.

transf.ipft: inverse of the Freeman-Tukey (double arcsine) transformation for proportions. See Miller (1978).

transf.ipft.hm: inverse of the Freeman-Tukey (double arcsine) transformation for proportions using the harmonic mean of the sample sizes for the back-transformation. See Miller (1978). The sample sizes are specified via the targs argument (the list element should be called ni).

transf.isqrt: inverse of the square root transformation (i.e., function to square a number).

transf.irft: Freeman-Tukey transformation for incidence rates. See Freeman & Tukey (1950). The xi argument is used to specify the incidence rates and the ti argument the corresponding person-times at risk.

transf.iirft: inverse of the Freeman-Tukey transformation for incidence rates.

transf.ahw: transformation of coefficient alpha as suggested by Hakstian & Whalen (1976).

transf.iahw: inverse of the transformation of coefficient alpha as suggested by Hakstian & Whalen (1976).

transf.abt: transformation of coefficient alpha as suggested by Bonett (2002).

transf.iabt: inverse of the transformation of coefficient alpha as suggested by Bonett (2002).

transf.ztor.int: integral transformation method for the z-to-r transformation.

transf.exp.int: integral transformation method for the exponential transformation.

transf.ilogit.int: integral transformation method for the inverse of the logit transformation.

transf.dtou1: transformation of standardized mean differences to Cohen's U1 values (Cohen, 1988).

transf.dtou2: transformation of standardized mean differences to Cohen's U2 values (Cohen, 1988).

transf.dtou3: transformation of standardized mean differences to Cohen's U3 values (Cohen, 1988).

transf.dtocles: transformation of standardized mean differences to common language effect size values (McGraw & Wong, 1992).

transf.dtorpb: transformation of standardized mean differences to point-biserial correlations. If n1i and n2i are not specified, the function assumes n1i=n2i and uses an approximate formula. If n1i and n2i are specified, the exact transformation formula is used.

transf.dtobesd: transformation of standardized mean differences to binomial effect size display values (Rosenthal & Rubin, 1982). Note that the function only provides the proportion in the first group scoring above the median (the proportion in the second group scoring above the median is simply one minus the proportion in the first group scoring above the median).

transf.dtomd: transformation of standardized mean differences to mean differences given a known standard deviation, which must be specified via the targs argument.

transf.logortord: transformation of log odds ratios to risk differences, assuming a particular value for the control group risk (which needs to be specified via the pc argument).

transf.logortorr: transformation of log odds ratios to risk ratios, assuming a particular value for the control group risk (which needs to be specified via the pc argument).

Value

A vector with the transformed values.

Note

The integral transformation method for a transformation function \(h(z)\) integrates \(h(z) f(z)\) over \(z\) using the limits targs$lower and targs$upper, where \(f(z)\) is the density of a normal distribution with mean equal to xi and variance equal to targs$tau2. An example is provided below.

Author

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Bonett, D. G. (2002). Sample size requirements for testing and estimating coefficient alpha. Journal of Educational and Behavioral Statistics, 27(4), 335--340. https://doi.org/10.3102/10769986027004335

Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). Hillsdale, NJ: Lawrence Erlbaum Associates.

Fisher, R. A. (1921). On the “probable error” of a coefficient of correlation deduced from a small sample. Metron, 1, 1--32. http://hdl.handle.net/2440/15169

Freeman, M. F., & Tukey, J. W. (1950). Transformations related to the angular and the square root. Annals of Mathematical Statistics, 21(4), 607--611. https://doi.org/10.1214/aoms/1177729756

Hakstian, A. R., & Whalen, T. E. (1976). A k-sample significance test for independent alpha coefficients. Psychometrika, 41(2), 219--231. https://doi.org/10.1007/BF02291840

McGraw, K. O., & Wong, S. P. (1992). A common language effect size statistic. Psychological Bulletin, 111(2), 361--365. https://doi.org/10.1037/0033-2909.111.2.361

Miller, J. J. (1978). The inverse of the Freeman-Tukey double arcsine transformation. American Statistician, 32(4), 138. https://doi.org/10.1080/00031305.1978.10479283

Rosenthal, R., & Rubin, D. B. (1982). A simple, general purpose display of magnitude of experimental effect. Journal of Educational Psychology, 74(2), 166--169. https://doi.org/10.1037/0022-0663.74.2.166

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://doi.org/10.18637/jss.v036.i03

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### fit random-effects model
res <- rma(yi, vi, data=dat)

### average risk ratio with 95% CI (but technically, this provides an
### estimate of the median risk ratio, not the mean risk ratio!)
predict(res, transf=exp)
#> 
#>    pred  ci.lb  ci.ub  pi.lb  pi.ub 
#>  0.4894 0.3441 0.6962 0.1546 1.5490 
#> 

### average risk ratio with 95% CI using the integral transformation
predict(res, transf=transf.exp.int, targs=list(tau2=res$tau2, lower=-4, upper=4))
#> 
#>    pred  ci.lb  ci.ub  pi.lb  pi.ub 
#>  0.5724 0.4024 0.8142 0.1809 1.8117 
#>