Compute Weights for 'rma' Objects

The function computes the weights given to the observed effect sizes or outcomes during the model fitting for objects of class "rma.uni", "rma.mh", "rma.peto", and "rma.mv".

Usage

# S3 method for rma.uni
weights(object, type="diagonal", ...)
# S3 method for rma.mh
weights(object, type="diagonal", ...)
# S3 method for rma.peto
weights(object, type="diagonal", ...)
# S3 method for rma.glmm
weights(object, ...)
# S3 method for rma.mv
weights(object, type="diagonal", ...)

Arguments

object

an object of class "rma.uni", "rma.mh", "rma.peto", or "rma.mv". The method is not yet implemented for objects of class "rma.glmm".

type

character string to specify whether to return only the diagonal of the weight matrix ("diagonal") or the entire weight matrix ("matrix"). For "rma.mv", this can also be "rowsum" for ‘row-sum weights’ (for intercept-only models).

...

other arguments.

Value

Either a vector with the diagonal elements of the weight matrix or the entire weight matrix. When only the diagonal elements are returned, they are given in % (and they add up to 100%). When the entire weight matrix is requested, this is always a diagonal matrix for objects of class "rma.uni", "rma.mh", "rma.peto". For "rma.mv", the structure of the weight matrix depends on the model fitted (i.e., the random effects included and the variance-covariance matrix of the sampling errors) but is often more complex and not just diagonal. For "rma.mv" intercept-only models, one can also take the sum over the rows in the weight matrix, which are actually the weights assigned to the observed effect sizes or outcomes when estimating the model intercept. These weights can be obtained with type="rowsum" (as with type="diagonal", they are also given in %).

Author

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://doi.org/10.18637/jss.v036.i03

See also

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### fit mixed-effects model with absolute latitude and publication year as moderators
res <- rma(yi, vi, mods = ~ ablat + year, data=dat)

### extract the model fitting weights (in %)
weights(res)
#>         1         2         3         4         5         6         7         8         9        10 
#>  3.366420  4.810621  2.791974 11.231203  9.068114 12.481731  4.400814 12.801967  8.784824  7.991920 
#>        11        12        13 
#> 11.923841  2.283581  8.062988 

### extract the weight matrix
round(weights(res, type="matrix"), 4)
#>         1      2      3      4      5      6      7      8      9     10     11     12     13
#> 1  2.2916 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#> 2  0.0000 3.2747 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#> 3  0.0000 0.0000 1.9006 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#> 4  0.0000 0.0000 0.0000 7.6454 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#> 5  0.0000 0.0000 0.0000 0.0000 6.1729 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#> 6  0.0000 0.0000 0.0000 0.0000 0.0000 8.4967 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#> 7  0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 2.9958 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000
#> 8  0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 8.7147 0.0000 0.0000 0.0000 0.0000 0.0000
#> 9  0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 5.9801 0.0000 0.0000 0.0000 0.0000
#> 10 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 5.4403 0.0000 0.0000 0.0000
#> 11 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 8.1169 0.0000 0.0000
#> 12 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 1.5545 0.0000
#> 13 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 5.4887