addpoly {metafor}	R Documentation

Add Polygons to Forest Plot

Description

The function addpoly is generic. It can be used to add polygons to a forest plot, for example, to indicate summary estimates and fitted/predicted values.

Usage

addpoly(x, ...)

Arguments

x 	 either an object of class "rma" or the values at which polygons should be drawn. See ‘Details’.

... 	 other arguments.

Details

Currently, methods exist for two types of situations.

In the first case, object x is a fitted model coming from the rma.uni, rma.mh, or rma.peto functions. The model must either be a fixed- or random-effects model, that is, the model should not contain any moderators. The corresponding method is called addpoly.rma. It is used to add a polygon to the bottom of an existing forest plot, showing the summary estimate with corresponding confidence interval based on the fitted model.

Alternatively, object x can be a vector with values at which one or more polygons should be drawn. The corresponding method is then addpoly.default.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

addpoly.rma, addpoly.default, forest.rma, forest.default

[Package metafor version 0.5-7 Index]