blup {metafor}	R Documentation

Best Linear Unbiased Predictions

Description

The function blup is generic. It extracts the best linear unbiased predictions (BLUPs) for specific classes of objects. BLUPs combine the fitted values based on the fixed effects in the model and the estimated contributions of the random effects.

Usage

blup(x, ...)

Arguments

x 	 an object for which BLUPs are meaningful. See ‘Details’.

... 	 other arguments.

Details

Currently, there is only a specific method for objects of class "rma.uni" created by the rma.uni function. Accordingly, the corresponding method is called blup.rma.uni. See the documentation for that function for more details.

Value

Best linear unbiased predictions and possibly corresponding standard errors and prediction interval bounds.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

References

Robinson, G. K. (1991) That BLUP is a good thing: The estimation of random effects. Statistical Science, 6, 15–32.

See Also

blup.rma.uni

[Package metafor version 0.5-7 Index]