Calculate the Bias Correction Factor for SMDs and Related Measures

Function to calculate the bias correction factor the standardized mean differences and related measures.

cmicalc(mi, method="exact")

Arguments

mi

numeric vector.

method

character string to specify if the exact ("exact") or approximate ("approx") formula should be used.

Details

Standardized mean differences (Cohen's d values) have a slight positive bias (i.e., they tend to be too large on average). Hedges (1981) derived the bias correction factor to make make standardized mean differences unbiased (yielding what is commonly referred to as Hedges' g values).

The function computes the bias correction factor, using \(m_i\) as input. For standardized mean differences, this is equal to \(m_i = n_{1i} + n_{2i} - 2\), where \(n_{1i}\) and \(n_{2i}\) are the number of subjects in the two groups. The same bias correction factor can be used for various other related measures, such as standardized mean changes, except that \(m_i\) may need to be calculated differently.

By default (i.e., when method="exact"), the method uses the exact formula to calculate the bias correction factor (equation 6e in Hedges, 1981), that is, \[c(m_i) = \frac{\Gamma(\frac{m_i}{2})}{\sqrt{\frac{m_i}{2}}\Gamma(\frac{m_i-1}{2})}.\] When method="approx", the (very accurate) approximation \[c(m_i) = 1 - \frac{3}{4m_i -1}\] is used instead (first equation on page 114 in Hedges, 1981).

Value

A numeric vector of length equal to the length of mi.

Note

The notation \(c(m_i)\) is based on that used in Hedges (1981). Later publications and references often denote the bias correction factor \(J\) or \(J(m_i)\).

Author

Wolfgang Viechtbauer (wvb@metafor-project.org, https://www.metafor-project.org).

References

Hedges, L. V. (1981). Distribution theory for Glass's estimator of effect size and related estimators. Journal of Educational Statistics, 6(2), 107–128. https://doi.org/10.3102/10769986006002107

See also

Examples

### copy dataset to dat
dat <- dat.furukawa2003

### calculate the bias correction factor for each study
dat$mi <- dat$Ne + dat$Nc - 2
dat$cmi <- cmicalc(dat$mi)
dat
#>               author  Ne    Me    Se Nc    Mc    Sc                    measure  mi       cmi
#> 1    Blashki(75&150)  13  6.40  5.40 18 11.40  9.60                    HRSD-17  29 0.9738750
#> 2     Hormazabal(86)  17 11.00  8.20 16 19.00  8.20                    HRSD-21  31 0.9755765
#> 3   Jacobson(75-100)  10 17.50  8.80  6 23.00  8.80                    HRSD-24  14 0.9452877
#> 4        Jenkins(75)   7 12.30  9.90  7 20.00 10.50                        BDI  12 0.9359418
#> 5     Lecrubier(100)  73 15.70 10.60 73 18.70 10.60                      MADRS 144 0.9947811
#> 6        Murphy(100)  26  8.50 11.00 28 14.50 11.00     ad hoc physician scale  52 0.9854955
#> 7          Nandi(97)  17 25.50 24.00 10 53.20 11.20                       HRSD  25 0.9696456
#> 8      Petracca(100)  11  6.20  7.60 10 10.00  7.60                    HRSD-17  19 0.9599104
#> 9       Philipp(100) 105 -8.10  3.90 46 -8.50  5.20                    HRSD-17 149 0.9949566
#> 10     Rampello(100)  22 13.40  2.30 19 19.70  1.30                       HRSD  39 0.9806242
#> 11       Reifler(83)  13 12.50  7.60 15 12.50  7.60                    HRSD-17  26 0.9708263
#> 12       Rickels(70)  29  1.99  0.77 39  2.54  0.77 Physician Depression Scale  66 0.9885859
#> 13     Robertson(75)  13 11.00  8.20 13 15.00  8.20                    HRSD-21  24 0.9683652
#> 14      Rouillon(98)  78 15.80  6.80 71 17.10  7.20                      MADRS 147 0.9948878
#> 15           Tan(70)  23 -8.50  8.60 23 -8.30  6.00                      MADRS  44 0.9828407
#> 16 Tetreault(50-100)  11 51.90 18.50 11 74.30 18.50             Wechsler scale  20 0.9619445
#> 17      Thompson(75)  11  8.00  8.10 18 10.00  9.70                    HRSD-17  27 0.9719186