coef.permutest.rma.uni {metafor}	R Documentation

Extract the Model Coefficient Table from 'permutest.rma.uni' Objects

Description

The function extracts the estimated model coefficients, corresponding standard errors, test statistics, p-values (based on the permutation tests), and confidence interval bounds from objects of class "permutest.rma.uni".

Usage

## S3 method for class 'permutest.rma.uni'
coef(object, ...)

Arguments

object
an object of class "permutest.rma.uni".

...
other arguments.

Value

A data frame with the following elements:

estimate
estimated model coefficient(s).

se
corresponding standard error(s).

zval
corresponding test statistic(s).

pval
p-value(s) based on the permutation test(s).

ci.lb
lower bound of the (permutation-based) confidence interval(s).

ci.ub
upper bound of the (permutation-based) confidence interval(s).

When the model was fitted with test="t" or test="knha", then zval is called tval in the data frame that is returned by the function.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

See Also

permutest.rma.uni for the function to conduct permutation tests and rma.uni for the function to fit models for which permutation tests can be conducted.

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### fit mixed-effects model with absolute latitude and publication year as moderators
res <- rma(yi, vi, mods = ~ ablat + year, data=dat)

### carry out permutation test
## Not run: 
set.seed(1234) # for reproducibility
sav <- permutest(res)
coef(sav)

## End(Not run)

[Package metafor version 3.4-0 Index]