Extract the Model Coefficient Table from 'permutest.rma.uni' Objects

The function extracts the estimated model coefficients, corresponding standard errors, test statistics, p-values (based on the permutation test), and confidence interval bounds from objects of class "permutest.rma.uni".

# S3 method for permutest.rma.uni
coef(object, ...)

Arguments

object
an object of class "permutest.rma.uni".

...
other arguments.

Value

A data frame with the following elements:

estimate

estimated model coefficient(s).

se

corresponding standard error(s).

zval

corresponding test statistic(s).

pval

p-value(s) based on the permutation test.

ci.lb

lower bound of the (permutation-based) confidence interval(s).

ci.ub

upper bound of the (permutation-based) confidence interval(s).

When the model was fitted with the Knapp and Hartung (2003) method (i.e., test="knha" in the rma.uni function), then zval is called tval in the data frame that is returned by the function.
References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://www.jstatsoft.org/v036/i03.

See also

Examples

### meta-analysis of the log risk ratios using a mixed-effects model
### with two moderators (absolute latitude and publication year)
res <- rma(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg,
           mods = ~ ablat + year, data=dat.bcg)

### permutation test
if (FALSE) {
pres <- permutest(res)
coef(pres)}