coef.rma {metafor}	R Documentation

Model Coefficients for rma Objects

Description

The function extracts the estimated model coefficients, corresponding standard errors, test statistics, p-values, and confidence interval bounds from objects of class "rma".

Usage

## S3 method for class 'rma':
coef(object, ...)

Arguments

object 	 an object of class "rma".

... 	 other arguments.

Value

A data frame with the following elements:

estimate 	 estimated model coefficient(s).

se 	 corresponding standard error(s).

zval 	 corresponding test statistic(s).

pval 	 corresponding p-value(s).

ci.lb 	 corresponding lower bound of the confidence interval(s).

ci.ub 	 corresponding upper bound of the confidence interval(s).

When the model was fitted with the Knapp and Hartung (2003) method (i.e., knha=TRUE in the rma.uni function), then zval is called tval in the data frame that is returned by the function.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

rma.uni, rma.mh, rma.peto

Examples

### load BCG vaccine data
data(dat.bcg)

### meta-analysis of the log risk ratios using a mixed-effects model
### with two moderators (absolute latitude and publication year)
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, mods=cbind(ablat, year),
           data=dat.bcg, measure="RR", method="REML")
coef(res)

[Package metafor version 0.5-7 Index]