cumul {metafor}	R Documentation

Cumulative Meta-Analysis

Description

The function cumul is generic. For suitable model objects, it repeatedly fits the model, adding one observation at a time to the model.

Usage

cumul(x, ...)

Arguments

x 	 an object of class "rma.uni", "rma.mh", or "rma.peto".

... 	 other arguments.

Details

Currently, there are methods for handling objects of class "rma.uni", "rma.mh", and "rma.peto" with the cumul function. Accordingly, the corresponding methods are called cumul.rma.uni, cumul.rma.mh, and cumul.rma.peto. See the documentation for those functions for more details.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

cumul.rma.uni, cumul.rma.mh, cumul.rma.peto

[Package metafor version 0.5-7 Index]