cumul {metafor}	R Documentation

Cumulative Meta-Analysis for 'rma' Objects

Description

The functions repeatedly fit the specified model, adding one observation/study at a time to the model.

Usage

cumul(x, ...)

## S3 method for class 'rma.uni'
cumul(x, order, digits, transf, targs, progbar=FALSE, ...)
## S3 method for class 'rma.mh'
cumul(x, order, digits, transf, targs, progbar=FALSE, ...)
## S3 method for class 'rma.peto'
cumul(x, order, digits, transf, targs, progbar=FALSE, ...)

Arguments

x
an object of class "rma.mh", "rma.peto", "rma.uni".

order
optional vector with indices giving the desired order for the cumulative meta-analysis.

digits
integer specifying the number of decimal places to which the printed results should be rounded (if unspecified, the default is to take the value from the object).

transf
optional argument specifying the name of a function that should be used to transform the model coefficients and interval bounds (e.g., transf=exp; see also transf). If unspecified, no transformation is used.

targs
optional arguments needed by the function specified under transf.

progbar
logical indicating whether a progress bar should be shown (the default is FALSE).

...
other arguments.

Details

For "rma.uni" objects, the model specified by x must be a model without moderators (i.e., either a fixed- or a random-effects model).

Value

An object of class c("list.rma","cumul.rma"). The object is a list containing the following components:

estimate
estimated coefficients of the model.

se
standard errors of the coefficients.

zval
test statistics of the coefficients.

pval
p-values for the test statistics.

ci.lb
lower bounds of the confidence intervals for the coefficients.

ci.ub
upper bounds of the confidence intervals for the coefficients.

QE
test statistics for the tests of heterogeneity.

QEp
p-values for the tests of heterogeneity.

tau2
estimated amounts of (residual) heterogeneity (only for random-effects models).

I2
values of I² (only for random-effects models).

H2
values of H² (only for random-effects models).

The object is formatted and printed with print.list.rma. A forest plot showing the results from the cumulative meta-analysis can be obtained with forest.cumul.rma. For random-effects models, plot.cumul.rma can also be used to visualize the results.

Note

When using the transf option, the transformation is applied to the estimated coefficients and the corresponding interval bounds. The standard errors are then set equal to NA and are omitted from the printed output.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Chalmers, T. C., & Lau, J. (1993). Meta-analytic stimulus for changes in clinical trials. Statistical Methods in Medical Research, 2, 161–172.

Lau, J., Schmid, C. H., & Chalmers, T. C. (1995). Cumulative meta-analysis of clinical trials builds evidence for exemplary medical care. Journal of Clinical Epidemiology, 48, 45–57.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

forest.cumul.rma, plot.cumul.rma

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### random-effects model
res <- rma(yi, vi, data=dat)

### cumulative meta-analysis (in the order of publication year)
cumul(res, transf=exp, order=order(dat$year))

### meta-analysis of the (log) risk ratios using the Mantel-Haenszel method
res <- rma.mh(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### cumulative meta-analysis
cumul(res, order=order(dat.bcg$year))
cumul(res, order=order(dat.bcg$year), transf=TRUE)

### meta-analysis of the (log) odds ratios using Peto's method
res <- rma.mh(ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### cumulative meta-analysis
cumul(res, order=order(dat.bcg$year))
cumul(res, order=order(dat.bcg$year), transf=TRUE)

[Package metafor version 2.4-0 Index]