Studies on the Effectiveness of Writing-to-Learn Interventions

Results from 48 studies on the effectiveness of school-based writing-to-learn interventions on academic achievement.

dat.bangertdrowns2004

Format

The data frame contains the following columns:

id	numeric	study number

author	character	study author(s)

year	numeric	publication year

grade	numeric	grade level (1 = elementary; 2 = middle; 3 = high-school; 4 = college)

length	numeric	treatment length (in weeks)

minutes	numeric	minutes per assignment

wic	numeric	writing tasks were completed in class (0 = no; 1 = yes)

feedback	numeric	feedback on writing was provided (0 = no; 1 = yes)

info	numeric	writing contained informational components (0 = no; 1 = yes)

pers	numeric	writing contained personal components (0 = no; 1 = yes)

imag	numeric	writing contained imaginative components (0 = no; 1 = yes)

meta	numeric	prompts for metacognitive reflection (0 = no; 1 = yes)

subject	character	subject matter

ni	numeric	total sample size of the study

yi	numeric	standardized mean difference

vi	numeric	corresponding sampling variance

Details

In each of the studies included in this meta-analysis, an experimental group (i.e., a group of students that received instruction with increased emphasis on writing tasks) was compared against a control group (i.e., a group of students that received conventional instruction) with respect to some content-related measure of academic achievement (e.g., final grade, an exam/quiz/test score). The outcome measure for this meta-analysis was the standardized mean difference (with positive values indicating a higher mean level of academic achievement in the intervention group).

The standardized mean differences given here are bias-corrected and therefore differ slightly from the values reported in the article. Also, since only the total sample size is given in the article, the sampling variances were computed under the assumption that \(n_{i1} = n_{i2} = n_i / 2\).

Source

Bangert-Drowns, R. L., Hurley, M. M., & Wilkinson, B. (2004). The effects of school-based writing-to-learn interventions on academic achievement: A meta-analysis. Review of Educational Research, 74(1), 29--58. https://doi.org/10.3102/00346543074001029

Examples

### copy data into 'dat' and examine data
dat <- dat.bangertdrowns2004
dat

#>    id            author year grade length minutes wic feedback info pers imag 
#> 1   1          Ashworth 1992     4     15      NA   1        1    1    1    0 
#> 2   2             Ayers 1993     2     10      NA   1       NA    1    1    1 
#> 3   3            Baisch 1990     2      2      NA   1        0    1    1    0 
#> 4   4             Baker 1994     4      9      10   1        1    1    0    0 
#> 5   5            Bauman 1992     1     14      10   1        1    1    1    0 
#> 6   6            Becker 1996     4      1      20   1        0    0    1    0 
#> 7   7       Bell & Bell 1985     3      4      NA   1        1    1    1    0 
#> 8   8           Brodney 1994     1     15      NA   1        1    1    1    0 
#> 9   9            Burton 1986     4      4      NA   0        1    1    0    0 
#> 10 10         Davis, BH 1990     1      9      10   1        0    1    1    0 
#> 11 11         Davis, JJ 1996     4     15      NA   0        1    1    0    0 
#> 12 12               Day 1994     4     15      NA   0        1    1    1    0 
#> 13 13           Dipillo 1994     1      8       6   1        1    1    1    0 
#> 14 14           Ganguli 1989     4      4       4   1        0    1    0    0 
#> 15 15        Giovinazzo 1996     4     14      NA   0        1    1    0    0 
#> 16 16              Goss 1998     4     15      10   1        1    1    0    0 
#> 17 17     Greene et al. 1991     4      4      10   1        1    1    1    0 
#> 18 18            Guckin 1992     4     10       8   1        1   NA   NA    0 
#> 19 19            Guckin 1992     4     10       8   1        1   NA   NA    0 
#> 20 20     Horton et al. 1985     4      3      NA   0        1    1    0    0 
#> 21 21             Hyser 1992     1     24      10   1        0    1    1    1 
#> 22 22       Johnson, LA 1991     3     19      NA   0        1    1    0    0 
#> 23 23       Johnson, VM 1998     1      4      15   1        0    1    0    0 
#> 24 24          Kasparek 1993     3     12      NA   1        1    1    1    0 
#> 25 25   Konopack et al. 1990     2      1      NA   1        0    1    0    0 
#> 26 26 Langer & Applebee 1987     3      1      35   1        0    1    0    0 
#> 27 27 Langer & Applebee 1987     3      1      20   1        0    1    0    0 
#> 28 28            Licata 1993     3      1      15   1        1    1    0    0 
#> 29 29           Lodholz 1980     1     14      15   1        1    1    0    0 
#> 30 30            Madden 1993     1     20       8   1        0    1    1    0 
#> 31 31          Millican 1994     1     10      NA   1        1    1    1    1 
#> 32 32          Moynihan 1994     1      7      10   1        1    1    1    0 
#> 33 33          Mulvaney 1991     4     11      NA   1        1    1    0    0 
#> 34 34         Nieswandt 1997     3     NA      NA   0        1    1    0    0 
#> 35 35         Radmacher 1995     4     NA      NA   0        1    1    0    0 
#> 36 36            Reaves 1991     3      1      NA   1        0    1    0    0 
#> 37 37            Rivard 1996     2      6      55   1        0    1    0    0 
#> 38 38           Rodgers 1996     4     15      NA   0        1    1    0    0 
#> 39 39   Ross & Faucette 1994     4     15      NA   0        1    1    0    0 
#> 40 40             Sharp 1987     4      2      15   1        0    0    1    0 
#> 41 41           Shepard 1992     2      4      NA   0        1    1    0    0 
#> 42 42           Stewart 1992     3     24       5   1        1    1    1    0 
#> 43 43           Ullrich 1926     4     11      NA   0        1    1    0    0 
#> 44 44   Weiss & Walters 1980     4     15       3   1        0    1    0    0 
#> 45 45             Wells 1986     1      8      15   1        0    1    1    0 
#> 46 46            Willey 1988     3     15      NA  NA        0    1    1    0 
#> 47 47            Willey 1988     2     15      NA  NA        0    1    1    0 
#> 48 48         Youngberg 1989     4     15      10   1        1    1    0    0 
#>    meta                    subject  ni     yi    vi 
#> 1     1                    Nursing  60  0.650 0.070 
#> 2     0              Earth Science  34 -0.750 0.126 
#> 3     1                       Math  95 -0.210 0.042 
#> 4     0                    Algebra 209 -0.040 0.019 
#> 5     1                       Math 182  0.230 0.022 
#> 6     0                 Literature 462  0.030 0.009 
#> 7     1                       Math  38  0.260 0.106 
#> 8     1                       Math 542  0.060 0.007 
#> 9     0                       Math  99  0.060 0.040 
#> 10    0             Social Studies  77  0.120 0.052 
#> 11    0                 Statistics  40  0.770 0.107 
#> 12    1                  Sociology 190  0.000 0.021 
#> 13    1                       Math 113  0.520 0.037 
#> 14    0                       Math  50  0.540 0.083 
#> 15    0                    Algebra  47  0.200 0.086 
#> 16    0                   Calculus  44  0.200 0.091 
#> 17    0 Comp Science and Chemistry  24 -0.160 0.167 
#> 18   NA                    Algebra  78  0.420 0.052 
#> 19   NA                    Algebra  46  0.600 0.091 
#> 20    0                  Chemistry  64  0.510 0.065 
#> 21    0             Social Studies  57  0.580 0.073 
#> 22    0                    Algebra  68  0.540 0.061 
#> 23    0                       Math  40  0.090 0.100 
#> 24    1                    Algebra  68  0.370 0.060 
#> 25    0              World History  48 -0.010 0.083 
#> 26    0             Social Studies 107 -0.130 0.037 
#> 27    0             Social Studies  58  0.180 0.069 
#> 28    0            Math in Science 225  0.270 0.018 
#> 29    0                       Math 446 -0.020 0.009 
#> 30    1                       Math  77  0.330 0.053 
#> 31    1                       Math 243  0.590 0.017 
#> 32    1                       Math  39  0.840 0.112 
#> 33    0                 Management  67 -0.320 0.060 
#> 34    0                  Chemistry  91  0.120 0.044 
#> 35    1                 Psychology  36  1.120 0.129 
#> 36    0          Natural Resources 177 -0.120 0.023 
#> 37    0                    Science  20 -0.440 0.205 
#> 38    0                    Algebra 120 -0.070 0.033 
#> 39    0                    Algebra  16  0.700 0.265 
#> 40    1                    Biology 105  0.490 0.039 
#> 41    0                       Math 195  0.200 0.021 
#> 42    1                    Algebra  62  0.580 0.067 
#> 43    0     Educational Psychology 289  0.150 0.014 
#> 44    0                 Statistics  25  0.630 0.168 
#> 45    1                       Math 250  0.040 0.016 
#> 46    1                    Biology  51  1.460 0.099 
#> 47    1             Social Studies  46  0.040 0.087 
#> 48    0                    Algebra  56  0.250 0.072 

### fit random-effects model
res <- rma(yi, vi, data=dat)
res

#> 
#> Random-Effects Model (k = 48; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0499 (SE = 0.0197)
#> tau (square root of estimated tau^2 value):      0.2235
#> I^2 (total heterogeneity / total variability):   58.37%
#> H^2 (total variability / sampling variability):  2.40
#> 
#> Test for Heterogeneity:
#> Q(df = 47) = 107.1061, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.2219  0.0460  4.8209  <.0001  0.1317  0.3122  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>