Studies on Bone-Marrow Transplantation versus Chemotherapy for the Treatment of Leukemia

Results from controlled and uncontrolled studies on the effectiveness of allogeneic bone-marrow transplantation (BMT) and conventional chemotherapy (CMO) in the treatment of acute nonlymphocytic leukemia.

dat.berkey1989

Format

The data frame contains the following columns:

study	numeric	study number

trt	character	treatment (BMT or CMO)

arms	numeric	number of arms in the study (1 = uncontrolled studies; 2 = controlled studies)

yi	numeric	2-year disease-free survival

sei	numeric	corresponding standard error

vi	numeric	corresponding sampling variance

Details

The dataset includes the results from controlled and uncontrolled studies on the 2-year disease-free survival in patients with acute nonlymphocytic leukemia receiving either allogeneic bone-marrow transplantation (BMT) or conventional chemotherapy (CMO). In the controlled (two-arm) studies (studies 1-4), a cohort of patients in complete remission and potentially eligible for BMT was assembled, and those who consented and for whom a donor could be found received BMT, with the remaining patients used as controls (receiving CMO). In the uncontrolled (one-arm) studies (studies 5-16), only a single group was studied, receiving either BMT or CMO.

The data in this dataset were obtained from Table 1 in Begg & Pilote (1991, p. 902).

Source

Begg, C. B., & Pilote, L. (1991). A model for incorporating historical controls into a meta-analysis. Biometrics, 47, 899--906.

References

Begg, C. B., Pilote, L., & McGlave, P. B. (1989). Bone marrow transplantation versus chemotherapy in acute non-lymphocytic leukemia: A meta-analytic review. European Journal of Cancer and Clinical Oncology, 25, 1519--1523.

Examples

### load data
dat <- get(data(dat.begg1989))

### turn trt and arms into factors and set reference levels
dat$trt  <- relevel(factor(dat$trt), ref="CMO")
dat$arms <- relevel(factor(dat$arms), ref="2")

### create data frame with the treatment differences for the controlled studies
dat2 <- data.frame(yi = dat$yi[c(1,3,5,7)] - dat$yi[c(2,4,6,8)],
                   vi = dat$vi[c(1,3,5,7)] + dat$vi[c(2,4,6,8)])

### DerSimonian and Laird method using the treatment differences
res <- rma(yi, vi, data=dat2, method="DL", digits=2)
res

#> 
#> Random-Effects Model (k = 4; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0 (SE = 0.02)
#> tau (square root of estimated tau^2 value):      0
#> I^2 (total heterogeneity / total variability):   0.00%
#> H^2 (total variability / sampling variability):  1.00
#> 
#> Test for Heterogeneity:
#> Q(df = 3) = 1.28, p-val = 0.73
#> 
#> Model Results:
#> 
#> estimate    se  zval  pval  ci.lb  ci.ub 
#>     0.20  0.08  2.59  <.01   0.05   0.34  ** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### Begg & Pilote (1991) model incorporating the uncontrolled studies
res <- rma.mv(yi, vi, mods = ~ trt, random = ~ 1 | study,
              data=dat, method="ML", digits=2)
res

#> 
#> Multivariate Meta-Analysis Model (k = 20; method: ML)
#> 
#> Variance Components:
#> 
#>            estim  sqrt  nlvls  fixed  factor 
#> sigma^2     0.00  0.04     16     no   study 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 18) = 27.65, p-val = 0.07
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 16.50, p-val < .01
#> 
#> Model Results:
#> 
#>          estimate    se   zval  pval  ci.lb  ci.ub 
#> intrcpt      0.32  0.02  16.47  <.01   0.28   0.36  *** 
#> trtBMT       0.15  0.04   4.06  <.01   0.08   0.22  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### model involving bias terms for the uncontrolled studies
res <- rma.mv(yi, vi, mods = ~ trt + trt:arms, random = ~ 1 | study,
              data=dat, method="ML", digits=2)
res

#> 
#> Multivariate Meta-Analysis Model (k = 20; method: ML)
#> 
#> Variance Components:
#> 
#>            estim  sqrt  nlvls  fixed  factor 
#> sigma^2     0.00  0.03     16     no   study 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 16) = 26.72, p-val = 0.04
#> 
#> Test of Moderators (coefficients 2:4):
#> QM(df = 3) = 17.64, p-val < .01
#> 
#> Model Results:
#> 
#>               estimate    se   zval  pval  ci.lb  ci.ub 
#> intrcpt           0.30  0.05   6.63  <.01   0.21   0.39  *** 
#> trtBMT            0.20  0.08   2.64  <.01   0.05   0.35   ** 
#> trtCMO:arms1      0.03  0.05   0.52  0.60  -0.07   0.12      
#> trtBMT:arms1     -0.04  0.07  -0.56  0.58  -0.19   0.10      
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### model with random treatment effect
res <- rma.mv(yi, vi, mods = ~ trt, random = list(~ 1 | study, ~ trt | study),
              struct="UN", tau2=c(0,NA), rho=0, data=dat, method="ML", digits=2)
res

#> 
#> Multivariate Meta-Analysis Model (k = 20; method: ML)
#> 
#> Variance Components:
#> 
#>            estim  sqrt  nlvls  fixed  factor 
#> sigma^2     0.00  0.04     16     no   study 
#> 
#> outer factor: study (nlvls = 16)
#> inner factor: trt   (nlvls = 2)
#> 
#>            estim  sqrt  k.lvl  fixed  level 
#> tau^2.1     0.00  0.00      8    yes    BMT 
#> tau^2.2     0.00  0.00     12     no    CMO 
#> 
#>      rho.BMT  rho.CMO    BMT  CMO 
#> BMT        1     0.00      -  yes 
#> CMO     0.00        1      4    - 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 18) = 27.65, p-val = 0.07
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 16.50, p-val < .01
#> 
#> Model Results:
#> 
#>          estimate    se   zval  pval  ci.lb  ci.ub 
#> intrcpt      0.32  0.02  16.47  <.01   0.28   0.36  *** 
#> trtBMT       0.15  0.04   4.06  <.01   0.08   0.22  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### model with random treatment effect, but with equal variances in both arms
res <- rma.mv(yi, vi, mods = ~ trt, random = list(~ 1 | study, ~ trt | study),
              struct="CS", rho=0, data=dat, method="ML", digits=2)
res

#> 
#> Multivariate Meta-Analysis Model (k = 20; method: ML)
#> 
#> Variance Components:
#> 
#>            estim  sqrt  nlvls  fixed  factor 
#> sigma^2     0.00  0.00     16     no   study 
#> 
#> outer factor: study (nlvls = 16)
#> inner factor: trt   (nlvls = 2)
#> 
#>            estim  sqrt  fixed 
#> tau^2       0.00  0.04     no 
#> rho         0.00          yes 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 18) = 27.65, p-val = 0.07
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 15.60, p-val < .01
#> 
#> Model Results:
#> 
#>          estimate    se   zval  pval  ci.lb  ci.ub 
#> intrcpt      0.32  0.02  16.49  <.01   0.28   0.36  *** 
#> trtBMT       0.15  0.04   3.95  <.01   0.07   0.22  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>