Studies on the Relationship between the Competitive State Anxiety Inventory-2 and Sport Performance

Results from 10 studies on the relationship between the Competitive State Anxiety Inventory-2 (CSAI-2) and sport performance.

dat.craft2003

Format

The data frame contains the following columns:

study	numeric	study number

ni	numeric	sample size

sport	character	type of sport (T = team sport, I = individual sport)

ri	numeric	correlation coefficient

var1	character	variable 1 of the correlation coefficient (see ‘Details’)

var2	character	variable 2 of the correlation coefficient (see ‘Details’)

Details

The 10 studies included in this dataset are a subset of the studies included in the meta-analysis by Craft et al. (2003) on the relationship between the Competitive State Anxiety Inventory-2 (CSAI-2) and sport performance.

The CSAI-2 has three subscales: cognitive anxiety (acog), somatic anxiety (asom), and self-confidence (conf). The studies included in this dataset administered the CSAI-2 prior to some sport competition and then measured sport performance based on the competition. Most studies provided all 6 correlations (3 for the correlations among the 3 subscales and 3 for the correlations between the subscales and sport performance), but 2 studies (with study numbers 6 and 17) only provided a subset.

Source

Becker, B. J., & Aloe, A. M. (2019). Model-based meta-analysis and related approaches. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (3nd ed., pp. 339--363). New York: Russell Sage Foundation.

References

Craft, L. L., Magyar, T. M., Becker, B. J., & Feltz, D. L. (2003). The relationship between the Competitive State Anxiety Inventory-2 and sport performance: A meta-analysis. Journal of Sport and Exercise Psychology, 25(1), 44--65. https://doi.org/10.1123/jsep.25.1.44

Examples

### copy data into 'dat'
dat <- dat.craft2003

### construct dataset and var-cov matrix of the correlations
tmp <- rcalc(ri ~ var1 + var2 | study, ni=ni, data=dat)
V <- tmp$V
dat <- tmp$dat

### examine data for study 1
dat[dat$study == 1,]

#>   study var1 var2 var1.var2    yi  ni
#> 1     1 acog perf acog.perf -0.55 142
#> 2     1 asom perf asom.perf -0.48 142
#> 3     1 conf perf conf.perf  0.66 142
#> 4     1 acog asom acog.asom  0.47 142
#> 5     1 acog conf acog.conf -0.38 142
#> 6     1 asom conf asom.conf -0.46 142

V[dat$study == 1, dat$study == 1]

#>               acog.perf     asom.perf     conf.perf     acog.asom    acog.conf
#> acog.perf  0.0034503989  0.0013265149 -0.0005545798 -0.0013967848  0.002501895
#> asom.perf  0.0013265149  0.0042005969 -0.0009521407 -0.0019433591  0.001264856
#> conf.perf -0.0005545798 -0.0009521407  0.0022592011  0.0005791091 -0.001533798
#> acog.asom -0.0013967848 -0.0019433591  0.0005791091  0.0043049419 -0.001802689
#> acog.conf  0.0025018954  0.0012648562 -0.0015337979 -0.0018026891  0.005191854
#> asom.conf  0.0009322372  0.0025160783 -0.0010692460 -0.0012050560  0.001884405
#>               asom.conf
#> acog.perf  0.0009322372
#> asom.perf  0.0025160783
#> conf.perf -0.0010692460
#> acog.asom -0.0012050560
#> acog.conf  0.0018844047
#> asom.conf  0.0044083302

### examine data for study 6
dat[dat$study == 6,]

#>    study var1 var2 var1.var2   yi ni
#> 13     6 acog perf acog.perf 0.44 16
#> 14     6 asom perf asom.perf 0.46 16
#> 15     6 conf perf conf.perf   NA 16
#> 16     6 acog asom acog.asom 0.67 16
#> 17     6 acog conf acog.conf   NA 16
#> 18     6 asom conf asom.conf   NA 16

V[dat$study == 6, dat$study == 6]

#>             acog.perf   asom.perf conf.perf   acog.asom acog.conf asom.conf
#> acog.perf 0.043352064 0.025583395        NA 0.009529623        NA        NA
#> asom.perf 0.025583395 0.041438304        NA 0.008459787        NA        NA
#> conf.perf          NA          NA        NA          NA        NA        NA
#> acog.asom 0.009529623 0.008459787        NA 0.020247414        NA        NA
#> acog.conf          NA          NA        NA          NA        NA        NA
#> asom.conf          NA          NA        NA          NA        NA        NA

### examine data for study 17
dat[dat$study == 17,]

#>    study var1 var2 var1.var2    yi ni
#> 25    17 acog perf acog.perf  0.10 45
#> 26    17 asom perf asom.perf  0.31 45
#> 27    17 conf perf conf.perf -0.17 45
#> 28    17 acog asom acog.asom    NA 45
#> 29    17 acog conf acog.conf    NA 45
#> 30    17 asom conf asom.conf    NA 45

V[dat$study == 17, dat$study == 17]

#>           acog.perf  asom.perf  conf.perf acog.asom acog.conf asom.conf
#> acog.perf  0.022275         NA         NA        NA        NA        NA
#> asom.perf        NA 0.01856898         NA        NA        NA        NA
#> conf.perf        NA         NA 0.02143262        NA        NA        NA
#> acog.asom        NA         NA         NA        NA        NA        NA
#> acog.conf        NA         NA         NA        NA        NA        NA
#> asom.conf        NA         NA         NA        NA        NA        NA

### multivariate random-effects model
res <- rma.mv(yi, V, mods = ~ var1.var2 - 1, random = ~ var1.var2 | study, struct="UN", data=dat)

#> Warning: Rows with NAs omitted from model fitting.

res

#> 
#> Multivariate Meta-Analysis Model (k = 51; method: REML)
#> 
#> Variance Components:
#> 
#> outer factor: study     (nlvls = 9)
#> inner factor: var1.var2 (nlvls = 6)
#> 
#>             estim    sqrt  k.lvl  fixed      level 
#> tau^2.1    0.0047  0.0683      9     no  acog.asom 
#> tau^2.2    0.0125  0.1119      8     no  acog.conf 
#> tau^2.3    0.1611  0.4014      9     no  acog.perf 
#> tau^2.4    0.0111  0.1052      8     no  asom.conf 
#> tau^2.5    0.0604  0.2459      9     no  asom.perf 
#> tau^2.6    0.0468  0.2163      8     no  conf.perf 
#> 
#>            rho.acg.s  rho.acg.c  rho.acg.p  rho.asm.c  rho.asm.p  rho.cnf. 
#> acog.asom          1    -0.6961     0.5491    -0.0891     0.4604   -0.9345   
#> acog.conf    -0.6961          1     0.0432     0.4193    -0.0495    0.7023   
#> acog.perf     0.5491     0.0432          1     0.3532     0.9497   -0.6178   
#> asom.conf    -0.0891     0.4193     0.3532          1     0.2688   -0.1311   
#> asom.perf     0.4604    -0.0495     0.9497     0.2688          1   -0.5969   
#> conf.perf    -0.9345     0.7023    -0.6178    -0.1311    -0.5969         1   
#>            acg.s  acg.c  acg.p  asm.c  asm.p  cnf. 
#> acog.asom      -     no     no     no     no    no 
#> acog.conf      8      -     no     no     no    no 
#> acog.perf      9      8      -     no     no    no 
#> asom.conf      8      8      8      -     no    no 
#> asom.perf      9      8      9      8      -    no 
#> conf.perf      8      8      8      8      8     - 
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 45) = 334.8358, p-val < .0001
#> 
#> Test of Moderators (coefficients 1:6):
#> QM(df = 6) = 596.7709, p-val < .0001
#> 
#> Model Results:
#> 
#>                     estimate      se     zval    pval    ci.lb    ci.ub 
#> var1.var2acog.asom    0.5671  0.0367  15.4640  <.0001   0.4953   0.6390  *** 
#> var1.var2acog.conf   -0.4888  0.0509  -9.6047  <.0001  -0.5886  -0.3891  *** 
#> var1.var2acog.perf   -0.0600  0.1408  -0.4264  0.6698  -0.3359   0.2159      
#> var1.var2asom.conf   -0.4750  0.0506  -9.3901  <.0001  -0.5741  -0.3758  *** 
#> var1.var2asom.perf   -0.1423  0.0917  -1.5527  0.1205  -0.3220   0.0373      
#> var1.var2conf.perf    0.3167  0.0847   3.7394  0.0002   0.1507   0.4827  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>