Studies on Standard Care Quality and HAART-Adherence

Results from 13 trials providing information about standard care quality and HAART-adherence in control groups.

dat.debruin2009

Format

The data frame contains the following columns:

author	character	(first) author of study

year	numeric	publication year

scq	numeric	standard care quality

ni	numeric	number of patients in the standard care group

xi	numeric	number of patients with an undetectable viral load in standard care group

mi	numeric	number of patients with a detectable viral load in standard care group

ethnicity	character	dominant ethnicity of the patients in the standard care group

patients	character	inclusion of patients continuing or starting (a new) treatment

select	character	baseline selection of patients with adherence problems or no selection

sens	character	sensitivity of viral load assessments (<400 vs. >=400 copies/ml)

Details

Highly active antiretroviral therapy (HAART) refers to a combination of multiple antiretroviral drugs that can effectively suppress the HIV virus. However, achieving viral suppression (to the point that the virus becomes essentially undetectable in a blood sample) requires high levels of adherence to an often complicated medication regimen. A number of trials have examined various interventions that aim to increase adherence levels. In each trial, patients receiving the intervention are compared to patients in a control group receiving standard care (often referred to as ‘care as usual’). However, the quality of standard care can vary substantially between these studies. de Bruin et al. (2009) assessed the quality of standard care provided (based on a quantification of the number of behavior change techniques applied) and examined to what extent the quality of standard care was related to the proportion of patients achieving effective viral suppression in the control groups.

Source

de Bruin, M., Viechtbauer, W., Hospers, H. J., Schaalma, H. P., & Kok, G. (2009). Standard care quality determines treatment outcomes in control groups of HAART-adherence intervention studies: Implications for the interpretation and comparison of intervention effects. Health Psychology, 28(6), 668--674. https://doi.org/10.1037/a0015989

Examples

### copy data into 'dat'
dat <- dat.debruin2009

### calculate proportions and corresponding sampling variances
dat <- escalc(measure="PR", xi=xi, ni=ni, data=dat)
dat

#>             author year   scq  ni  xi  mi ethnicity   patients select sens 
#> 1   Andrade et al. 2005 21.00  29  11  18     other   starting     no <400 
#> 2   Fairley et al. 2003 22.67  33  24   9 caucasian continuing    yes <400 
#> 3   Goujard et al. 2003 15.07 326 179 147 caucasian continuing     no <400 
#> 4  Holzemer et al. 2006  3.67 240  82 158     other continuing     no  400 
#> 5    Knobel et al. 1999  8.60 110  60  50 caucasian   starting     no <400 
#> 6    Murphy et al. 2002 27.80  29  11  18     other continuing    yes  400 
#> 7   Pradier et al. 2003  4.80 244  99 145 caucasian continuing     no <400 
#> 8   Rathbun et al. 2005 12.07  17   9   8 caucasian   starting     no <400 
#> 9    Remien et al. 2005 20.00 215  88 127     other continuing    yes <400 
#> 10   Tuldra et al. 2000  7.47  26  17   9 caucasian   starting     no  400 
#> 11   van Servellen 2005  7.80 138  26 112     other continuing    yes  400 
#> 12   Wagner et al. 2006 10.47  55  23  32     other   starting    yes  400 
#> 13     Wohl et al. 2006 22.80  54  40  14     other   starting     no  400 
#>        yi     vi 
#> 1  0.3793 0.0081 
#> 2  0.7273 0.0060 
#> 3  0.5491 0.0008 
#> 4  0.3417 0.0009 
#> 5  0.5455 0.0023 
#> 6  0.3793 0.0081 
#> 7  0.4057 0.0010 
#> 8  0.5294 0.0147 
#> 9  0.4093 0.0011 
#> 10 0.6538 0.0087 
#> 11 0.1884 0.0011 
#> 12 0.4182 0.0044 
#> 13 0.7407 0.0036 

### random-effects model
res <- rma(yi, vi, data=dat)
print(res, digits=2)

#> 
#> Random-Effects Model (k = 13; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.02 (SE = 0.01)
#> tau (square root of estimated tau^2 value):      0.15
#> I^2 (total heterogeneity / total variability):   91.77%
#> H^2 (total variability / sampling variability):  12.14
#> 
#> Test for Heterogeneity:
#> Q(df = 12) = 135.51, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>     0.48  0.05  10.45  <.01   0.39   0.57  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### mixed-effects meta-regression model with all predictors/covariates
res <- rma(yi, vi, mods = ~ scq + ethnicity + patients + select + sens, data=dat)
print(res, digits=3)

#> 
#> Mixed-Effects Model (k = 13; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.007 (SE = 0.006)
#> tau (square root of estimated tau^2 value):             0.085
#> I^2 (residual heterogeneity / unaccounted variability): 72.43%
#> H^2 (unaccounted variability / sampling variability):   3.63
#> R^2 (amount of heterogeneity accounted for):            68.55%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 7) = 22.309, p-val = 0.002
#> 
#> Test of Moderators (coefficients 2:6):
#> QM(df = 5) = 21.884, p-val < .001
#> 
#> Model Results:
#> 
#>                   estimate     se    zval   pval   ci.lb   ci.ub 
#> intrcpt              0.458  0.087   5.243  <.001   0.287   0.629  *** 
#> scq                  0.014  0.005   3.168  0.002   0.005   0.023   ** 
#> ethnicityother      -0.231  0.084  -2.759  0.006  -0.395  -0.067   ** 
#> patientsstarting     0.081  0.066   1.230  0.219  -0.048   0.209      
#> selectyes           -0.073  0.075  -0.983  0.325  -0.219   0.073      
#> sens<400            -0.104  0.080  -1.307  0.191  -0.261   0.052      
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### mixed-effects meta-regression model with scq and ethnicity as predictors/covariates
res <- rma(yi, vi, mods = ~ scq + ethnicity, data=dat)
print(res, digits=3)

#> 
#> Mixed-Effects Model (k = 13; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.012 (SE = 0.007)
#> tau (square root of estimated tau^2 value):             0.109
#> I^2 (residual heterogeneity / unaccounted variability): 83.24%
#> H^2 (unaccounted variability / sampling variability):   5.97
#> R^2 (amount of heterogeneity accounted for):            48.39%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 10) = 53.525, p-val < .001
#> 
#> Test of Moderators (coefficients 2:3):
#> QM(df = 2) = 10.333, p-val = 0.006
#> 
#> Model Results:
#> 
#>                 estimate     se    zval   pval   ci.lb   ci.ub 
#> intrcpt            0.430  0.075   5.698  <.001   0.282   0.577  *** 
#> scq                0.011  0.005   2.324  0.020   0.002   0.020    * 
#> ethnicityother    -0.196  0.072  -2.725  0.006  -0.337  -0.055   ** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>