dat.hackshaw1998 {metafor}	R Documentation

Studies on Lung Cancer Risk from ETS Exposure

Description

Results from 37 studies on the risk of lung cancer from environmental tobacco smoke (ETS) exposure.

Usage

dat.hackshaw1998

Format

The data frame contains the following columns:

study 	 numeric 	 study number

author 	 character 	 first author of study

year 	 numeric 	 publication year

country 	 character 	 country where study was conducted

design 	 character 	 study design (either cohort or case-control)

cases 	 numeric 	 number of lung cancer cases

yi 	 numeric 	 log odds ratio

vi 	 numeric 	 corresponding sampling variance

Details

The dataset includes the results from 37 studies (4 cohort, 33 case-control) on the risk of lung cancer from environmental tobacco smoke (ETS) exposure from the spouse in women who are lifelong nonsmokers. Values of the log odds ratio greater than 0 indicate an increased risk of cancer in exposed women compared to women not exposed to ETS from their spouse.

Source

Hackshaw, A. K., Law, M. R., & Wald, N. J. (1997). The accumulated evidence on lung cancer and environmental tobacco smoke. British Medical Journal, 315, 980–988.

Hackshaw, A. K. (1998). Lung cancer and passive smoking. Statistical Methods in Medical Research, 7, 119–136.

Examples

### copy data into 'dat' and examine data
dat <- dat.hackshaw1998
dat

### random-effects model using the log odds ratios
res <- rma(yi, vi, data=dat, method="DL")
res

### estimated average odds ratio with CI (and credibility/prediction interval)
predict(res, transf=exp, digits=2)

[Package metafor version 2.4-0 Index]