Studies on the Effectiveness of Azithromycin for Treating Lower Respiratory Tract Infections

Results from 15 studies on the effectiveness of azithromycin versus amoxycillin or amoxycillin/clavulanic acid (amoxyclav) in the treatment of acute lower respiratory tract infections.

dat.laopaiboon2015

Format

The data frame contains the following columns:

author	character	author(s)

year	numeric	publication year

ai	numeric	number of clinical failures in the group treated with azithromycin

n1i	numeric	number of patients in the group treated with azithromycin

ci	numeric	number of clinical failures in the group treated with amoxycillin or amoxyclav

n2i	numeric	number of patients in the group treated with amoxycillin or amoxyclav

age	character	whether the trial included adults or children

diag.ab	numeric	trial included patients with a diagnosis of acute bacterial bronchitis

diag.cb	numeric	trial included patients with a diagnosis of chronic bronchitis with acute exacerbation

diag.pn	numeric	trial included patients with a diagnosis of pneumonia

ctrl	character	antibiotic in control group (amoxycillin or amoxyclav)

Details

Azithromycin is an antibiotic useful for the treatment of a number of bacterial infections. Laopaiboon et al. (2015) conducted a meta-analysis of trials comparing the effectiveness of azithromycin versus amoxycillin or amoxycillin/clavulanic acid (amoxyclav) in the treatment of acute lower respiratory tract infections, including acute bacterial bronchitis, acute exacerbations of chronic bronchitis, and pneumonia. The results from 15 trials are included in this dataset.

Source

Laopaiboon, M., Panpanich, R., & Swa Mya, K. (2015). Azithromycin for acute lower respiratory tract infections. Cochrane Database of Systematic Reviews, 3, CD001954. https://doi.org/10.1002/14651858.CD001954.pub4

Examples

### copy data into 'dat' and examine data
dat <- dat.laopaiboon2015
dat

#>       author year ai n1i ci n2i      age diag.ab diag.cb diag.pn        ctrl
#> 1     Balmes 1991  4  48  7  56   adults       1       0       0   amoxyclav
#> 2      Beghi 1995 22  69  2  73   adults       0       1       0   amoxyclav
#> 3   Biebuyck 1996 53 497 53 257   adults       1       1       0   amoxyclav
#> 4     Daniel 1991  5 121 10 120   adults       1       0       0 amoxycillin
#> 5   Ferwerda 2001  5  55  7  53 children       0       0       1   amoxyclav
#> 6       Gris 1996  6  34  2  33   adults       1       1       1   amoxyclav
#> 7     Harris 1998 11 125  4  63 children       0       0       1   amoxyclav
#> 8  Hoepelman 1993  4  48  4  51   adults       1       0       0   amoxyclav
#> 9  Hoepelman 1998  3  62  5  61   adults       0       1       0   amoxyclav
#> 10   Mertens 1992  1  25  5  25   adults       0       1       0 amoxycillin
#> 11   Sevieri 1993  5  25  2  25   adults       0       1       0   amoxyclav
#> 12  Whitlock 1995  0  29  2  27   adults       0       1       0   amoxyclav
#> 13    Wubbel 1999  1  39  2  49 children       0       0       1   amoxyclav
#> 14 Zachariah 1996  8 173  7 173   adults       1       1       1   amoxyclav
#> 15     Zheng 2002 12  38  2  42   adults       0       1       0   amoxyclav

### analysis using the Mantel-Haenszel method
rma.mh(measure="RR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat, digits=3)

#> 
#> Fixed-Effects Model (k = 15)
#> 
#> I^2 (total heterogeneity / total variability):  65.30%
#> H^2 (total variability / sampling variability): 2.88
#> 
#> Test for Heterogeneity: 
#> Q(df = 14) = 40.348, p-val < .001
#> 
#> Model Results (log scale):
#> 
#> estimate     se    zval   pval   ci.lb  ci.ub 
#>   -0.083  0.117  -0.709  0.479  -0.311  0.146 
#> 
#> Model Results (RR scale):
#> 
#> estimate  ci.lb  ci.ub 
#>    0.921  0.733  1.157 
#> 

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=ai, n1i=n1i, ci=ci, n2i=n2i, data=dat)

### random-effects model
res <- rma(yi, vi, data=dat)
res

#> 
#> Random-Effects Model (k = 15; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.5874 (SE = 0.3897)
#> tau (square root of estimated tau^2 value):      0.7664
#> I^2 (total heterogeneity / total variability):   64.85%
#> H^2 (total variability / sampling variability):  2.84
#> 
#> Test for Heterogeneity:
#> Q(df = 14) = 38.4438, p-val = 0.0004
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval    ci.lb   ci.ub 
#>   0.0852  0.2673  0.3186  0.7501  -0.4388  0.6092    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### average risk ratio with 95% CI
predict(res, transf=exp)

#> 
#>    pred  ci.lb  ci.ub  pi.lb  pi.ub 
#>  1.0889 0.6448 1.8389 0.2218 5.3447 
#>