dat.linde2005 {metafor}	R Documentation

Studies on the Effectiveness of St. John's Wort for Treating Depression

Description

Results from 26 studies on the effectiveness of Hypericum perforatum extracts (St. John's wort) for treating depression.

Usage

dat.linde2005

Format

The data frame contains the following columns:

id 	 numeric 	 study number

study 	 character 	 study author(s)

year 	 numeric 	 publication year

country 	 character 	 study location

ni 	 numeric 	 total sample size

major 	 numeric 	 sample restricted to patients who met criteria for major depression

baseline 	 numeric 	 HRSD baseline score

version 	 numeric 	 HRSD version (17 or 21 items)

duration 	 numeric 	 study duration (in weeks)

prep 	 character 	 Hypericum extract preparation

dosage 	 numeric 	 dosage (in mg)

response 	 numeric 	 definition of response (1 = HRSD score reduction of at least 50% or HRSDscore after therapy <10; 2 = HRSDreduction of at least 50%; 3 = based on HRSD scale but exact definition not reported; 4 = global patient assessment of efficacy; 5 = at least 'much improved' on the Clinical Global Impression sub-scale global improvement)

ai 	 numeric 	 number of responses in treatment group

n1i 	 numeric 	 number of patients in treatment group

ci 	 numeric 	 number of responses in placebo group

n2i 	 numeric 	 number of patients in placebo group

group 	 numeric 	 stratification variable used by the authors (1 = smaller trials restricted to major depression; 2 = larger trials restricted to major depression; 3 = smaller trials not restricted to major depression; 4 = larger trials not restricted to major depression)

Details

The dataset includes the results from 26 double-blind placebo-controlled trials on the effectiveness of Hypericum perforatum extracts (St. John's wort) for treating depression (note that 2 studies did not provide sufficient response information).

Data were extracted from Table 1 and Figure 3. For study duration, the assessment week (instead of the total study duration) was coded for Philipp et al. (1999) and Montgomery et al. (2000). For dosage, the midpoint was coded when a range of values was given.

Source

Linde, K., Berner, M., Egger, M., & Mulrow, C. (2005). St John's wort for depression: Meta-analysis of randomised controlled trials. British Journal of Psychiatry, 186, 99–107.

References

Viechtbauer, W. (2007). Accounting for heterogeneity via random-effects models and moderator analyses in meta-analysis. Zeitschrift für Psychologie / Journal of Psychology, 215, 104–121.

Examples

### copy data into 'dat'
dat <- dat.linde2005

### remove studies with no response information and study with no responses in either group
dat <- dat[-c(5,6,26),]

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=ai, ci=ci, n1i=n1i, n2i=n2i, data=dat)
dat

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(yi, vi, data=dat, method="DL")
res

### mixed-effects meta-regression model with stratification variable
res <- rma(yi, vi, mods = ~ factor(group) - 1, data=dat, method="DL")
res

### predicted average risk ratio for each level of the stratification variable
predict(res, newmods=diag(4), transf=exp, digits=2)

[Package metafor version 2.4-0 Index]