Studies on the Effectiveness of St. John's Wort for Treating Depression

Results from 26 studies on the effectiveness of Hypericum perforatum extracts (St. John's wort) for treating depression.

dat.linde2005

Format

The data frame contains the following columns:

id	numeric	study number

study	character	study author(s)

year	numeric	publication year

country	character	study location

ni	numeric	total sample size

major	numeric	sample restricted to patients who met criteria for major depression

baseline	numeric	HRSD baseline score

version	numeric	HRSD version (17 or 21 items)

duration	numeric	study duration (in weeks)

prep	character	Hypericum extract preparation

dosage	numeric	dosage (in mg)

response	numeric	definition of response (1 = HRSD score reduction of at least 50% or HRSDscore after therapy <10; 2 = HRSDreduction of at least 50%; 3 = based on HRSD scale but exact definition not reported; 4 = global patient assessment of efficacy; 5 = at least 'much improved' on the Clinical Global Impression sub-scale global improvement)

ai	numeric	number of responses in treatment group

n1i	numeric	number of patients in treatment group

ci	numeric	number of responses in placebo group

n2i	numeric	number of patients in placebo group

group	numeric	stratification variable used by the authors (1 = smaller trials restricted to major depression; 2 = larger trials restricted to major depression; 3 = smaller trials not restricted to major depression; 4 = larger trials not restricted to major depression)

Details

The dataset includes the results from 26 double-blind placebo-controlled trials on the effectiveness of Hypericum perforatum extracts (St. John's wort) for treating depression (note that 2 studies did not provide sufficient response information).

Data were extracted from Table 1 and Figure 3. For study duration, the assessment week (instead of the total study duration) was coded for Philipp et al. (1999) and Montgomery et al. (2000). For dosage, the midpoint was coded when a range of values was given.

Source

Linde, K., Berner, M., Egger, M., & Mulrow, C. (2005). St John's wort for depression: Meta-analysis of randomised controlled trials. British Journal of Psychiatry, 186(2), 99--107. https://doi.org/10.1192/bjp.186.2.99

References

Viechtbauer, W. (2007). Accounting for heterogeneity via random-effects models and moderator analyses in meta-analysis. Zeitschrift für Psychologie / Journal of Psychology, 215(2), 104--121. https://doi.org/10.1027/0044-3409.215.2.104

Examples

### copy data into 'dat'
dat <- dat.linde2005

### remove studies with no response information and study with no responses in either group
dat <- dat[-c(5,6,26),]

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=ai, ci=ci, n1i=n1i, n2i=n2i, data=dat)
dat

#>    id             study year         country  ni major baseline version 
#> 1   1   Hoffman & Kuehl 1979         Germany  60     0       NA      NA 
#> 2   2    Schlich et al. 1987         Germany  49     0     31.3      21 
#> 3   3    Schmidt et al. 1989         Germany  40     0     29.4      21 
#> 4   4            Halama 1991         Germany  50     0     18.2      17 
#> 7   7        Reh et al. 1992         Germany  50     0     19.5      21 
#> 8   8    Huebner et al. 1993         Germany  40     0     12.5      17 
#> 9   9     Lehrl & Woelk 1993         Germany  50     1     22.7      21 
#> 10 10  Schmidt & Sommer 1993         Germany  65     0     16.5      21 
#> 11 11     Quandt et al. 1993         Germany  88     0     17.6      21 
#> 12 12            Koenig 1993     Switzerland 112     0       NA      NA 
#> 13 13   Sommer & Harrer 1994 Germany/Austria 105     0     15.8      21 
#> 14 14      Witte et al. 1995         Germany  97     1     23.6      21 
#> 15 15 Haensgen & Vesper 1996         Germany 197     1     20.7      21 
#> 16 16   Laakmann et al. 1998         Germany 147     1     21.1      17 
#> 17 17   Schrader et al. 1998         Germany 162     1     19.4      21 
#> 18 18    Philipp et al. 1999         Germany 263     1     22.7      17 
#> 19 19     Winkel et al. 2000         Germany 119     0     16.7      21 
#> 20 20       Volz et al. 2000         Germany 140     1     20.9      21 
#> 21 21 Montgomery et al. 2000              UK 247     1     21.5      17 
#> 22 22       Kalb et al. 2001         Germany  72     1     19.9      17 
#> 23 23    Shelton et al. 2001             USA 200     1     22.5      17 
#> 24 24             HDTSG 2002             USA 340     1     22.9      17 
#> 25 25  Lecrubier et al. 2002          France 375     1     21.9      17 
#>    duration         prep dosage response ai n1i ci n2i group      yi     vi 
#> 1         6   Hyperforat     NA        4 19  30  3  30     3  1.8458 0.3193 
#> 2         4 PsychotoninM    350        1 15  25  3  24     3  1.5686 0.3183 
#> 3         4 PsychotoninM    500        1 10  20  4  20     3  0.9163 0.2500 
#> 4         4        LI160    900        1 10  25  0  25     3  3.0445 2.0183 
#> 7         8   Neuroplant    380        1 20  25 11  25     4  0.5978 0.0609 
#> 8         4        LI160    900        1 14  20  9  20     4  0.4418 0.0825 
#> 9         4        LI160    900        1  4  25  2  25     1  0.6931 0.6700 
#> 10        6        LI160    900        1 20  32  6  33     4  1.2347 0.1551 
#> 11        4 PsychotoninM    500        1 29  44  3  44     3  2.2687 0.3224 
#> 12        6       Z90017    750        4 29  55 31  57     4 -0.0310 0.0310 
#> 13        4        LI160    900        1 28  50 13  55     4  0.8626 0.0745 
#> 14        6 Psychotoninf    240        1 34  48 25  49     2  0.3281 0.0282 
#> 15        4        LI160    900        1 35  53 12  54     1  1.0891 0.0745 
#> 16        6       WS5572    900        2 24  49 16  49     1  0.4055 0.0634 
#> 17        6        ZE117    500        1 45  80 12  79     1  1.3092 0.0804 
#> 18        6     STEI3000   1050        1 67 106 22  47     2  0.3004 0.0297 
#> 19        6        LI160    900        3 34  60 17  59     4  0.6763 0.0546 
#> 20        6        D0496    500        5 46  70 34  70     2  0.3023 0.0226 
#> 21        6        LI160    900        1 55 123 57 124     2 -0.0276 0.0195 
#> 22        6       WS5572    900        2 23  37 15  35     1  0.3719 0.0545 
#> 23        8        LI160   1050        2 26  98 19 102     1  0.3537 0.0711 
#> 24        8        LI160   1200        1 46 113 56 116     2 -0.1705 0.0221 
#> 25        6       WS5570    900        2 98 186 80 189     2  0.2189 0.0120 

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(yi, vi, data=dat, method="DL")
res

#> 
#> Random-Effects Model (k = 23; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.1413 (SE = 0.0697)
#> tau (square root of estimated tau^2 value):      0.3759
#> I^2 (total heterogeneity / total variability):   73.94%
#> H^2 (total variability / sampling variability):  3.84
#> 
#> Test for Heterogeneity:
#> Q(df = 22) = 84.4155, p-val < .0001
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval   ci.lb   ci.ub 
#>   0.5574  0.1002  5.5631  <.0001  0.3610  0.7537  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### mixed-effects meta-regression model with stratification variable
res <- rma(yi, vi, mods = ~ factor(group) - 1, data=dat, method="DL")
res

#> 
#> Mixed-Effects Model (k = 23; tau^2 estimator: DL)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.0555 (SE = 0.0377)
#> tau (square root of estimated tau^2 value):             0.2357
#> I^2 (residual heterogeneity / unaccounted variability): 53.18%
#> H^2 (unaccounted variability / sampling variability):   2.14
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 19) = 40.5792, p-val = 0.0027
#> 
#> Test of Moderators (coefficients 1:4):
#> QM(df = 4) = 68.4988, p-val < .0001
#> 
#> Model Results:
#> 
#>                 estimate      se    zval    pval    ci.lb   ci.ub 
#> factor(group)1    0.6826  0.1546  4.4138  <.0001   0.3795  0.9857  *** 
#> factor(group)2    0.1548  0.1136  1.3622  0.1731  -0.0679  0.3774      
#> factor(group)3    1.6681  0.2918  5.7166  <.0001   1.0962  2.2401  *** 
#> factor(group)4    0.5442  0.1430  3.8056  0.0001   0.2639  0.8245  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### predicted average risk ratio for each level of the stratification variable
predict(res, newmods=diag(4), transf=exp, digits=2)

#> 
#>   pred ci.lb ci.ub pi.lb pi.ub 
#> 1 1.98  1.46  2.68  1.14  3.44 
#> 2 1.17  0.93  1.46  0.70  1.95 
#> 3 5.30  2.99  9.39  2.54 11.06 
#> 4 1.72  1.30  2.28  1.00  2.96 
#>