Studies on Anti-Infective-Treated Central Venous Catheters for Prevention of Catheter-Related Bloodstream Infections

Results from 18 studies comparing the risk of catheter-related bloodstream infection when using anti-infective-treated versus standard catheters for total parenteral nutrition or chemotherapy.

dat.nielweise2008

Format

The data frame contains the following columns:

study

numeric

study number		authors

character

study authors

year

numeric		publication year

x1i

numeric

number of CRBSIs in patients receiving an anti-infective catheter		t1i

numeric

total number of catheter days for patients receiving an anti-infective catheter

x2i

numeric		number of CRBSIs in patients receiving a standard catheter

t2i

numeric

total number of catheter days for patients receiving a standard catheter		study

Details

The use of a central venous catheter may lead to a catheter-related bloodstream infection (CRBSI), which in turn increases the risk of morbidity and mortality. Anti-infective-treated catheters have been developed that are meant to reduce the risk of CRBSIs. Niel-Weise et al. (2008) conducted a meta-analysis of studies comparing infection risk when using anti-infective-treated versus standard catheters for total parenteral nutrition or chemotherapy. The results from 9 such studies are included in this dataset.

The dataset was used in the article by Stijnen et al. (2010) to illustrate various generalized linear mixed-effects models for the meta-analysis of incidence rates and incidence rate ratios (see ‘References’).

Source

Niel-Weise, B. S., Stijnen, T., & van den Broek, P. J. (2008). Anti-infective-treated central venous catheters for total parenteral nutrition or chemotherapy: A systematic review. Journal of Hospital Infection, 69, 114--123.

References

Stijnen, T., Hamza, T. H., & Ozdemir, P. (2010). Random effects meta-analysis of event outcome in the framework of the generalized linear mixed model with applications in sparse data. Statistics in Medicine, 29, 3046--3067.

Examples

### copy data into 'dat' and examine data
dat <- dat.nielweise2008
dat

#>   study          authors year x1i   t1i x2i   t2i
#> 1     1      Bong et al. 2003   7  1344  11  1988
#> 2     2    Ciresi et al. 1996   8  1600   8  1461
#> 3     3     Hanna et al. 2004   3 12012  14 10962
#> 4     4    Harter et al. 2002   6  1536  10  1503
#> 5     5    Jaeger et al. 2001   1   370   1   483
#> 6     6    Jaeger et al. 2005   1   729   8   913
#> 7     7    Logghe et al. 1997  17  6760  15  6840
#> 8     8 Ostendorf et al. 2005   3  1107   7  1015
#> 9     9 Pemberton et al. 1996   2   320   3   440

### standard (inverse-variance) random-effects model
res <- rma(measure="IRR", x1i=x1i, t1i=t1i, x2i=x2i, t2i=t2i, data=dat)
print(res, digits=3)

#> 
#> Random-Effects Model (k = 9; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.094 (SE = 0.212)
#> tau (square root of estimated tau^2 value):      0.306
#> I^2 (total heterogeneity / total variability):   20.88%
#> H^2 (total variability / sampling variability):  1.26
#> 
#> Test for Heterogeneity:
#> Q(df = 8) = 9.698, p-val = 0.287
#> 
#> Model Results:
#> 
#> estimate     se    zval   pval   ci.lb  ci.ub 
#>   -0.396  0.227  -1.747  0.081  -0.841  0.048  . 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp, digits=2)

#> 
#>  pred ci.lb ci.ub cr.lb cr.ub 
#>  0.67  0.43  1.05  0.32  1.42 
#> 

### random-effects conditional Poisson model

# NOT RUN {
res <- rma.glmm(measure="IRR", x1i=x1i, t1i=t1i, x2i=x2i, t2i=t2i, data=dat, model="CM.EL")
print(res, digits=3)
predict(res, transf=exp, digits=2)
# }