Studies on the Effectiveness of Locoregional Treatment plus Chemotherapy for Head and Neck Squamous-Cell Carcinoma

Results from studies examining mortality risk in patients with nonmetastatic head and neck squamous-cell carcinoma receiving either locoregional treatment plus chemotherapy versus locoregional treatment alone.

dat.pignon2000

Format

The data frame contains the following columns:

id	numeric	study id number

trial	character	trial abbreviation

OmE	numeric	observed minus expected number of deaths in the locoregional treatment plus chemotherapy group

V	numeric	corresponding variance

grp	numeric	timing of chemotherapy: 1 = adjuvant, 2 = neoadjuvant, 3 = concomitant

Details

The purpose of this meta-analysis was to examine the mortality risk in patients with nonmetastatic head and neck squamous-cell carcinoma receiving either locoregional treatment plus chemotherapy versus locoregional treatment alone. For 65 trials, the dataset provides the observed minus expected number of deaths and corresponding variances in the locoregional treatment plus chemotherapy group. Based on these values, we can estimate the log hazard ratios with OmE/V and the corresponding sampling variance with 1/V.

The trials were also divided according to the timing of the chomotherapy: (1) adjuvant, after the locoregional treatment, (2) neoadjuvant, before the locoregional treatment, and (3) concomitant, chemotherapy given concomitantly or alternating with radiotherapy.

Source

Pignon, J. P., Bourhis, J., Domenge, C., & Designe, L. (2000). Chemotherapy added to locoregional treatment for head and neck squamous-cell carcinoma: Three meta-analyses of updated individual data. Lancet, 355(9208), 949--955. https://doi.org/10.1016/S0140-6736(00)90011-4

Examples

### copy data into 'dat'
dat <- dat.pignon2000

### compute estimated log hazard ratios and sampling variances
dat$yi <- with(dat, OmE/V)
dat$vi <- with(dat, 1/V)
dat

#>       id          trial   OmE     V grp           yi          vi
#> 1   44.0       Pitie-74  -1.8  14.5   1 -0.124137931 0.068965517
#> 2   45.0      GETTECadj  11.8  57.1   1  0.206654991 0.017513135
#> 3   46.0       Int 0034  -7.4  80.6   1 -0.091811414 0.012406948
#> 4  102.0         JHCFUS  -5.6   8.5   1 -0.658823529 0.117647059
#> 5   97.0        TMH R-4   0.4   6.4   1  0.062500000 0.156250000
#> 6  103.0         KKD-86   1.8   8.7   1  0.206896552 0.114942529
#> 7   10.0        HNU-87a  -0.2   5.0   1 -0.040000000 0.200000000
#> 8   47.0        HNU-87b  -3.9  30.2   1 -0.129139073 0.033112583
#> 9   14.0         IGR-65   0.0   7.8   2  0.000000000 0.128205128
#> 10  17.0      RTOG 6801  -0.8 125.9   2 -0.006354249 0.007942812
#> 11  33.0    EORTC 24771   2.2  27.9   2  0.078853047 0.035842294
#> 12  28.0      DENVER-77  -3.3  13.5   2 -0.244444444 0.074074074
#> 13  29.0           HNCP  -0.6  63.1   2 -0.009508716 0.015847861
#> 14  34.0   EORTC 78-OCP  -9.2  29.7   2 -0.309764310 0.033670034
#> 15  52.0          MCW-1   6.3  16.6   2  0.379518072 0.060240964
#> 16  31.0      SWOG 8006   2.6  36.6   2  0.071038251 0.027322404
#> 17  22.0       Pitie-81  -2.0  25.8   2 -0.077519380 0.038759690
#> 18  32.0   Buenos Aires  -2.1  18.1   2 -0.116022099 0.055248619
#> 19  36.0     Creteil-82   7.2  19.8   2  0.363636364 0.050505051
#> 20  23.0       HNCGICO2   2.7  22.2   2  0.121621622 0.045045045
#> 21  30.0          MCW-2   1.2   9.0   2  0.133333333 0.111111111
#> 22  26.1 AC CAMARGO (a)   0.4  11.9   2  0.033613445 0.084033613
#> 23  92.1   SECOG II (a)  -3.6  33.2   2 -0.108433735 0.030120482
#> 24  41.0    EORTC 24844  -0.3  13.4   2 -0.022388060 0.074626866
#> 25  90.0        SHNG 85  -4.2  92.2   2 -0.045553145 0.010845987
#> 26  24.0       HNCGICO3   1.6  19.6   2  0.081632653 0.051020408
#> 27  35.0     Creteil-86  -5.7  20.2   2 -0.282178218 0.049504950
#> 28  37.0       GSTTC-86  -8.5  47.4   2 -0.179324895 0.021097046
#> 29  27.0     GETTECneo1  -7.5  24.8   2 -0.302419355 0.040322581
#> 30  11.0     GETTECneo2  -6.5  15.9   2 -0.408805031 0.062893082
#> 31  96.0          AHNTG  -5.7  46.5   2 -0.122580645 0.021505376
#> 32  25.0     Las Palmas  -4.0   5.2   2 -0.769230769 0.192307692
#> 33  40.0      Rennes-87  -2.9  25.8   2 -0.112403101 0.038759690
#> 34  39.0          Parma   5.4   9.2   2  0.586956522 0.108695652
#> 35  78.0          CFHNS  -8.1  46.6   2 -0.173819742 0.021459227
#> 36 101.0        Cologne  -1.1   6.7   2 -0.164179104 0.149253731
#> 37  99.0        Songkla   0.6   9.0   2  0.066666667 0.111111111
#> 38  38.0        HNAP-02   4.6   5.9   2  0.779661017 0.169491525
#> 39  93.0        BNH 003  -7.1  16.4   2 -0.432926829 0.060975610
#> 40  55.0         MDA-70   1.8   8.0   3  0.225000000 0.125000000
#> 41  56.0       WIA-OC5a  -4.3   8.4   3 -0.511904762 0.119047619
#> 42 104.0       WIA-OC5b -17.9  14.1   3 -1.269503546 0.070921986
#> 43  62.0    EORTC 73-OC   1.2  40.0   3  0.030000000 0.025000000
#> 44  71.0         Bergen  -3.1   6.7   3 -0.462686567 0.149253731
#> 45  68.0      RT-BLM-73   2.1   5.5   3  0.381818182 0.181818182
#> 46 109.0       WIA-OC5c  -4.4   7.0   3 -0.628571429 0.142857143
#> 47  60.0          Turku  -2.0   9.9   3 -0.202020202 0.101010101
#> 48  48.0          UW-77 -12.0  10.4   3 -1.153846154 0.096153846
#> 49  59.0         NRH-78   4.5  45.6   3  0.098684211 0.021929825
#> 50  64.0      Barcelona -25.3 121.2   3 -0.208745875 0.008250825
#> 51  58.0          UW-79  -6.8   4.6   3 -1.478260870 0.217391304
#> 52  61.0     Manchester  -2.4  66.2   3 -0.036253776 0.015105740
#> 53  72.0        Yale 80   2.0  24.3   3  0.082304527 0.041152263
#> 54  69.0         PMHCGS  -4.3  41.2   3 -0.104368932 0.024271845
#> 55  65.0      ECOG 2382   5.4  79.4   3  0.068010076 0.012594458
#> 56  26.2 AC CAMARGO (b)   2.1  11.9   3  0.176470588 0.084033613
#> 57  73.0       Toulouse -13.5  16.6   3 -0.813253012 0.060240964
#> 58  92.2   SECOG II (b)  -8.7  31.7   3 -0.274447950 0.031545741
#> 59  74.0        CH-7401  -2.2  11.2   3 -0.196428571 0.089285714
#> 60  75.0        Yale 86  -0.8  12.2   3 -0.065573770 0.081967213
#> 61  70.0      INRC HN-8 -14.3  30.0   3 -0.476666667 0.033333333
#> 62  66.0        Ontario  -5.1  28.3   3 -0.180212014 0.035335689
#> 63 108.0     Kragujevac -13.3  21.9   3 -0.607305936 0.045662100
#> 64  91.0     Bavaria 89 -16.8  29.8   3 -0.563758389 0.033557047
#> 65  98.0       LOHNG 91  -5.3  12.3   3 -0.430894309 0.081300813

### meta-analysis based on all 65 trials
res <- rma(yi, vi, data=dat, method="FE", digits=2)
res

#> 
#> Fixed-Effects Model (k = 65)
#> 
#> I^2 (total heterogeneity / total variability):   48.25%
#> H^2 (total variability / sampling variability):  1.93
#> 
#> Test for Heterogeneity:
#> Q(df = 64) = 123.67, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.11  0.02  -4.67  <.01  -0.16  -0.06  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp)

#> 
#>  pred ci.lb ci.ub 
#>  0.90  0.85  0.94 
#> 

### only adjuvant trials
res <- rma(yi, vi, data=dat, method="FE", subset=grp==1, digits=2)
res

#> 
#> Fixed-Effects Model (k = 8)
#> 
#> I^2 (total heterogeneity / total variability):   10.56%
#> H^2 (total variability / sampling variability):  1.12
#> 
#> Test for Heterogeneity:
#> Q(df = 7) = 7.83, p-val = 0.35
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.02  0.07  -0.34  0.74  -0.16   0.11    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp)

#> 
#>  pred ci.lb ci.ub 
#>  0.98  0.85  1.12 
#> 

### only neoadjuvant trials
res <- rma(yi, vi, data=dat, method="FE", subset=grp==2, digits=2)
res

#> 
#> Fixed-Effects Model (k = 31)
#> 
#> I^2 (total heterogeneity / total variability):   4.96%
#> H^2 (total variability / sampling variability):  1.05
#> 
#> Test for Heterogeneity:
#> Q(df = 30) = 31.57, p-val = 0.39
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.06  0.03  -1.64  0.10  -0.12   0.01    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp)

#> 
#>  pred ci.lb ci.ub 
#>  0.95  0.88  1.01 
#> 

### only concomitant trials
res <- rma(yi, vi, data=dat, method="FE", subset=grp==3, digits=2)
res

#> 
#> Fixed-Effects Model (k = 26)
#> 
#> I^2 (total heterogeneity / total variability):   66.13%
#> H^2 (total variability / sampling variability):  2.95
#> 
#> Test for Heterogeneity:
#> Q(df = 25) = 73.81, p-val < .01
#> 
#> Model Results:
#> 
#> estimate    se   zval  pval  ci.lb  ci.ub 
#>    -0.21  0.04  -5.43  <.01  -0.28  -0.13  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

predict(res, transf=exp)

#> 
#>  pred ci.lb ci.ub 
#>  0.81  0.76  0.88 
#>