Studies on Assessing the Effects of Teacher Expectations on Pupil IQ

Results from 19 studies examining how teachers' expectations about their pupils can influence actual IQ levels.

dat.raudenbush1985

Format

The data frame contains the following columns:

study	numeric	study number

author	character	study author(s)

year	numeric	publication year

weeks	numeric	weeks of contact prior to expectancy induction

setting	character	whether tests were group or individually administered

tester	character	whether test administrator was aware or blind

n1i	numeric	sample size of experimental group

n2i	numeric	sample size of control group

yi	numeric	standardized mean difference

vi	numeric	corresponding sampling variance

Details

In the so-called ‘Pygmalion study’ (Rosenthal & Jacobson, 1968), “all of the predominantly poor children in the so-called Oak elementary school were administered a test pretentiously labeled the ‘Harvard Test of Inflected Acquisition.’ After explaining that this newly designed instrument had identified those children most likely to show dramatic intellectual growth during the coming year, the experimenters gave the names of these ‘bloomers’ to the teachers. In truth, the test was a traditional IQ test and the ‘bloomers’ were a randomly selected 20% of the student population. After retesting the children 8 months later, the experimenters reported that those predicted to bloom had in fact gained significantly more in total IQ (nearly 4 points) and reasoning IQ (7 points) than the control group children. Further, at the end of the study, the teachers rated the experimental children as intellectually more curious, happier, better adjusted, and less in need of approval than their control group peers” (Raudenbush, 1984).

In the following years, a series of studies were conducted attempting to replicate this rather controversial finding. However, the great majority of those studies were unable to demonstrate a statistically significant difference between the two experimental groups in terms of IQ scores. Raudenbush (1984) conducted a meta-analysis based on 19 such studies to further examine the evidence for the existence of the ‘Pygmalion effect’. The dataset includes the results from these studies.

The outcome measure used for the meta-analysis was the standardized mean difference (yi), with positive values indicating that the supposed ‘bloomers’ had, on average, higher IQ scores than those in the control group. The weeks variable indicates the number of weeks of prior contact between teachers and students before the expectancy induction. Testing was done either in a group setting or individually, which is indicated by the setting variable. Finally, the tester variable indicates whether the test administrators were either aware or blind to the researcher-provided designations of the children's intellectual potential.

The data in this dataset were obtained from Raudenbush and Bryk (1985) with information on the setting and tester variables extracted from Raudenbush (1984).

Source

Raudenbush, S. W. (1984). Magnitude of teacher expectancy effects on pupil IQ as a function of the credibility of expectancy induction: A synthesis of findings from 18 experiments. Journal of Educational Psychology, 76(1), 85--97. https://doi.org/10.1037/0022-0663.76.1.85

Raudenbush, S. W., & Bryk, A. S. (1985). Empirical Bayes meta-analysis. Journal of Educational Statistics, 10(2), 75--98. https://doi.org/10.3102/10769986010002075

Examples

### copy data into 'dat' and examine data
dat <- dat.raudenbush1985
dat

#>    study               author year weeks setting tester n1i n2i      yi     vi 
#> 1      1     Rosenthal et al. 1974     2   group  aware  77 339  0.0300 0.0156 
#> 2      2          Conn et al. 1968    21   group  aware  60 198  0.1200 0.0216 
#> 3      3          Jose & Cody 1971    19   group  aware  72  72 -0.1400 0.0279 
#> 4      4   Pellegrini & Hicks 1972     0   group  aware  11  22  1.1800 0.1391 
#> 5      5   Pellegrini & Hicks 1972     0   group  blind  11  22  0.2600 0.1362 
#> 6      6    Evans & Rosenthal 1969     3   group  aware 129 348 -0.0600 0.0106 
#> 7      7       Fielder et al. 1971    17   group  blind 110 636 -0.0200 0.0106 
#> 8      8             Claiborn 1969    24   group  aware  26  99 -0.3200 0.0484 
#> 9      9               Kester 1969     0   group  aware  75  74  0.2700 0.0269 
#> 10    10              Maxwell 1970     1   indiv  blind  32  32  0.8000 0.0630 
#> 11    11               Carter 1970     0   group  blind  22  22  0.5400 0.0912 
#> 12    12              Flowers 1966     0   group  blind  43  38  0.1800 0.0497 
#> 13    13              Keshock 1970     1   indiv  blind  24  24 -0.0200 0.0835 
#> 14    14            Henrikson 1970     2   indiv  blind  19  32  0.2300 0.0841 
#> 15    15                 Fine 1972    17   group  aware  80  79 -0.1800 0.0253 
#> 16    16              Grieger 1970     5   group  blind  72  72 -0.0600 0.0279 
#> 17    17 Rosenthal & Jacobson 1968     1   group  aware  65 255  0.3000 0.0193 
#> 18    18   Fleming & Anttonen 1971     2   group  blind 233 224  0.0700 0.0088 
#> 19    19             Ginsburg 1970     7   group  aware  65  67 -0.0700 0.0303 

### random-effects model
res <- rma(yi, vi, data=dat)
res

#> 
#> Random-Effects Model (k = 19; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of total heterogeneity): 0.0188 (SE = 0.0155)
#> tau (square root of estimated tau^2 value):      0.1372
#> I^2 (total heterogeneity / total variability):   41.86%
#> H^2 (total variability / sampling variability):  1.72
#> 
#> Test for Heterogeneity:
#> Q(df = 18) = 35.8295, p-val = 0.0074
#> 
#> Model Results:
#> 
#> estimate      se    zval    pval    ci.lb   ci.ub 
#>   0.0837  0.0516  1.6208  0.1051  -0.0175  0.1849    
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

### create weeks variable where values larger than 3 are set to 3
dat$weeks.c <- ifelse(dat$week > 3, 3, dat$week)

### mixed-effects model with weeks.c variable as moderator
res <- rma(yi, vi, mods=~weeks.c, data=dat, digits=3)
res

#> 
#> Mixed-Effects Model (k = 19; tau^2 estimator: REML)
#> 
#> tau^2 (estimated amount of residual heterogeneity):     0.000 (SE = 0.007)
#> tau (square root of estimated tau^2 value):             0.001
#> I^2 (residual heterogeneity / unaccounted variability): 0.00%
#> H^2 (unaccounted variability / sampling variability):   1.00
#> R^2 (amount of heterogeneity accounted for):            100.00%
#> 
#> Test for Residual Heterogeneity:
#> QE(df = 17) = 16.571, p-val = 0.484
#> 
#> Test of Moderators (coefficient 2):
#> QM(df = 1) = 19.258, p-val < .001
#> 
#> Model Results:
#> 
#>          estimate     se    zval   pval   ci.lb   ci.ub 
#> intrcpt     0.407  0.087   4.678  <.001   0.237   0.578  *** 
#> weeks.c    -0.157  0.036  -4.388  <.001  -0.227  -0.087  *** 
#> 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#>