fitted.rma {metafor}	R Documentation

Fitted Values for rma Objects

Description

The function calculates the fitted values for objects of class "rma".

Usage

## S3 method for class 'rma':
fitted(object, ...)

Arguments

object 	 an object of class "rma".

... 	 other arguments.

Value

A vector with the fitted values.

Note

For rma.uni objects, the predict.rma.uni function also provides standard errors and confidence intervals for the fitted values. Best linear unbiased predictions (BLUPs) that combine the fitted values based on the fixed effects and the estimated contributions of the random effects can be obtained with blup.rma.uni.

For objects not involving moderators, the fitted values are all identical to the estimated value of the model intercept.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

predict.rma.uni, blup.rma.uni

Examples

### load BCG vaccine data
data(dat.bcg)

### meta-analysis of the log risk ratios using a mixed-effects model
### with two moderators (absolute latitude and publication year)
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, mods=cbind(ablat, year),
           data=dat.bcg, measure="RR", method="REML")
fitted(res)

[Package metafor version 0.5-7 Index]