fitted.rma {metafor}	R Documentation

Fitted Values for 'rma' Objects

Description

The function computes the fitted values for objects of class "rma".

Usage

## S3 method for class 'rma'
fitted(object, ...)

Arguments

object
an object of class "rma".

...
other arguments.

Value

A vector with the fitted values.

Note

The predict function also provides standard errors and confidence intervals for the fitted values. Best linear unbiased predictions (BLUPs) that combine the fitted values based on the fixed effects and the estimated contributions of the random effects can be obtained with blup (only for objects of class "rma.uni").

For objects not involving moderators, the fitted values are all identical to the estimated value of the model intercept.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

See Also

predict.rma for a function to computed predicted values and blup.rma.uni for a function to compute BLUPs that combine the fitted values and predicted random effects.

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### fit mixed-effects model with absolute latitude and publication year as moderators
res <- rma(yi, vi, mods = ~ ablat + year, data=dat)

### compute the fitted values
fitted(res)

[Package metafor version 3.4-0 Index]