fitted.rma {metafor}	R Documentation

Fitted Values for 'rma' Objects

Description

The function computes the fitted values for objects of class "rma".

Usage

## S3 method for class 'rma'
fitted(object, ...)

Arguments

object
an object of class "rma".

...
other arguments.

Value

A vector with the fitted values.

Note

The predict.rma function also provides standard errors and confidence intervals for the fitted values. Best linear unbiased predictions (BLUPs) that combine the fitted values based on the fixed effects and the estimated contributions of the random effects can be obtained with blup.rma.uni (only for objects of class "rma.uni").

For objects not involving moderators, the fitted values are all identical to the estimated value of the model intercept.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

predict.rma, blup.rma.uni

Examples

### meta-analysis of the log risk ratios using a mixed-effects model
### with two moderators (absolute latitude and publication year)
res <- rma(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg,
           mods = ~ ablat + year, data=dat.bcg)
fitted(res)

[Package metafor version 2.4-0 Index]