forest {metafor}	R Documentation

Forest Plots

Description

The function forest is generic. It can be used to create forest plots.

Usage

forest(x, ...)

Arguments

x 	 either an object of class "rma", a vector with the observed effect size or outcomes, or an object of class "cumul.rma". See ‘Details’.

... 	 other arguments.

Details

Currently, methods exist for three types of situations.

In the first case, object x is a fitted model object coming from the rma.uni, rma.mh, or rma.peto functions. The corresponding method is then forest.rma.

Alternatively, object x can be a vector with observed effect size or outcomes. The corresponding method is then forest.default.

Finally, object x could be an object coming from the cumul.rma.uni, cumul.rma.mh, or cumul.rma.peto functions. The corresponding method is then forest.cumul.rma.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

forest.rma, forest.default, forest.cumul.rma

[Package metafor version 0.5-7 Index]