Forest Plots

The forest function can be used to create forest plots.

forest(x, ...)

Arguments

x
either an object of class "rma", a vector with the observed effect sizes or outcomes, or an object of class "cumul.rma". See ‘Details’.

...
other arguments.

Details

Currently, methods exist for three types of situations.

In the first case, object x is a fitted model object coming from the rma.uni, rma.mh, or rma.peto functions. The corresponding method is then forest.rma.

Alternatively, object x can be a vector with observed effect sizes or outcomes. The corresponding method is then forest.default.

Finally, object x can be an object coming from the cumul.rma.uni, cumul.rma.mh, or cumul.rma.peto functions. The corresponding method is then forest.cumul.rma.

Author

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Lewis, S., & Clarke, M. (2001). Forest plots: Trying to see the wood and the trees. British Medical Journal, 322(7300), 1479--1480. https://doi.org/10.1136/bmj.322.7300.1479

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://doi.org/10.18637/jss.v036.i03

See also