funnel {metafor}	R Documentation

Funnel Plots

Description

The function funnel is generic. It can be used to create funnel plots.

Usage

funnel(x, ...)

Arguments

x 	 an object of class "rma".

... 	 other arguments.

Details

Currently, there is only a method for handling objects of class "rma" with the funnel function. Accordingly, the corresponding method is called funnel.rma. See the documentation for that function for more details.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

funnel.rma

[Package metafor version 0.5-7 Index]