leave1out {metafor}	R Documentation

Leave-One-Out Diagnostics

Description

The function leave1out is generic. For suitable model objects, it repeatedly fits the model, leaving out one observation at a time.

Usage

leave1out(x, ...)

Arguments

x 	 an object of class "rma.uni", "rma.mh", or "rma.peto".

... 	 other arguments.

Details

Currently, there are methods for handling objects of class "rma.uni", "rma.mh", and "rma.peto" with the leave1out function. Accordingly, the corresponding methods are called leave1out.rma.uni, leave1out.rma.mh, and leave1out.rma.peto. See the documentation for those functions for more details.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

leave1out.rma.uni, leave1out.rma.mh, leave1out.rma.peto

[Package metafor version 0.5-7 Index]