leave1out {metafor}	R Documentation

Leave-One-Out Diagnostics for 'rma' Objects

Description

The functions repeatedly fit the specified model, leaving out one observation/study at a time.

Usage

leave1out(x, ...)

## S3 method for class 'rma.uni'
leave1out(x, digits, transf, targs, progbar=FALSE, ...)
## S3 method for class 'rma.mh'
leave1out(x, digits, transf, targs, progbar=FALSE, ...)
## S3 method for class 'rma.peto'
leave1out(x, digits, transf, targs, progbar=FALSE, ...)

Arguments

x
an object of class "rma.mh", "rma.peto", or "rma.uni".

digits
integer specifying the number of decimal places to which the printed results should be rounded (if unspecified, the default is to take the value from the object).

transf
an optional argument specifying the name of a function that should be used to transform the model coefficients and interval bounds (e.g., transf=exp; see also transf). If unspecified, no transformation is used.

targs
optional arguments needed by the function specified under transf.

progbar
logical indicating whether a progress bar should be shown (the default is FALSE).

...
other arguments.

Details

The model specified via x must be a model without moderators (i.e., either a fixed- or a random-effects model and not a fixed-effects with moderators or mixed-effects model).

Value

An object of class "list.rma". The object is a list containing the following components:

estimate
estimated coefficients of the model.

se
standard errors of the coefficients.

zval
test statistics of the coefficients.

pval
p-values for the test statistics.

ci.lb
lower bounds of the confidence intervals for the coefficients.

ci.ub
upper bounds of the confidence intervals for the coefficients.

Q
test statistics for the tests of heterogeneity.

Qp
p-values for the tests of heterogeneity.

tau2
estimated amounts of (residual) heterogeneity (only for random-effects models).

I2
values of I² (only for random-effects models).

H2
values of H² (only for random-effects models).

The "list.rma" object is formatted and printed with print.list.rma.

Note

When using the transf option, the transformation is applied to the estimated coefficients and the corresponding interval bounds. The standard errors are then set equal to NA and are omitted from the printed output.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

Viechtbauer, W., & Cheung, M. W.-L. (2010). Outlier and influence diagnostics for meta-analysis. Research Synthesis Methods, 1, 112–125.

See Also

rma.uni, rma.mh, rma.peto

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### random-effects model
res <- rma(yi, vi, data=dat)

### leave-one-out analysis
leave1out(res)
leave1out(res, transf=exp)

### meta-analysis of the (log) risk ratios using the Mantel-Haenszel method
res <- rma.mh(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### leave-one-out analysis
leave1out(res)
leave1out(res, transf=exp)

### meta-analysis of the (log) odds ratios using Peto's method
res <- rma.peto(ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### leave-one-out analysis
leave1out(res)
leave1out(res, transf=exp)

[Package metafor version 2.4-0 Index]