permutest {metafor}	R Documentation

Permutation Tests

Description

The function permutest is generic. It can be used to carry out permutation tests for specific classes of objects.

Usage

permutest(x, ...)

Arguments

x 	 an object for which permutation tests can be carried out. See ‘Details’.

... 	 other arguments.

Details

Currently, there is only a method for objects of class "rma.uni" created by the rma.uni function. Accordingly, the corresponding method is called permutest.rma.uni. See the documentation for that function for more details.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

permutest.rma.uni

[Package metafor version 0.5-7 Index]