plot.infl.rma.uni {metafor}	R Documentation

Plot Method for infl.rma.uni Objects

Description

Plot method for objects of class "infl.rma.uni".

Usage

## S3 method for class 'infl.rma.uni':
plot(x, plotdfb=FALSE, dfbnew=FALSE, logcov=TRUE, las=0,
     pch=21, bg="black", bg.infl="red", col.na="lightgray", ...)

Arguments

x 	 an object of class "infl.rma.uni".

plotdfb 	 logical indicating whether the DFBETAS values should be plotted (default is FALSE).

dfbnew 	 logical indicating whether a new device should be opened for plotting the DFBETAS values (default is FALSE). If set to TRUE, then plotdfb is automatically set to TRUE as well.

logcov 	 logical indicating whether the covariance ratios should be plotted on a log scale (default is FALSE).

las 	 numeric (0,1,2,3). Defines style of axis labels (see par). The most useful alternative to 0 is 3, so that the x axis labels are drawn vertical to the axis.

pch 	 plotting symbol to use. By default, a filled circle is used. See points for other options.

bg 	 color to use for filling the plotting symbol (default is "black").

bg.infl 	 color to use for filling the plotting symbol when the point is considered influential (default is "red").

col.na 	 color to use for lines connecting two points with NAs in between (default is "lightgray").

... 	 other arguments.

Details

The function plots the externally standardized residuals, DFFITS values, Cook's distances, covariance ratios, leave-one-out tau^2 estimates, leave-one-out (residual) heterogeneity test statistics, hat values, and weights. If plotdfb=TRUE, the DFBETAS values are also plotted either after confirming the page change (if newdfb=FALSE) or on a separate device (if newdfb=TRUE). Points are considered ‘influential’ if at least one of the following is true:

The absolute DFFITS value is larger than 3sqrt(p/(k-p)), where p is the number of model coefficients and k the number of studies.
The lower tail area of a chi-square distribution with p degrees of freedom cut off by the Cook's distance is larger than 50%.
The hat value is larger than 3(p/k).
Any DFBETAS value is larger than 1.

These cut-offs are also indicated in the plot with horizontal reference lines. In addition, on the plot of the externally standardized residuals, horizontal reference lines are drawn at -1.96, 0, and 1.96. On the plot of the hat values, a horizontal reference line is drawn at p/k. Since the sum of the hat values is equal to p, the value p/k indicates equal hat values for all k studies. Finally, on the plot of weights, a horizontal reference line is drawn at 100/k, corresponding to the value for equal weights (in %) for all k studies. Note that all weights will automatically be equal to each other when using unweighted model fitting. Also, except for their scaling, the hat values and weights will be equal to each other in models without moderators.

The chosen cut-offs are (somewhat) arbitrary. Substantively informed judgment should always be used when examining the influence of each study on the results.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

influence.rma.uni

Examples

### load BCG vaccine data
data(dat.bcg)

### meta-analysis of the log risk ratios using a mixed-effects model
### with two moderators (absolute latitude and publication year)
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, mods=cbind(ablat, year),
           data=dat.bcg, measure="RR", method="REML")
plot(influence(res))
plot(influence(res), plotdfb=TRUE)

[Package metafor version 0.5-7 Index]