predict.rma {metafor}	R Documentation

Predicted Values for 'rma' Objects

Description

The function computes predicted values, corresponding standard errors, confidence intervals, and prediction intervals for objects of class "rma".

Usage

## S3 method for class 'rma'
predict(object, newmods, intercept, tau2.levels, gamma2.levels, addx=FALSE,
        level, digits, transf, targs, vcov=FALSE, ...)

## S3 method for class 'rma.ls'
predict(object, newmods, intercept, addx=FALSE, newscale, addz=FALSE,
        level, digits, transf, targs, vcov=FALSE, ...)

Arguments

object
an object of class "rma" or "rma.ls".

newmods
optional vector or matrix to specify the values of the moderator values for which the predicted values should be calculated. See ‘Details’.

intercept
logical to specify whether the intercept should be included when calculating the predicted values for newmods. If unspecified, the intercept is automatically added when the original model also included an intercept.

tau2.levels
vector to specify the levels of the inner factor when computing prediction intervals. Only relevant for models of class "rma.mv" (see rma.mv) and when the model includes more than a single \(\tau^2\) value. See ‘Details’.

gamma2.levels
vector to specify the levels of the inner factor when computing prediction intervals. Only relevant for models of class "rma.mv" (see rma.mv) and when the model includes more than a single \(\gamma^2\) value. See ‘Details’.

addx
logical to specify whether the values of the moderator variables should be added to the returned object. See ‘Examples’.

newscale
optional vector or matrix to specify the values of the scale variables for which the predicted values should be calculated. Only relevant for location-scale models (see rma.uni). See ‘Details’.

addz
logical to specify whether the values of the scale variables should be added to the returned object.

level
numeric value between 0 and 100 to specify the confidence and prediction interval level. If unspecified, the default is to take the value from the object.

digits
optional integer to specify the number of decimal places to which the printed results should be rounded. If unspecified, the default is to take the value from the object.

transf
optional argument to specify a function that should be used to transform the predicted values and interval bounds (e.g., transf=exp; see also transf). If unspecified, no transformation is used.

targs
optional arguments needed by the function specified under transf.

vcov
logical to specify whether the variance-covariance matrix of the predicted values should also be returned (the default is FALSE).

...
other arguments.

Details

For an equal-effects model, predict(object) returns the estimated (average) outcome in the set of studies included in the meta-analysis. This is the same as the estimated intercept in the equal-effects model (i.e., \(\hat{\theta}\)).

For a random-effects model, predict(object) returns the estimated (average) outcome in the hypothetical population of studies from which the set of studies included in the meta-analysis are assumed to be a random selection. This is the same as the estimated intercept in the random-effects model (i.e., \(\hat{\mu}\)).

For models including one or more moderators, predict(object) returns the estimated (average) outcomes for values of the moderator(s) equal to those of the \(k\) studies included in the meta-analysis (i.e., the ‘fitted values’ for the \(k\) studies).

For models including \(p'\) moderator variables, new moderator values (for \(k_{new}\) hypothetical new studies) can be specified by setting newmods equal to a \(k_{new} \times p'\) matrix with the corresponding new moderator values. If the model object included an intercept, then it should not be explicitly specified under newmods, as it will be added by default (unless one sets intercept=FALSE). Also, any factors in the original model get turned into the appropriate contrast variables within the rma function, so that newmods should actually include the values for the contrast variables. Examples are shown below.

For random/mixed-effects models, an approximate prediction interval is also constructed (Riley et al., 2011). The interval estimates where level % of the true effect sizes or outcomes fall in the hypothetical population of studies (and hence where the true effect or outcome of a new study from the population of studies should fall in level % of the cases).

For random-effects models that were fitted with the rma.mv function, the model may actually include multiple \(\tau^2\) values (i.e., when the random argument includes an ‘~ inner | outer’ term and struct="HCS", struct="DIAG", struct="HAR", or struct="UN"). In that case, the function will provide prediction intervals for each level of the inner factor (since the prediction intervals differ depending on the \(\tau^2\) value). Alternatively, one can use the tau2.levels argument to specify for which level(s) the prediction interval should be provided. If the model includes a second ‘~ inner | outer’ term with multiple \(\gamma^2\) values, prediction intervals for each combination of levels of the inner factors will be provided. Alternatively, one can use the tau2.levels and gamma2.levels arguments to specify for which level combination(s) the prediction interval should be provided.

When using the newmods argument for mixed-effects models that were fitted with the rma.mv function, if the model includes multiple \(\tau^2\) (and multiple \(\gamma^2\)) values, then one must use the tau2.levels (and gamma2.levels) argument to specify the levels of the inner factor(s) (i.e., a vector of length \(k_{new}\)) to obtain the appropriate prediction interval(s).

For location-scale models fitted with the rma.uni function, one can use newmods to specify the values of the \(p'\) moderator variables included in the model and newscale to specify the values of the \(q'\) scale variables included in the model. Whenever newmods is specified, the function computes predicted effects/outcomes for the specified moderators values. To obtain the corresponding prediction intervals, one must also specify the corresponding newscale values. If only newscale is specified (and not newmods), the function computes the predicted log-transformed \(\tau^2\) values (when using a log link) for the specified scale values. By setting transf=exp, one can then obtain the predicted \(\tau^2\) values.

Value

An object of class c("predict.rma","list.rma"). The object is a list containing the following components:

pred
predicted value(s).

se
corresponding standard error(s).

ci.lb
lower bound of the confidence interval(s).

ci.ub
upper bound of the confidence interval(s).

pi.lb
lower bound of the prediction interval(s) (only for random/mixed-effects models).

pi.ub
upper bound of the prediction interval(s) (only for random/mixed-effects models).

tau2.level
the level(s) of the inner factor (only for models of class "rma.mv" with multiple \(\tau^2\) values).

gamma2.level
the level(s) of the inner factor (only for models of class "rma.mv" with multiple \(\gamma^2\) values).

X
the moderator value(s) used to calculate the predicted values (only when addx=TRUE).

Z
the scale value(s) used to calculate the predicted values (only when addz=TRUE and only for location-scale models).

...
some additional elements/values.

If vcov=TRUE, then the returned object is a list with the first element equal to the one as described above and the second element equal to the variance-covariance matrix of the predicted values.

The object is formatted and printed with the print function.

Note

Confidence and prediction intervals are constructed based on the critical values from a standard normal distribution (i.e., \(\pm 1.96\) for level=95). When the model was fitted with test="t" or test="knha", then a t-distribution with \(k-p\) degrees of freedom is used.

For a random-effects model (where \(p=1\)) fitted with the rma.uni function, note that this differs slightly from Riley et al. (2001), who suggest to use a t-distribution with \(k-2\) degrees of freedom for constructing the prediction interval. Neither a normal, nor a t-distribution with \(k-1\) or \(k-2\) degrees of freedom is correct; all of these are approximations. The computations are done in the way described above, so that the prediction interval is identical to the confidence interval when \(\hat{\tau}^2 = 0\), which could be argued is the logical thing that should happen. If the prediction interval should be computed exactly as described by Riley et al. (2001), then one can use argument pi.type="riley".

The predicted values are based only on the fixed effects of the model. Best linear unbiased predictions (BLUPs) that combine the fitted values based on the fixed effects and the estimated contributions of the random effects can be obtained with blup (currently only for objects of class "rma.uni").

When using the transf option, the transformation is applied to the predicted values and the corresponding interval bounds. The standard errors are omitted from the printed output. Also, vcov=TRUE is ignored when using the transf option.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Hedges, L. V., & Olkin, I. (1985). Statistical methods for meta-analysis. San Diego, CA: Academic Press.

Riley, R. D., Higgins, J. P. T., & Deeks, J. J. (2011). Interpretation of random effects meta-analyses. British Medical Journal, 342, d549. https://doi.org/10.1136/bmj.d549

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

See Also

fitted.rma for a function to extract the fitted values, blup.rma.uni for a function to compute BLUPs that combine the fitted values and predicted random effects, and addpoly.predict.rma to add polygons based on predicted values to a forest plot.

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### fit random-effects model
res <- rma(yi, vi, data=dat)

### average risk ratio with 95% CI
predict(res, transf=exp)

### fit mixed-effects model with absolute latitude as a moderator
res <- rma(yi, vi, mods = ~ ablat, data=dat)

### predicted average risk ratios for given absolute latitude values
predict(res, transf=exp, addx=TRUE)

### predicted average risk ratios for 10-60 degrees absolute latitude
predict(res, newmods=c(10, 20, 30, 40, 50, 60), transf=exp, addx=TRUE)

### fit mixed-effects model with absolute latitude and publication year as moderators
res <- rma(yi, vi, mods = ~ ablat + year, data=dat)

### predicted average risk ratios for 10 and 60 degrees latitude in 1950 and 1980
predict(res, newmods=cbind(c(10,60,10,60),c(1950,1950,1980,1980)), transf=exp, addx=TRUE)

### fit mixed-effects model with two moderators (one of which is a factor)
res <- rma(yi, vi, mods = ~ ablat + factor(alloc), data=dat)

### examine how the factor was actually coded for the studies in the dataset
predict(res, addx=TRUE)

### predictd average risk ratios at 30 degrees for the three factor levels
### note: the contrast (dummy) variables need to specified explicitly here
predict(res, newmods=c(30, 0, 0), addx=TRUE)   # for alternate  allocation
predict(res, newmods=c(30, 1, 0), addx=TRUE)   # for random     allocation
predict(res, newmods=c(30, 0, 1), addx=TRUE)   # for systematic allocation

### can also use named vector with arbitrary order and abbreviated variable names
predict(res, newmods=c(sys=0, ran=0, abl=30))
predict(res, newmods=c(sys=0, ran=1, abl=30))
predict(res, newmods=c(sys=1, ran=0, abl=30))

[Package metafor version 3.4-0 Index]