print.anova.rma {metafor}	R Documentation

Print Method for 'anova.rma' Objects

Description

Print method for objects of class "anova.rma".

Usage

## S3 method for class 'anova.rma'
print(x, digits=x$digits, ...)

Arguments

x
an object of class "anova.rma" obtained with anova.

digits
integer to specify the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

...
other arguments.

Details

The output includes:

the number of parameters in the full and the reduced model.

the AIC, BIC, AICc, and log-likelihood of the full and the reduced model.

the value of the likelihood ratio test statistic.

the corresponding p-value.

the test statistic of the test for (residual) heterogeneity for the full and the reduced model.

the estimate of \(\tau^2\) from the full and the reduced model. Suppressed for equal-effects models.

amount (in percent) of heterogeneity in the reduced model that is accounted for in the full model (NA for "rma.mv" objects). This can be regarded as a pseudo \(R^2\) statistic (Raudenbush, 2009). Note that the value may not be very accurate unless \(k\) is large (Lopez-Lopez et al., 2014).

The last two items are not provided when comparing "rma.mv" models.

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

López-López, J. A., Marín-Martínez, F., Sánchez-Meca, J., Van den Noortgate, W., & Viechtbauer, W. (2014). Estimation of the predictive power of the model in mixed-effects meta-regression: A simulation study. British Journal of Mathematical and Statistical Psychology, 67(1), 30–48. https://doi.org/10.1111/bmsp.12002

Raudenbush, S. W. (2009). Analyzing effect sizes: Random effects models. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (2nd ed., pp. 295–315). New York: Russell Sage Foundation.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

See Also

anova.rma for the function to create anova.rma objects.

[Package metafor version 3.4-0 Index]