print.anova.rma {metafor}	R Documentation

Print Method for 'anova.rma' Objects

Description

Print method for objects of class "anova.rma".

Usage

## S3 method for class 'anova.rma'
print(x, digits=x$digits, ...)

Arguments

x
an object of class "anova.rma".

digits
integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

...
other arguments.

Details

The output includes:

the number of parameters in the full and the reduced model.

the AIC, BIC, AICc, and log-likelihood of the full and the reduced model.

the value of the likelihood ratio test statistic.

the p-value for the likelihood ratio test.

the test statistic for the test of (residual) heterogeneity for the full and the reduced model.

the estimate of τ² from the full and the reduced model. Suppressed for fixed-effects models.

R2amount (in percent) of heterogeneity in the reduced model that is accounted for in the full model (NA for fixed-effects models or for "rma.mv" objects). This can be regarded as a pseudo R² statistic (Raudenbush, 2009). Note that the value may not be very accurate unless k is large (Lopez-Lopez et al., 2014).

The last two items are not provided when comparing "rma.mv" models.

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

López-López, J. A., Marín-Martínez, F., Sánchez-Meca, J., Van den Noortgate, W., & Viechtbauer, W. (2014). Estimation of the predictive power of the model in mixed-effects meta-regression: A simulation study. British Journal of Mathematical and Statistical Psychology, 67, 30–48.

Raudenbush, S. W. (2009). Analyzing effect sizes: Random effects models. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (2nd ed., pp. 295–315). New York: Russell Sage Foundation.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

anova.rma

[Package metafor version 2.4-0 Index]