print.escalc {metafor}	R Documentation

Print and Summary Methods for 'escalc' Objects

Description

Print and summary methods for objects of class "escalc".

Usage

## S3 method for class 'escalc'
print(x, digits=attr(x,"digits"), ...)

## S3 method for class 'escalc'
summary(object, out.names=c("sei","zi","ci.lb","ci.ub"), var.names,
        H0=0, append=TRUE, replace=TRUE, level=95, clim, digits, transf, ...)

Arguments

x
an object of class "escalc".

object
an object of class "escalc".

digits
integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

out.names
character string with four elements, specifying the variable names for the standard errors, test statistics, and lower/upper confidence interval bounds.

var.names
character string with two elements, specifying the variable names for the observed outcomes and the sampling variances (the default is to take the value from the object if possible).

H0
numeric value specifying the value of the outcome under the null hypothesis.

append
logical indicating whether the data frame specified via the object argument should be returned together with the additional variables that are calculated by the summary function (the default is TRUE).

replace
logical indicating whether existing values for sei, zi, ci.lb, and ci.ub in the data frame should be replaced or not. Only relevant when the data frame already contains these variables. If replace=TRUE (the default), all of the existing values will be overwritten. If replace=FALSE, only NA values will be replaced.

level
numerical value between 0 and 100 specifying the confidence interval level (the default is 95).

clim
limits for the confidence intervals. If unspecified, no limits are used.

transf
optional argument specifying the name of a function that should be used to transform the observed outcomes and interval bounds (e.g., transf=exp; see also transf). If unspecified, no transformation is used. Any additional arguments needed for the function specified here can be passed via ....

...
other arguments.

Value

The print.escalc function formats and prints the data frame, so that the observed outcomes and sampling variances are rounded (to the number of digits specified).

The summary.escalc function creates an object that is a data frame containing the original data (if append=TRUE) and the following components:

yi
observed outcomes or effect size estimates (transformed if transf is specified).

vi
corresponding (estimated) sampling variances.

sei
standard errors of the observed outcomes or effect size estimates.

zi
test statistics for testing H₀: θᵢ = H0 (i.e., (yi-H0)/sei).

ci.lb
lower confidence interval bounds (transformed if transf is specified).

ci.ub
upper confidence interval bounds (transformed if transf is specified).

Note that the actual variable names above depend on the out.names (and var.names) arguments. If the data frame already contains variables with names as specified by the out.names argument, the values for these variables will be overwritten when replace=TRUE (which is the default). By setting replace=FALSE, only values that are NA will be replaced.

The print.escalc function again formats and prints the data frame, rounding the added variables to the number of digits specified.

Note

If some transformation function has been specified for the transf argument, then yi, ci.lb, and ci.ub will be transformed accordingly. However, vi and sei then still reflect the sampling variances and standard errors of the untransformed values.

The summary.escalc function computes level % Wald-type confidence intervals, which may or may not be the most accurate method for computing confidence intervals for the chosen outcome or effect size measure.

If the outcome measure used is bounded (e.g., correlations are bounded between -1 and 1, proportions are bounded between 0 and 1), one can use the clim argument to enforce those limits (confidence intervals cannot exceed those bounds then).

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

escalc

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)
dat

### apply summary function
summary(dat)
summary(dat, transf=exp)

[Package metafor version 2.4-0 Index]