print.hc.rma.uni {metafor}	R Documentation

Print Method for 'hc.rma.uni' Objects

Description

Print method for objects of class "hc.rma.uni".

Usage

## S3 method for class 'hc.rma.uni'
print(x, digits=x$digits, ...)

Arguments

x
an object of class "hc.rma.uni".

digits
integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

...
other arguments.

Details

The output is a data frame with two rows, the first (labeled rma) corresponding to the results based on the usual estimation method, the second (labeled hc) corresponding to the results based on the method by Henmi and Copas (2010). The data frame includes the following variables:

the method used to estimate τ² (always DL for hc)

the estimated amount of heterogeneity

the estimated average true effect

the corresponding standard errror (NA when transf argument has been used)

the lower and upper confidence interval bounds

Value

The function returns the data frame invisibly.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

hc.rma.uni

[Package metafor version 2.4-0 Index]