print.list.rma {metafor}	R Documentation

Print method for 'list.rma' Objects

Description

Print method for objects of class "list.rma".

Usage

## S3 method for class 'list.rma'
print(x, digits=x$digits, ...)

Arguments

x
an object of class "list.rma".

digits
integer to specify the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

...
other arguments.

Value

See the documentation of the function that creates the "list.rma" object for details on what is printed. Regardless of what is printed, a data frame with the results is also returned invisibly.

See methods.list.rma for some additional method functions for "list.rma" objects.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

[Package metafor version 3.4-0 Index]