print.permutest.rma.uni {metafor}	R Documentation

Print Method for permutest.rma.uni Objects

Description

Print method for objects of class "permutest.rma.uni".

Usage

## S3 method for class 'permutest.rma.uni':
print(x, digits=x$digits, signif.legend=TRUE, ...)

Arguments

x 	 an object of class "permutest.rma.uni".

digits 	 an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

signif.legend 	 logical indicating whether the legend for the ‘significance stars’ should be printed.

... 	 other arguments.

Details

The output includes:

the results of the omnibus test of the coefficients in the model. Suppressed if the model includes only one coefficient (e.g., only an intercept, like in the fixed- and random-effects model). The p-value based on the permutation test is given by p-val*.
a table with the estimated coefficients, corresponding standard errors, test statistics, p-values, and confidence interval bounds. The p-values based on the permutation test are given by pval*.

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

permutest.rma.uni

[Package metafor version 0.5-7 Index]