print.permutest.rma.uni {metafor}	R Documentation

Print Method for 'permutest.rma.uni' Objects

Description

Print method for objects of class "permutest.rma.uni".

Usage

## S3 method for class 'permutest.rma.uni'
print(x, digits=x$digits, signif.stars=getOption("show.signif.stars"),
      signif.legend=signif.stars, ...)

Arguments

x
an object of class "permutest.rma.uni".

digits
integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

signif.stars
logical indicating whether p-values should be encoded visually with ‘significance stars’. Defaults to the show.signif.stars slot of options.

signif.legend
logical indicating whether the legend for the ‘significance stars’ should be printed. Defaults to the value for signif.stars.

...
other arguments.

Details

The output includes:

the results of the omnibus test of the coefficients in the model. Suppressed if the model includes only one coefficient (e.g., only an intercept, like in the fixed- and random-effects model). The p-value based on the permutation test is indicated by p-val*.

a table with the estimated coefficients, corresponding standard errors, test statistics, p-values, and confidence interval bounds. The p-values based on the permutation test are indicated by pval*. When permutation-based CIs have been obtained, then the CI bounds are indicated with ci.lb* and ci.ub*.

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

permutest.rma.uni

[Package metafor version 2.4-0 Index]