print.regtest.rma {metafor}	R Documentation

Print Method for 'regtest.rma' Objects

Description

Print method for objects of class "regtest.rma".

Usage

## S3 method for class 'regtest.rma'
print(x, digits=x$digits, ret.fit=x$ret.fit, ...)

Arguments

x
an object of class "regtest.rma".

digits
integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

ret.fit
logical indicating whether the full results from the fitted model should also be returned (if unspecified, the default is to take the value from the object).

...
other arguments.

Details

The output includes:

the model used for the regression test

the predictor used for the regression test

the results from the fitted model (only when ret.fit=TRUE)

the value of the test statistic for the test that the predictor is unreleated to the outcomes

the degrees of freedom of the test statistic (only if the test statistic follows a t-distribution)

the p-value for the test statistic

Value

The function does not return an object.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

regtest.rma

[Package metafor version 2.4-0 Index]