print.rma {metafor}	R Documentation

Print and Summary Methods for 'rma' Objects

Description

Print and summary methods for objects of class "rma.uni", "rma.mh", "rma.peto", "rma.glmm", and "rma.glmm".

Usage

## S3 method for class 'rma.uni'
print(x, digits, showfit=FALSE, signif.stars=getOption("show.signif.stars"),
      signif.legend=signif.stars, ...)

## S3 method for class 'rma.mh'
print(x, digits, showfit=FALSE, ...)

## S3 method for class 'rma.peto'
print(x, digits, showfit=FALSE, ...)

## S3 method for class 'rma.glmm'
print(x, digits, showfit=FALSE, signif.stars=getOption("show.signif.stars"),
      signif.legend=signif.stars, ...)

## S3 method for class 'rma.mv'
print(x, digits, showfit=FALSE, signif.stars=getOption("show.signif.stars"),
      signif.legend=signif.stars, ...)

## S3 method for class 'rma'
summary(object, digits, showfit=TRUE, ...)

## S3 method for class 'summary.rma'
print(x, digits, showfit=TRUE, signif.stars=getOption("show.signif.stars"),
      signif.legend=signif.stars, ...)

Arguments

x
an object of class "rma.uni", "rma.mh", "rma.peto", "rma.glmm", "rma.mv", or "summary.rma" (for print).

object
an object of class "rma" (for summary).

digits
integer specifying the number of decimal places to which the printed results should be rounded (if unspecified, the default is to take the value from the object).

showfit
logical indicating whether the fit statistics and information criteria should be printed (the default is FALSE for print and TRUE for summary).

signif.stars
logical indicating whether p-values should be encoded visually with ‘significance stars’. Defaults to the show.signif.stars slot of options.

signif.legend
logical indicating whether the legend for the ‘significance stars’ should be printed. Defaults to the value for signif.stars.

...
other arguments.

Details

The output includes:

the log-likelihood, deviance, AIC, BIC, and AICc value (when setting showfit=TRUE or by default for summary).

for objects of class "rma.uni" and "rma.glmm", the amount of (residual) heterogeneity in the random/mixed-effects model (i.e., the estimate of τ² and its square root). Suppressed for fixed-effects models. The (asymptotic) standard error of the estimate of τ² is also provided (where possible).

for objects of "rma.mv", a table providing information about the variance components and correlations in the model. For σ² components, the estimate and its square root are provided, in addition to the number of values/levels, whether the component was fixed or estimated, and the name of the grouping variable/factor. If the R argument was used to specify known correlation matrices, this is also indicated. For models with an ~ inner | outer formula term, the name of the inner and outer grouping variable/factor are given and the number of values/levels of these variables/factors. In addition, for each τ² component, the estimate and its square root are provided, the number of effects or outcomes observed at each level of the inner grouping variable/factor (only for struct="HCS", struct="DIAG", struct="HAR", and struct="UN"), and whether the component was fixed or estimated. Finally, either the estimate of ρ (for struct="CS", struct="AR", struct="CAR", struct="HAR", or struct="HCS") or the entire estimated correlation matrix (for struct="UN") between the levels of the inner grouping variable/factor is provided, again with information whether a particular correlation was fixed or estimated, and how often each combination of levels of the inner grouping variable/factor was observed across the levels of the outer grouping variable/factor. If there is a second ~ inner | outer formula term, the same information as described above will be provided, but now for the γ² and φ components.

the I² statistic, which estimates (in percent) how much of the total variability in the effect size estimates (which is composed of heterogeneity plus sampling variability) can be attributed to heterogeneity among the true effects. For a meta-regression model, I² estimates how much of the unaccounted variability (which is composed of residual heterogeneity plus sampling variability) can be attributed to residual heterogeneity. See ‘Note’ for how I² is computed.

the H² statistic, which estimates the ratio of the total amount of variability in the effect size estimates to the amount of sampling variability. For a meta-regression model, H² estimates the ratio of the unaccounted variability in the effect size estimates to the amount of sampling variability. See ‘Note’ for how H² is computed.

for objects of class "rma.uni", the R² statistic, which estimates the amount of heterogeneity accounted for by the moderators included in the model and can be regarded as a pseudo R² statistic (Raudenbush, 2009). Only provided when fitting a mixed-effects models (i.e., for models including moderators). This is suppressed (and set to NULL) for models without moderators, fixed-effects models, or if the model does not contain an intercept. See ‘Note’ for how R² is computed.

for objects of class "rma.glmm", the amount of study level variability (only when using a model that models study level differences as a random effect).

the results of the test for (residual) heterogeneity. This is the usual Q-test for heterogeneity when not including moderators in the model and the Q_E-test for residual heterogeneity when moderators are included. For objects of class "rma.glmm", the results from a Wald-type test and a likelihood ratio test are provided (see rma.glmm for more details).

the results of the omnibus (Wald-type) test of the coefficients in the model (the indices of the coefficients tested are also indicated). Suppressed if the model includes only one coefficient (e.g., only an intercept, like in the fixed- and random-effects model).

a table with the estimated coefficients, corresponding standard errors, test statistics, p-values, and confidence interval bounds.

the Cochran-Mantel-Haenszel test and Tarone's test for heterogeneity (only when analyzing odds ratios using the Mantel-Haenszel method, i.e., "rma.mh").

Value

The print functions do not return an object. The summary function returns the object passed to it (with additional class "summary.rma").

Note

For random-effects models, the I² statistic is computed with

I² = 100% hat(τ)² / (hat(τ)² + s²),

where hat(τ)² is the estimated value of τ² and

s² = ((k-1) ∑ wᵢ) / ((∑ wᵢ)² - ∑ wᵢ²),

where wᵢ is the inverse of the sampling variance of the ith study (s² is equation 9 in Higgins & Thompson, 2002, and can be regarded as the ‘typical’ within-study variance of the observed effect sizes or outcomes). The H² statistic is computed with

H² = (hat(τ)² + s²) / s².

Analogous equations are used for mixed-effects models.

Therefore, depending on the estimator of τ² used, the values of I² and H² will change. For random-effects models, I² and H² are typically computed in practice with I² = (Q-(k-1))/Q and H² = Q/(k-1), where Q denotes the statistic for the test of heterogeneity and k the number of studies (i.e., observed effects or outcomes) included in the meta-analysis. The equations used in the metafor package to compute these statistics are more general and have the advantage that the values of I² and H² will be consistent with the estimated value of τ² (i.e., if hat(τ)² = 0, then I² = 0 and H² = 1 and if hat(τ)² > 0, then I² > 0 and H² > 1).

The two definitions of I² and H² actually coincide when using the DerSimonian-Laird estimator of τ² (i.e., the commonly used equations are actually special cases of the more general definitions given above). Therefore, if you prefer the more conventional definitions of these statistics, use method="DL" when fitting the random/mixed-effects model with the rma.uni function. The conventional definitions are also automatically used when fitting fixed-effects models.

The pseudo R² statistic (Raudenbush, 2009) is computed with

R² = (hat(τ)²_RE - hat(τ)²_ME) / hat(τ)²_RE,

where hat(τ)²_RE denotes the estimated value of τ² based on the random-effects model (i.e., the total amount of heterogeneity) and hat(τ)²_ME denotes the estimated value of τ² based on the mixed-effects model (i.e., the residual amount of heterogeneity). It can happen that hat(τ)²_RE < hat(τ)²_ME, in which case R² is set to zero (also if hat(τ)²_RE = 0). Again, the value of R² will change depending on the estimator of τ² used. Also note that this statistic is only computed when the mixed-effects model includes an intercept (so that the random-effects model is clearly nested within the mixed-effects model). You can also use the anova.rma function to compute R² for any two models that are known to be nested. Note that the pseudo R² statistic may not be very accurate unless k is large (Lopez-Lopez et al., 2014).

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Higgins, J. P. T., & Thompson, S. G. (2002). Quantifying heterogeneity in a meta-analysis. Statistics in Medicine, 21, 1539–1558.

López-López, J. A., Marín-Martínez, F., Sánchez-Meca, J., Van den Noortgate, W., & Viechtbauer, W. (2014). Estimation of the predictive power of the model in mixed-effects meta-regression: A simulation study. British Journal of Mathematical and Statistical Psychology, 67, 30–48.

Raudenbush, S. W. (2009). Analyzing effect sizes: Random effects models. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (2nd ed., pp. 295–315). New York: Russell Sage Foundation.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

See Also

rma.uni, rma.mh, rma.peto, rma.glmm, rma.mv

[Package metafor version 2.4-0 Index]