radial {metafor}	R Documentation

Radial (Galbraith) Plots for 'rma' Objects

Description

Function to create radial (also called Galbraith) plots for objects of class "rma".

Usage

radial(x, ...)
galbraith(x, ...)

## S3 method for class 'rma'
radial(x, center=FALSE, xlim, zlim, xlab, zlab,
       atz, aty, steps=7, level=x$level, digits=2, back="lightgray",
       transf, targs, pch=19, arc.res=100, cex, ...)

Arguments

x
an object of class "rma".

center
logical to indicate whether the plot should be centered horizontally at the model estimate (the default is FALSE).

xlim
x-axis limits. If unspecified, the function tries to set the x-axis limits to some sensible values.

zlim
z-axis limits. If unspecified, the function tries to set the z-axis limits to some sensible values (note that the z-axis limits are the actual vertical limit of the plotting region).

xlab
title for the x-axis. If unspecified, the function tries to set an appropriate axis title.

zlab
title for the z-axis. If unspecified, the function tries to set an appropriate axis title.

atz
position for the z-axis tick marks and labels. If unspecified, these values are set by the function.

aty
position for the y-axis tick marks and labels. If unspecified, these values are set by the function.

steps
the number of tick marks for the y-axis (the default is 7). Ignored when argument aty is used.

level
numeric value between 0 and 100 to specify the level of the z-axis error region (the default is to take the value from the object).

digits
integer to specify the number of decimal places to which the tick mark labels of the y-axis should be rounded (the default is 2).

back
color of the z-axis error region. Set to NA to suppress shading of the region.

transf
optional argument to specify a function that should be used to transform the y-axis labels (e.g., transf=exp; see also transf). If unspecified, no transformation is used.

targs
optional arguments needed by the function specified via transf.

pch
plotting symbol. By default, a filled circle is used. See points for other options.

arc.res
integer to specify the number of line segments to use when drawing the y-axis and confidence interval arcs (the default is 100).

cex
optional character and symbol expansion factor. If unspecified, the function tries to set this to a sensible value.

...
other arguments.

Details

For an equal-effects model, the plot shows the inverse of the standard errors on the horizontal axis against the observed effect sizes or outcomes standardized by their corresponding standard errors on the vertical axis. Since the vertical axis corresponds to standardized values, it is referred to as the z-axis within this function. On the right hand side of the plot, an arc is drawn (referred to as the y-axis within this function) corresponding to the observed effect sizes or outcomes. A line projected from (0,0) through a particular point within the plot onto this arc indicates the value of the observed effect size or outcome for that point.

For a random-effects model, the function uses \(1/\sqrt{v_i + \tau^2}\) for the horizontal axis, where \(v_i\) is the sampling variance of the observed effect size or outcome and \(\tau^2\) is the amount of heterogeneity as estimated based on the model. For the z-axis, \(\sqrt{v_i + \tau^2}\) is used to standardize the observed effect sizes or outcomes.

If the model contains moderators, the function returns an error.

Value

A data frame with components:

x
the x-axis coordinates of the points that were plotted.

y
the y-axis coordinates of the points that were plotted.

ids
the study id numbers.

slab
the study labels.

Note that the data frame is returned invisibly.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Galbraith, R. F. (1988). Graphical display of estimates having differing standard errors. Technometrics, 30(3), 271–281. https://doi.org/10.1080/00401706.1988.10488400

Galbraith, R. F. (1988). A note on graphical presentation of estimated odds ratios from several clinical trials. Statistics in Medicine, 7(8), 889–894. https://doi.org/10.1002/sim.4780070807

Galbraith, R. F (1994). Some applications of radial plots. Journal of the American Statistical Association, 89(428), 1232–1242. https://doi.org/10.1080/01621459.1994.10476864

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

See Also

rma.uni, rma.mh, rma.peto, rma.glmm, and rma.mv for functions to fit models for which radial plots can be drawn.

Examples

### calculate log risk ratios and corresponding sampling variances
dat <- escalc(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)
dat

### fit equal-effects model
res <- rma(yi, vi, data=dat, method="EE")

### draw radial plot
radial(res)

### line from (0,0) with slope equal to the log risk ratio from the 4th study
abline(a=0, b=dat$yi[4], lty="dotted")

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(yi, vi, data=dat)

### draw radial plot
radial(res)

[Package metafor version 3.4-0 Index]