ranktest {metafor}	R Documentation

Rank Correlation Test for Funnel Plot Asymmetry

Description

The function ranktest is generic. It can be used to carry out the rank correlation test for funnel plot asymmetry as described by Begg and Mazumdar (1994).

Usage

ranktest(x, ...)

Arguments

x 	 an object of class "rma".

... 	 other arguments.

Details

Currently, there is only a method for handling objects of class "rma" with the ranktest function. Accordingly, the corresponding method is called ranktest.rma. See the documentation for that function for more details.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

References

Begg, C. B. & Mazumdar, M. (1994) Operating characteristics of a rank correlation test for publication bias. Biometrics, 50, 1088–1101.

See Also

ranktest.rma

[Package metafor version 0.5-7 Index]