regtest {metafor}	R Documentation

Regression Tests for Funnel Plot Asymmetry

Description

The function regtest is generic. It can be used to carry out various tests for funnel plot asymmetry, including Egger's regression test and variations thereof.

Usage

regtest(x, ...)

Arguments

x 	 an object of class "rma".

... 	 other arguments.

Details

Currently, there is only a method for handling objects of class "rma" with the regtest function. Accordingly, the corresponding method is called regtest.rma. See the documentation for that function for more details.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

regtest.rma

[Package metafor version 0.5-7 Index]