residuals.rma {metafor}	R Documentation

Residual Values based on rma Objects

Description

The residuals, rstandard, and rstudent functions can be used to extract residuals, corresponding standard errors, and standardized residuals for models fitted with the rma.uni, rma.mh, and rma.peto functions.

Usage

## S3 method for class 'rma':
residuals(object, ...)
## S3 method for class 'rma.uni':
rstandard(model, digits=model$digits, ...)
## S3 method for class 'rma.mh':
rstandard(model, digits=model$digits, ...)
## S3 method for class 'rma.peto':
rstandard(model, digits=model$digits, ...)
## S3 method for class 'rma.uni':
rstudent(model, digits=model$digits, ...)
## S3 method for class 'rma.mh':
rstudent(model, digits=model$digits, ...)
## S3 method for class 'rma.peto':
rstudent(model, digits=model$digits, ...)

Arguments

object 	 an object of class "rma" (for residuals).

model 	 an object of class "rma.uni", "rma.mh", or "rma.peto" (for rstandard and rstudent).

digits 	 an integer specifying the number of decimal places to which the printed results should be rounded (the default is to take the value from the object).

... 	 other arguments.

Details

The observed residuals (obtained with residuals) are simply equal to the ‘observed - fitted’ values.

Dividing the observed residuals by their corresponding standard errors yields (internally) standardized residuals. These can be obtained with rstandard.

The rstudent function calculates externally standardized residuals (studentized deleted residuals). The externally standardized residual for the ith case is obtained by deleting the ith case from the dataset, fitting the model based on the remaining cases, calculating the predicted value for the ith case based on the fitted model, taking the difference between the observed and the predicted value for the ith case (the deleted residual), and then standardizing the deleted residual. The standard error of the deleted residual is equal to the square root of the sampling variance of the ith case plus the variance of the predicted value plus the amount of (residual) heterogeneity from the fitted model (for fixed-effects models, this last part is always equal to zero).

If a particular study fits the model, its standardized residual follows (asymptotically) a standard normal distribution. A large standardized residual for a study therefore may suggest that the study does not fit the assumed model (i.e., it may be an outlier).

See also influence.rma.uni for other leave-one-out diagnostics that are useful for detecting influential cases in models fitted with the rma.uni function.

Value

Either a vector with the observed residuals (for residuals) or an object of class "list.rma", which is a list containing the following components:

resid 	 observed residuals (for rstandard) or deleted residuals (for rstudent).

se 	 corresponding standard errors.

z 	 standardized residuals (internally standardized for rstandard or externally standardized for rstudent).

The "list.rma" object is formated and printed with print.list.rma.

Note

Right now, the externally standardized residuals are calculated by refitting the model k times. Depending on how large k is, it may take a few moments to finish the calculations.

It may not be possible to fit the model after deletion of the ith case from the dataset. This will result in NA values for that case when calling rstudent.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

References

Hedges, L. V. & Olkin, I. (1985) Statistical methods for meta-analysis. San Diego, CA: Academic Press.

See Also

rma.uni, rma.mh, rma.peto, influence.rma.uni

Examples

### load BCG vaccine data
data(dat.bcg)

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, 
           data=dat.bcg, measure="RR", method="REML")
rstudent(res)

### mixed-effects model with absolute latitude as a moderator
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, mods=ablat, 
           measure="RR", data=dat.bcg, method="REML")
rstudent(res)

[Package metafor version 0.5-7 Index]