Meta-Analysis via Peto's Method

Function to fit fixed-effects models to \(2 \times 2\) table data via Peto's method. See below and the documentation of the metafor-package for more details on these models.

rma.peto(ai, bi, ci, di, n1i, n2i,
         data, slab, subset,
         add=1/2, to="only0", drop00=TRUE,
         level=95, digits, verbose=FALSE, ...)

Arguments

ai
vector to specify the \(2 \times 2\) table frequencies (upper left cell). See below and the documentation of the escalc function for more details.

bi
vector to specify the \(2 \times 2\) table frequencies (upper right cell). See below and the documentation of the escalc function for more details.

ci
vector to specify the \(2 \times 2\) table frequencies (lower left cell). See below and the documentation of the escalc function for more details.

di
vector to specify the \(2 \times 2\) table frequencies (lower right cell). See below and the documentation of the escalc function for more details.

n1i
vector to specify the group sizes or row totals (first group). See below and the documentation of the escalc function for more details.

n2i
vector to specify the group sizes or row totals (second group). See below and the documentation of the escalc function for more details.

data
optional data frame containing the data supplied to the function.

slab
optional vector with labels for the \(k\) studies.

subset
optional (logical or numeric) vector indicating the subset of studies that should be used for the analysis.

add
non-negative number indicating the amount to add to zero cells, counts, or frequencies when calculating the observed outcomes of the individual studies. Can also be a vector of two numbers, where the first number is used in the calculation of the observed outcomes and the second number is used when applying Peto's method. See below and the documentation of the escalc function for more details.

to
character string indicating when the values under add should be added (either "only0", "all", "if0all", or "none"). Can also be a character vector, where the first string again applies when calculating the observed outcomes and the second string when applying Peto's method. See below and the documentation of the escalc function for more details.

drop00
logical indicating whether studies with no cases (or only cases) in both groups should be dropped when calculating the observed outcomes (the outcomes for such studies are set to NA). Can also be a vector of two logicals, where the first applies to the calculation of the observed outcomes and the second when applying Peto's method. See below and the documentation of the escalc function for more details.

level
numerical value between 0 and 100 specifying the confidence interval level (the default is 95).

digits
integer specifying the number of decimal places to which the printed results should be rounded (if unspecified, the default is 4).

verbose
logical indicating whether output should be generated on the progress of the model fitting (the default is FALSE).

...
additional arguments.

Details

Specifying the Data

The studies are assumed to provide data in terms of \(2 \times 2\) tables of the form:

outcome 1	outcome 2	total

group 1	ai	bi	n1i

group 2	ci	di	n2i

where ai, bi, ci, and di denote the cell frequencies and n1i and n2i the row totals. For example, in a set of randomized clinical trials (RCTs) or cohort studies, group 1 and group 2 may refer to the treatment (exposed) and placebo/control (not exposed) group, with outcome 1 denoting some event of interest (e.g., death) and outcome 2 its complement. In a set of case-control studies, group 1 and group 2 may refer to the group of cases and the group of controls, with outcome 1 denoting, for example, exposure to some risk factor and outcome 2 non-exposure.

Peto's Method

An approach for aggregating \(2 \times 2\) table data of this type was suggested by Peto (see Yusuf et al., 1985). The method provides a weighted estimate of the (log) odds ratio under a fixed-effects model. The method is particularly advantageous when the event of interest is rare, but it should only be used when the group sizes within the individual studies are not too dissimilar and effect sizes are generally small (Greenland & Salvan, 1990; Sweeting et al., 2004; Bradburn et al., 2007). Note that the printed results are given both in terms of the log and the raw units (for easier interpretation).

Observed Outcomes of the Individual Studies

Peto's method itself does not require the calculation of the observed (log) odds ratios of the individual studies and directly makes use of the \(2 \times 2\) table counts. Zero cells are not a problem (except in extreme cases, such as when one of the two outcomes never occurs in any of the tables). Therefore, it is unnecessary to add some constant to the cell counts when there are zero cells.

However, for plotting and various other functions, it is necessary to calculate the observed (log) odds ratios for the \(k\) studies. Here, zero cells can be problematic, so adding a constant value to the cell counts ensures that all \(k\) values can be calculated. The add and to arguments are used to specify what value should be added to the cell frequencies and under what circumstances when calculating the observed (log) odds ratios and when applying Peto's method. Similarly, the drop00 argument is used to specify how studies with no cases (or only cases) in both groups should be handled. The documentation of the escalc function explains how the add, to, and drop00 arguments work. If only a single value for these arguments is specified (as per default), then these values are used when calculating the observed (log) odds ratios and no adjustment to the cell counts is made when applying Peto's method. Alternatively, when specifying two values for these arguments, the first value applies when calculating the observed (log) odds ratios and the second value when applying Peto's method.

Note that drop00 is set to TRUE by default. Therefore, the observed (log) odds ratios for studies where ai=ci=0 or bi=di=0 are set to NA. When applying Peto's method, such studies are not explicitly dropped (unless the second value of drop00 argument is also set to TRUE), but this is practically not necessary, as they do not actually influence the results (assuming no adjustment to the cell/event counts are made when applying Peto's method).

Value

An object of class c("rma.peto","rma"). The object is a list containing the following components:

beta

aggregated log odds ratio.

se

standard error of the aggregated value.

zval

test statistics of the aggregated value.

pval

p-value for the test statistic.

ci.lb

lower bound of the confidence interval.

ci.ub

upper bound of the confidence interval.

QE

test statistic for the test of heterogeneity.

QEp

p-value for the test of heterogeneity.

k

number of tables included in the analysis.

yi, vi

the vector of individual log odds ratios and corresponding sampling variances.

fit.stats

a list with the log-likelihood, deviance, AIC, BIC, and AICc values under the unrestricted and restricted likelihood.

...

some additional elements/values.

Methods

The results of the fitted model are formatted and printed with the print.rma.peto function. If fit statistics should also be given, use summary.rma (or use the fitstats.rma function to extract them).

The residuals.rma, rstandard.rma.peto, and rstudent.rma.peto functions extract raw and standardized residuals. Leave-one-out diagnostics can be obtained with leave1out.rma.peto.

Forest, funnel, radial, L'Abbé, and Baujat plots can be obtained with forest.rma, funnel.rma, radial.rma, labbe.rma, and baujat. The qqnorm.rma.peto function provides normal QQ plots of the standardized residuals. One can also just call plot.rma.peto on the fitted model object to obtain various plots at once.

A cumulative meta-analysis (i.e., adding one observation at a time) can be obtained with cumul.rma.peto.

Other extractor functions include coef.rma, vcov.rma, logLik.rma, deviance.rma, AIC.rma, and BIC.rma.

References

Bradburn, M. J., Deeks, J. J., Berlin, J. A., & Localio, A. R. (2007). Much ado about nothing: A comparison of the performance of meta-analytical methods with rare events. Statistics in Medicine, 26, 53--77.

Greenland, S., & Salvan, A. (1990). Bias in the one-step method for pooling study results. Statistics in Medicine, 9, 247--252.

Sweeting, M. J., Sutton, A. J., & Lambert, P. C. (2004). What to add to nothing? Use and avoidance of continuity corrections in meta-analysis of sparse data. Statistics in Medicine, 23, 1351--1375.

Yusuf, S., Peto, R., Lewis, J., Collins, R., & Sleight, P. (1985). Beta blockade during and after myocardial infarction: An overview of the randomized trials. Progress in Cardiovascular Disease, 27, 335--371.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://www.jstatsoft.org/v036/i03.

See also

Examples

### meta-analysis of the (log) odds ratios using Peto's method
rma.peto(ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

#> 
#> Fixed-Effects Model (k = 13)
#> 
#> I^2 (total heterogeneity / total variability):  92.85%
#> H^2 (total variability / sampling variability): 13.98
#> 
#> Test for Heterogeneity: 
#> Q(df = 12) = 167.7302, p-val < .0001
#> 
#> Model Results (log scale):
#> 
#> estimate      se      zval    pval    ci.lb    ci.ub 
#>  -0.4744  0.0407  -11.6689  <.0001  -0.5541  -0.3948 
#> 
#> Model Results (OR scale):
#> 
#> estimate   ci.lb   ci.ub 
#>   0.6222  0.5746  0.6738 
#>