to.table {metafor}	R Documentation

Convert Data from Vector to Table Format

Description

The function converts summary data in vector format to the corresponding table format.

Usage

to.table(measure, ai, bi, ci, di, n1i, n2i, x1i, x2i, t1i, t2i,
         m1i, m2i, sd1i, sd2i, xi, mi, ri, ti, sdi, ni, data, slab, subset,
         add=1/2, to="none", drop00=FALSE, rows, cols)

Arguments

measure
a character string to specify the effect size or outcome measure corresponding to the summary data supplied. See ‘Details’ and the documentation of the escalc function for possible options.

ai
vector to specify the \(2 \times 2\) table frequencies (upper left cell).

bi
vector to specify the \(2 \times 2\) table frequencies (upper right cell).

ci
vector to specify the \(2 \times 2\) table frequencies (lower left cell).

di
vector to specify the \(2 \times 2\) table frequencies (lower right cell).

n1i
vector to specify the group sizes or row totals (first group/row).

n2i
vector to specify the group sizes or row totals (second group/row).

x1i
vector to specify the number of events (first group).

x2i
vector to specify the number of events (second group).

t1i
vector to specify the total person-times (first group).

t2i
vector to specify the total person-times (second group).

m1i
vector to specify the means (first group or time point).

m2i
vector to specify the means (second group or time point).

sd1i
vector to specify the standard deviations (first group or time point).

sd2i
vector to specify the standard deviations (second group or time point).

xi
vector to specify the frequencies of the event of interest.

mi
vector to specify the frequencies of the complement of the event of interest or the group means.

ri
vector to specify the raw correlation coefficients.

ti
vector to specify the total person-times.

sdi
vector to specify the standard deviations.

ni
vector to specify the sample/group sizes.

data
optional data frame containing the variables given to the arguments above.

slab
optional vector with labels for the studies.

subset
optional (logical or numeric) vector to specify the subset of studies that should be included in the array returned by the function.

add
see the documentation of the escalc function.

to
see the documentation of the escalc function.

drop00
see the documentation of the escalc function.

rows
optional vector with row/group names.

cols
optional vector with column/outcome names.

Details

The escalc function describes a wide variety of effect size or outcome measures that can be computed for a meta-analysis. The summary data used to compute those measures are typically contained in vectors, each element corresponding to a study. The to.table function takes this information and constructs an array of \(k\) tables from these data.

For example, in various fields (such as the health and medical sciences), the response variable measured is often dichotomous (binary), so that the data from a study comparing two different groups can be expressed in terms of a \(2 \times 2\) table, such as:

outcome 1 	 outcome 2 	 total

group 1 	 ai 	 bi 	 n1i

group 2 	 ci 	 di 	 n2i

where ai, bi, ci, and di denote the cell frequencies (i.e., the number of people falling into a particular category) and n1i and n2i the row totals (i.e., the group sizes).

The cell frequencies in \(k\) such \(2 \times 2\) tables can be specified via the ai, bi, ci, and di arguments (or alternatively, via the ai, ci, n1i, and n2i arguments). The function then creates the corresponding \(2 \times 2 \times k\) array of tables. The measure argument should then be set equal to one of the outcome measures that can be computed based on this type of data, such as "RR", "OR", "RD" (it is not relevant which specific measure is chosen, as long as it corresponds to the specified summary data). See the documentation of the escalc function for more details on the types of data formats available.

The examples below illustrate the use of this function.

Value

An array with \(k\) elements each consisting of either 1 or 2 rows and an appropriate number of columns.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

See Also

escalc for a function to compute observed effect sizes or outcomes (and corresponding sampling variances) based on similar inputs.

to.long for a function to turn similar inputs into a long format dataset.

Examples

### create tables
dat <- to.table(measure="OR", ai=tpos, bi=tneg, ci=cpos, di=cneg,
                data=dat.bcg, slab=paste(author, year, sep=", "),
                rows=c("Vaccinated", "Not Vaccinated"), cols=c("TB+", "TB-"))
dat

### create tables
dat <- to.table(measure="IRR", x1i=x1i, x2i=x2i, t1i=t1i, t2i=t2i,
                data=dat.hart1999, slab=paste(study, year, sep=", "),
                rows=c("Warfarin Group", "Placebo/Control Group"))
dat

### create tables
dat <- to.table(measure="MD", m1i=m1i, sd1i=sd1i, n1i=n1i,
                m2i=m2i, sd2i=sd2i, n2i=n2i, data=dat.normand1999,
                slab=source, rows=c("Specialized Care", "Routine Care"))
dat

[Package metafor version 3.4-0 Index]