transf {metafor}	R Documentation

Transformation Functions

Description

A set of transformation functions useful for meta-analyses.

Usage

transf.rtoz(x, ...)
transf.ztor(x, ...)
transf.logit(x, ...)
transf.ilogit(x, ...)
transf.ztor.int(x, targs, ...)
transf.exp.int(x, targs, ...)

Arguments

x 	 a vector of values to be transformed.

targs 	 a list with additional arguments for the transformation function. See ‘Details’.

... 	 other arguments.

Details

The following transformation functions are currently implemented:

transf.rtoz: Fisher's r-to-z transformation.
transf.ztor: inverse of the Fisher's r-to-z transformation.
transf.logit: logit (log odds) transformation.
transf.ilogit: inverse of the logit transformation.
transf.ztor.int: integral transformation method for the z-to-r transformation.
transf.exp.int: integral transformation method for the exponential transformation.

The integral transformation method for a transformation function h(z) integrates h(z) f(z) over z using the limits targs$lower and targs$upper, where f(z) is the density of a normal distribution with mean equal to x and variance equal to targs$tau2.

Value

A vector with the transformed values.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

Examples

### load BCG vaccine data
data(dat.bcg)

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, 
           measure="RR", data=dat.bcg, method="REML")

### average risk ratio with 95% CI
predict(res, transf=exp)

### average risk ratio with 95% CI using integral transformation
predict(res, transf=transf.exp.int, targs=list(tau2=res$tau2, lower=-4, upper=4))

[Package metafor version 0.5-7 Index]