transf {metafor}	R Documentation

Transformation Function

Description

A set of transformation functions useful for meta-analyses.

Usage

transf.rtoz(xi, ...)
transf.ztor(xi, ...)
transf.logit(xi, ...)
transf.ilogit(xi, ...)
transf.arcsin(xi, ...)
transf.iarcsin(xi, ...)
transf.pft(xi, ni, ...)
transf.ipft(xi, ni, ...)
transf.ipft.hm(xi, targs, ...)
transf.isqrt(xi, ...)
transf.irft(xi, ti, ...)
transf.iirft(xi, ti, ...)
transf.ahw(xi, ...)
transf.iahw(xi, ...)
transf.abt(xi, ...)
transf.iabt(xi, ...)
transf.ztor.int(xi, targs, ...)
transf.exp.int(xi, targs, ...)
transf.ilogit.int(xi, targs, ...)

Arguments

xi
vector of values to be transformed.

ni
vector of sample sizes.

ti
vector of person-times at risk.

targs
list with additional arguments for the transformation function. See ‘Details’.

...
other arguments.

Details

The following transformation functions are currently implemented:

transf.rtoz: Fisher's r-to-z transformation for correlations.

transf.ztor: inverse of the Fisher's r-to-z transformation.

transf.logit: logit (log odds) transformation for proportions.

transf.ilogit: inverse of the logit transformation.

transf.arcsin: arcsine square root transformation for proportions.

transf.iarcsin: inverse of the arcsine transformation.

transf.pft: Freeman-Tukey (double arcsine) transformation for proportions. See Freeman & Tukey (1950). The xi argument is used to specify the proportions and the ni argument the corresponding sample sizes.

transf.ipft: inverse of the Freeman-Tukey (double arcsine) transformation for proportions. See Miller (1978).

transf.ipft.hm: inverse of the Freeman-Tukey (double arcsine) transformation for proportions using the harmonic mean of the sample sizes for the back-transformation. See Miller (1978). The sample sizes are specified via the targs argument (the list element should be called ni).

transf.isqrt: inverse of the square root transformation (i.e., function to square a number).

transf.irft: Freeman-Tukey transformation for incidence rates. See Freeman & Tukey (1950). The xi argument is used to specify the incidence rates and the ti argument the corresponding person-times at risk.

transf.iirft: inverse of the Freeman-Tukey transformation for incidence rates.

transf.ahw: Transformation of coefficient alpha as suggested by Hakstian & Whalen (1976).

transf.iahw: Inverse of the transformation of coefficient alpha as suggested by Hakstian & Whalen (1976).

transf.abt: Transformation of coefficient alpha as suggested by Bonett (2002).

transf.iabt: Inverse of the transformation of coefficient alpha as suggested by Bonett (2002).

transf.ztor.int: integral transformation method for the z-to-r transformation.

transf.exp.int: integral transformation method for the exponential transformation.

transf.ilogit.int: integral transformation method for the inverse of the logit transformation.

The integral transformation method for a transformation function h(z) integrates h(z) f(z) over z using the limits targs$lower and targs$upper, where f(z) is the density of a normal distribution with mean equal to xi and variance equal to targs$tau2. An example is provided below.

Value

A vector with the transformed values.

Author(s)

Wolfgang Viechtbauer wvb@metafor-project.org http://www.metafor-project.org/

References

Bonett, D. G. (2002). Sample size requirements for testing and estimating coefficient alpha. Journal of Educational and Behavioral Statistics, 27, 335–340.

Fisher, R. A. (1921). On the “probable error” of a coefficient of correlation deduced from a small sample. Metron, 1, 1–32.

Freeman, M. F., & Tukey, J. W. (1950). Transformations related to the angular and the square root. Annals of Mathematical Statistics, 21, 607–611.

Hakstian, A. R., & Whalen, T. E. (1976). A k-sample significance test for independent alpha coefficients. Psychometrika, 41, 219–231.

Miller, J. J. (1978). The inverse of the Freeman-Tukey double arcsine transformation. American Statistician, 32, 138.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1–48. https://www.jstatsoft.org/v036/i03.

Examples

### meta-analysis of the log risk ratios using a random-effects model
res <- rma(measure="RR", ai=tpos, bi=tneg, ci=cpos, di=cneg, data=dat.bcg)

### average risk ratio with 95% CI (but technically, this provides an
### estimate of the median risk ratio, not the mean risk ratio!)
predict(res, transf=exp)

### average risk ratio with 95% CI using the integral transformation
predict(res, transf=transf.exp.int, targs=list(tau2=res$tau2, lower=-4, upper=4))

[Package metafor version 2.4-0 Index]