trimfill {metafor}	R Documentation

Trim and Fill Method

Description

The function trimfill is generic. It can be used to apply the trim and fill method proposed by Duval and Tweedie (2000a, 2000b; see also Duval, 2005) to specific classes of objects.

Usage

trimfill(x, ...)

Arguments

x 	 an object of class "rma.uni".

... 	 other arguments.

Details

Currently, there is only a method for objects of class "rma.uni" created by the rma.uni function. Accordingly, the corresponding method is called trimfill.rma.uni. See the documentation for that function for more details.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

References

Duval, S. J. & Tweedie, R. L. (2000a) Trim and fill: A simple funnel-plot-based method of testing and adjusting for publication bias in meta-analysis. Biometrics, 56, 455–463.

Duval, S. J. & Tweedie, R. L. (2000b) A nonparametric "trim and fill" method of accounting for publication bias in meta-analysis. Journal of the American Statistical Association, 95, 89–98.

Duval, S. J. (2005). The trim and fill method. In H. R. Rothstein, A. J. Sutton & M. Borenstein (Eds.) (2005) Publication bias in meta-analysis: Prevention, assessment, and adjustments (pp. 127–144). Chichester, England: Wiley.

See Also

trimfill.rma.uni

[Package metafor version 0.5-7 Index]