Construct or Approximate the Variance-Covariance Matrix of Dependent Effect Sizes or Outcomes

The function can be used to construct or approximate the variance-covariance matrix of dependent effect sizes or outcomes, or more precisely, of their sampling errors (i.e., the V matrix in rma.mv).

vcalc(vi, cluster, subgroup, obs, type, time1, time2, grp1, grp2, w1, w2,
      data, rho, phi, rvars, checkpd=TRUE, nearpd=FALSE, ...)

Arguments

vi
numeric vector to specify the sampling variances of the observed effect sizes or outcomes.

cluster
vector to specify the clustering variable (e.g., study).

subgroup
optional vector to specify different (independent) subgroups within clusters.

obs
optional vector to distinguish different observed effect sizes or outcomes corresponding to the same construct or response/dependent variable.

type
optional vector to distinguish different types of constructs or response/dependent variables underlying the observed effect sizes or outcomes.

time1
optional numeric vector to specify the time points when the observed effect sizes or outcomes were obtained (in the first condition if the observed effect sizes or outcomes represent contrasts between two conditions).

time2
optional numeric vector to specify the time points when the observed effect sizes or outcomes were obtained in the second condition (only relevant when the observed effect sizes or outcomes represent contrasts between two conditions).

grp1
optional vector to specify the group of the first condition when the observed effect sizes or outcomes represent contrasts between two conditions.

grp2
optional vector to specify the group of the second condition when the observed effect sizes or outcomes represent contrasts between two conditions.

w1
optional numeric vector to specify the size of the group (or more generally, the inverse-sampling variance weight) of the first condition when the observed effect sizes or outcomes represent contrasts between two conditions.

w2
optional numeric vector to specify the size of the group (or more generally, the inverse-sampling variance weight) of the second condition when the observed effect sizes or outcomes represent contrasts between two conditions.

data
optional data frame containing the variables given to the arguments above.

rho
argument to specify the correlation(s) of observed effect sizes or outcomes measured concurrently. See ‘Details’.

phi
argument to specify the autocorrelation of observed effect sizes or outcomes measured at different time points. See ‘Details’.

rvars
unused argument for the moment.

checkpd
logical to specify whether to check that the variance-covariance matrices within clusters are positive definite (the default is TRUE). See ‘Note’.

nearpd
logical to specify whether the nearPD function from the Matrix package should be used on variance-covariance matrices that are not positive definite. See ‘Note’.

...
other arguments.

Details

Standard meta-analytic models (such as those that can be fitted with the rma.uni function) assume that the observed effect sizes or outcomes (or more precisely, their sampling errors) are independent. This assumption is typically violated whenever multiple observed effect sizes or outcomes are computed based on the same sample of subjects (or whatever the unit of analysis is) or if there is at least partial overlap of subjects that contribute information to the computation of multiple effect sizes or outcomes.

The present function can be used to construct or approximate the variance-covariance matrix of the sampling errors of dependent effect sizes or outcomes for a wide variety of circumstances (this variance-covariance matrix is the so-called V matrix that may be needed as input for multilevel/multivariate meta-analytic models as can be fitted with the rma.mv function).

Argument cluster is used to specify the clustering variable. Rows with the same value of this variable are allowed to be dependent, while rows with different values are assumed to be independent. Typically, cluster will be a study identifier.

Within the same cluster, there may be different subgroups with no overlap of subjects across subgroups. Argument subgroup can be used to distinguish such subgroups. Rows with the same value of this variable within a cluster are allowed to be dependent, while rows with different values are assumed to be independent even if they come from the same cluster.

Multiple effect sizes or outcomes belonging to the same cluster and subgroup may be dependent due to a variety of reasons:

The same construct of interest (e.g., severity of depression) may have been measured using different scales or instruments within a study (e.g., using the Beck Depression Inventory (BDI) and the Hamilton Depression Rating Scale (HDRS)) based on which multiple effect sizes can be computed in the same group of subjects (e.g., contrasting a treated versus a control group with respect to each scale). In this case, we have multiple effect sizes that are different ‘observations’ of the effect with respect to the same type of construct.

Argument obs is then used to distinguish different effect sizes corresponding to the same construct. If obs is specified, then argument rho must also be specified to indicate the degree of correlation among the sampling errors of the different effect sizes. Since this correlation is typically not known, the correlation among the various scales (or a rough ‘guestimate’ thereof) can be used as a proxy (i.e., the (typical) correlation between BDI and HDRS measurements).

One can also specify an entire correlation matrix via rho to indicate, for each possible pair of obs values, the corresponding correlation. The row/column names of the matrix must then correspond to the unique values of the obs variable.

Multiple types of constructs (or more generally, types of response/dependent variables) may have been measured in the same group of subjects (e.g., severity of depression as measured with the Beck Depression Inventory (BDI) and severity of anxiety as measured with the State-Trait Anxiety Inventory (STAI)). If this is of interest for a meta-analysis, effect sizes can then be computed with respect to each ‘type’ of construct.

Argument type is then used to distinguish effect sizes corresponding to these different types of constructs. If type is specified, then argument rho must also be specified to indicate the degree of correlation among the sampling errors of effect sizes belonging to these different types. As above, the correlation among the various scales is typically used here as a proxy (i.e., the (typical) correlation between BDI and STAI measurements).

One can also specify an entire correlation matrix via rho to indicate, for each possible pair of type values, the corresponding correlation. The row/column names of the matrix must then correspond to the unique values of the type variable.

If there are multiple constructs, multiple scales or instruments may also have been used (in at least some of the studies) to measure the same construct and hence there may again be multiple effect sizes that are ‘observations’ of the same type of construct. Arguments type and obs should then be used together to indicate the various construct types and observations thereof. In this case, argument rho must be a vector of two values, the first to specify the within-construct correlation and the second to specify the between-construct correlation.

One can also specify a list with two elements for rho, the first element being either a scalar or an entire correlation matrix for the within-construct correlation(s) and the second element being a scalar or an entire correlation matrix for the between-construct correlation(s). As above, any matrices specified must have row/column names corresponding to the unique values of the obs and/or type variables.

The same construct and scale may have been assessed/used multiple times, allowing the computation of multiple effect sizes for the same group of subjects at different time points (e.g., right after the end of a treatment, at a short-term follow-up, and at a long-term follow-up). Argument time1 is then used to specify the time points when the observed effect sizes were obtained. Argument phi must then also be specified to indicate the autocorrelation among the sampling errors of two effect sizes that differ by one unit on the time1 variable. As above, the autocorrelation of the measurements themselves can be used here as a proxy.

If multiple constructs and/or multiple scales have also been assessed at the various time points, then arguments type and/or obs (together with argument rho) should be used as needed to differentiate effect sizes corresponding to the different constructs and/or scales.

Many effect size or outcome measures (e.g., raw or standardized mean differences, log-transformed ratios of means, log risk/odds ratios and risk differences) reflect the difference between two conditions (i.e., a contrast). Within a study, there may be more than two conditions, allowing the computation of multiple such contrasts (e.g., treatment A versus a control condition and treatment B versus the same control condition) and hence corresponding effect sizes. The reuse of information from the ‘shared’ condition (in this example, the control condition) then induces correlation among the effect sizes.

To account for this, arguments grp1 and grp2 should be specified to indicate (within a cluster or subgroup thereof) which two groups were compared in the computation of each effect size (e.g., in the example above, the coding could be grp1=c(1,2) and grp2=c(3,3); whether numbers or strings are used as identifiers is irrelevant).

The degree of correlation between two contrast-type effect sizes that is induced by the use of a shared condition is a function of the size of the groups involved in the computation of the two effect sizes (or, more generally, the inverse-sampling variance weights of the condition-specific outcomes). By default, the group sizes (weights) are assumed to be identical across conditions, which implies a correlation of 0.5. If the group sizes (weights) are known, they can be specified via arguments w1 and w2.

In fact, a contrast-type effect size can be based on a between- or a within-subjects design. When at least one or more of the contrast-type effect sizes are based on a within-subjects design, then time1 and time2 should be used in combination with grp1 and grp2 to indicate for each effect size the group(s) and time point(s) involved.

For example, grp1=c(1,2) and grp2=c(3,3) as above in combination with time1=c(1,1) and time2=c(1,1) would imply a between-subjects design involving three groups where two effect size were computed contrasting groups 1 and 2 versus group 3 at a single time point. On the other hand, grp1=c(1,1) and grp2=c(1,1) in combination with time1=c(2,1) and time2=c(3,1) would imply a within-subjects design where two effect size were computed contrasting time points 2 and 3 versus time point 1 in a single group. Argument phi is then used as above to indicate the autocorrelation of the measurements within groups (i.e., for the within-subjects design above, it would be the autocorrelation between time points 2 and 1 or equivalently, between time points 3 and 2).

Value

A \(k \times k\) variance-covariance matrix, where \(k\) denotes the length of the vi variable (i.e., the number of rows in the dataset).

Note

Depending on the data structure, the specified variables, and the specified values for rho and/or phi, it is possible that the constructed variance-covariance matrix is not positive definite within one or more clusters (this is checked when checkpd=TRUE, which is the default). If such non-positive definite submatrices are found, the reasons for this should be carefully checked since this might indicate misapplication of the function and/or the specification of implausible values for rho and/or phi.

When setting nearpd=TRUE, the nearPD function from the Matrix package is used on variance-covariance submatrices that are not positive definite. This should only be used cautiously and after understanding why these matrices are not positive definite.

Author

Wolfgang Viechtbauer wvb@metafor-project.org https://www.metafor-project.org

References

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. https://doi.org/10.18637/jss.v036.i03

See also

Examples

############################################################################

# see help(dat.assink2016) for further details on this dataset

dat <- dat.assink2016
dat

#>     study esid  id      yi     vi pubstatus  year deltype
#> 1       1    1   1  0.9066 0.0740         1   4.5 general
#> 2       1    2   2  0.4295 0.0398         1   4.5 general
#> 3       1    3   3  0.2679 0.0481         1   4.5 general
#> 4       1    4   4  0.2078 0.0239         1   4.5 general
#> 5       1    5   5  0.0526 0.0331         1   4.5 general
#> 6       1    6   6 -0.0507 0.0886         1   4.5 general
#> 7       2    1   7  0.5117 0.0115         1   1.5 general
#> 8       2    2   8  0.4738 0.0076         1   1.5 general
#> 9       2    3   9  0.3544 0.0065         1   1.5 general
#> 10      3    1  10  2.2844 0.3325         1  -8.5 general
#> 11      3    2  11  2.1771 0.3073         1  -8.5 general
#> 12      3    3  12  1.7777 0.2697         1  -8.5 general
#> 13      3    4  13  1.5480 0.4533         1  -8.5 general
#> 14      3    5  14  1.4855 0.1167         1  -6.5 general
#> 15      3    6  15  1.4836 0.1706         1  -6.5 general
#> 16      3    7  16  1.2777 0.1538         1  -8.5 general
#> 17      3    8  17  1.0311 0.3132         1  -8.5 general
#> 18      3    9  18  0.9409 0.1487         1  -6.5 general
#> 19      3   10  19  0.6263 0.2139         1  -6.5 general
#> 20      4    1  20 -0.0447 0.0331         1  -7.5 general
#> 21      5    1  21  1.5490 0.1384         1 -11.5 general
#> 22      6    1  22  0.7516 0.0477         1   5.5 general
#> 23      6    2  23  0.3979 0.0449         1   5.5 general
#> 24      6    3  24  0.2062 0.3533         1   5.5 general
#> 25      6    4  25 -0.2927 0.0758         1   5.5 general
#> 26      6    5  26 -0.5067 0.0314         1   5.5 general
#> 27      7    1  27  1.7083 0.0729         1   3.5 general
#> 28      7    2  28  0.8003 0.0815         1   3.5 general
#> 29      7    3  29  0.7268 0.1394         1   3.5 general
#> 30      7    4  30  0.6569 0.0943         1   3.5 general
#> 31      7    5  31  0.5569 0.1355         1   3.5 general
#> 32      7    6  32  0.3504 0.0763         1   3.5 general
#> 33      8    1  33  0.3695 0.0199         1   1.5 general
#> 34      9    1  34  0.1048 0.0331         1   5.5 general
#> 35      9    2  35  0.1748 0.0346         1   5.5 general
#> 36     10    1  36  0.4549 0.0278         1   4.5 general
#> 37     10    2  37  0.3161 0.0185         1   4.5 general
#> 38     10    3  38  0.0826 0.0328         1   4.5 general
#> 39     10    4  39  0.0221 0.0612         1   4.5 general
#> 40     10    5  40 -0.0199 0.0127         1   4.5 general
#> 41     11    1  41  2.3580 0.1226         0  -6.5 general
#> 42     11    2  42  2.3580 0.1226         0  -6.5 general
#> 43     11    3  43  2.3157 0.2416         0  -6.5 general
#> 44     11    4  44  2.0765 0.0844         0  -6.5 general
#> 45     11    5  45  2.0034 0.0780         0  -6.5 general
#> 46     11    6  46  1.7716 0.0723         0  -6.5 general
#> 47     11    7  47  1.4918 0.0701         0  -6.5 general
#> 48     11    8  48  1.4476 0.0422         0  -6.5 general
#> 49     11    9  49  1.2897 0.0663         0  -6.5   overt
#> 50     11   10  50  1.2637 0.0935         0  -6.5   overt
#> 51     11   11  51  1.2423 0.0392         0  -6.5 general
#> 52     11   12  52  1.1769 0.0387         0  -6.5 general
#> 53     11   13  53  1.0760 0.0391         0  -6.5 general
#> 54     11   14  54  1.0653 0.0380         0  -6.5   overt
#> 55     11   15  55  1.0379 0.1651         0  -6.5 general
#> 56     11   16  56  1.0101 0.0586         0  -6.5   overt
#> 57     11   17  57  1.0087 0.0386         0  -6.5   overt
#> 58     11   18  58  0.8765 0.0893         0  -6.5 general
#> 59     11   19  59  0.8615 0.1591         0  -6.5   overt
#> 60     11   20  60  0.8531 0.0350         0  -6.5 general
#> 61     11   21  61  0.6999 0.0353         0  -6.5 general
#> 62     11   22  62  0.6947 0.0353         0  -6.5 general
#> 63     12    1  63  0.2994 0.0041         0   4.5   overt
#> 64     12    2  64  0.2992 0.0042         0   4.5 general
#> 65     12    3  65  0.2989 0.0041         0   4.5   overt
#> 66     12    4  66  0.2910 0.0042         0   4.5   overt
#> 67     12    5  67  0.2170 0.0041         0   4.5   overt
#> 68     13    1  68  0.3874 0.0138         1  -3.5 general
#> 69     13    2  69  0.2178 0.0137         1  -3.5 general
#> 70     14    1  70  0.4571 0.0100         1  -1.5 general
#> 71     14    2  71  0.1947 0.0059         1  -1.5 general
#> 72     14    3  72  0.0239 0.0025         1  -1.5 general
#> 73     14    4  73 -0.0575 0.0032         1  -1.5 general
#> 74     14    5  74 -0.1190 0.0789         1  -1.5 general
#> 75     14    6  75 -0.2945 0.0179         1  -1.5 general
#> 76     15    1  76  0.7130 0.1537         1 -13.5 general
#> 77     15    2  77  0.3990 0.2334         1 -13.5 general
#> 78     15    3  78 -0.7645 0.1674         1 -13.5 general
#> 79     16    1  79  0.7156 0.0914         1   2.5   overt
#> 80     16    2  80  0.7067 0.0875         1   2.5  covert
#> 81     16    3  81  0.6475 0.0330         1   2.5 general
#> 82     16    4  82  0.6428 0.0861         1   2.5  covert
#> 83     16    5  83  0.6271 0.0400         1   2.5 general
#> 84     16    6  84  0.6238 0.0680         1   2.5 general
#> 85     16    7  85  0.6025 0.1287         1   2.5   overt
#> 86     16    8  86  0.5763 0.0332         1   2.5 general
#> 87     16    9  87  0.5171 0.0517         1   2.5  covert
#> 88     16   10  88 -0.3797 0.0390         1   2.5  covert
#> 89     16   11  89 -0.4228 0.0664         1   2.5  covert
#> 90     16   12  90 -0.4245 0.0809         1   2.5  covert
#> 91     16   13  91 -0.4671 0.0667         1   2.5  covert
#> 92     16   14  92 -0.5230 0.0988         1   2.5   overt
#> 93     16   15  93 -0.5675 0.0340         1   2.5  covert
#> 94     16   16  94 -0.7586 0.0437         1   2.5  covert
#> 95     17    1  95  0.3453 0.0340         1   5.5 general
#> 96     17    2  96  0.1221 0.0158         1   5.5 general
#> 97     17    3  97  0.0906 0.0107         1   5.5 general
#> 98     17    4  98  0.0040 0.0208         1   5.5 general
#> 99     17    5  99 -0.0207 0.0123         1   5.5 general
#> 100    17    6 100 -0.0660 0.0100         1   5.5 general

# assume that the effect sizes within studies are correlated with rho=0.6
V <- vcalc(vi, cluster=study, obs=esid, data=dat, rho=0.6)
V[dat$study %in% c(1,2), dat$study %in% c(1,2)]

#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.07400000 0.03256182 0.03579642 0.02523284 0.02969485 0.04858296
#>  [2,] 0.03256182 0.03980000 0.02625218 0.01850511 0.02177744 0.03562949
#>  [3,] 0.03579642 0.02625218 0.04810000 0.02034336 0.02394075 0.03916883
#>  [4,] 0.02523284 0.01850511 0.02034336 0.02390000 0.01687579 0.02761004
#>  [5,] 0.02969485 0.02177744 0.02394075 0.01687579 0.03310000 0.03249242
#>  [6,] 0.04858296 0.03562949 0.03916883 0.02761004 0.03249242 0.08860000
#>  [7,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [8,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [9,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>              [,7]        [,8]        [,9]
#>  [1,] 0.000000000 0.000000000 0.000000000
#>  [2,] 0.000000000 0.000000000 0.000000000
#>  [3,] 0.000000000 0.000000000 0.000000000
#>  [4,] 0.000000000 0.000000000 0.000000000
#>  [5,] 0.000000000 0.000000000 0.000000000
#>  [6,] 0.000000000 0.000000000 0.000000000
#>  [7,] 0.011500000 0.005609278 0.005187485
#>  [8,] 0.005609278 0.007600000 0.004217108
#>  [9,] 0.005187485 0.004217108 0.006500000

# use a correlation of 0.7 for effect sizes corresponding to the same type of
# delinquent behavior and a correlation of 0.5 for effect sizes corresponding
# to different types of delinquent behavior
V <- vcalc(vi, cluster=study, type=deltype, obs=esid, data=dat, rho=c(0.7, 0.5))
V[dat$study %in% 16, dat$study %in% 16]

#>             [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.09140000 0.04471437 0.02745997 0.04435521 0.03023243 0.03941827
#>  [2,] 0.04471437 0.08750000 0.02686773 0.06075802 0.02958040 0.03856812
#>  [3,] 0.02745997 0.02686773 0.03300000 0.02665192 0.02543226 0.03315961
#>  [4,] 0.04435521 0.06075802 0.02665192 0.08610000 0.02934280 0.03825833
#>  [5,] 0.03023243 0.02958040 0.02543226 0.02934280 0.04000000 0.03650753
#>  [6,] 0.03941827 0.03856812 0.03315961 0.03825833 0.03650753 0.06800000
#>  [7,] 0.07592074 0.05305952 0.03258489 0.05263333 0.03587478 0.04677499
#>  [8,] 0.02754306 0.02694903 0.02316989 0.02673256 0.02550921 0.03325995
#>  [9,] 0.03437070 0.04708118 0.02065248 0.04670301 0.02273763 0.02964625
#> [10,] 0.02985214 0.04089163 0.01793739 0.04056317 0.01974842 0.02574879
#> [11,] 0.03895177 0.05335635 0.02340513 0.05292778 0.02576820 0.03359762
#> [12,] 0.04299494 0.05889472 0.02583457 0.05842166 0.02844293 0.03708504
#> [13,] 0.03903966 0.05347675 0.02345794 0.05304721 0.02582634 0.03367343
#> [14,] 0.06651960 0.04648925 0.02854996 0.04611583 0.03143247 0.04098292
#> [15,] 0.02787293 0.03818049 0.01674813 0.03787382 0.01843909 0.02404163
#> [16,] 0.03159976 0.04328553 0.01898750 0.04293785 0.02090454 0.02725619
#>             [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,] 0.07592074 0.02754306 0.03437070 0.02985214 0.03895177 0.04299494
#>  [2,] 0.05305952 0.02694903 0.04708118 0.04089163 0.05335635 0.05889472
#>  [3,] 0.03258489 0.02316989 0.02065248 0.01793739 0.02340513 0.02583457
#>  [4,] 0.05263333 0.02673256 0.04670301 0.04056317 0.05292778 0.05842166
#>  [5,] 0.03587478 0.02550921 0.02273763 0.01974842 0.02576820 0.02844293
#>  [6,] 0.04677499 0.03325995 0.02964625 0.02574879 0.03359762 0.03708504
#>  [7,] 0.12870000 0.03268348 0.04078538 0.03542351 0.04622142 0.05101919
#>  [8,] 0.03268348 0.03320000 0.02071497 0.01799166 0.02347595 0.02591274
#>  [9,] 0.04078538 0.02071497 0.05170000 0.03143226 0.04101355 0.04527074
#> [10,] 0.03542351 0.01799166 0.03143226 0.03900000 0.03562168 0.03931919
#> [11,] 0.04622142 0.02347595 0.04101355 0.03562168 0.06640000 0.05130460
#> [12,] 0.05101919 0.02591274 0.04527074 0.03931919 0.05130460 0.08090000
#> [13,] 0.04632572 0.02352892 0.04110610 0.03570206 0.04658488 0.05142037
#> [14,] 0.07893430 0.02863634 0.03573500 0.03103707 0.04049790 0.04470157
#> [15,] 0.03307491 0.01679881 0.02934829 0.02549000 0.03325995 0.03671231
#> [16,] 0.03749730 0.01904495 0.03327239 0.02889822 0.03770707 0.04162105
#>            [,13]      [,14]      [,15]      [,16]
#>  [1,] 0.03903966 0.06651960 0.02787293 0.03159976
#>  [2,] 0.05347675 0.04648925 0.03818049 0.04328553
#>  [3,] 0.02345794 0.02854996 0.01674813 0.01898750
#>  [4,] 0.05304721 0.04611583 0.03787382 0.04293785
#>  [5,] 0.02582634 0.03143247 0.01843909 0.02090454
#>  [6,] 0.03367343 0.04098292 0.02404163 0.02725619
#>  [7,] 0.04632572 0.07893430 0.03307491 0.03749730
#>  [8,] 0.02352892 0.02863634 0.01679881 0.01904495
#>  [9,] 0.04110610 0.03573500 0.02934829 0.03327239
#> [10,] 0.03570206 0.03103707 0.02549000 0.02889822
#> [11,] 0.04658488 0.04049790 0.03325995 0.03770707
#> [12,] 0.05142037 0.04470157 0.03671231 0.04162105
#> [13,] 0.06670000 0.04058928 0.03333500 0.03779216
#> [14,] 0.04058928 0.09880000 0.02897930 0.03285407
#> [15,] 0.03333500 0.02897930 0.03400000 0.02698225
#> [16,] 0.03779216 0.03285407 0.02698225 0.04370000

# examine the correlation matrix for study 16
cov2cor(V[dat$study %in% 16, dat$study %in% 16])

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]  1.0  0.5  0.5  0.5  0.5  0.5  0.7  0.5  0.5   0.5   0.5   0.5   0.5
#>  [2,]  0.5  1.0  0.5  0.7  0.5  0.5  0.5  0.5  0.7   0.7   0.7   0.7   0.7
#>  [3,]  0.5  0.5  1.0  0.5  0.7  0.7  0.5  0.7  0.5   0.5   0.5   0.5   0.5
#>  [4,]  0.5  0.7  0.5  1.0  0.5  0.5  0.5  0.5  0.7   0.7   0.7   0.7   0.7
#>  [5,]  0.5  0.5  0.7  0.5  1.0  0.7  0.5  0.7  0.5   0.5   0.5   0.5   0.5
#>  [6,]  0.5  0.5  0.7  0.5  0.7  1.0  0.5  0.7  0.5   0.5   0.5   0.5   0.5
#>  [7,]  0.7  0.5  0.5  0.5  0.5  0.5  1.0  0.5  0.5   0.5   0.5   0.5   0.5
#>  [8,]  0.5  0.5  0.7  0.5  0.7  0.7  0.5  1.0  0.5   0.5   0.5   0.5   0.5
#>  [9,]  0.5  0.7  0.5  0.7  0.5  0.5  0.5  0.5  1.0   0.7   0.7   0.7   0.7
#> [10,]  0.5  0.7  0.5  0.7  0.5  0.5  0.5  0.5  0.7   1.0   0.7   0.7   0.7
#> [11,]  0.5  0.7  0.5  0.7  0.5  0.5  0.5  0.5  0.7   0.7   1.0   0.7   0.7
#> [12,]  0.5  0.7  0.5  0.7  0.5  0.5  0.5  0.5  0.7   0.7   0.7   1.0   0.7
#> [13,]  0.5  0.7  0.5  0.7  0.5  0.5  0.5  0.5  0.7   0.7   0.7   0.7   1.0
#> [14,]  0.7  0.5  0.5  0.5  0.5  0.5  0.7  0.5  0.5   0.5   0.5   0.5   0.5
#> [15,]  0.5  0.7  0.5  0.7  0.5  0.5  0.5  0.5  0.7   0.7   0.7   0.7   0.7
#> [16,]  0.5  0.7  0.5  0.7  0.5  0.5  0.5  0.5  0.7   0.7   0.7   0.7   0.7
#>       [,14] [,15] [,16]
#>  [1,]   0.7   0.5   0.5
#>  [2,]   0.5   0.7   0.7
#>  [3,]   0.5   0.5   0.5
#>  [4,]   0.5   0.7   0.7
#>  [5,]   0.5   0.5   0.5
#>  [6,]   0.5   0.5   0.5
#>  [7,]   0.7   0.5   0.5
#>  [8,]   0.5   0.5   0.5
#>  [9,]   0.5   0.7   0.7
#> [10,]   0.5   0.7   0.7
#> [11,]   0.5   0.7   0.7
#> [12,]   0.5   0.7   0.7
#> [13,]   0.5   0.7   0.7
#> [14,]   1.0   0.5   0.5
#> [15,]   0.5   1.0   0.7
#> [16,]   0.5   0.7   1.0

############################################################################

# see help(dat.kalaian1996) for further details on this dataset

dat <- dat.kalaian1996
dat

#>    id                   study year n1i n2i outcome    yi     vi  hrs ets
#> 1   1   Alderman & Powers (A) 1980  28  22  verbal  0.22 0.0817  7.0   1
#> 2   2   Alderman & Powers (B) 1980  39  40  verbal  0.09 0.0507 10.0   1
#> 3   3   Alderman & Powers (C) 1980  22  17  verbal  0.14 0.1045 10.5   1
#> 4   4   Alderman & Powers (D) 1980  48  43  verbal  0.14 0.0442 10.0   1
#> 5   5   Alderman & Powers (E) 1980  25  74  verbal -0.01 0.0535  6.0   1
#> 6   6   Alderman & Powers (F) 1980  37  35  verbal  0.14 0.0557  5.0   1
#> 7   7   Alderman & Powers (G) 1980  24  70  verbal  0.18 0.0561 11.0   1
#> 8   8   Alderman & Powers (H) 1980  16  19  verbal  0.01 0.1151 45.0   1
#> 9   9        Evans & Pike (A) 1973 145 129  verbal  0.13 0.0147 21.0   1
#> 10 10        Evans & Pike (A) 1973 145 129    math  0.12 0.0147 21.0   1
#> 11 11        Evans & Pike (B) 1973  72 129  verbal  0.25 0.0218 21.0   1
#> 12 12        Evans & Pike (B) 1973  72 129    math  0.06 0.0216 21.0   1
#> 13 13        Evans & Pike (C) 1973  71 129  verbal  0.31 0.0221 21.0   1
#> 14 14        Evans & Pike (C) 1973  71 129    math  0.09 0.0219 21.0   1
#> 15 15               Laschewer 1986  13  14  verbal  0.00 0.1484  8.9   0
#> 16 16               Laschewer 1986  13  14    math  0.07 0.1484  8.9   0
#> 17 17 Roberts & Oppenheim (A) 1966  43  37  verbal  0.01 0.0503  7.5   1
#> 18 18 Roberts & Oppenheim (B) 1966  19  13  verbal  0.67 0.1366  7.5   1
#> 19 19 Roberts & Oppenheim (D) 1966  16  11  verbal -0.38 0.1561  7.5   1
#> 20 20 Roberts & Oppenheim (E) 1966  20  12  verbal -0.24 0.1342  7.5   1
#> 21 21 Roberts & Oppenheim (F) 1966  39  28  verbal  0.29 0.0620  7.5   1
#> 22 22 Roberts & Oppenheim (G) 1966  38  25    math  0.26 0.0669  7.5   1
#> 23 23 Roberts & Oppenheim (H) 1966  18  13    math -0.41 0.1352  7.5   1
#> 24 24 Roberts & Oppenheim (I) 1966  19  13    math  0.08 0.1297  7.5   1
#> 25 25 Roberts & Oppenheim (J) 1966  37  22    math  0.30 0.0732  7.5   1
#> 26 26 Roberts & Oppenheim (K) 1966  19  11    math -0.53 0.1482  7.5   1
#> 27 27 Roberts & Oppenheim (L) 1966  17  13    math  0.13 0.1360  7.5   1
#> 28 28 Roberts & Oppenheim (M) 1966  20  12    math  0.26 0.1344  7.5   1
#> 29 29 Roberts & Oppenheim (N) 1966  20  13    math  0.47 0.1303  7.5   1
#> 30 30               Zuman (B) 1988  16  17  verbal  0.13 0.1216 24.0   0
#> 31 31               Zuman (B) 1988  16  17    math  0.48 0.1248 24.0   0
#> 32 32               Burke (A) 1986  25  25  verbal  0.50 0.0825 50.0   0
#> 33 33               Burke (B) 1986  25  25  verbal  0.74 0.0855 50.0   0
#> 34 34                  Coffin 1987   8   8  verbal -0.23 0.2517 18.0   0
#> 35 35                  Coffin 1987   8   8    math  0.33 0.2534 18.0   0
#> 36 36                   Davis 1985  22  21  verbal  0.13 0.0933 15.0   0
#> 37 37                   Davis 1985  22  21    math  0.13 0.0933 15.0   0
#> 38 38                 Frankel 1960  45  45  verbal  0.13 0.0445 30.0   0
#> 39 39                 Frankel 1960  45  45    math  0.34 0.0451 30.0   0
#> 40 40                Kintisch 1979  38  38  verbal  0.06 0.0527 20.0   0
#> 41 41                  Whitla 1962  52  52  verbal  0.09 0.0385 10.0   1
#> 42 42                  Whitla 1962  52  52    math -0.11 0.0385 10.0   1
#> 43 43              Curran (A) 1988  21  17  verbal -0.10 0.1066  6.0   0
#> 44 44              Curran (A) 1988  21  17    math -0.08 0.1065  6.0   0
#> 45 45              Curran (B) 1988  24  17  verbal -0.14 0.1007  6.0   0
#> 46 46              Curran (B) 1988  24  17    math -0.29 0.1015  6.0   0
#> 47 47              Curran (C) 1988  20  17  verbal -0.16 0.1092  6.0   0
#> 48 48              Curran (C) 1988  20  17    math -0.34 0.1104  6.0   0
#> 49 49              Curran (D) 1988  20  17  verbal -0.07 0.1089  6.0   0
#> 50 50              Curran (D) 1988  20  17    math -0.06 0.1089  6.0   0
#> 51 51                    Dear 1958  60 526  verbal -0.02 0.0186 15.0   1
#> 52 52                    Dear 1958  60 526    math  0.21 0.0186 15.0   1
#> 53 53                    Dyer 1953 225 193  verbal  0.06 0.0096 15.0   1
#> 54 54                    Dyer 1953 225 193    math  0.17 0.0097 15.0   1
#> 55 55              French (B) 1955 110 158  verbal  0.06 0.0154  4.5   1
#> 56 56              French (C) 1955 161 158    math  0.20 0.0126 15.0   1
#> 57 57                     FTC 1978 192 684  verbal  0.15 0.0067 40.0   0
#> 58 58                     FTC 1978 192 684    math  0.03 0.0067 40.0   0
#> 59 59               Keefauver 1976  16  25  verbal  0.17 0.1029 14.0   0
#> 60 60               Keefauver 1976  16  25    math -0.19 0.1029 14.0   0
#> 61 61                    Lass 1961  38  82  verbal  0.02 0.0385   NA   1
#> 62 62                    Lass 1961  38  82    math  0.10 0.0386   NA   1
#> 63 63      Reynolds & Oberman 1987  93  47  verbal -0.04 0.0320 63.0   0
#> 64 64      Reynolds & Oberman 1987  93  47    math  0.60 0.0333 63.0   0
#> 65 65                  Teague 1992  10  15  verbal  0.40 0.1699 18.0   0
#> 66 66               Zuman (A) 1988  21  34  verbal  0.54 0.0797 27.0   0
#> 67 67               Zuman (A) 1988  21  34    math  0.57 0.0800 27.0   0
#>    homework type
#> 1         1    1
#> 2         1    1
#> 3         1    1
#> 4         1    1
#> 5         1    1
#> 6         1    1
#> 7         1    1
#> 8         1    1
#> 9         1    1
#> 10        1    1
#> 11        1    1
#> 12        1    1
#> 13        1    1
#> 14        1    1
#> 15        0    1
#> 16        0    1
#> 17        0    1
#> 18        0    1
#> 19        0    1
#> 20        0    1
#> 21        0    1
#> 22        0    1
#> 23        0    1
#> 24        0    1
#> 25        0    1
#> 26        0    1
#> 27        0    1
#> 28        0    1
#> 29        0    1
#> 30        1    1
#> 31        1    1
#> 32        1    2
#> 33        1    2
#> 34        0    2
#> 35        0    2
#> 36        0    2
#> 37        0    2
#> 38        0    2
#> 39        0    2
#> 40        1    2
#> 41        1    2
#> 42        1    2
#> 43        0    3
#> 44        0    3
#> 45        0    3
#> 46        0    3
#> 47        0    3
#> 48        0    3
#> 49        0    3
#> 50        0    3
#> 51        1    3
#> 52        1    3
#> 53        1    3
#> 54        1    3
#> 55        1    3
#> 56        1    3
#> 57        0    3
#> 58        0    3
#> 59        0    3
#> 60        0    3
#> 61        1    3
#> 62        1    3
#> 63        l    3
#> 64        l    3
#> 65        0    3
#> 66        1    3
#> 67        1    3

# construct the variance-covariance matrix assuming rho = 0.66 for effect sizes
# corresponding to the 'verbal' and 'math' outcome types
V <- vcalc(vi, cluster=dat$study, type=outcome, data=dat, rho=0.66)
V[1:14,1:14]

#>         [,1]   [,2]   [,3]   [,4]   [,5]   [,6]   [,7]   [,8]     [,9]    [,10]
#>  [1,] 0.0817 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.000000 0.000000
#>  [2,] 0.0000 0.0507 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.000000 0.000000
#>  [3,] 0.0000 0.0000 0.1045 0.0000 0.0000 0.0000 0.0000 0.0000 0.000000 0.000000
#>  [4,] 0.0000 0.0000 0.0000 0.0442 0.0000 0.0000 0.0000 0.0000 0.000000 0.000000
#>  [5,] 0.0000 0.0000 0.0000 0.0000 0.0535 0.0000 0.0000 0.0000 0.000000 0.000000
#>  [6,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0557 0.0000 0.0000 0.000000 0.000000
#>  [7,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0561 0.0000 0.000000 0.000000
#>  [8,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.1151 0.000000 0.000000
#>  [9,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.014700 0.009702
#> [10,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.009702 0.014700
#> [11,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.000000 0.000000
#> [12,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.000000 0.000000
#> [13,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.000000 0.000000
#> [14,] 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.0000 0.000000 0.000000
#>            [,11]      [,12]      [,13]      [,14]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.00000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.00000000 0.00000000 0.00000000
#>  [4,] 0.00000000 0.00000000 0.00000000 0.00000000
#>  [5,] 0.00000000 0.00000000 0.00000000 0.00000000
#>  [6,] 0.00000000 0.00000000 0.00000000 0.00000000
#>  [7,] 0.00000000 0.00000000 0.00000000 0.00000000
#>  [8,] 0.00000000 0.00000000 0.00000000 0.00000000
#>  [9,] 0.00000000 0.00000000 0.00000000 0.00000000
#> [10,] 0.00000000 0.00000000 0.00000000 0.00000000
#> [11,] 0.02180000 0.01432185 0.00000000 0.00000000
#> [12,] 0.01432185 0.02160000 0.00000000 0.00000000
#> [13,] 0.00000000 0.00000000 0.02210000 0.01451985
#> [14,] 0.00000000 0.00000000 0.01451985 0.02190000

############################################################################