vcov.rma {metafor}	R Documentation

Variance-Covariance Matrix of Parameter Estimates for rma Objects

Description

The function extracts the (estimated) variance-covariance matrix of the fixed effects parameter estimates from objects of class "rma".

Usage

## S3 method for class 'rma':
vcov(object, ...)

Arguments

object 	 an object of class "rma".

... 	 other arguments.

Value

A matrix corresponding to the variance-covariance matrix.

Author(s)

Wolfgang Viechtbauer; wvb@www.wvbauer.com; http://www.wvbauer.com/

See Also

rma.uni, rma.mh, rma.peto

Examples

### load BCG vaccine data
data(dat.bcg)

### meta-analysis of the log risk ratios using a mixed-effects model
### with two moderators (absolute latitude and publication year)
res <- rma(ai=tpos, bi=tneg, ci=cpos, di=cneg, mods=cbind(ablat, year),
           data=dat.bcg, measure="RR", method="REML")
vcov(res)

[Package metafor version 0.5-7 Index]