Variance Inflation Factors for 'rma' Objects

Compute variance inflation factors (VIFs) for objects of class "rma".

vif(x, …)

# S3 method for rma
vif(x, intercept=FALSE, table=FALSE, digits, …)

Arguments

x
an object of class "rma".

intercept
logical indicating whether to include the intercept (if the model includes one) in the computation of the VIFs (the default is FALSE). See ‘Note’.

table
logical indicating whether the VIFs should be added to the model coefficient table (the default is FALSE).

digits
integer specifying the number of decimal places to which the printed results should be rounded (if unspecified, the default is to take the value from the object).

…
other arguments.

Details

The function computes variance inflation factors (VIFs) for meta-regression models. Hence, the model specified via x must include moderator variables. VIFs indicate the inflation in the variance of a model coefficient due to collinearity among the predictor variables.

Value

Either a vector (if table=FALSE) with the VIFs or a data frame (if table=TRUE) with the following elements:

estimate

estimated model coefficients.

se

corresponding standard errors.

zval

corresponding test statistics.

pval

corresponding p-values.

ci.lb

corresponding lower bound of the confidence intervals.

ci.ub

corresponding upper bound of the confidence intervals.

vif

corresponding variance inflation factors.

Note

The values of the VIFs are invariant to the scaling of the predictor variables if the model includes an intercept that is removed when inverting the correlation matrix of the model coefficients to compute the VIFs. This is the default behavior. See ‘Examples’.

References

Davis, C. E., Hyde, J. E., Bangdiwala, S. I., & Nelson, J. J. (1986). An example of dependencies among variables in a conditional logistic regression. In S. H. Moolgavkar & R. L. Prentice (Eds.), Modern statistical methods in chronic disease epidemiology (pp. 140--147). New York: Wiley.

Wax, Y. (1992). Collinearity diagnosis for a relative risk regression-analysis: An application to assessment of diet cancer relationship in epidemiologic studies. Statistics in Medicine, 11, 1273--1287.

Viechtbauer, W. (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1--48. http://www.jstatsoft.org/v36/i03/.

See also

Examples

### copy data from Bangert-Drowns et al. (2004) into 'dat'
dat <- dat.bangertdrowns2004

### fit mixed-effects meta-regression model
res <- rma(yi, vi, mods = ~ length + wic + feedback + info + pers + imag + meta, data=dat)

#> Warning: Studies with NAs omitted from model fitting.

### get variance inflation factors
vif(res)

#>   length      wic feedback     info     pers     imag     meta 
#>   1.5371   1.3860   1.6490   1.8340   5.6780   1.1554   4.5333 

### show that VIFs are not influenced by scaling of the predictors
u <- scale # to standardize the predictors
res <- rma(yi, vi, mods = ~ u(length) + u(wic) + u(feedback) + u(info) + u(pers) + u(imag) + u(meta), data=dat)

#> Warning: Studies with NAs omitted from model fitting.

vif(res)

#>   u(length)      u(wic) u(feedback)     u(info)     u(pers)     u(imag) 
#>      1.5371      1.3860      1.6490      1.8340      5.6780      1.1554 
#>     u(meta) 
#>      4.5333 

### get full table
vif(res, table=TRUE)

#>             estimate     se    zval   pval   ci.lb  ci.ub    vif
#> intrcpt       0.1825 0.0406  4.4898 0.0000  0.1028 0.2621     NA
#> u(length)     0.0458 0.0496  0.9240 0.3555 -0.0514 0.1431 1.5371
#> u(wic)       -0.0210 0.0487 -0.4308 0.6666 -0.1164 0.0744 1.3860
#> u(feedback)   0.0329 0.0524  0.6265 0.5310 -0.0699 0.1357 1.6490
#> u(info)      -0.0460 0.0418 -1.1006 0.2711 -0.1280 0.0360 1.8340
#> u(pers)      -0.0573 0.0956 -0.5992 0.5490 -0.2446 0.1301 5.6780
#> u(imag)       0.1004 0.0452  2.2233 0.0262  0.0119 0.1890 1.1554
#> u(meta)       0.0981 0.0850  1.1537 0.2486 -0.0685 0.2647 4.5333